Process data for autoplot

Process data for autoplot

Usage

autoplot_data_processor(x, newdata, x_str, ranef_plot = NULL)

Arguments

x

cglmm object.

newdata

new dataset (if used).

x_str

A character vector naming variable(s) to be plotted. Default has no value and plots all groups.

ranef_plot

Specify the random effects variables that you wish to plot. If not specified, only the fixed effects will be visualized.

Value

a data.frame.

Examples

object <- cglmm(
  vit_d ~ amp_acro(time_col = time, group = "X", period = 12),
  data = vitamind
)
GLMMcosinor:::autoplot_data_processor(
  object,
  vitamind[ "time"],
  x_str = "X"
)
#>           time X   main_rrr1   main_sss1 levels
#> 1   11.4395249 0  0.95724757 -0.28926992      0
#> 2    5.8071039 0 -0.99490382  0.10082851      0
#> 3    1.0454920 0  0.85387107  0.52048457      0
#> 4    4.0829835 0 -0.53714510  0.84348986      0
#> 5   10.6062469 0  0.74532948 -0.66669631      0
#> 6    5.1260544 0 -0.89711678  0.44179349      0
#> 7    6.6579942 0 -0.94123600 -0.33774960      0
#> 8   10.3629108 0  0.65457359 -0.75599829      0
#> 9    1.7815771 0  0.59556171  0.80330956      0
#> 10  11.8814943 0  0.99807555 -0.06200961      0
#> 11   2.9245825 0  0.03947825  0.99922043      0
#> 12   3.3690848 0 -0.19205173  0.98138480      0
#> 13  10.8085739 0  0.81164756 -0.58414745      0
#> 14   3.2512541 0 -0.13117721  0.99135894      0
#> 15   7.5673160 0 -0.68174968 -0.73158552      0
#> 16   0.2577064 0  0.99091011  0.13452565      0
#> 17   9.8401312 0  0.42584144 -0.90479781      0
#> 18   4.6852044 0 -0.77224739  0.63532194      0
#> 19   6.0399353 0 -0.99978139 -0.02090856      0
#> 20   6.7275242 0 -0.92831899 -0.37178469      0
#> 21   3.7244321 0 -0.37028124  0.92891970      0
#> 22   5.5853785 0 -0.97652722  0.21539401      0
#> 23  10.7046169 0  0.77866500 -0.62743989      0
#> 24   1.3904045 0  0.74649728  0.66538846      0
#> 25   8.4112493 0 -0.30340982 -0.95286016      0
#> 26   8.5015419 0 -0.25803915 -0.96613446      0
#> 27   4.3498771 0 -0.64939910  0.76044777      0
#> 28   9.1817036 0  0.09499630 -0.99547763      0
#> 29   7.7634308 0 -0.60316727 -0.79761472      0
#> 30   1.7527889 0  0.60760227  0.79424145      0
#> 31  10.4465596 0  0.68704678 -0.72661318      0
#> 32   0.9275858 0  0.88435642  0.46681229      0
#> 33  11.2710373 0  0.92803869 -0.37248380      0
#> 34   4.3221571 0 -0.63829384  0.76979281      0
#> 35   3.6329666 0 -0.32538660  0.94558107      0
#> 36   1.8921467 0  0.54808301  0.83642394      0
#> 37   8.6489376 0 -0.18278248 -0.98315338      0
#> 38   6.1360675 0 -0.99746317 -0.07118451      0
#> 39  10.4307690 0  0.68101578 -0.73226874      0
#> 40   8.9012058 0 -0.05170547 -0.99866238      0
#> 41   0.4951951 0  0.96657392  0.25638810      0
#> 42   8.0993826 0 -0.45427850 -0.89085972      0
#> 43  10.2183575 0  0.59553424 -0.80332993      0
#> 44   7.7760288 0 -0.59789288 -0.80157601      0
#> 45  10.3292839 0  0.64116194 -0.76740560      0
#> 46   7.5231146 0 -0.69849723 -0.71561276      0
#> 47  11.0692944 0  0.88359268 -0.46825631      0
#> 48   6.1031769 0 -0.99854110 -0.05399702      0
#> 49   3.0982198 0 -0.05140507  0.99867789      0
#> 50   4.4038373 0 -0.67062238  0.74179891      0
#> 51  11.5279052 0  0.96960422 -0.24467868      0
#> 52   5.1019377 0 -0.89146668  0.45308626      0
#> 53   4.5678583 0 -0.73177910  0.68154189      0
#> 54   7.9679893 0 -0.51444432 -0.85752378      0
#> 55  10.0343099 0  0.51547630 -0.85690384      0
#> 56   2.5610647 0  0.22780811  0.97370605      0
#> 57   6.5487480 0 -0.95900571 -0.28338676      0
#> 58   7.3569442 0 -0.75803958 -0.65220856      0
#> 59   3.1453298 0 -0.07602108  0.99710621      0
#> 60   4.4634024 0 -0.69342787  0.72052605      0
#> 61   2.0455533 0  0.47920362  0.87770376      0
#> 62  11.1689572 0  0.90681400 -0.42153099      0
#> 63   1.6958015 0  0.63102729  0.77576063      0
#> 64   6.9089820 0 -0.88886158 -0.45817582      0
#> 65  11.5031885 0  0.96635658 -0.25720606      0
#> 66  11.1687746 0  0.90677368 -0.42161770      0
#> 67   1.6151464 0  0.66321601  0.74842804      0
#> 68  10.0439404 0  0.51979068 -0.85429366      0
#> 69   2.1029540 0  0.45261181  0.89170766      0
#> 70   8.2447904 0 -0.38520212 -0.92283223      0
#> 71  11.0759045 0  0.88520805 -0.46519535      0
#> 72   0.5661396 0  0.95638540  0.29210779      0
#> 73   5.4569622 0 -0.95984871  0.28051818      0
#> 74   6.8317477 0 -0.90665836 -0.42186564      0
#> 75   7.7781662 0 -0.59699544 -0.80224463      0
#> 76   5.7107668 0 -0.98855454  0.15086394      0
#> 77   8.9569165 0 -0.02255655 -0.99974557      0
#> 78   9.7245118 0  0.37032002 -0.92890424      0
#> 79   2.4632888 0  0.27733706  0.96077269      0
#> 80   0.2719170 0  0.98988173  0.14189491      0
#> 81   9.2108872 0  0.11019604 -0.99390987      0
#> 82   5.6088367 0 -0.97909911  0.20338371      0
#> 83   2.9166005 0  0.04365402  0.99904671      0
#> 84   1.6001840 0  0.66905899  0.74320930      0
#> 85   0.2651303 0  0.99037971  0.13837644      0
#> 86   5.2304744 0 -0.91991892  0.39210863      0
#> 87   9.3877762 0  0.20164695 -0.97945827      0
#> 88   1.0857651 0  0.84270664  0.53837304      0
#> 89  11.5959746 0  0.97770722 -0.20997286      0
#> 90   7.7737138 0 -0.59886409 -0.80085067      0
#> 91   1.5423938 0  0.69123799  0.72262718      0
#> 92   1.9894992 0  0.50475400  0.86326322      0
#> 93   7.5896403 0 -0.67315180 -0.73950433      0
#> 94  11.2741562 0  0.92864575 -0.37096776      0
#> 95   9.2511895 0  0.13114368 -0.99136337      0
#> 96  11.2519963 0  0.92427902 -0.38171755      0
#> 97   7.2059514 0 -0.80718143 -0.59030343      0
#> 98   9.0635547 0  0.03327101 -0.99944637      0
#> 99   5.3530543 0 -0.94317412  0.33229893      0
#> 100  7.8744052 0 -0.55582916 -0.83129654      0
#> 101  0.6597983 0  0.94091654  0.33863855      0
#> 102  3.9734737 0 -0.48792380  0.87288623      0
#> 103  9.5177903 0  0.26780527 -0.96347306      0
#> 104 10.2846713 0  0.62306273 -0.78217187      0
#> 105  0.5738949 0  0.95519137  0.29598892      0
#> 106  5.3576077 0 -0.94396368  0.33004934      0
#> 107  3.4417425 0 -0.22923903  0.97337016      0
#> 108  2.0337626 0  0.48461301  0.87472866      0
#> 109  4.4686183 0 -0.69539307  0.71862958      0
#> 110  6.9162051 0 -0.88712239 -0.46153425      0
#> 111  6.5154807 0 -0.96379622 -0.26663993      0
#> 112  6.5441738 0 -0.95968168 -0.28108908      0
#> 113  6.2534499 0 -0.99120747 -0.13231686      0
#> 114  5.1860105 0 -0.91054167  0.41341730      0
#> 115  7.7564274 0 -0.60608806 -0.79539755      0
#> 116  4.6450448 0 -0.75871838  0.65141878      0
#> 117  6.0767892 0 -0.99919182 -0.04019590      0
#> 118  3.2056626 0 -0.10747667  0.99420761      0
#> 119  5.4271527 0 -0.95535358  0.29546495      0
#> 120  8.4787094 0 -0.26957063 -0.96298062      0
#> 121  8.5037112 0 -0.25694160 -0.96642693      0
#> 122  8.6898143 0 -0.16169975 -0.98684000      0
#> 123  9.4197148 0  0.21799752 -0.97594932      0
#> 124  0.6626884 0  0.94040301  0.34006203      0
#> 125  1.9570880 0  0.51933061  0.85457341      0
#> 126 11.4610895 0  0.96045269 -0.27844323      0
#> 127 11.2714325 0  0.92811575 -0.37229175      0
#> 128  5.0558515 0 -0.88027492  0.47446398      0
#> 129  3.0940964 0 -0.04924884  0.99878654      0
#> 130  3.2206602 0 -0.11528051  0.99333298      0
#> 131  7.2350871 0 -0.79808254 -0.60254813      0
#> 132  6.7114849 0 -0.93140852 -0.36397552      0
#> 133  1.0527236 0  0.85189418  0.52371395      0
#> 134 10.2724142 0  0.61803011 -0.78615442      0
#> 135  9.6374544 0  0.32760760 -0.94481388      0
#> 136  1.1373819 0  0.82785032  0.56094906      0
#> 137  3.6084338 0 -0.31321378  0.94968264      0
#> 138  7.3370613 0 -0.76478829 -0.64428167      0
#> 139  0.5868602 0  0.95316003  0.30246647      0
#> 140 11.8426146 0  0.99660648 -0.08231358      0
#> 141 11.3080327 0  0.93507943 -0.35443822      0
#> 142 10.9439328 0  0.85097599 -0.52520459      0
#> 143 10.1840199 0  0.58099559 -0.81390670      0
#> 144 11.8900364 0  0.99834291 -0.05754501      0
#> 145  2.7117812 0  0.15033883  0.98863453      0
#> 146  4.2648370 0 -0.61490626  0.78860021      0
#> 147  9.1592249 0  0.08327342 -0.99652674      0
#> 148  9.2066595 0  0.10799560 -0.99415137      0
#> 149  7.7649399 0 -0.60253687 -0.79809105      0
#> 150  1.5667978 0  0.68194815  0.73140052      0
#> 151  2.8420136 0  0.08262716  0.99658053      0
#> 152  8.1446614 0 -0.43303241 -0.90137835      0
#> 153  0.9796193 0  0.87131164  0.49073010      0
#> 154  3.4164948 0 -0.21635174  0.97631548      0
#> 155  9.8643172 0  0.43726516 -0.89933263      0
#> 156  0.3859391 0  0.97965178  0.20070474      0
#> 157 10.8437720 0  0.82227479 -0.56909065      0
#> 158  1.7374482 0  0.61396225  0.78933539      0
#> 159  0.3269302 0  0.98538441  0.17034545      0
#> 160  2.4498268 0  0.28410229  0.95879398      0
#> 161  6.0675886 0 -0.99937386 -0.03538191      0
#> 162  1.0313966 0  0.85768915  0.51416857      0
#> 163  9.5200400 0  0.26894001 -0.96315693      0
#> 164  1.7798364 0  0.59629362  0.80276641      0
#> 165  2.3164199 0  0.35032840  0.93662693      0
#> 166  4.6050395 0 -0.74490785  0.66716737      0
#> 167 10.0334543 0  0.51509235 -0.85713469      0
#> 168  9.9608719 0  0.48215369 -0.87608665      0
#> 169  0.2804965 0  0.98923433  0.14634018      0
#> 170  6.4606393 0 -0.97105438 -0.23885852      0
#> 171 10.9284610 0  0.84669343 -0.53208104      0
#> 172  9.2508363 0  0.13096033 -0.99138761      0
#> 173 11.8349258 0  0.99626702 -0.08632508      0
#> 174  6.8121552 0 -0.91093831 -0.41254259      0
#> 175  1.5355029 0  0.69384077  0.72012845      0
#> 176  9.1843438 0  0.09637239 -0.99534535      0
#> 177  6.9982817 0 -0.86647491 -0.49922061      0
#> 178  3.0565049 0 -0.02958159  0.99956237      0
#> 179  7.5438536 0 -0.69068543 -0.72315533      0
#> 180  4.7525195 0 -0.79415573  0.60771431      0
#> 181  5.3122749 0 -0.93586441  0.35236033      0
#> 182  0.7526566 0  0.92334634  0.38396815      0
#> 183  6.8593106 0 -0.90047582 -0.43490608      0
#> 184  9.2079802 0  0.10868305 -0.99407645      0
#> 185  1.2901631 0  0.78037701  0.62530930      0
#> 186  4.4004538 0 -0.66930716  0.74298582      0
#> 187  1.4844200 0  0.71285152  0.70131499      0
#> 188  8.2069239 0 -0.40342206 -0.91501401      0
#> 189  4.8795922 0 -0.83280305  0.55356940      0
#> 190  9.7335898 0  0.37473111 -0.92713354      0
#> 191  0.4040244 0  0.97770733  0.20997233      0
#> 192 10.6911255 0  0.77421334 -0.63292472      0
#> 193  6.2686548 0 -0.99012266 -0.14020387      0
#> 194 10.6517445 0  0.76099885 -0.64875322      0
#> 195 10.8544010 0  0.82542920 -0.56450566      0
#> 196  5.5682845 0 -0.97456027  0.22412557      0
#> 197  7.5465960 0 -0.68964633 -0.72414635      0
#> 198  1.9715145 0  0.51286065  0.85847187      0
#> 199  6.1340190 0 -0.99753894 -0.07011463      0
#> 200  6.0856508 0 -0.99899456 -0.04483160      0
#> 201 11.4395249 1  0.95724757 -0.28926992      1
#> 202  5.8071039 1 -0.99490382  0.10082851      1
#> 203  1.0454920 1  0.85387107  0.52048457      1
#> 204  4.0829835 1 -0.53714510  0.84348986      1
#> 205 10.6062469 1  0.74532948 -0.66669631      1
#> 206  5.1260544 1 -0.89711678  0.44179349      1
#> 207  6.6579942 1 -0.94123600 -0.33774960      1
#> 208 10.3629108 1  0.65457359 -0.75599829      1
#> 209  1.7815771 1  0.59556171  0.80330956      1
#> 210 11.8814943 1  0.99807555 -0.06200961      1
#> 211  2.9245825 1  0.03947825  0.99922043      1
#> 212  3.3690848 1 -0.19205173  0.98138480      1
#> 213 10.8085739 1  0.81164756 -0.58414745      1
#> 214  3.2512541 1 -0.13117721  0.99135894      1
#> 215  7.5673160 1 -0.68174968 -0.73158552      1
#> 216  0.2577064 1  0.99091011  0.13452565      1
#> 217  9.8401312 1  0.42584144 -0.90479781      1
#> 218  4.6852044 1 -0.77224739  0.63532194      1
#> 219  6.0399353 1 -0.99978139 -0.02090856      1
#> 220  6.7275242 1 -0.92831899 -0.37178469      1
#> 221  3.7244321 1 -0.37028124  0.92891970      1
#> 222  5.5853785 1 -0.97652722  0.21539401      1
#> 223 10.7046169 1  0.77866500 -0.62743989      1
#> 224  1.3904045 1  0.74649728  0.66538846      1
#> 225  8.4112493 1 -0.30340982 -0.95286016      1
#> 226  8.5015419 1 -0.25803915 -0.96613446      1
#> 227  4.3498771 1 -0.64939910  0.76044777      1
#> 228  9.1817036 1  0.09499630 -0.99547763      1
#> 229  7.7634308 1 -0.60316727 -0.79761472      1
#> 230  1.7527889 1  0.60760227  0.79424145      1
#> 231 10.4465596 1  0.68704678 -0.72661318      1
#> 232  0.9275858 1  0.88435642  0.46681229      1
#> 233 11.2710373 1  0.92803869 -0.37248380      1
#> 234  4.3221571 1 -0.63829384  0.76979281      1
#> 235  3.6329666 1 -0.32538660  0.94558107      1
#> 236  1.8921467 1  0.54808301  0.83642394      1
#> 237  8.6489376 1 -0.18278248 -0.98315338      1
#> 238  6.1360675 1 -0.99746317 -0.07118451      1
#> 239 10.4307690 1  0.68101578 -0.73226874      1
#> 240  8.9012058 1 -0.05170547 -0.99866238      1
#> 241  0.4951951 1  0.96657392  0.25638810      1
#> 242  8.0993826 1 -0.45427850 -0.89085972      1
#> 243 10.2183575 1  0.59553424 -0.80332993      1
#> 244  7.7760288 1 -0.59789288 -0.80157601      1
#> 245 10.3292839 1  0.64116194 -0.76740560      1
#> 246  7.5231146 1 -0.69849723 -0.71561276      1
#> 247 11.0692944 1  0.88359268 -0.46825631      1
#> 248  6.1031769 1 -0.99854110 -0.05399702      1
#> 249  3.0982198 1 -0.05140507  0.99867789      1
#> 250  4.4038373 1 -0.67062238  0.74179891      1
#> 251 11.5279052 1  0.96960422 -0.24467868      1
#> 252  5.1019377 1 -0.89146668  0.45308626      1
#> 253  4.5678583 1 -0.73177910  0.68154189      1
#> 254  7.9679893 1 -0.51444432 -0.85752378      1
#> 255 10.0343099 1  0.51547630 -0.85690384      1
#> 256  2.5610647 1  0.22780811  0.97370605      1
#> 257  6.5487480 1 -0.95900571 -0.28338676      1
#> 258  7.3569442 1 -0.75803958 -0.65220856      1
#> 259  3.1453298 1 -0.07602108  0.99710621      1
#> 260  4.4634024 1 -0.69342787  0.72052605      1
#> 261  2.0455533 1  0.47920362  0.87770376      1
#> 262 11.1689572 1  0.90681400 -0.42153099      1
#> 263  1.6958015 1  0.63102729  0.77576063      1
#> 264  6.9089820 1 -0.88886158 -0.45817582      1
#> 265 11.5031885 1  0.96635658 -0.25720606      1
#> 266 11.1687746 1  0.90677368 -0.42161770      1
#> 267  1.6151464 1  0.66321601  0.74842804      1
#> 268 10.0439404 1  0.51979068 -0.85429366      1
#> 269  2.1029540 1  0.45261181  0.89170766      1
#> 270  8.2447904 1 -0.38520212 -0.92283223      1
#> 271 11.0759045 1  0.88520805 -0.46519535      1
#> 272  0.5661396 1  0.95638540  0.29210779      1
#> 273  5.4569622 1 -0.95984871  0.28051818      1
#> 274  6.8317477 1 -0.90665836 -0.42186564      1
#> 275  7.7781662 1 -0.59699544 -0.80224463      1
#> 276  5.7107668 1 -0.98855454  0.15086394      1
#> 277  8.9569165 1 -0.02255655 -0.99974557      1
#> 278  9.7245118 1  0.37032002 -0.92890424      1
#> 279  2.4632888 1  0.27733706  0.96077269      1
#> 280  0.2719170 1  0.98988173  0.14189491      1
#> 281  9.2108872 1  0.11019604 -0.99390987      1
#> 282  5.6088367 1 -0.97909911  0.20338371      1
#> 283  2.9166005 1  0.04365402  0.99904671      1
#> 284  1.6001840 1  0.66905899  0.74320930      1
#> 285  0.2651303 1  0.99037971  0.13837644      1
#> 286  5.2304744 1 -0.91991892  0.39210863      1
#> 287  9.3877762 1  0.20164695 -0.97945827      1
#> 288  1.0857651 1  0.84270664  0.53837304      1
#> 289 11.5959746 1  0.97770722 -0.20997286      1
#> 290  7.7737138 1 -0.59886409 -0.80085067      1
#> 291  1.5423938 1  0.69123799  0.72262718      1
#> 292  1.9894992 1  0.50475400  0.86326322      1
#> 293  7.5896403 1 -0.67315180 -0.73950433      1
#> 294 11.2741562 1  0.92864575 -0.37096776      1
#> 295  9.2511895 1  0.13114368 -0.99136337      1
#> 296 11.2519963 1  0.92427902 -0.38171755      1
#> 297  7.2059514 1 -0.80718143 -0.59030343      1
#> 298  9.0635547 1  0.03327101 -0.99944637      1
#> 299  5.3530543 1 -0.94317412  0.33229893      1
#> 300  7.8744052 1 -0.55582916 -0.83129654      1
#> 301  0.6597983 1  0.94091654  0.33863855      1
#> 302  3.9734737 1 -0.48792380  0.87288623      1
#> 303  9.5177903 1  0.26780527 -0.96347306      1
#> 304 10.2846713 1  0.62306273 -0.78217187      1
#> 305  0.5738949 1  0.95519137  0.29598892      1
#> 306  5.3576077 1 -0.94396368  0.33004934      1
#> 307  3.4417425 1 -0.22923903  0.97337016      1
#> 308  2.0337626 1  0.48461301  0.87472866      1
#> 309  4.4686183 1 -0.69539307  0.71862958      1
#> 310  6.9162051 1 -0.88712239 -0.46153425      1
#> 311  6.5154807 1 -0.96379622 -0.26663993      1
#> 312  6.5441738 1 -0.95968168 -0.28108908      1
#> 313  6.2534499 1 -0.99120747 -0.13231686      1
#> 314  5.1860105 1 -0.91054167  0.41341730      1
#> 315  7.7564274 1 -0.60608806 -0.79539755      1
#> 316  4.6450448 1 -0.75871838  0.65141878      1
#> 317  6.0767892 1 -0.99919182 -0.04019590      1
#> 318  3.2056626 1 -0.10747667  0.99420761      1
#> 319  5.4271527 1 -0.95535358  0.29546495      1
#> 320  8.4787094 1 -0.26957063 -0.96298062      1
#> 321  8.5037112 1 -0.25694160 -0.96642693      1
#> 322  8.6898143 1 -0.16169975 -0.98684000      1
#> 323  9.4197148 1  0.21799752 -0.97594932      1
#> 324  0.6626884 1  0.94040301  0.34006203      1
#> 325  1.9570880 1  0.51933061  0.85457341      1
#> 326 11.4610895 1  0.96045269 -0.27844323      1
#> 327 11.2714325 1  0.92811575 -0.37229175      1
#> 328  5.0558515 1 -0.88027492  0.47446398      1
#> 329  3.0940964 1 -0.04924884  0.99878654      1
#> 330  3.2206602 1 -0.11528051  0.99333298      1
#> 331  7.2350871 1 -0.79808254 -0.60254813      1
#> 332  6.7114849 1 -0.93140852 -0.36397552      1
#> 333  1.0527236 1  0.85189418  0.52371395      1
#> 334 10.2724142 1  0.61803011 -0.78615442      1
#> 335  9.6374544 1  0.32760760 -0.94481388      1
#> 336  1.1373819 1  0.82785032  0.56094906      1
#> 337  3.6084338 1 -0.31321378  0.94968264      1
#> 338  7.3370613 1 -0.76478829 -0.64428167      1
#> 339  0.5868602 1  0.95316003  0.30246647      1
#> 340 11.8426146 1  0.99660648 -0.08231358      1
#> 341 11.3080327 1  0.93507943 -0.35443822      1
#> 342 10.9439328 1  0.85097599 -0.52520459      1
#> 343 10.1840199 1  0.58099559 -0.81390670      1
#> 344 11.8900364 1  0.99834291 -0.05754501      1
#> 345  2.7117812 1  0.15033883  0.98863453      1
#> 346  4.2648370 1 -0.61490626  0.78860021      1
#> 347  9.1592249 1  0.08327342 -0.99652674      1
#> 348  9.2066595 1  0.10799560 -0.99415137      1
#> 349  7.7649399 1 -0.60253687 -0.79809105      1
#> 350  1.5667978 1  0.68194815  0.73140052      1
#> 351  2.8420136 1  0.08262716  0.99658053      1
#> 352  8.1446614 1 -0.43303241 -0.90137835      1
#> 353  0.9796193 1  0.87131164  0.49073010      1
#> 354  3.4164948 1 -0.21635174  0.97631548      1
#> 355  9.8643172 1  0.43726516 -0.89933263      1
#> 356  0.3859391 1  0.97965178  0.20070474      1
#> 357 10.8437720 1  0.82227479 -0.56909065      1
#> 358  1.7374482 1  0.61396225  0.78933539      1
#> 359  0.3269302 1  0.98538441  0.17034545      1
#> 360  2.4498268 1  0.28410229  0.95879398      1
#> 361  6.0675886 1 -0.99937386 -0.03538191      1
#> 362  1.0313966 1  0.85768915  0.51416857      1
#> 363  9.5200400 1  0.26894001 -0.96315693      1
#> 364  1.7798364 1  0.59629362  0.80276641      1
#> 365  2.3164199 1  0.35032840  0.93662693      1
#> 366  4.6050395 1 -0.74490785  0.66716737      1
#> 367 10.0334543 1  0.51509235 -0.85713469      1
#> 368  9.9608719 1  0.48215369 -0.87608665      1
#> 369  0.2804965 1  0.98923433  0.14634018      1
#> 370  6.4606393 1 -0.97105438 -0.23885852      1
#> 371 10.9284610 1  0.84669343 -0.53208104      1
#> 372  9.2508363 1  0.13096033 -0.99138761      1
#> 373 11.8349258 1  0.99626702 -0.08632508      1
#> 374  6.8121552 1 -0.91093831 -0.41254259      1
#> 375  1.5355029 1  0.69384077  0.72012845      1
#> 376  9.1843438 1  0.09637239 -0.99534535      1
#> 377  6.9982817 1 -0.86647491 -0.49922061      1
#> 378  3.0565049 1 -0.02958159  0.99956237      1
#> 379  7.5438536 1 -0.69068543 -0.72315533      1
#> 380  4.7525195 1 -0.79415573  0.60771431      1
#> 381  5.3122749 1 -0.93586441  0.35236033      1
#> 382  0.7526566 1  0.92334634  0.38396815      1
#> 383  6.8593106 1 -0.90047582 -0.43490608      1
#> 384  9.2079802 1  0.10868305 -0.99407645      1
#> 385  1.2901631 1  0.78037701  0.62530930      1
#> 386  4.4004538 1 -0.66930716  0.74298582      1
#> 387  1.4844200 1  0.71285152  0.70131499      1
#> 388  8.2069239 1 -0.40342206 -0.91501401      1
#> 389  4.8795922 1 -0.83280305  0.55356940      1
#> 390  9.7335898 1  0.37473111 -0.92713354      1
#> 391  0.4040244 1  0.97770733  0.20997233      1
#> 392 10.6911255 1  0.77421334 -0.63292472      1
#> 393  6.2686548 1 -0.99012266 -0.14020387      1
#> 394 10.6517445 1  0.76099885 -0.64875322      1
#> 395 10.8544010 1  0.82542920 -0.56450566      1
#> 396  5.5682845 1 -0.97456027  0.22412557      1
#> 397  7.5465960 1 -0.68964633 -0.72414635      1
#> 398  1.9715145 1  0.51286065  0.85847187      1
#> 399  6.1340190 1 -0.99753894 -0.07011463      1
#> 400  6.0856508 1 -0.99899456 -0.04483160      1