Fit cosinor model

Given an outcome and time variable, fit the cosinor model with optional covariate effects.

Usage

cosinor.glmm.default(formula, ...)

Arguments

formula

Formula specifying the model. Indicate the time variable with time() and covariate effects on the amplitude and acrophase with amp.acro(). See details.

...

other arguments

Details

This defines special functions that are used in the formula to indicate the time variable and which covariates effect the amplitude. To indicate the time variable wrap the name of it in the function time(). To indicate a variable which affects the acrophase/amplitude, wrap the name in amp.acro(). This will then do all the transformations for you. See examples for usage.

Examples

cosinor.glmm(Y ~ time(time) + X + amp.acro(X), data = vitamind)
#> Warning: fixed effects in conditional model are rank deficient
#> Warning: Model convergence problem; non-positive-definite Hessian matrix. See vignette('troubleshooting')
#> 
#>  Conditional Model 
#> 
#>  Raw formula: 
#> Y ~ time(time) + X + main_rrr1 + main_sss1
#> <environment: 0x55eaa8fa3850>
#> 
#>  Raw Coefficients: 
#> (Intercept)  time(time)           X   main_rrr1   main_sss1 
#>  15.8334076  -0.0182114   2.5942937  15.6198127   1.7971469 
#> 
#>  Transformed Coefficients: 
#> (Intercept)  time(time)           X         amp         acr 
#>  15.8334076  -0.0182114   2.5942937  15.7228587  -0.1145519