Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The MESOR is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times group
#> 1    3.53798552  6.5115495     0
#> 2    3.30731543  5.4278403     0
#> 3    0.36341157 12.3632877     0
#> 4   -0.54355670 13.5990957     0
#> 5    1.19922068  7.5229932     0
#> 6   -0.75654914 15.7588858     0
#> 7    1.00535048 23.0827339     0
#> 8    1.05600811 12.2047004     0
#> 9    0.50774221  8.9022858     0
#> 10   0.49600457 13.2442903     0
#> 11   4.06171612  2.2996008     0
#> 12   0.90533563 14.8227453     0
#> 13   0.96734709 19.8341097     0
#> 14   0.84939592  0.1400204     0
#> 15   0.67242228 18.9342633     0
#> 16   0.84256067 16.4632837     0
#> 17   1.02465555 19.7201552     0
#> 18  -0.70154969 13.9276257     0
#> 19   1.03745949  9.6916067     0
#> 20   2.18038193  7.0121329     0
#> 21  -1.36895386 14.9841146     0
#> 22   0.97095140  4.4965177     0
#> 23  -0.94068833 21.8048822     0
#> 24   2.66718596  3.8756766     0
#> 25   2.81712026  4.1507986     0
#> 26  -0.18638514 17.3027961     0
#> 27   1.05611064 19.2883503     0
#> 28   1.21441871 18.4357950     0
#> 29  -0.45151924  9.0105376     0
#> 30   2.14937369  0.3177142     0
#> 31   2.06502319 22.8276000     0
#> 32   2.00604963  8.8820790     0
#> 33   1.09238895 23.0785806     0
#> 34   3.09402107 21.6618412     0
#> 35   2.97960013  3.1287853     0
#> 36   1.39935145  1.5024329     0
#> 37  -0.14323532 10.7023848     0
#> 38   0.56573490  0.3166253     0
#> 39   0.84600405 14.2946060     0
#> 40   0.12479535 10.7117829     0
#> 41   0.01879581 13.0073356     0
#> 42  -0.10546078 12.5498352     0
#> 43   3.41806399  1.8384680     0
#> 44   1.84593270 18.4446227     0
#> 45   1.33105216  1.2539905     0
#> 46  -0.04920444 22.2498254     0
#> 47   2.72097079  2.8848051     0
#> 48  -0.07428984 13.0134173     0
#> 49   0.58572213  9.3894425     0
#> 50   0.25507066 16.2635867     0
#> 51   1.90701429  1.2402553     0
#> 52   1.61103679 22.0489569     0
#> 53  -0.13434364 12.2650008     0
#> 54  -1.13093314 13.4048545     0
#> 55   1.15939859  0.2044530     0
#> 56   2.61134190 22.3435340     0
#> 57  -1.15858970 14.8922867     0
#> 58   1.27719707 11.2809113     0
#> 59  -0.17728924 21.5921824     0
#> 60   0.57223703 16.7222153     0
#> 61   0.59875930 20.9495877     0
#> 62  -0.69681935 15.7040330     0
#> 63   0.82946728 14.5975821     0
#> 64   1.98150329  2.3746773     0
#> 65   0.90008256 16.8845224     0
#> 66  -1.25164945 18.3013775     0
#> 67   0.92816849  5.6651303     0
#> 68   1.53130252  4.8901295     0
#> 69   0.84086348  0.3171551     0
#> 70   1.02247859  7.2720812     0
#> 71  -1.08560917 18.1714938     0
#> 72   0.10850542 13.3917519     0
#> 73   0.51628627  9.9443075     0
#> 74   2.34270162  3.9302023     0
#> 75   1.69975094  6.4567211     0
#> 76   3.50859169  1.7829947     0
#> 77   1.89572173 23.7160083     0
#> 78   2.04718922  1.6967664     0
#> 79   2.02191942  2.6823915     0
#> 80  -0.46799247 18.7104305     0
#> 81   1.45446034 11.9258614     0
#> 82  -0.20616732 21.7356753     0
#> 83  -0.38314060 15.9650285     0
#> 84   3.14503962  8.0481251     0
#> 85   1.81648803  7.1060571     0
#> 86  -0.55496167 16.9521888     0
#> 87   0.16368918 12.1998975     0
#> 88  -0.25019594 15.6190755     0
#> 89  -0.03225310 10.6239619     0
#> 90   0.73355889  9.9599649     0
#> 91   1.46640768 20.8230345     0
#> 92   0.74459285 12.8908417     0
#> 93  -0.50075179 18.4103133     0
#> 94   2.43534856  3.2352851     0
#> 95   1.88106693 18.4004588     0
#> 96   1.61699704  9.9059143     0
#> 97   0.57128321  4.8113105     0
#> 98   0.67926563 11.7286590     0
#> 99   2.52766083  1.7220612     0
#> 100  1.92345266 10.6600294     0