Test for differences in a cosinor model

Given a time variable and optional covariates, generate inference a cosinor fit. For the covariate named (or vector of covariates), this function performs a Wald test comparing the group with covariates equal to 1 to the group with covariates equal to 0. This may not be the desired result for continuous covariates.

Usage

test_cosinor(
  x,
  x_str,
  param = "amp",
  comparison_A = 0,
  comparison_B = 1,
  comparison_type = "levels",
  component_index = 1,
  level_index = 0,
  ci_level = 0.95
)

Arguments

x

An object of class cosinor.glmm

x_str

Character naming the covariate whose amplitude/acrophase will be tested

param

Character string naming the parameter to test, either "amp" for amplitude or "acr" for acrophase

comparison_A

A number, referring to the level (within a group) or component number that is to be compared to comparison_B

comparison_B

A number, referring to the level (within a group) or component number that is to be compared to comparison_A

comparison_type

A string that is either: "levels" (default), or "components". If "levels", then comparison_A and comparison_B will refer to the two levels that are being compared. If comparison_type = "components", then comparison_A component will be compared to comparison_B component

component_index

If comparison_type = "levels", this controls which single component the levels are being compared to. Note that component_index must be an integer, and must refer to a component within the model

level_index

If comparison_type = "components", this controls which single level the components are being compared to. Note that level_index must be an integer, and must refer to a level within the model

ci_level

The level for calculated confidence intervals. Defaults to 0.95.

Examples

fit <- cosinor.glmm(Y ~ X + amp.acro(time,
  group = "X",
  n_components = 1,
  period = 12
), data = vitamind)
test_cosinor(fit, "X", "amp")
#> Global test: 
#> Statistic: 
#> [1] 3.22
#> 
#> 
#>  P-value: 
#> [1] 0.0728
#> 
#>  Individual tests: 
#> Statistic: 
#> [1] 1.79
#> 
#> 
#>  P-value: 
#> [1] 0.0728
#> 
#>  Estimate and confidence interval[1] "1.83 (-0.17 to 3.83)"