Checks the validity of user inputs and creates formula and modifies dataframe

Checks the validity of user inputs and creates formula and modifies dataframe

Usage

amp_acro(time_col, n_components = 1, group, period, ...)

Arguments

time_col

A numeric column within the data.frame() passed by via the data arg containing the time values.

n_components

The Number of cosinor components in the model.

group

A vector of the names for the group factors (column names within the data.frame() passed by via the data arg).

period

A numeric value or vector containing the period. The number of values should be equal to n_components.

...

Extra arguments for use within GLMMcosinor.

Value

A data.frame and formula appropriate for use by data_processor().

Examples

# Single component cosinor model
cosinor.glmm(
  Y ~ amp_acro(time_col = time, group = "X", period = 12),
  data = vitamind
)
#> Warning: Package version inconsistency detected.
#> TMB was built with Matrix version 1.5.4.1
#> Current Matrix version is 1.5.4
#> Please re-install 'TMB' from source using install.packages('TMB', type = 'source') or ask CRAN for a binary version of 'TMB' matching CRAN's 'Matrix' package
#> 
#>  Conditional Model 
#> 
#>  Raw formula: 
#> Y ~ X:main_rrr1 + X:main_sss1 
#> 
#>  Raw Coefficients: 
#>              Estimate
#> (Intercept)  30.32687
#> X0:main_rrr1  0.86520
#> X1:main_rrr1  6.47625
#> X0:main_sss1  6.24438
#> X1:main_sss1  4.66701
#> 
#>  Transformed Coefficients: 
#>             Estimate
#> (Intercept) 30.32687
#> [X=0]:amp    6.30404
#> [X=1]:amp    7.98266
#> [X=0]:acr   -1.43312
#> [X=1]:acr   -0.62444

# 2-component cosinor model with simulated data
sim_data <- simulate_cosinor(
  n = 500,
  mesor = 5,
  amp = c(2, 1),
  acro = c(1, 1.5),
  beta.mesor = 2,
  beta.amp = c(2, 1),
  beta.acro = c(1, 1.5),
  family = "gaussian",
  period = c(12, 6),
  n_components = 2,
  beta.group = TRUE,
)

cosinor.glmm(
  Y ~ group + amp_acro(times,
                       n_components = 2,
                       group = "group",
                       period = c(12, 6)),
  data = sim_data,
  family = gaussian
)
#> 
#>  Conditional Model 
#> 
#>  Raw formula: 
#> Y ~ group + group:main_rrr1 + group:main_sss1 + group:main_rrr2 +      group:main_sss2 
#> 
#>  Raw Coefficients: 
#>                  Estimate
#> (Intercept)       4.99387
#> group1           -2.96801
#> group0:main_rrr1  1.13785
#> group1:main_rrr1  1.08017
#> group0:main_sss1 -1.60315
#> group1:main_sss1 -1.65632
#> group0:main_rrr2  0.12819
#> group1:main_rrr2  0.01933
#> group0:main_sss2 -0.99023
#> group1:main_sss2 -1.01564
#> 
#>  Transformed Coefficients: 
#>                Estimate
#> (Intercept)     4.99387
#> [group=1]      -2.96801
#> [group=0]:amp1  1.96590
#> [group=1]:amp1  1.97741
#> [group=0]:amp2  0.99849
#> [group=1]:amp2  1.01582
#> [group=0]:acr1  0.95355
#> [group=1]:acr1  0.99291
#> [group=0]:acr2  1.44206
#> [group=1]:acr2  1.55177