Fit cosinor model with {glmmTMB}

Given an outcome and time variable, fit the cosinor model with optional covariate effects.

Usage

cosinor.glmm(
  formula,
  data,
  family = stats::gaussian(),
  quietly = TRUE,
  dispformula = ~1,
  ziformula = ~0,
  ...
)

Arguments

formula

Formula specifying the model. Indicate the time variable with time() and covariate effects on the amplitude and acrophase with amp.acro(). See details for more information.

data

Data frame where variable can be found

family

a family function, see ?family and ?glmmTMB::nbinom2

quietly

controls whether messages from amp.acro are displayed. TRUE by default

dispformula

Formula specifying a dispersion model (optional). Use the same format as the main formula

ziformula

Formula specifying a zero-inflation model (optional). Use the same format as the main formula

...

optional additional arguments passed to glmmTMB::glmmTMB()

Details

This defines special functions that are used in the formula to indicate the time variable and which covariates effect the amplitude. To indicate the time variable wrap the name of it in the function time(). To indicate a variable which affects the acrophase/amplitude, wrap the name in amp.acro(). This will then do all the transformations for you. See examples for usage.

References

Tong, YL. Parameter Estimation in Studying Circadian Rhythms, Biometrics (1976). 32(1):85--94.

Examples

cosinor.glmm(Y ~ X + amp.acro(time,
  n_components = 3,
  group = "X",
  period = c(12, 8, 9)
), data = vitamind)
#> Warning: Package version inconsistency detected.
#> TMB was built with Matrix version 1.5.3
#> Current Matrix version is 1.5.1
#> Please re-install 'TMB' from source using install.packages('TMB', type = 'source') or ask CRAN for a binary version of 'TMB' matching CRAN's 'Matrix' package
#> 
#>  Conditional Model 
#> 
#>  Raw formula: 
#> Y ~ X + X:main_rrr1 + X:main_sss1 + X:main_rrr2 + X:main_sss2 + 
#>     X:main_rrr3 + X:main_sss3
#> <environment: 0x5578bce70b10>
#> 
#>  Raw Coefficients: 
#>  (Intercept)           X1 X0:main_rrr1 X1:main_rrr1 X0:main_sss1 X1:main_sss1 
#>   29.7207882    5.0427216    0.5449958   25.7817398    2.1843787    4.2067454 
#> X0:main_rrr2 X1:main_rrr2 X0:main_sss2 X1:main_sss2 X0:main_rrr3 X1:main_rrr3 
#>    8.3651420   -0.4844171   -3.6044821   24.2788845   -8.3800373  -20.4904692 
#> X0:main_sss3 X1:main_sss3 
#>    8.7132257  -34.3472358 
#> 
#>  Transformed Coefficients: 
#> (Intercept)       [X=1]  [X=0]:amp1  [X=1]:amp1  [X=0]:amp2  [X=1]:amp2 
#>  29.7207882   5.0427216   2.2513398  26.1226877   9.1086713  24.2837166 
#>  [X=0]:amp3  [X=1]:amp3  [X=0]:acr1  [X=1]:acr1  [X=0]:acr2  [X=1]:acr2 
#>  12.0890582  39.9948989  -1.3262912  -0.1617423   0.4068516  -1.5907459 
#>  [X=0]:acr3  [X=1]:acr3 
#>  -2.3367046   2.1086886