Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The MESOR is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>                Y        times group
#> 1    1.118715357 23.307355938     0
#> 2   -0.103878502 15.588657975     0
#> 3    1.137270195 10.604167404     0
#> 4    2.333031014  7.563537106     0
#> 5    0.483430955 14.966297764     0
#> 6    2.644885834  1.654893642     0
#> 7    0.534209643 15.833686542     0
#> 8    1.195359956 21.883315204     0
#> 9    2.954548170  6.645005967     0
#> 10   1.181192468 18.290936150     0
#> 11   0.799680537 19.497729890     0
#> 12   0.100115711 16.197078612     0
#> 13   0.112808009 11.399509326     0
#> 14   0.422277682 23.045049638     0
#> 15   1.706684547  3.206493750     0
#> 16   0.931573362 12.579467336     0
#> 17   1.275477937 23.547824521     0
#> 18   2.256725850  8.578690790     0
#> 19  -0.735112665 16.202093868     0
#> 20   2.258243767  3.804089466     0
#> 21  -0.488510522 16.814833162     0
#> 22   0.029378958  9.145869942     0
#> 23   0.238153189 21.294674428     0
#> 24   0.498477115 17.682378810     0
#> 25   1.017148997 19.917992977     0
#> 26   2.043932175  4.162056757     0
#> 27   2.706275363  0.009387754     0
#> 28   1.480402546  0.124466887     0
#> 29   1.050169589 11.110656537     0
#> 30  -0.167066260 15.970299717     0
#> 31   1.031590058  6.014957964     0
#> 32   2.049689470  0.211856833     0
#> 33   1.444088258 21.473751588     0
#> 34   1.454099626  8.225948609     0
#> 35   1.218879379  7.955259478     0
#> 36   0.594966960 22.406219764     0
#> 37   1.314305192  1.463257598     0
#> 38   0.114106593  7.674848182     0
#> 39   2.426139764  4.646033389     0
#> 40   0.570013644  7.682109922     0
#> 41   1.297105840  2.887757096     0
#> 42   0.443359144 19.037282171     0
#> 43   0.068314088 22.085168296     0
#> 44   0.942157423 23.514017500     0
#> 45   3.008671379  3.401864223     0
#> 46   3.110033761  1.977848159     0
#> 47   2.027656068 23.027781023     0
#> 48   1.029207446  8.760611694     0
#> 49   3.010900930  7.397289271     0
#> 50  -1.201106367  7.753264518     0
#> 51   2.389963933 10.901369076     0
#> 52  -0.002781056  9.614635941     0
#> 53   2.780717150  0.589050656     0
#> 54   2.189764113 19.763693249     0
#> 55   1.264715251 19.187301317     0
#> 56   0.167317552 16.495629301     0
#> 57   2.449068165  2.991735443     0
#> 58   1.033000595 20.265071908     0
#> 59   1.655212388 23.045284716     0
#> 60   0.561741538 20.752476130     0
#> 61  -0.168152686 16.745413972     0
#> 62   0.228059213 17.375500018     0
#> 63   1.092890319  5.768056529     0
#> 64   1.388555239 18.178668480     0
#> 65   1.618727509  3.981270500     0
#> 66  -0.807613211 13.259617494     0
#> 67   0.464685115 12.028695133     0
#> 68  -0.484045103  6.400715081     0
#> 69   1.871010206  4.470947256     0
#> 70   0.373026758 12.930216389     0
#> 71   1.741990984  8.680496424     0
#> 72  -0.811167855 15.634385247     0
#> 73   0.645078277 19.965143675     0
#> 74   1.879615940  3.119572073     0
#> 75   1.157253530  4.776534911     0
#> 76   1.460004918  6.119932909     0
#> 77   1.811156039  5.655370943     0
#> 78  -0.525621368 18.169909755     0
#> 79   0.994488402 23.808048304     0
#> 80   1.452241433 23.111432713     0
#> 81   2.538605857  6.793954181     0
#> 82   3.961007480  1.416229092     0
#> 83   1.315096841  1.461871438     0
#> 84  -1.682676337 20.075207910     0
#> 85   2.451869374 20.796379214     0
#> 86  -0.073015184 22.922577174     0
#> 87   0.989029162  9.454076707     0
#> 88   1.112152774  0.422925163     0
#> 89  -0.329303693 11.563132446     0
#> 90   1.623649855  4.854863930     0
#> 91   1.469600032 15.425612541     0
#> 92   1.727940510 10.239455771     0
#> 93   0.365675425  0.089700032     0
#> 94   1.903002615 11.234568033     0
#> 95   1.345766344  6.460895944     0
#> 96   0.542535943 12.426812131     0
#> 97  -0.269859878 15.560159896     0
#> 98   1.484828511  3.247671805     0
#> 99   1.992770229  2.128893837     0
#> 100  1.231726927 11.899557687     0