Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The MESOR is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times group
#> 1    2.57970533  5.7580473     0
#> 2    1.00498811  9.5521061     0
#> 3    1.13008351 19.3264005     0
#> 4    1.14282797 23.4794543     0
#> 5    0.28888349 23.5951444     0
#> 6    2.66832220 23.2900599     0
#> 7    0.94095433 20.1158381     0
#> 8    0.41936717  5.6023091     0
#> 9    1.18973982 21.9937280     0
#> 10   1.69303326  9.2055057     0
#> 11   1.36025664 18.0650365     0
#> 12  -0.59174656 16.5583126     0
#> 13   2.80232804 23.3523202     0
#> 14   1.28534960  3.7268924     0
#> 15   0.62809795 16.1766986     0
#> 16   2.10615208  7.4226476     0
#> 17  -0.43054202 21.5016592     0
#> 18   3.39295313  5.9459088     0
#> 19   1.47115194 20.4153398     0
#> 20   0.21589853 18.9030560     0
#> 21   1.80024557  3.2837151     0
#> 22   2.69306600 23.6922216     0
#> 23   2.66076967  6.7127390     0
#> 24   2.73988061  3.7350080     0
#> 25   2.67114003  8.7541906     0
#> 26   1.57072611 21.1672753     0
#> 27  -0.41497852  7.0371320     0
#> 28  -1.09663285 12.9189456     0
#> 29   1.95405626  3.2544525     0
#> 30   3.23675620  2.9883294     0
#> 31   0.84819625  9.1642114     0
#> 32   1.13780128 22.6426770     0
#> 33   1.02938360  4.9199464     0
#> 34   0.73663423  4.1597285     0
#> 35   1.11124139  7.9434590     0
#> 36   2.77765484  5.3006211     0
#> 37   2.56641372  4.4823807     0
#> 38   0.19763462  6.0551415     0
#> 39   0.75982895 16.9706592     0
#> 40  -0.65792606 10.4436032     0
#> 41   0.72631480 17.9649951     0
#> 42   0.43370060 17.1668933     0
#> 43  -0.85921753 20.4211476     0
#> 44  -0.04207306 15.7018660     0
#> 45   1.71916873  9.4559514     0
#> 46   0.51283084 17.7411486     0
#> 47  -0.65716056 17.9470592     0
#> 48  -0.04736022 15.2729492     0
#> 49  -0.29186043 20.6678852     0
#> 50   0.71729122 18.4737465     0
#> 51  -0.73458620 20.7394341     0
#> 52   0.87398110 15.8846318     0
#> 53   0.58796092 14.3995265     0
#> 54   0.42185396  2.3806772     0
#> 55  -0.82162237 20.3227558     0
#> 56   1.02941563 22.3573243     0
#> 57  -0.07611193  4.0682940     0
#> 58   1.83104990  8.0920128     0
#> 59   0.07133756 19.0063578     0
#> 60  -0.06885650 14.3762609     0
#> 61   2.37287768 20.4753186     0
#> 62   1.97924494  3.5514172     0
#> 63   1.79693521  7.1444596     0
#> 64  -0.50957009 14.5977605     0
#> 65   1.09029080 12.5298473     0
#> 66  -0.29185220 16.9958312     0
#> 67  -0.17812174 19.7492676     0
#> 68   1.72445200 21.5595623     0
#> 69  -0.59958396 11.2102054     0
#> 70   1.99659536 11.0486347     0
#> 71   0.88062431 15.9190839     0
#> 72   1.71172545  7.1078273     0
#> 73   2.60775683  8.2871988     0
#> 74   0.39084507 15.5500810     0
#> 75   1.64015917  7.5951363     0
#> 76  -0.78113180 21.2816229     0
#> 77   2.40818654  9.6296166     0
#> 78   1.55423568  9.8946647     0
#> 79   0.57738509 21.6915728     0
#> 80   0.93255660 16.3095282     0
#> 81   0.11721898 15.5395257     0
#> 82   2.23318652  4.8157048     0
#> 83   2.80197461 12.7233320     0
#> 84  -0.60553755 18.9735624     0
#> 85   3.10310397  4.3465106     0
#> 86  -0.97750825 15.6928445     0
#> 87   0.90150599  9.5350265     0
#> 88   1.96904882  2.1695398     0
#> 89  -0.41936359 18.9846835     0
#> 90   0.75343441 10.2472998     0
#> 91   1.57236011  7.3114209     0
#> 92   0.49654354  2.8403551     0
#> 93   2.22483362 23.2141114     0
#> 94   1.03369289  8.6983211     0
#> 95   1.67972238  0.4566284     0
#> 96  -0.22546400 18.8354546     0
#> 97   2.82223762  0.2343327     0
#> 98   0.97810596 12.8369478     0
#> 99   1.62289285 21.0952354     0
#> 100  3.08322555  2.2577937     0