Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The MESOR is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times group
#> 1    0.23926334  8.1895274     0
#> 2    0.97926844 19.4187960     0
#> 3    0.87599723 16.0558038     0
#> 4    0.66108155 12.7094059     0
#> 5    1.03165921  2.5391512     0
#> 6    0.01378953 10.9423951     0
#> 7    0.74982248  5.2575055     0
#> 8   -1.40811315  8.4587611     0
#> 9   -0.56292552 22.5829187     0
#> 10   2.20764044 11.0287348     0
#> 11   3.65597583  6.5032001     0
#> 12   1.68695662  0.9718913     0
#> 13   0.70883531  0.4434958     0
#> 14   0.35521062  9.3359375     0
#> 15   1.15821555  0.3880835     0
#> 16  -0.02859380 12.1668656     0
#> 17   2.95661761 21.6209097     0
#> 18   0.79404712 22.7766431     0
#> 19  -0.60337625 12.5618542     0
#> 20   0.47258903 23.0075372     0
#> 21   1.21433245  5.5187392     0
#> 22   0.88160002 21.6228768     0
#> 23   1.25491762 10.8395902     0
#> 24   1.53701778 18.7278172     0
#> 25   1.63232088 22.3784011     0
#> 26  -1.16678907 18.4277357     0
#> 27   0.98723160 11.7498221     0
#> 28   1.30119565 21.9349466     0
#> 29   1.94469618  6.1905360     0
#> 30   0.96845244 21.4638350     0
#> 31   1.21426255 20.8974198     0
#> 32   1.49748646 10.9058642     0
#> 33   2.84710646  0.7083484     0
#> 34  -0.40256359 14.9768154     0
#> 35   2.91879246  3.3135888     0
#> 36   1.53207384 12.3814956     0
#> 37  -0.08491005 14.4673719     0
#> 38   0.78648278 13.8244003     0
#> 39   1.62332288 21.5486748     0
#> 40   2.78729247  4.6306019     0
#> 41   0.62546310 16.2142209     0
#> 42  -0.30745589 17.5784638     0
#> 43   2.08548290  7.0582283     0
#> 44   1.86568912 21.7471747     0
#> 45   2.32697961 22.4023040     0
#> 46   0.88177539 11.6235125     0
#> 47   1.68039565 19.9780484     0
#> 48  -0.89523507 12.6792195     0
#> 49   2.77862293 11.2421502     0
#> 50   0.85280065 20.3840734     0
#> 51   1.30804828 16.2604412     0
#> 52   0.11948729 15.9817484     0
#> 53   1.77009817  3.3673148     0
#> 54   0.93200971 16.2189238     0
#> 55   1.00066411 11.0699819     0
#> 56   0.70990761  5.5065691     0
#> 57   0.70747301 14.0720926     0
#> 58   2.34555403  1.6918183     0
#> 59  -1.08616487 15.2207552     0
#> 60   1.93478945  0.6237983     0
#> 61   0.66051046 18.6613474     0
#> 62   1.62084492  2.3691987     0
#> 63  -1.47414884 12.7179314     0
#> 64   0.67615289 11.8799118     0
#> 65   0.11336657  7.8063253     0
#> 66   1.08679374  6.7924879     0
#> 67  -0.90546063 19.9158442     0
#> 68   0.25837296 17.1974702     0
#> 69   3.36292392  2.9252072     0
#> 70   0.66107107 14.6280877     0
#> 71   0.47005187 15.1579072     0
#> 72  -1.15330517 19.9627793     0
#> 73   1.44219539 21.9033227     0
#> 74   1.76333951  8.5037751     0
#> 75  -0.30749645 13.2640575     0
#> 76  -0.47270201 19.3485403     0
#> 77  -1.87218982 16.5547652     0
#> 78  -0.13656787 18.0102980     0
#> 79   2.47802273  5.1199255     0
#> 80   0.26189518  6.5895729     0
#> 81   3.37721293  1.3177446     0
#> 82  -0.90012083 21.1447460     0
#> 83   1.93468913  0.8213415     0
#> 84  -0.51673267 23.0407149     0
#> 85   1.72027153  1.1547553     0
#> 86   1.00705074  9.6338877     0
#> 87  -0.42397346  5.3410213     0
#> 88  -1.59988711 13.6414174     0
#> 89   1.01488734 11.9833092     0
#> 90   0.55908878 18.2843837     0
#> 91  -0.72232778 18.9246688     0
#> 92  -0.14410376 19.9048306     0
#> 93   2.14768207  1.6694659     0
#> 94   1.50773424  9.2040931     0
#> 95  -0.43932875 19.1003354     0
#> 96   1.62893207 16.7778680     0
#> 97   0.23439233 10.8147840     0
#> 98   0.70314885 14.6208471     0
#> 99  -3.09596063 15.2733241     0
#> 100  0.61114266 10.1992652     0