Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1.

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times group
#> 1    2.16177689 20.6619165     0
#> 2    0.22610369 18.9987341     0
#> 3    0.66230880 23.2095503     0
#> 4    1.29358460 21.5270081     0
#> 5    0.77419555  8.8357118     0
#> 6   -0.61239350 15.3787253     0
#> 7    0.96989672 23.1710933     0
#> 8    2.73781469  6.6420810     0
#> 9    0.82817827  0.5695319     0
#> 10   3.20068931  3.5246054     0
#> 11   0.88736070 12.6988034     0
#> 12   0.86137499 23.5490651     0
#> 13   1.91114746  4.7509709     0
#> 14   2.13589284  1.9791413     0
#> 15   0.35538445 19.0797668     0
#> 16   0.32374432 18.5924408     0
#> 17   0.69394585 22.5492875     0
#> 18   1.57146273  9.0152308     0
#> 19   3.32148523  1.5736005     0
#> 20   1.11532310  8.8694494     0
#> 21   0.24763282 22.5808187     0
#> 22   1.67991219 10.5096839     0
#> 23   0.55429183 18.3373553     0
#> 24  -0.20777849 19.8813665     0
#> 25   1.06783976  5.4677245     0
#> 26  -0.73786831 11.5895283     0
#> 27   0.50929140 17.6709571     0
#> 28  -1.23423061 13.8322242     0
#> 29   0.96202148 12.6630170     0
#> 30   0.58049935  4.1698456     0
#> 31   2.99428191  0.1511985     0
#> 32   2.14893341  1.4472890     0
#> 33   1.37442339  6.1298031     0
#> 34   2.32672642  8.5258949     0
#> 35  -1.18260645 16.8058039     0
#> 36  -0.36996083 17.9087781     0
#> 37  -0.05866179 10.2817774     0
#> 38   1.88403241 11.3664799     0
#> 39   4.51787513  3.2715571     0
#> 40   1.07889697 22.6460454     0
#> 41   1.71012595 22.0064264     0
#> 42   0.61837086 18.5041110     0
#> 43  -0.44389792 15.6766827     0
#> 44   0.80153463  5.1363262     0
#> 45   2.12264804  2.6553196     0
#> 46   0.24867222  8.0715807     0
#> 47   4.43786205 23.6403116     0
#> 48  -2.08932277 18.3867989     0
#> 49  -1.34368325 14.9765454     0
#> 50   2.53083821 19.9988820     0
#> 51   2.21231808 21.8991499     0
#> 52   1.01036185  9.2068432     0
#> 53   1.32488896 23.1399051     0
#> 54   1.61381607  2.3187542     0
#> 55  -0.20735354 13.6384872     0
#> 56   0.31163477 22.3032544     0
#> 57   1.91162643 19.4839388     0
#> 58  -0.74742717 16.1190456     0
#> 59   1.19016398  6.7753898     0
#> 60   0.19950438 13.5787747     0
#> 61  -0.37079720 18.9916865     0
#> 62   0.56692834 10.5958047     0
#> 63   1.12460440  3.1735473     0
#> 64   0.19102312  8.7338028     0
#> 65  -0.68474029 15.3216616     0
#> 66   0.05001478 19.3540121     0
#> 67   1.90807621  5.2032756     0
#> 68  -1.99060804 16.1020611     0
#> 69   0.45416957 22.2295630     0
#> 70   0.89028561 10.0794340     0
#> 71   0.40242009 23.3481769     0
#> 72  -1.90587398 12.2366305     0
#> 73   1.78540015 22.7383215     0
#> 74   0.73670488  3.5154651     0
#> 75  -1.16049311 21.0797251     0
#> 76   0.17413901 21.4266434     0
#> 77   1.44813621 16.9547560     0
#> 78   0.86229878  0.6670497     0
#> 79   2.70237152  0.2373419     0
#> 80   2.68204089  1.7016259     0
#> 81   2.58971334  0.5592352     0
#> 82   1.34952382 13.1809054     0
#> 83   0.39285650 23.0763614     0
#> 84   2.76014159  6.0898540     0
#> 85   0.96268519  9.1579621     0
#> 86  -0.39143561 17.6875624     0
#> 87   0.12602024 16.5213627     0
#> 88   3.29358410  1.4311437     0
#> 89   1.63364882 23.7989794     0
#> 90   1.90438751  5.8074880     0
#> 91   2.78698313 23.4345458     0
#> 92   0.08198808 16.1160765     0
#> 93   1.01239011  0.2789891     0
#> 94   0.66559694 16.4712163     0
#> 95   2.72768292 10.3762871     0
#> 96   0.65274555  6.8259320     0
#> 97   1.52158712  1.5803623     0
#> 98   2.51688179  3.8485685     0
#> 99   1.95751187  2.8789346     0
#> 100  1.29582731  4.8598499     0