Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The MESOR is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>              Y      times group
#> 1    0.8043414 21.5982982     0
#> 2    0.2495838 20.6356364     0
#> 3    2.2215251  3.1987670     0
#> 4    0.2998532 19.8534591     0
#> 5    2.2832527  9.3898278     0
#> 6    1.9074412  1.8989769     0
#> 7    1.3862551 18.9500792     0
#> 8    1.0856388 18.5725407     0
#> 9    1.6552255  3.1769269     0
#> 10   0.5499579  8.5680985     0
#> 11   3.3091639  6.9310916     0
#> 12   1.5402826 16.3764046     0
#> 13  -0.6734292 17.5866830     0
#> 14   0.2957660  7.6149926     0
#> 15  -0.9340002 18.4892873     0
#> 16   1.0948949 22.0894148     0
#> 17  -0.4719248 13.2453963     0
#> 18   0.4911591 23.8363995     0
#> 19  -0.9128556 19.7834706     0
#> 20  -1.9067043 14.4962468     0
#> 21   0.1070160 12.8548367     0
#> 22  -0.7707715 14.3987289     0
#> 23   0.9099690  9.5264648     0
#> 24   2.2706209 21.4841940     0
#> 25   1.4304164 11.6101130     0
#> 26   1.5690705  0.5136127     0
#> 27   0.9011493  8.7040177     0
#> 28   1.5279545  0.5850845     0
#> 29   2.1502058 10.2517036     0
#> 30   2.7557499 20.4031108     0
#> 31   1.2485708  4.8508008     0
#> 32   4.1136479  7.4632424     0
#> 33  -0.5619389 19.5698965     0
#> 34   2.6060267 22.7163945     0
#> 35   1.8003660 10.1436854     0
#> 36   1.6835689  6.6530983     0
#> 37   2.1762864  3.4124814     0
#> 38  -1.5867242 12.5705302     0
#> 39  -0.9146548 18.1121169     0
#> 40   1.0031845  7.9838280     0
#> 41   1.3947962  1.6902572     0
#> 42   1.9474283  4.6091181     0
#> 43  -0.7880009 12.0545445     0
#> 44   2.2209170  7.4195259     0
#> 45   2.3798139 20.7232957     0
#> 46   0.5171460 16.8889182     0
#> 47   1.5922796 15.9607435     0
#> 48   2.3566702  5.7927608     0
#> 49   0.4690497 13.8067812     0
#> 50   1.1112462  8.6925096     0
#> 51   0.6907750  8.2900418     0
#> 52   0.6335003 20.9216756     0
#> 53   1.3421023 21.2047497     0
#> 54   2.0705191 23.1402780     0
#> 55   2.2443678 10.4268556     0
#> 56   0.5589113  9.7461010     0
#> 57   0.4476833  7.0769419     0
#> 58   1.2126329 20.6757397     0
#> 59   2.6975845  2.7778975     0
#> 60  -0.9917728 18.5516058     0
#> 61   0.6287663 19.3190353     0
#> 62   1.6864940 22.3560637     0
#> 63  -0.7548671 14.3290143     0
#> 64   0.6271585  5.6869987     0
#> 65   2.7379628  4.0182723     0
#> 66   1.7346138  3.2701020     0
#> 67  -0.3072520 14.0859314     0
#> 68   0.6866176 13.3151993     0
#> 69   1.3814777  7.8139891     0
#> 70   0.9311461  5.5247954     0
#> 71  -0.1348380 15.2407496     0
#> 72   1.8902430 23.5224412     0
#> 73   0.6545317  1.5661247     0
#> 74   2.2845131  4.1930941     0
#> 75   0.0513578 21.4065482     0
#> 76   2.8954751  9.3225344     0
#> 77   1.4832213 22.4861062     0
#> 78   0.1985430  8.9111688     0
#> 79   1.9517426  3.0238191     0
#> 80   0.5472955 11.7516118     0
#> 81   0.4354748 11.0185712     0
#> 82   1.6920768 19.1627116     0
#> 83  -1.3437120 13.6853225     0
#> 84   1.1806748  5.0646212     0
#> 85   0.4368060 11.8976234     0
#> 86   0.5593880 10.6524301     0
#> 87  -0.2912780 19.3500084     0
#> 88   2.0386985 23.9352898     0
#> 89   1.9735560 23.8388182     0
#> 90   2.2814249  6.1232602     0
#> 91  -0.9158422 12.6251285     0
#> 92   1.1117623 20.5380638     0
#> 93   0.8762875 20.0959561     0
#> 94   2.2055856  9.4613249     0
#> 95   0.9164745  8.3787616     0
#> 96   2.2238823 11.6198900     0
#> 97  -0.8535245 19.1150294     0
#> 98   0.7725850 19.2639701     0
#> 99   4.1441852  2.5602866     0
#> 100  0.8913666  3.8711140     0