Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1.

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1).

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times
#> 1    1.45507694 23.4111744
#> 2   -0.48620511  6.8251317
#> 3    2.29562810  5.8272642
#> 4    0.39678529  4.9528162
#> 5    0.61754243 16.2080800
#> 6    1.45862679 20.7365065
#> 7    1.51215461 18.0602373
#> 8   -0.21942108  8.1567022
#> 9    1.21092717 15.2528741
#> 10   0.10230056  7.7202326
#> 11   0.86550535 22.6086356
#> 12   2.46274076 22.1145927
#> 13   0.73185529 17.9410304
#> 14   0.56461565  3.9251835
#> 15   0.67809347  4.1991279
#> 16   1.75830736  1.6288436
#> 17   0.42927062 22.1009919
#> 18   1.22616612 16.7539158
#> 19   0.62511435 15.8566712
#> 20  -0.66667529  7.6538533
#> 21   0.65884765 18.0082859
#> 22   2.19749779 22.9956011
#> 23   0.89732642  6.8080832
#> 24   1.87546642 15.2671386
#> 25   0.64687833  3.5967546
#> 26   1.32951408 11.4711226
#> 27   1.91033549 23.0702532
#> 28  -0.75286932 12.0764650
#> 29   3.90901593 18.8746834
#> 30   2.32250455 23.4679399
#> 31   0.37047721 10.0691603
#> 32   0.41750510  2.5053500
#> 33   1.88709849 18.2956321
#> 34  -0.11299979 11.1230464
#> 35  -0.25955448 13.7777700
#> 36   1.30196050  7.0241591
#> 37   0.72627256 15.6806670
#> 38   1.32708099 17.3955246
#> 39  -1.13722863 16.5737643
#> 40  -0.26971190  3.4534786
#> 41   1.59442700 19.7168337
#> 42  -0.56753088  2.7766934
#> 43   0.33721682  8.9309511
#> 44  -0.43560666 17.0634276
#> 45   1.20150210 17.5385816
#> 46  -0.73943500  7.9752103
#> 47   0.93117559 23.6005014
#> 48  -1.23148211  2.7632282
#> 49   0.16217042 10.6446223
#> 50  -1.23010220  7.6931894
#> 51   0.11254378 17.3665491
#> 52  -0.88416975  7.0822536
#> 53   0.66831614 15.2051913
#> 54   1.26269204  1.5874438
#> 55   1.23312137 15.5695442
#> 56   3.40026508 16.5891604
#> 57   0.65581054  2.6765299
#> 58  -0.05384297  6.8847277
#> 59  -1.45995683  1.7682500
#> 60  -0.31186821  7.2940697
#> 61  -0.66179415  8.7391214
#> 62   2.61039424 14.5308551
#> 63   1.92281643  9.5885363
#> 64   1.15002605 22.3333324
#> 65   1.76333614 15.7453210
#> 66   0.15495886  2.4099837
#> 67   0.98778202  8.9080245
#> 68   1.88520453  1.5106107
#> 69   1.94027922 18.8165968
#> 70  -0.63844884  4.3972733
#> 71   0.86219962 11.8842580
#> 72   0.09787110 10.5377044
#> 73  -0.29447105  9.1611488
#> 74   0.68073914 10.9600765
#> 75   2.91847286 15.9196043
#> 76  -0.16077893 11.6795634
#> 77   1.29041818  0.6776369
#> 78   2.97104920 20.5263431
#> 79   0.54532861 16.6846398
#> 80   0.15106740  5.9744799
#> 81   2.34532685 15.7561761
#> 82   0.29898554 14.1597182
#> 83   0.15523038  4.8217970
#> 84  -1.18323678  9.2772690
#> 85  -0.96274268 23.8617031
#> 86   3.62236657 19.7832813
#> 87  -0.14805834 10.4287100
#> 88   1.93504388  8.9329237
#> 89   2.59716745 14.5334690
#> 90   0.88418575 10.4417778
#> 91   0.90602902 19.2377194
#> 92  -1.43018282 10.4270332
#> 93   0.80146980  8.9075057
#> 94   2.98552232 15.9765723
#> 95   1.41217500 19.3586989
#> 96   1.71570910  8.1915971
#> 97  -0.11299975  7.1575126
#> 98   1.79610675  0.4647344
#> 99   2.24130388  0.8908404
#> 100  1.27992922 16.0771058