Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The MESOR is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times group
#> 1    1.33053353  8.1584779     0
#> 2    2.18429934 21.5440220     0
#> 3    0.78939423 18.0479531     0
#> 4    0.73962878  0.2661837     0
#> 5   -0.31128686 14.8939224     0
#> 6    3.35336038  2.3776547     0
#> 7   -0.27838833 15.0332382     0
#> 8    1.42454522 19.0016298     0
#> 9   -0.55687510 11.6311602     0
#> 10   0.88576071 23.1057277     0
#> 11  -0.54563297 21.4104188     0
#> 12   1.70069735 22.2183711     0
#> 13   1.49634426 12.2004068     0
#> 14   3.43260299  1.5278441     0
#> 15   2.00316854  4.8441309     0
#> 16   0.62752785 12.3337590     0
#> 17   1.97727566 11.4344429     0
#> 18   0.63379286  0.1572320     0
#> 19   1.18801197  3.5995518     0
#> 20   1.79575742  6.2758171     0
#> 21   2.41983540  3.5808582     0
#> 22   0.65179781 14.0093093     0
#> 23   1.00489748  0.9746472     0
#> 24   0.05193273 23.3159818     0
#> 25   1.98455354  0.8397676     0
#> 26   0.93120407 22.7000228     0
#> 27   3.19902379 23.5560010     0
#> 28   0.87378727  3.7021794     0
#> 29  -0.69409929 11.7705908     0
#> 30   0.41736398  2.8104537     0
#> 31  -1.09420147 15.6023273     0
#> 32   1.38638097  8.7201044     0
#> 33   2.60628654  6.6722741     0
#> 34  -0.83772409 12.7346681     0
#> 35  -0.32758448 12.2032383     0
#> 36   0.19585173 22.0883991     0
#> 37   2.31736532 22.3211672     0
#> 38   0.65690774 20.4645768     0
#> 39   0.81337050 15.5761700     0
#> 40   2.50028171  4.3955931     0
#> 41  -0.49362572 14.1648723     0
#> 42   0.74033220 17.7017891     0
#> 43   0.25892800 20.4138483     0
#> 44   3.25997078  0.2246742     0
#> 45   0.69212763  4.3133901     0
#> 46  -0.57056907 16.8235254     0
#> 47   1.53040150  5.4040393     0
#> 48   0.15986890  8.3155850     0
#> 49  -0.96178334 17.7814098     0
#> 50   3.23711479  0.7183574     0
#> 51  -1.11881899 22.8254133     0
#> 52   0.23470192 18.4359192     0
#> 53   1.27079234 20.2694658     0
#> 54  -1.18515783 16.2253208     0
#> 55   0.37872291 13.6392963     0
#> 56   1.56488216  2.3911443     0
#> 57  -1.96957555 19.2850162     0
#> 58  -0.39554010 17.9422651     0
#> 59   0.70142411  5.9081830     0
#> 60   1.31206925  4.6793299     0
#> 61   1.21886795  7.3618984     0
#> 62   0.07072782 12.0410886     0
#> 63  -0.67787639 21.0345177     0
#> 64  -0.56071085 22.0797271     0
#> 65   1.18036141 18.3964871     0
#> 66   1.07554443  1.8299380     0
#> 67   1.74400651 16.2362762     0
#> 68  -0.26588419 11.8232123     0
#> 69  -0.42997132 18.7828986     0
#> 70   2.64039108  2.6922319     0
#> 71   2.49266586 10.7157223     0
#> 72   3.86833188  4.0958549     0
#> 73   1.54591532  7.2604174     0
#> 74   0.35487357 23.4110365     0
#> 75   1.31082325 11.4233446     0
#> 76  -1.16351585 13.3483617     0
#> 77   2.91646765  3.4029446     0
#> 78   0.68731540 15.4017510     0
#> 79   0.17880881 23.7024347     0
#> 80   0.66388400  6.8890630     0
#> 81   0.94452903 20.1509485     0
#> 82  -1.03546918 23.2717120     0
#> 83   0.57149704 15.1445581     0
#> 84   0.04246989 22.2223893     0
#> 85   0.67623295 10.9567472     0
#> 86   1.06105778 21.4412344     0
#> 87  -0.79537693  9.2575428     0
#> 88  -0.78486115 17.7056430     0
#> 89   1.29630952  4.6241655     0
#> 90   0.18408294 19.7820966     0
#> 91   1.18719109 17.5680419     0
#> 92   0.17813714  8.5895897     0
#> 93   2.32275911 21.8240460     0
#> 94   1.68196822  1.4690826     0
#> 95   1.88762467  6.1480449     0
#> 96   0.58158894  1.1189284     0
#> 97   2.78505926  2.5778620     0
#> 98   2.74167704  6.9140974     0
#> 99   2.64266783  2.9586705     0
#> 100  0.67575063 15.5309885     0