Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1.

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1).

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times
#> 1    1.50471936 12.1291275
#> 2   -1.12559108  9.7404552
#> 3    1.95742520 14.5974748
#> 4    0.08892411  3.3684568
#> 5    2.40159993  0.1211957
#> 6    1.84008919 17.2727802
#> 7    0.51286071  9.3461538
#> 8   -0.50261542  6.4572863
#> 9   -0.99439253  4.5895083
#> 10   0.69010418  3.0328455
#> 11   3.25077668 20.6971497
#> 12   0.60905951  2.6832085
#> 13   0.21119979  9.2841636
#> 14   1.76043459 17.0498620
#> 15   2.16106313 20.1513647
#> 16   0.61499555  3.2605411
#> 17   1.67246213  2.3158520
#> 18   1.59936370 15.9009273
#> 19   1.75009802 18.6422349
#> 20   2.11288991 19.4762891
#> 21   0.56468016 21.0049344
#> 22   0.18226858  2.1010317
#> 23  -0.01240114 20.8702538
#> 24   0.83957103  8.5173329
#> 25   0.52532772 17.5554479
#> 26   2.83162062 20.0836937
#> 27   0.56221104  8.6783221
#> 28   1.34099980  8.8037639
#> 29   2.77097805 21.1108058
#> 30   2.38038094 14.5110319
#> 31   1.10660483 22.7382162
#> 32   1.58648582  8.5497225
#> 33   0.30645175  5.7574900
#> 34   1.01705361  8.6261066
#> 35   0.18920121 21.1260672
#> 36  -0.69774492  4.6858840
#> 37   1.91681994 18.3795112
#> 38   2.85014210 18.2704795
#> 39   1.77213460  5.4078975
#> 40   1.75756654 23.2544206
#> 41   1.82463857 15.1699698
#> 42  -1.43693880  3.7108051
#> 43   1.02219176 23.3868283
#> 44   2.12548960  2.0262156
#> 45   0.76997864 13.7156138
#> 46   1.66321723  0.8334854
#> 47   1.38924574 21.8796654
#> 48   0.56262558 10.1120305
#> 49   0.14396607 13.4896197
#> 50  -0.13754240 12.2485590
#> 51   2.18859789 20.2975756
#> 52  -1.06007361  7.5676209
#> 53   1.84100211 16.6994684
#> 54   1.65320300  2.9376436
#> 55   1.21732212 11.9354713
#> 56  -1.11005267  8.1475362
#> 57   0.25304947  9.5196638
#> 58   0.98019708  8.8451849
#> 59  -0.95384526  4.8213818
#> 60  -0.26011219 16.8330452
#> 61  -0.88383241  9.7726872
#> 62   2.34802348 18.1411139
#> 63   1.67954353 23.7952538
#> 64   1.14363301  7.4937812
#> 65  -0.58508950 10.5132090
#> 66   1.41904726 13.2908872
#> 67   1.08426106 16.0241175
#> 68   1.21207735  6.6963382
#> 69   0.97701554  9.1126239
#> 70   2.55449645 12.9644277
#> 71   2.55087575  1.6662015
#> 72   0.79398141  2.7259291
#> 73   0.41588617 19.0870498
#> 74   1.61689414 17.9070273
#> 75   2.48838620 20.6381512
#> 76  -0.54880612 14.8069260
#> 77  -0.52582815 11.0786079
#> 78   1.84931018 16.6819621
#> 79   0.12723196  3.0651787
#> 80   1.73321122 23.5459700
#> 81   3.57189858 19.1117966
#> 82   0.64152458  9.5902203
#> 83   2.89641358 13.7548861
#> 84   3.36046673 21.5026406
#> 85  -1.09307408  1.0654281
#> 86   2.92185751 21.5888610
#> 87   1.82446629  0.4465577
#> 88   0.46717880 17.4173417
#> 89  -0.18357263  2.4277624
#> 90   0.12224587 18.4761985
#> 91   0.24374210  4.5477895
#> 92   0.55688908  6.0861958
#> 93   0.15258815  2.0175569
#> 94   1.06704932  0.4500178
#> 95   0.68578446 20.4424405
#> 96   1.32875908  1.5977471
#> 97   1.53947097  1.0615273
#> 98   3.13304992 23.3087228
#> 99   0.80120221  6.7134634
#> 100  1.69874108 20.6773432