Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1.

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1).

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times
#> 1    1.67085399 19.2464120
#> 2    0.16033275  4.6925167
#> 3   -0.63657350 10.2426484
#> 4    2.30622095 15.7120462
#> 5    1.61708591  1.8712421
#> 6    1.22525750  6.8977973
#> 7    2.61393191 20.3451795
#> 8    1.03147672  9.4855683
#> 9    1.59932157 20.2263897
#> 10   0.85636753  4.5102585
#> 11   1.15068457  6.8379855
#> 12   0.05498600 19.0811422
#> 13  -0.79774155  9.1096277
#> 14   0.41384867 11.8218869
#> 15   3.16368623 20.4377774
#> 16   1.76406864 16.6211619
#> 17   2.65212138  0.3838610
#> 18  -1.58022360  4.2305582
#> 19   3.49383844 18.4941811
#> 20  -0.05784236  7.7665661
#> 21   0.83461964 15.7548629
#> 22   0.18080188  7.9437999
#> 23  -0.11586387 12.7832760
#> 24  -0.11415677 10.1456607
#> 25  -1.38256160 12.1019321
#> 26  -0.13387976  7.7476278
#> 27  -0.67336886  5.2847799
#> 28   3.44688681 20.8059784
#> 29   1.80917496 15.7662738
#> 30   2.99924872 20.0674216
#> 31   0.14726457 12.5719483
#> 32   0.61600763  4.3515683
#> 33   2.12125354 21.0806441
#> 34   1.39758862  1.4184440
#> 35   0.11478363 20.7224936
#> 36   2.30980306  0.2978766
#> 37  -1.01324395 13.5084049
#> 38   4.11854215 20.1090340
#> 39   0.02177285 14.5215098
#> 40   1.02528313  1.7852056
#> 41   2.11703220 22.1292183
#> 42   1.72562427 14.3698856
#> 43   1.11955179  4.7630710
#> 44  -0.47717304 12.8184715
#> 45   2.23232237 12.2459943
#> 46  -0.56212395  4.4894329
#> 47   2.35134878  3.2333715
#> 48   0.93899370  6.7963850
#> 49   3.25068846 20.9423397
#> 50   2.38881733  2.4278736
#> 51   0.61944817  9.3143238
#> 52  -0.27435120  7.0580138
#> 53   1.38559256 13.4621700
#> 54   3.76931978 19.4901374
#> 55  -0.04047466 10.1268861
#> 56   1.06137709  3.5810709
#> 57  -0.49739594  1.0860725
#> 58  -0.81649785  8.9662816
#> 59   0.31045769 14.2066226
#> 60   0.51641129 11.1155991
#> 61   0.39982333 12.5030651
#> 62   4.31599373 15.6837344
#> 63  -0.41385973  0.3515654
#> 64  -0.84669741  1.2529341
#> 65   0.27356322  8.1315753
#> 66   0.79432669 13.7258836
#> 67   1.43352746 15.0472683
#> 68   1.60226995  3.1815470
#> 69   0.27978960 23.9838995
#> 70   0.02409894  7.1154487
#> 71   2.21020731 20.7907345
#> 72   0.72713225 23.8145413
#> 73   1.89861488  3.1867537
#> 74   0.30308222  3.2015045
#> 75   1.63119095 20.3686719
#> 76  -1.64861008  4.7627092
#> 77  -0.10699470  5.1173993
#> 78   1.45938518 12.4947621
#> 79  -0.44672616  8.7696507
#> 80   1.28960848 21.3111121
#> 81   1.72512854  3.0356585
#> 82   2.85327129 23.0213927
#> 83   1.64656396  1.0847350
#> 84   2.17901231 22.0802133
#> 85  -1.03946291  8.0591182
#> 86   0.75646084  0.3547113
#> 87   0.51319713  0.9493336
#> 88   2.63550193 22.4003092
#> 89  -2.03907482  8.9862858
#> 90   1.04864857  7.7732821
#> 91   0.99765493 15.0469494
#> 92  -0.41540676  4.9404073
#> 93   2.30675530 21.7131084
#> 94  -0.85349628  7.9624436
#> 95  -1.33002349 12.7989570
#> 96  -0.91182220  7.1825220
#> 97   0.88624389  4.1471541
#> 98   3.03556785 23.3312810
#> 99   0.90605335 23.5830954
#> 100 -0.69657901 10.8630110