Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1.

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1).

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y     times
#> 1   -0.52519436  8.640308
#> 2    0.21354913  9.267151
#> 3    2.05280701 14.599319
#> 4    4.07766805 13.938028
#> 5    3.28205251 21.514907
#> 6   -1.07275044  4.589580
#> 7    0.99888560 15.311957
#> 8   -0.29082441  8.010538
#> 9   -1.20078152  6.067739
#> 10   2.69520372 18.022353
#> 11   1.74592708 18.159884
#> 12   0.47444758 12.136294
#> 13   2.97032794 17.125863
#> 14   1.27795811  3.165354
#> 15  -0.22298219  6.784760
#> 16   1.34799428  9.823548
#> 17   0.48408515 21.840788
#> 18  -0.82283321  5.649726
#> 19   3.31585244 21.606304
#> 20   1.84316428 18.699911
#> 21   1.49470222 18.998966
#> 22   0.08008278 21.923316
#> 23   0.40570946 10.946997
#> 24  -1.80643492  8.205675
#> 25   1.87464165 14.602004
#> 26   1.77481705 21.098739
#> 27   0.42090405  5.025329
#> 28   2.04168602 19.142055
#> 29   0.81628646  6.900409
#> 30   2.29888256 23.119188
#> 31   3.17091619 15.929544
#> 32  -0.02086018 10.904665
#> 33   0.99338204 10.154702
#> 34   0.90305455 17.955471
#> 35   2.24033511 22.496711
#> 36   0.38080551 16.492852
#> 37  -0.26436270  7.923237
#> 38   1.77565504  4.586168
#> 39   0.51479034 16.402577
#> 40   0.35445160  2.719365
#> 41  -0.95021293  8.506398
#> 42   2.15368253 22.812793
#> 43   3.03502715 17.597181
#> 44  -0.11125406 19.482823
#> 45   1.22826958 22.126501
#> 46   1.49574247 12.831551
#> 47   1.32569032  7.277394
#> 48   2.04885284 18.758251
#> 49   0.92503977 18.989339
#> 50   1.51835684 11.681800
#> 51   0.67281551 10.034969
#> 52   2.36332490 23.630628
#> 53   0.26344845  5.976074
#> 54   2.61533539 23.148911
#> 55   0.28992010  5.588013
#> 56   1.49011267 18.823182
#> 57  -0.03105706  5.476851
#> 58   3.92328464 21.087346
#> 59   2.10802441 17.308383
#> 60  -2.11393342  9.117132
#> 61   1.30017947  1.502970
#> 62   1.33773472 19.928545
#> 63  -0.57638962 13.582499
#> 64   0.74370289  5.362620
#> 65  -0.34121772 12.517073
#> 66   0.89220417  7.836749
#> 67   2.13391391 20.090779
#> 68   0.83701496 16.358319
#> 69   2.46233505 21.637122
#> 70   0.95795639 23.865472
#> 71  -0.56763166 10.092753
#> 72  -2.03382932  7.660806
#> 73  -0.69183017  1.824416
#> 74  -1.51056170  7.995696
#> 75   0.41237438 17.733378
#> 76   2.58674722 20.999553
#> 77  -1.06692714  9.579622
#> 78   0.31573090  6.676543
#> 79   1.89894061 19.427980
#> 80   1.53439266  7.667177
#> 81   1.61066643 15.571948
#> 82  -0.36508107  9.417047
#> 83  -0.07768396  6.986403
#> 84   1.80857092 19.977714
#> 85   1.03205192 15.400604
#> 86  -0.25519441  4.061720
#> 87   1.45281928 23.286216
#> 88   1.53330264  2.181939
#> 89   0.20157610 15.029772
#> 90  -0.44184908  9.345348
#> 91   0.54185361  9.713795
#> 92   2.80082049 18.686714
#> 93   0.87619696  1.204392
#> 94  -0.40833304  4.050310
#> 95   0.28785801  1.823686
#> 96   2.88404382 18.964251
#> 97  -0.60289186  6.455839
#> 98   0.32226890 11.167609
#> 99   1.78472556  7.644146
#> 100 -1.00315534  8.669194