Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1.

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1).

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y       times
#> 1    1.93212920  9.21377585
#> 2    1.91239649 22.93126080
#> 3    0.54177416 12.70332036
#> 4   -0.34444959 11.39292146
#> 5    0.23226219 10.30884232
#> 6   -0.82738360  8.58652093
#> 7    1.82070685 20.14825840
#> 8    0.35165381 18.71609670
#> 9    3.16567951 18.76104430
#> 10   0.49669197 16.68091779
#> 11   2.90571642 15.68447875
#> 12   2.26871473  1.88326723
#> 13   1.58840014 14.88435299
#> 14   2.87936758 17.26025676
#> 15   2.73995744 16.51827183
#> 16   0.26103131 13.59812578
#> 17   0.11004970 11.75158250
#> 18   1.18802282 20.86612866
#> 19   3.97859081  0.16321406
#> 20  -0.23639258  3.55731693
#> 21   0.82292727  2.26208596
#> 22   2.53321082 21.56163448
#> 23   1.54950951 11.60762617
#> 24   2.82126325 20.66705161
#> 25  -0.37466439  8.28598513
#> 26   1.92841954 13.62903811
#> 27   0.62680044 22.05942452
#> 28   2.32543817 16.92317135
#> 29  -0.95433510  9.80712212
#> 30   0.44014192 12.17930507
#> 31   1.13866572 13.54622390
#> 32   0.29557132  8.70943041
#> 33   2.50176522 18.53386495
#> 34   0.16819364  2.52725401
#> 35   1.46735520  0.30132977
#> 36   2.38449459 12.47640638
#> 37  -0.07981914  6.21091392
#> 38   0.16976009  0.29851886
#> 39   1.98611066 13.15808635
#> 40   0.61056942 23.66929045
#> 41   1.57206201 19.79705711
#> 42   0.23403252 15.58793829
#> 43   1.53183806 21.82533244
#> 44   1.78652415  1.09000475
#> 45   0.00215298  6.00486325
#> 46   3.07533651 20.06192222
#> 47   0.63475052  8.23773187
#> 48   0.89969348  0.02019603
#> 49   1.25212440  8.61608313
#> 50   0.44896593 16.41455954
#> 51   1.82757534 13.24397947
#> 52   1.42870910 21.12457816
#> 53   1.61158826 23.88673006
#> 54   0.78321106  8.36217972
#> 55   2.62988625 20.36945841
#> 56  -0.28784580  3.11540744
#> 57   1.37117364  7.08464037
#> 58   1.35248119 15.27504273
#> 59   2.69772784 19.65041793
#> 60   1.40390914  1.58660350
#> 61   1.60023916  3.65602511
#> 62   0.12740116  8.44947081
#> 63   1.24943913  8.01933722
#> 64   1.40074277  9.08050064
#> 65   0.24160735 14.08912876
#> 66   0.10508003  4.82876846
#> 67   0.56122779 12.81294828
#> 68   1.29702587  1.12359339
#> 69   0.62368086  1.52799089
#> 70   1.36779163 16.05983378
#> 71   0.29637104 13.26753477
#> 72   0.62294431 10.64081909
#> 73  -0.11519227  6.79018283
#> 74   1.83614831 14.15530652
#> 75   1.68292625 23.63019001
#> 76   2.52222919 22.24908215
#> 77  -0.99150550  6.62483370
#> 78   2.65211788 20.27847100
#> 79   1.69317769  5.89830377
#> 80  -0.54383270 10.52785971
#> 81   4.19620217 18.86226398
#> 82   0.43552396 10.12163527
#> 83   1.95296704 15.42473128
#> 84  -1.30974795  9.70299321
#> 85  -0.31568829  5.23272279
#> 86   2.32769623  0.93091289
#> 87   0.61139098  2.41017719
#> 88   1.85599465 21.70154031
#> 89   0.46929621 12.67747791
#> 90   0.86810212  9.23283251
#> 91   0.89792697  5.84815359
#> 92   1.38906133 13.80703069
#> 93   2.98108119 20.07265998
#> 94   2.04400857  0.78807704
#> 95   1.44609753 14.32628759
#> 96   0.02934849  8.60927553
#> 97   1.18819795  1.46845422
#> 98   0.53784121 16.00755291
#> 99   2.93517226 18.36790809
#> 100  2.61563562  0.26756111