Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1.

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1).

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>                Y      times
#> 1    1.671850059  4.9720910
#> 2   -0.541129835  8.3929971
#> 3    2.597480703 19.5322915
#> 4   -0.350801884  1.4428209
#> 5    1.406230799  2.3123474
#> 6    0.542239192 11.6768852
#> 7    1.768576177 19.7099987
#> 8    0.980254875 10.9932208
#> 9   -1.060954288  5.0988306
#> 10  -0.194029689 16.7841944
#> 11   0.157113113  2.2789915
#> 12   1.042771258  2.4130097
#> 13  -0.262143666 10.0318386
#> 14   2.265954355 15.0546605
#> 15   1.147785015 21.7657419
#> 16   0.098709042  4.0637839
#> 17  -0.651387156 10.4045950
#> 18   1.074784783 21.9688969
#> 19   1.064437973  0.8453551
#> 20   2.688256543 21.5787216
#> 21  -0.354052162  8.1527753
#> 22   0.493139424  5.0043678
#> 23   0.829193072 11.9793274
#> 24   0.885562865  4.8553534
#> 25   2.393501984 20.7475726
#> 26   1.489581721 21.1487832
#> 27   2.016242198  2.0088238
#> 28   0.817418481 13.7604024
#> 29   3.498551878 22.7157033
#> 30   2.346555140 13.0875455
#> 31   0.452465391 21.8224539
#> 32   0.637353916 15.2916055
#> 33  -0.594936862 12.6250476
#> 34   4.019925029 18.5244287
#> 35   1.765140424 23.4100887
#> 36   3.497325668 18.5131742
#> 37   0.419234034  3.7553833
#> 38   0.015450453 22.6272623
#> 39   1.687767836  8.5220901
#> 40   0.014993069  9.9596886
#> 41   0.158200571 13.9876345
#> 42  -0.574369896  6.7556352
#> 43  -0.546012336  9.5758670
#> 44   0.813536714 10.6994271
#> 45   0.347387524 11.1897801
#> 46   2.394051349 17.3428253
#> 47   1.520749184 21.3044573
#> 48   0.859866237  3.6185861
#> 49   2.895535045 12.2932502
#> 50   1.302556109 14.2087995
#> 51   1.590940374 18.3460553
#> 52   0.303306998  4.0767449
#> 53  -0.069018073  5.4147735
#> 54   2.010029712 22.3384598
#> 55   0.002028003  2.3043211
#> 56  -0.290730853  7.9872453
#> 57   0.495791083 11.4934581
#> 58   2.057342697 22.6839193
#> 59   1.623323605 22.1566795
#> 60   1.760553461  0.2149867
#> 61  -0.007997947  6.8436575
#> 62   2.104780329 19.0323216
#> 63  -0.285739966  3.8736703
#> 64   2.626160286  0.5194995
#> 65   2.311515669  4.2820246
#> 66   2.981458772 16.8727696
#> 67   0.967348534  9.4675079
#> 68   2.745620110  3.8629260
#> 69   0.057289467  4.1570351
#> 70   0.745336896  5.5162054
#> 71   3.581433286  2.1657004
#> 72   0.674760247 13.1860910
#> 73  -0.088602165  7.5970278
#> 74   4.192014216  1.1574404
#> 75  -1.574901172  1.4022013
#> 76   2.635200414 20.6549729
#> 77   3.812064688 19.5200482
#> 78  -0.981001441  6.8905355
#> 79   0.887231923  0.6859579
#> 80   1.051025252 23.9693660
#> 81   3.709016841 20.2682863
#> 82   0.022635182 12.7684924
#> 83   0.617304290 23.1117037
#> 84   1.773491467 19.6284121
#> 85   3.314384170 17.2799379
#> 86   1.564379008  1.2456688
#> 87   0.851681846  3.0762835
#> 88   1.292830252 13.8555056
#> 89   0.019817338 16.0771155
#> 90   1.231154326 15.4923819
#> 91   0.648979127  3.6947801
#> 92   1.948436204  2.0980060
#> 93   1.147969287 20.2303414
#> 94   2.477622243 22.2881219
#> 95   1.969606700 17.4870377
#> 96  -1.678437821 11.9463146
#> 97   1.788972203 15.7880576
#> 98   1.169131360  1.3388023
#> 99   1.439397770 23.0200987
#> 100  0.973173208  3.5843428