Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y       times group
#> 1    1.05980546  1.39098028     0
#> 2    0.20265674 17.20659834     0
#> 3    0.89257649 20.98109275     0
#> 4    0.79669820  7.90144104     0
#> 5    1.54787613  7.21361947     0
#> 6    1.55179589  4.70688263     0
#> 7   -0.08670960 10.90050566     0
#> 8   -0.46054737 20.81271232     0
#> 9   -0.28159216 17.01959668     0
#> 10   2.63851319  3.15650479     0
#> 11   0.12922888 11.42674829     0
#> 12  -0.53656670 18.16866834     0
#> 13  -0.03051647 19.62522352     0
#> 14   1.23259894  9.19538820     0
#> 15  -1.21049794 13.81560688     0
#> 16   1.77159090 23.46975314     0
#> 17  -1.51927810 13.12155487     0
#> 18   1.82335410  4.66455470     0
#> 19   0.10557986 21.06875249     0
#> 20   1.25802895 23.60735032     0
#> 21   3.01133983  6.71712790     0
#> 22   1.84594197 15.97620755     0
#> 23  -0.80526804 13.44223895     0
#> 24  -1.24040911 17.43754138     0
#> 25   1.57244573  4.79201858     0
#> 26   0.78762403  8.09962206     0
#> 27   1.81082247  6.14893790     0
#> 28   0.02196197 14.62166261     0
#> 29   1.59471000 21.38358714     0
#> 30   1.74244106 18.25320921     0
#> 31   2.50018135  1.71948707     0
#> 32   2.40737299  2.13077476     0
#> 33  -0.43298191 16.02205975     0
#> 34   2.45943698  1.85682133     0
#> 35   1.88535128  5.45820522     0
#> 36   1.20633641  2.70875840     0
#> 37   0.55031881 18.77407603     0
#> 38   2.21151809  6.43665793     0
#> 39   0.18699794 13.91613421     0
#> 40   0.13815658 10.78629739     0
#> 41   1.34590576  0.42497820     0
#> 42   2.81385912  7.66630816     0
#> 43  -0.07480425 17.39159740     0
#> 44   2.17336281  8.63053034     0
#> 45   2.60504621 13.83544900     0
#> 46   1.22029233  6.96507866     0
#> 47   1.06646852 11.18382035     0
#> 48   3.11736851  6.36207894     0
#> 49   0.05211548 19.00278338     0
#> 50   2.23230800 20.39424319     0
#> 51  -0.52984980 16.67364235     0
#> 52   1.05262946 15.18109704     0
#> 53  -0.45187557 13.80771309     0
#> 54   0.35830970  0.92359206     0
#> 55   1.35195763  9.47792184     0
#> 56   1.37032270  2.47856673     0
#> 57   1.37463167 16.22368638     0
#> 58  -1.44879123 18.13724240     0
#> 59  -0.83655565 17.97990806     0
#> 60   2.44651364 23.46159960     0
#> 61   0.30558070 17.76731955     0
#> 62   1.37728323  2.79541581     0
#> 63   0.45565407 13.49876821     0
#> 64   3.05801312  7.81528093     0
#> 65   1.96050646 20.11506969     0
#> 66  -0.24965681 11.44181583     0
#> 67  -1.12310923 20.36988818     0
#> 68   2.05294766 17.84890481     0
#> 69   1.33283068  0.77472191     0
#> 70   0.84713528  0.09835094     0
#> 71   2.78344736 20.58478638     0
#> 72  -1.57561021 15.74761535     0
#> 73   1.60743868  6.54475375     0
#> 74   0.72760251 22.69056736     0
#> 75   3.06487514  1.19021808     0
#> 76   1.80589348 22.80033561     0
#> 77   1.08213193 16.29607677     0
#> 78   2.43747476  5.65934008     0
#> 79   1.05725658 10.08713574     0
#> 80   2.27178781  6.52353774     0
#> 81  -0.63555493 11.20759564     0
#> 82  -1.60580337 13.46745737     0
#> 83   2.50138995  0.60305900     0
#> 84   1.29309080  3.05454734     0
#> 85   2.84326345  1.03795785     0
#> 86   0.81570781 19.45930463     0
#> 87   0.46956257 19.97386724     0
#> 88   1.13316954  9.58283509     0
#> 89   1.06129304  9.43062898     0
#> 90   1.15129380  7.28593518     0
#> 91  -1.12023039 14.08039998     0
#> 92   2.46919820  1.68844450     0
#> 93   2.39259115  7.16669138     0
#> 94   1.67339722  5.76311315     0
#> 95   1.59411979  5.11249743     0
#> 96   1.95950479  0.87352256     0
#> 97   0.69204573 17.37940172     0
#> 98   1.17908929  7.31135244     0
#> 99   2.89119854  6.06478815     0
#> 100  1.68443370  1.08347318     0