Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times group
#> 1   -0.06703817 15.6609508     0
#> 2    2.16957253 19.5549790     0
#> 3    1.90391285 10.9967603     0
#> 4    2.82248796  9.2518632     0
#> 5    0.38370226 13.9133850     0
#> 6    1.89065911  6.3655320     0
#> 7   -0.22923799 11.3915795     0
#> 8    1.33760694 13.2297875     0
#> 9   -0.41244494 15.4846657     0
#> 10   1.09561935  9.7519315     0
#> 11   0.89933198  0.9640888     0
#> 12  -0.11494645 21.7516800     0
#> 13   0.26779269  7.8862635     0
#> 14   1.89968644  4.6669185     0
#> 15  -0.17757082 18.6408842     0
#> 16   0.46681433 21.3024451     0
#> 17   2.34606490  0.9830294     0
#> 18   3.62241753  7.7739893     0
#> 19   1.44959095  8.5371436     0
#> 20  -0.24748225 10.6937644     0
#> 21   0.84602711 14.2601629     0
#> 22   0.31868625  9.8719758     0
#> 23   0.44046263 14.9284486     0
#> 24   0.49901721 22.9624358     0
#> 25   2.08433260  3.8955597     0
#> 26  -0.39027064  9.7180108     0
#> 27   1.43597458  0.4753014     0
#> 28  -0.08042256 19.8321675     0
#> 29   3.22581701  5.8249888     0
#> 30   2.41049151 15.3855936     0
#> 31  -0.97376332 16.1266892     0
#> 32   0.28866094  8.8364945     0
#> 33   0.81947581 14.0356078     0
#> 34   1.93397474  1.5699056     0
#> 35   1.06733946 16.0685165     0
#> 36   1.98152730  6.7549057     0
#> 37   3.04824926  1.1609981     0
#> 38   1.47618638 15.0264758     0
#> 39   2.18800635  6.3789944     0
#> 40   1.09239947 19.8169484     0
#> 41   0.81100848 10.7690155     0
#> 42   3.44995115  6.8854798     0
#> 43   2.97483475  2.4680018     0
#> 44   4.01659035  4.0567438     0
#> 45   0.65740840 12.0942673     0
#> 46   2.36107486  5.8080781     0
#> 47   0.59009884 17.9815027     0
#> 48   1.74719903  9.4082278     0
#> 49   1.62660261 11.4570231     0
#> 50   1.13963571  3.5112384     0
#> 51   2.46680900 20.3037373     0
#> 52  -0.79572390 15.3815965     0
#> 53   2.47189357  1.9228213     0
#> 54   2.47681656  8.2743250     0
#> 55   1.72211403  5.2491668     0
#> 56   0.73929405 16.9326000     0
#> 57   1.14194464 22.6609093     0
#> 58   1.49648314 20.6886020     0
#> 59   1.64663223 23.4330686     0
#> 60   1.42524163 20.6432433     0
#> 61   1.99217849 22.4729270     0
#> 62   1.37123678  6.0280672     0
#> 63   0.46877730 18.0839261     0
#> 64   0.94983804 23.2751798     0
#> 65  -1.16737997  9.6608652     0
#> 66   1.79016709 22.9528603     0
#> 67   0.20333382 17.9596915     0
#> 68  -0.37398075 10.7687336     0
#> 69  -0.14755837 11.0459493     0
#> 70   1.22499791  5.9301108     0
#> 71   0.70703555 18.8890577     0
#> 72   0.86478316  2.1136568     0
#> 73   2.86659642  5.9970682     0
#> 74   3.77073812  3.0924337     0
#> 75  -0.53295532 20.4244021     0
#> 76  -0.12859248  0.5304422     0
#> 77   2.65582698  2.4335238     0
#> 78  -0.14794254 13.3861684     0
#> 79  -0.58894032 16.3638127     0
#> 80   0.27999036 17.5450991     0
#> 81   0.32210729 17.5190491     0
#> 82  -1.09644162 16.6910458     0
#> 83   2.65817345  5.8173031     0
#> 84  -0.97593415 12.8605081     0
#> 85   1.48835010  2.4846456     0
#> 86   1.39358000  6.0642174     0
#> 87   3.22308043  3.6296517     0
#> 88   2.92949059  2.3226992     0
#> 89   0.54027422 12.4555822     0
#> 90   1.81000051  3.2035019     0
#> 91  -0.02198614 10.8264130     0
#> 92   1.23824674 20.0601348     0
#> 93   0.72840808 22.3599651     0
#> 94   1.80717046  4.7125177     0
#> 95   2.70219257  7.5097897     0
#> 96   0.19026230 14.7066738     0
#> 97   0.63733692  4.7687136     0
#> 98   0.99074432 13.5876853     0
#> 99   0.93924591  8.7159860     0
#> 100  1.48907651 20.5074944     0