Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>                 Y       times group
#> 1    0.9905813287 12.82459184     0
#> 2    2.6705533795  0.23754841     0
#> 3   -0.1867408646 13.97693115     0
#> 4   -0.2005180352  9.36468526     0
#> 5    0.2096435929  9.80322208     0
#> 6    0.4157777397 11.95378525     0
#> 7    1.5888295263 19.16989850     0
#> 8    0.4894594334  0.85872490     0
#> 9    0.5793435819 15.32042761     0
#> 10   0.8961471951 14.14365727     0
#> 11   0.3348150430 14.05541559     0
#> 12   1.3543428814 23.66052846     0
#> 13   0.4479408009 15.86637031     0
#> 14   2.1861890910 23.94308848     0
#> 15   0.3635277217  9.06559121     0
#> 16   0.9206209224 22.16830461     0
#> 17   2.8124065996  6.80906665     0
#> 18   0.9423060519 16.95637988     0
#> 19  -1.1937582234 19.11749125     0
#> 20   0.4793010909 14.53013610     0
#> 21   1.6607953617  8.61076424     0
#> 22   1.1916641361 19.98132434     0
#> 23   1.8841808927 21.07797180     0
#> 24  -0.3400926948 12.23158627     0
#> 25   0.7432520107  1.83631318     0
#> 26   0.5353535657 13.32559683     0
#> 27   0.0285721821 10.41280529     0
#> 28   0.6629084701 11.34988382     0
#> 29   0.6196259942  1.40595442     0
#> 30   2.1666041556 23.60983817     0
#> 31   0.3323732705 11.41399387     0
#> 32   0.7421493423 17.23482125     0
#> 33   3.1795369214 22.82368269     0
#> 34   1.1301698989  6.44226851     0
#> 35   0.4467159093 10.03127017     0
#> 36  -1.0228135428 18.64108370     0
#> 37  -0.5930389175 14.04480488     0
#> 38  -1.7540548177 13.15228310     0
#> 39  -0.0623091147 10.13426609     0
#> 40   1.2800714268  5.96786239     0
#> 41   1.2854018826  9.22543167     0
#> 42   0.6403424142 22.27188414     0
#> 43   0.9483622314 19.63452574     0
#> 44   0.6807724287 17.69507981     0
#> 45  -0.4972024408 14.23707783     0
#> 46   2.1951065803 22.13583967     0
#> 47   0.6959875575 20.32783166     0
#> 48   1.4734833484  4.62112934     0
#> 49   1.0531521540 23.01592057     0
#> 50   2.3960370869  6.61205352     0
#> 51   0.7973978301 19.17295529     0
#> 52  -0.6984105056 13.93324257     0
#> 53  -0.6421747983 15.41652393     0
#> 54   0.2171546832  6.21655348     0
#> 55   1.2369775666  1.46150890     0
#> 56   0.4591022091 22.51876028     0
#> 57   2.6642897264  0.05782872     0
#> 58   1.3678491266  2.94273498     0
#> 59   1.0540363645 14.32513011     0
#> 60   1.4797326279  0.10315438     0
#> 61   1.5422386176 20.31731070     0
#> 62  -0.0008858511 11.59987388     0
#> 63  -0.2567951860 16.41005880     0
#> 64   2.2672205690  7.84256697     0
#> 65   1.3827063323  5.91000788     0
#> 66   0.0627908898 12.77428312     0
#> 67  -0.3223881623 23.79263623     0
#> 68   1.2106371503  6.75469655     0
#> 69   3.6782286906  2.78859119     0
#> 70   2.1610545098 17.99195000     0
#> 71   0.2923146855 12.04252447     0
#> 72   2.7287577488 23.90366379     0
#> 73   2.5624587093  6.54796822     0
#> 74   3.8940735476  5.43460594     0
#> 75   0.5865436919 20.37661124     0
#> 76   0.6130272137  9.21404760     0
#> 77   0.8421601099 21.52162859     0
#> 78  -0.2486519487 20.56920085     0
#> 79   1.5271613823  9.19516445     0
#> 80   1.8694962715  4.83754213     0
#> 81  -0.4911253359 15.50661537     0
#> 82   0.9982756314 16.93724701     0
#> 83   2.3216431651  6.55764186     0
#> 84   2.6349448528  3.64593054     0
#> 85  -0.0132628970 20.02617482     0
#> 86   0.5872688931 23.85001098     0
#> 87   0.0353953990 10.98040880     0
#> 88   0.4297605543 20.80500808     0
#> 89   1.8841471900 23.25697231     0
#> 90   0.8495871209 21.67460061     0
#> 91   0.8318301862 17.55120560     0
#> 92   2.4282008538 21.52676074     0
#> 93  -0.7706060819 14.82078027     0
#> 94   1.5518054604  6.19236354     0
#> 95  -0.0957341814  4.03918368     0
#> 96   0.8149753573 22.60209289     0
#> 97   1.9676872513  5.35672507     0
#> 98  -0.1808838954 13.67858609     0
#> 99   1.4683461434  3.20264254     0
#> 100  1.1809207869 19.80044894     0