Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y       times group
#> 1    0.84138413 17.49409030     0
#> 2    0.68328056 21.00092988     0
#> 3    1.09363143  1.14685495     0
#> 4    0.66855848  8.14892408     0
#> 5   -0.37007367 17.47553624     0
#> 6    1.40779520  2.83615034     0
#> 7    1.27052436  4.82437286     0
#> 8    1.67881942  9.33540490     0
#> 9   -0.86547543 15.01747469     0
#> 10   2.81744449  4.30481605     0
#> 11  -1.35050314 13.07034115     0
#> 12   0.64605529 15.23866620     0
#> 13   0.27492947  1.62231614     0
#> 14   0.70024732  8.94977230     0
#> 15   0.89601301 20.17657180     0
#> 16   3.55095667  3.99674089     0
#> 17   0.53533006  9.27932899     0
#> 18   0.96288834 21.51743870     0
#> 19  -0.49792695 14.02996938     0
#> 20   2.08171069 23.97818898     0
#> 21   1.76305614  1.96020759     0
#> 22   1.20316779 18.68212061     0
#> 23   2.63374193  2.63728980     0
#> 24   3.10897647  6.04917700     0
#> 25   0.49761207  9.21879769     0
#> 26  -0.10105066 17.32057937     0
#> 27   2.03636081 17.76031466     0
#> 28   1.13473819 22.28421528     0
#> 29   2.39075006  0.80130681     0
#> 30   0.03920748 19.82455889     0
#> 31   0.34210441 21.90988434     0
#> 32   2.39574496 13.64303623     0
#> 33   0.59377624 20.98420512     0
#> 34  -0.93812624 14.96665814     0
#> 35   2.12290894 23.27069714     0
#> 36   0.91156021  8.33813811     0
#> 37   0.98767321  8.28834549     0
#> 38   4.10559244  3.82811564     0
#> 39   1.23000965  3.53025363     0
#> 40   1.73098385 21.46586625     0
#> 41   1.00382877  7.10327318     0
#> 42  -0.85866676 16.46236562     0
#> 43   0.93287979  9.33759258     0
#> 44   2.73037750  3.98256992     0
#> 45   0.09579486 14.36197761     0
#> 46   1.96426052  0.06268717     0
#> 47   1.25064212 22.99990103     0
#> 48   1.87066498  5.29666349     0
#> 49   1.51375956  9.46035274     0
#> 50   0.84871120 16.25887748     0
#> 51   0.49009225  8.60765049     0
#> 52   0.83129809 18.66254472     0
#> 53   0.55784614 18.09092516     0
#> 54   2.82607659  5.82985441     0
#> 55  -0.43185284  3.51175804     0
#> 56   0.90282242  3.10709797     0
#> 57  -1.33938022 17.79854891     0
#> 58  -1.38502938 11.48267676     0
#> 59   2.81667288  8.73984252     0
#> 60   2.53781847  6.49306315     0
#> 61   0.01418226 17.20021787     0
#> 62   1.41156863 13.15805584     0
#> 63   0.82114939 19.85894280     0
#> 64   0.69924914 22.30521562     0
#> 65  -1.03051938 20.29834141     0
#> 66   0.95338335 12.36058106     0
#> 67   0.24477907 13.50408982     0
#> 68   2.54923417  9.17103946     0
#> 69  -1.02991293  3.14668906     0
#> 70   0.19361570 17.91367816     0
#> 71  -1.68784182 17.99553218     0
#> 72  -0.23610843 16.36933802     0
#> 73  -0.51455451 17.48122214     0
#> 74   1.80947989  8.41268802     0
#> 75   2.65787118  1.30488363     0
#> 76   2.91837820  6.38634674     0
#> 77   2.99738051  6.49370775     0
#> 78   2.32704324  9.41629204     0
#> 79   4.11234084  5.21159722     0
#> 80   1.19682507  0.05436511     0
#> 81   0.46852470  9.82048657     0
#> 82   2.13227352 10.03990951     0
#> 83  -1.53217127 18.43871436     0
#> 84   2.04601005  4.36873223     0
#> 85  -1.79556635 15.79544548     0
#> 86   1.16172810  4.45671370     0
#> 87   0.04578936 20.82635913     0
#> 88   1.61917295  5.61794972     0
#> 89   2.15086158  6.65301186     0
#> 90   1.35205436 10.64331042     0
#> 91   0.86506277 20.71810461     0
#> 92   2.18890424 11.87672224     0
#> 93   1.04026209 16.34720488     0
#> 94   1.19720371  4.21104943     0
#> 95  -1.71307637 16.26125796     0
#> 96   1.06122547 21.69911749     0
#> 97   1.56122819  5.73991916     0
#> 98   2.31014391  2.32177197     0
#> 99  -0.22641703  0.46178768     0
#> 100 -0.88534614 15.84736961     0