Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The MESOR is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>                Y      times group
#> 1    1.446550287  9.9517276     0
#> 2    1.560700300 23.2850782     0
#> 3    1.511049076  1.4943457     0
#> 4    0.852774265  8.2084453     0
#> 5    0.294611157 10.0882645     0
#> 6    3.025265589  2.3996394     0
#> 7    0.868681347 13.4518601     0
#> 8    1.071672597  3.9252707     0
#> 9    2.731858307  3.3446913     0
#> 10   1.271206098  0.1462221     0
#> 11   0.580164720 20.8108510     0
#> 12   0.029551943 14.9698661     0
#> 13   0.147237967  8.3049192     0
#> 14  -1.155259389 13.1126109     0
#> 15   1.022081233 12.9094378     0
#> 16   0.582835851  7.6383640     0
#> 17   1.883627200  3.6296653     0
#> 18   1.302864999 14.4791047     0
#> 19   1.382169657 13.9922427     0
#> 20  -0.081621418 19.5033094     0
#> 21  -0.004308559 18.9350419     0
#> 22   1.239219916  5.3308330     0
#> 23   2.749435332  8.1240713     0
#> 24   0.060417453 13.3623405     0
#> 25   1.567730549 11.3021159     0
#> 26  -1.133374681 14.4725752     0
#> 27   1.006876200  0.1407990     0
#> 28  -0.992280705 17.5949328     0
#> 29   1.409638488 21.5823499     0
#> 30  -0.371604937 18.1651649     0
#> 31   2.890053402 22.5539457     0
#> 32   3.638488804  3.6576549     0
#> 33   1.020240214  0.5834872     0
#> 34   1.653696663  9.9497634     0
#> 35  -0.774715528 15.7776494     0
#> 36  -0.311597322 13.9903605     0
#> 37   0.011348695 16.3594991     0
#> 38   2.826996748  2.9958626     0
#> 39   0.908052301 20.5803812     0
#> 40   0.236874535 19.0009301     0
#> 41   1.678740452 20.4715078     0
#> 42   0.374317435 16.6696611     0
#> 43  -0.449498120 19.5030994     0
#> 44  -0.075507638  9.2572321     0
#> 45   2.871051377  1.2261416     0
#> 46  -0.863606449 20.8770957     0
#> 47   0.085692126 18.5652325     0
#> 48   1.172942553  0.8534045     0
#> 49   3.623211971  6.9050319     0
#> 50   1.968785664  8.4497512     0
#> 51   0.998008678 10.7682213     0
#> 52  -2.396099964 15.0132031     0
#> 53   0.749647806 11.0965429     0
#> 54   1.351066281  9.3310684     0
#> 55  -0.184918653 19.5749091     0
#> 56   2.148197129 21.6953879     0
#> 57  -0.597345986 11.2295411     0
#> 58  -0.840237451 13.2551873     0
#> 59   1.790742974 22.6451699     0
#> 60   2.554688226  5.7652232     0
#> 61   1.262625271  2.0584915     0
#> 62  -1.107001261 19.7346858     0
#> 63   1.610479224  2.2941897     0
#> 64  -0.061491629 16.6907392     0
#> 65  -1.925766092 19.2354284     0
#> 66  -0.517833245 13.1645454     0
#> 67   0.167830994 23.3171473     0
#> 68   2.261304340  1.2166267     0
#> 69   1.988787953  6.3776847     0
#> 70   2.561609393  3.9072073     0
#> 71   2.594871729 10.7657480     0
#> 72   3.447596161  3.5719366     0
#> 73   3.058887966  6.5786193     0
#> 74   0.990966930  5.8404302     0
#> 75   1.907992508  9.6011739     0
#> 76   1.887346853  4.1274828     0
#> 77  -1.244285376 12.8819870     0
#> 78   0.720387287 11.1907648     0
#> 79  -0.674587476 12.8147734     0
#> 80   0.341140167 23.2839441     0
#> 81   4.273994409  4.6309442     0
#> 82   1.635676089 22.4894322     0
#> 83   0.438643428 13.0751524     0
#> 84  -0.128189532  9.1734010     0
#> 85   2.500907516  2.2369916     0
#> 86   1.453830123  7.0778516     0
#> 87   2.952362763  7.0857671     0
#> 88   0.470374499  5.7938283     0
#> 89   0.348065221 10.5804930     0
#> 90   0.971115437 10.2699293     0
#> 91   0.175563781 12.9276320     0
#> 92   1.515017242  8.7114165     0
#> 93   1.873929927  0.3341063     0
#> 94   1.218037585 22.1658499     0
#> 95  -0.487228564 21.1021880     0
#> 96   2.026183825  8.2297619     0
#> 97   3.486772464  0.1989667     0
#> 98   1.774739405  0.8875911     0
#> 99   1.012648029  4.8363112     0
#> 100  0.019875341 12.7531159     0