Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y       times group
#> 1    2.23025186 16.01430278     0
#> 2   -0.29197478 13.05700879     0
#> 3    1.42796929 23.89873892     0
#> 4    3.23151006  6.02764730     0
#> 5    2.63981543 22.21945708     0
#> 6    0.15802732 18.00187336     0
#> 7    0.14375329 15.36971757     0
#> 8    0.33840729 14.01815876     0
#> 9    0.11751042 18.89117638     0
#> 10   1.31577539 23.10667591     0
#> 11  -1.56179630 11.53186353     0
#> 12  -0.50803044  7.85421930     0
#> 13   1.97337866  7.00522918     0
#> 14   1.43172949  0.03846271     0
#> 15   1.26777354 10.11593671     0
#> 16   2.77523269 15.05371996     0
#> 17   0.05220488 10.51806387     0
#> 18   2.26550193  3.92896812     0
#> 19   0.57793113 17.18993947     0
#> 20  -1.37969016 13.32815032     0
#> 21   2.80436230 23.22092568     0
#> 22   2.17275502 11.63268180     0
#> 23   0.75893833  1.40518326     0
#> 24   1.34139452 22.35145305     0
#> 25   1.25739406 23.93383032     0
#> 26   0.20673346 14.58474945     0
#> 27   2.82520007  4.26579121     0
#> 28   3.31680746  2.52128808     0
#> 29   0.79107846 11.46963806     0
#> 30   1.69751192  4.73232660     0
#> 31  -0.04823798 15.51821319     0
#> 32   0.83900817 10.53686492     0
#> 33   0.12565826  6.49529866     0
#> 34   0.87885342 14.51360946     0
#> 35   2.84267568  2.94668384     0
#> 36   1.86755370 22.53902925     0
#> 37   2.13363047  7.41673121     0
#> 38   1.57731846 19.58389128     0
#> 39   3.31362698  6.68206790     0
#> 40   1.14251617 15.67134286     0
#> 41   0.64882280 18.08527702     0
#> 42   1.35037685 11.60443830     0
#> 43   1.47558252 11.55021174     0
#> 44   0.73155915 14.20751051     0
#> 45  -0.39393661 14.67095962     0
#> 46   1.07361866  5.95337529     0
#> 47   0.26031267 12.34029811     0
#> 48   1.68069383  2.92927034     0
#> 49   1.76906651  5.55101287     0
#> 50  -0.43044666 20.84939442     0
#> 51   1.25228554 23.54189593     0
#> 52   2.14914201  5.94409119     0
#> 53   0.40953694 18.91164064     0
#> 54   0.27607089 22.61491591     0
#> 55   0.36840715  7.41376367     0
#> 56   2.18222959  1.53747577     0
#> 57  -0.40439226 15.29514000     0
#> 58   0.28638010 12.73307316     0
#> 59   0.95457777 21.89736064     0
#> 60   2.94209792  3.52382098     0
#> 61  -0.09009272 14.17538763     0
#> 62  -1.11535631 22.45241670     0
#> 63   1.17175049 10.87678325     0
#> 64   1.06213082  3.00816654     0
#> 65  -0.02679626 20.77625735     0
#> 66   0.84716758 10.30708663     0
#> 67   1.80280842  9.09314861     0
#> 68  -0.15004688 22.75430299     0
#> 69   2.12836452 23.70385644     0
#> 70   0.82152685 12.91058670     0
#> 71   3.21953823  0.62280563     0
#> 72   0.31968974 11.95647341     0
#> 73   0.24638828 12.56091947     0
#> 74   0.39337656 13.56770242     0
#> 75  -1.17582931 16.08161703     0
#> 76   0.89791549  2.89219539     0
#> 77  -1.48222532 16.08741890     0
#> 78  -1.13145737 14.69247866     0
#> 79   0.89019928 16.04385505     0
#> 80   1.38636465 16.41222159     0
#> 81   3.07282283 23.46742653     0
#> 82   0.84086469 21.17777080     0
#> 83  -0.79628320 12.91151087     0
#> 84   0.89828326 11.91745420     0
#> 85   1.57539463 22.30587639     0
#> 86  -1.63535157 11.88986230     0
#> 87  -0.78631526 10.70627544     0
#> 88   0.48492704 22.59340999     0
#> 89   1.07262154 16.59122789     0
#> 90   1.38005723  5.71181529     0
#> 91   0.09657879 20.80416473     0
#> 92   0.81962754 20.78561928     0
#> 93   1.42568706 12.32664362     0
#> 94   1.28930252  7.15024940     0
#> 95   1.01848850 11.06421869     0
#> 96   3.32985591  1.90771352     0
#> 97   1.14588345 21.54878915     0
#> 98   0.94350823  9.42123343     0
#> 99   1.43705473  5.33429123     0
#> 100  1.96295642 14.87332592     0