Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

Sample size. Integer greater than 0.

mesor

Mesor parameter for group = 0.If multiple components, specify the parameters for each component as a vector. E.g: mesor = c(1,2) for two components.

amp

Amplitude parameter term for group = 0.If multiple components, specify the parameters for each component as a vector. E.g: amp = c(1,2) for two components.

acro

Acrophase parameter for group = 0. In units of radians.If multiple components, specify the parameters for each component as a vector. E.g: acr = c(1,2) for two components.

period

The period of data. If multiple components, specify the parameters for each component as a vector. E.g: period = c(12,6) for two components.

n_components

The number of components in the model. This must match the length of the parameter inputs.

beta.group

A logical argument. TRUE if you want to simulate second group dataset. If FALSE, beta. arguments will be ignored.

beta.mesor

Amplitude parameter term for group = 1.If multiple components, specify the parameters for each component as a vector. E.g: amp = c(1,2) for two components.

beta.amp

Acrophase parameter for group = 1. In units of radians.If multiple components, specify the parameters for each component as a vector. E.g: acr = c(1,2) for two components.

beta.acro

Mesor parameter for group = 1.If multiple components, specify the parameters for each component as a vector. E.g: mesor = c(1,2) for two components.

n_period

is the number of periods that are simulated.

family

family of simulated dataset expressed as a string. Can take values: "poisson", "binomial", "gamma", "gaussian".

...

Extra arguments, such as alpha parameter for the gamma simulation, or sd (standard deviation) for gaussian simulation.

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>                Y       times
#> 1    1.806846928 20.37906356
#> 2   -0.131529421 11.63203834
#> 3   -0.492932980 19.54159863
#> 4    0.473015623  1.46855393
#> 5    2.463481058 15.22721385
#> 6    1.221784132  3.97583162
#> 7    1.948502249  1.83464363
#> 8    2.222858536 17.31456315
#> 9    1.038172052  2.81439320
#> 10   1.466182077 16.45356967
#> 11  -1.201629050  2.98492514
#> 12  -0.381782678  9.30856781
#> 13  -0.886891321  9.97964321
#> 14   2.091338104 14.56959917
#> 15   0.636685370  0.40451426
#> 16  -0.025546344 12.75335062
#> 17   0.687116811  9.93180427
#> 18   0.556667100  7.43576845
#> 19   0.444133842  6.79754081
#> 20   1.676823222 15.72603954
#> 21   1.090927220 22.69101818
#> 22  -0.580572777  9.72175771
#> 23   3.211706646 18.99232319
#> 24   0.099931207 12.88249709
#> 25   0.086540636  8.61209419
#> 26   1.350750672 16.31582092
#> 27   2.705236482 10.40561584
#> 28   0.914377958 13.30059091
#> 29   0.774843002  4.33741239
#> 30   3.029761249 17.71671023
#> 31   1.002029483  8.56844154
#> 32  -0.661893753 12.87004665
#> 33   1.866557011 15.97193081
#> 34   1.525702080 15.66523439
#> 35  -0.009522979  4.69287011
#> 36   1.478110880 15.26044239
#> 37   1.090097607  6.51836332
#> 38   0.297933384  4.88519509
#> 39   3.060872993 14.50778271
#> 40  -0.040890837 12.77742430
#> 41   1.963298240 20.98601201
#> 42   1.639684285 19.24519583
#> 43  -0.565774777  7.63384116
#> 44   2.516103554 23.62854629
#> 45   0.600021420  2.36757468
#> 46  -1.040449334  2.85059850
#> 47   1.020722227 22.46634465
#> 48   1.894289458  9.99124569
#> 49   2.055586060 18.58463329
#> 50  -0.450849635  2.67321262
#> 51   1.319109179  3.90800169
#> 52   1.153192063 21.09262018
#> 53   3.805043956 19.67113402
#> 54   2.429300152  9.34066280
#> 55   0.604778625  0.39476790
#> 56   2.579360566  2.83638476
#> 57  -0.375362588 15.68306654
#> 58   1.639180742 23.01226696
#> 59   2.775568094 16.79051625
#> 60   1.019706960 11.51056006
#> 61   0.391976085 13.27844147
#> 62   0.098417168 12.12647711
#> 63   0.170051123  7.28373256
#> 64  -0.741113825  4.25984984
#> 65  -1.753330940 10.81010632
#> 66   0.889980323 18.42853168
#> 67   2.405705696  0.09476498
#> 68   0.309234026  2.57414163
#> 69   0.365682316  0.58610317
#> 70   2.553829352 21.70532659
#> 71   0.340414314 13.31389636
#> 72   1.858686101 23.36384864
#> 73   0.998709543 14.66980084
#> 74   2.005129606 18.64486621
#> 75   1.687932567 14.92163147
#> 76   0.508196991  5.74638224
#> 77  -0.064939035  8.60318781
#> 78  -0.077266603  9.69708782
#> 79  -0.876820736  6.72067929
#> 80  -1.479843067  9.88017253
#> 81   1.693711032 19.88179602
#> 82  -1.029453257 10.78758546
#> 83   0.503100730  1.32384423
#> 84   1.390320548 22.97343023
#> 85   1.353012978  2.94789455
#> 86   0.901674649  7.12001505
#> 87   1.381191362  4.13727524
#> 88   0.143687859  7.41840599
#> 89   0.375826383  2.23410264
#> 90   0.588512423  0.04398177
#> 91   0.732593229 17.67728523
#> 92   1.163747029 15.76039815
#> 93   0.323128746 14.48243475
#> 94  -0.931545275  2.06931536
#> 95   2.626386963 14.08453278
#> 96   2.964920278 21.99587584
#> 97   2.048800606  2.22847603
#> 98   0.529578805  4.01720832
#> 99   0.615811206  8.64421079
#> 100  1.029274638 20.96897896