Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The MESOR is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>                Y      times group
#> 1    0.365346090 23.3102461     0
#> 2    0.669542435 12.6946119     0
#> 3    1.798504872 12.0656838     0
#> 4    0.284835655 21.5527614     0
#> 5    1.712889352 17.6658013     0
#> 6    1.007038018 23.0600184     0
#> 7    1.012682734 11.6601834     0
#> 8    1.197148901 22.3334956     0
#> 9    0.480012209  7.7210914     0
#> 10   1.891142533 18.1339002     0
#> 11   0.253474750 13.1016547     0
#> 12  -0.120154087 17.5684058     0
#> 13   0.676950036  4.9607306     0
#> 14   1.929987546 20.4695884     0
#> 15   1.529068210 18.5691001     0
#> 16   0.934219500  2.7036064     0
#> 17   0.548647730 15.5791817     0
#> 18   2.430769923  3.0439170     0
#> 19   1.141866441  4.8926966     0
#> 20  -0.592181359 16.8506243     0
#> 21   1.179872222 20.3327610     0
#> 22   0.292129537 22.4515729     0
#> 23   0.859791283  4.7288863     0
#> 24   0.318328723 19.7217173     0
#> 25  -0.134983805 19.5543239     0
#> 26   0.569128422  2.3971108     0
#> 27   0.140563943 15.1371421     0
#> 28   1.211258773  6.4463865     0
#> 29   0.386642986  4.0128104     0
#> 30   0.557328249 23.7385847     0
#> 31   0.640316883 16.6659489     0
#> 32   1.687552856 10.2614389     0
#> 33  -1.094941911 16.0915182     0
#> 34   0.857831351 22.2029962     0
#> 35  -0.063978729  0.2211510     0
#> 36  -0.056303908 12.1083844     0
#> 37   0.008637896  7.0225357     0
#> 38  -0.968786911 18.7204764     0
#> 39   0.488114385  8.8668412     0
#> 40   2.028614291 22.8072926     0
#> 41   0.066748530 16.6216324     0
#> 42   2.867241585  1.7758939     0
#> 43   0.679001187 23.1726806     0
#> 44   2.421816748  8.6523700     0
#> 45  -0.279743514 13.2176564     0
#> 46   1.597320853 23.6679808     0
#> 47  -0.969337005 15.9428561     0
#> 48   2.831870417 10.5337576     0
#> 49   2.325329065 22.9930001     0
#> 50   0.012747974 18.3293569     0
#> 51   1.963081990 21.1368437     0
#> 52   2.379322298  7.6580374     0
#> 53  -1.257821623 13.9667689     0
#> 54   1.546748821  1.1970792     0
#> 55   1.050681156  1.1198317     0
#> 56   1.289066067 23.0430515     0
#> 57  -0.825152525 18.7205181     0
#> 58   0.933433459 10.8049871     0
#> 59  -0.917984306 15.3957025     0
#> 60   1.559888179 12.0599876     0
#> 61   0.698997583 23.6650787     0
#> 62   2.665297235  1.6963046     0
#> 63  -0.599206255 19.3133460     0
#> 64   0.421886592 19.4744359     0
#> 65   1.017839439 10.8586095     0
#> 66   0.380970079  0.7788109     0
#> 67  -0.126874059 15.3265322     0
#> 68   3.035515985  6.3307191     0
#> 69  -0.249362482 20.1233560     0
#> 70   0.809140387 12.5239424     0
#> 71   2.875701861  5.7443325     0
#> 72  -0.088014084 16.9518213     0
#> 73  -0.604903201 12.4408926     0
#> 74   0.660078148 16.3402244     0
#> 75  -1.227916379 19.1919960     0
#> 76   0.525259791 10.8518256     0
#> 77   2.236666814  8.1302037     0
#> 78   3.409486307  5.1256855     0
#> 79   1.971050395  2.3989516     0
#> 80   0.385155301 19.9908253     0
#> 81  -0.117425128  7.4130266     0
#> 82   0.752012165 17.5671649     0
#> 83  -1.777315767 13.3636408     0
#> 84   1.265100784  1.3406322     0
#> 85   2.013868121  4.4302533     0
#> 86   1.739516490  4.4041665     0
#> 87   2.145704884  1.1390723     0
#> 88  -0.070664623  5.4717761     0
#> 89   0.669079193 17.5945819     0
#> 90   3.395819898 22.4655099     0
#> 91  -1.468476619 18.3146082     0
#> 92   2.511923237  0.8894462     0
#> 93   1.340518652  5.4049367     0
#> 94   0.558089396  7.9916411     0
#> 95   0.563013964  8.5740596     0
#> 96   0.492661086 17.7979291     0
#> 97   1.645040854 22.1014263     0
#> 98   1.747626880 20.9726534     0
#> 99   1.557570598 11.2190520     0
#> 100  1.942788190  2.5743412     0