Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1.

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1).

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y       times
#> 1    1.72803041  0.06139751
#> 2    1.74783068  5.08984805
#> 3    1.09623155  7.41764396
#> 4    1.80029986  7.27719532
#> 5    0.42113977 10.15922916
#> 6    0.17499966 16.27536483
#> 7    2.63756861 10.53435589
#> 8    1.68315644  6.75932826
#> 9    1.03596461  6.19447104
#> 10   2.15235145 22.34358882
#> 11  -1.07646376 11.76313163
#> 12   1.89615706 21.18084882
#> 13   0.74310677 20.15292708
#> 14   1.13617115 15.70600107
#> 15  -0.10205984 10.33827893
#> 16  -0.46145347 10.75889399
#> 17  -0.33371863 18.18949814
#> 18   1.58756197  0.50464743
#> 19  -0.35928098 16.14704337
#> 20  -1.19802595 18.39828689
#> 21   1.13954239  2.28933859
#> 22   0.85657404 10.66109305
#> 23   1.60270662  6.76605703
#> 24   0.60392674  5.46617342
#> 25   0.54566370 10.56311408
#> 26   1.25356110 22.97661179
#> 27   0.06144750 20.42153487
#> 28   1.40989998  0.24807610
#> 29   2.28356693 13.44612900
#> 30   0.66575608 17.18679859
#> 31   2.93905513 20.80068627
#> 32   0.01861841 20.28190461
#> 33  -2.76339727 16.48475870
#> 34   2.52988600 23.28059779
#> 35   1.68864748  2.91270474
#> 36   1.16653650 14.82772671
#> 37   2.48713172 15.64278540
#> 38  -0.49749340 17.04770884
#> 39   0.94047167 22.42722134
#> 40   2.87898968  7.81250422
#> 41   0.05047166 13.38225840
#> 42   0.96269468  9.73545238
#> 43   2.35643066  4.37690568
#> 44   0.77089595 16.26508521
#> 45   1.67657794 10.15280242
#> 46   0.36346133 16.69487828
#> 47   1.86638461 21.26782397
#> 48   3.34986984  8.28545120
#> 49  -0.74060453 14.22790664
#> 50   0.90908592 10.29913471
#> 51  -0.17434954 17.53280389
#> 52   0.01081356 16.72237127
#> 53   1.09879859 10.83167718
#> 54   0.70334842 21.04071056
#> 55   2.57042700  9.99058220
#> 56  -0.40342949 14.04738713
#> 57   0.31746673 20.97873053
#> 58  -0.26868264 16.51063439
#> 59  -0.34461814  4.98031972
#> 60   3.41118744  6.66687215
#> 61   1.56195678  3.48559043
#> 62   0.29631207 15.73623599
#> 63   2.08109040 14.23171824
#> 64   0.66179080 20.60832590
#> 65   1.82040278  2.98471274
#> 66   2.94551114  5.15830772
#> 67   0.52023378  1.84031725
#> 68   1.97410637 13.19580076
#> 69   1.17820464 22.74969644
#> 70   2.97245016  2.08348996
#> 71   0.57790700  2.60506440
#> 72   2.16278580 23.75660384
#> 73   1.62271082  4.41934701
#> 74  -1.39736958 13.68120021
#> 75  -0.20386757 20.04717881
#> 76   2.86753493 22.19264130
#> 77   0.70110601  7.73476298
#> 78  -0.22106596 12.76549732
#> 79   2.46996782  5.43048121
#> 80   0.35026670 15.72724251
#> 81   0.62931557 11.07263758
#> 82   1.61748192 12.52296641
#> 83   0.85828834 18.97141171
#> 84  -0.46967645 19.53624023
#> 85   2.44969841 23.67419145
#> 86   0.09994982 11.37651986
#> 87   0.67536527 10.95839708
#> 88   2.60762692  7.89375864
#> 89  -1.93065793 10.93977524
#> 90  -2.41000740 13.16291630
#> 91   2.36275933 15.21399287
#> 92  -0.36786911 12.89630206
#> 93   3.79629133  6.69608477
#> 94  -0.04353940  8.97130419
#> 95  -0.38750257 17.85520253
#> 96  -1.35914206 14.83449083
#> 97  -0.59559953 11.05383584
#> 98   1.26939273  0.72829506
#> 99   1.01287640  8.16779245
#> 100  0.31757620  9.47253095