Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1.

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1).

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y       times
#> 1    1.82352478  2.39806541
#> 2    2.65685275 21.96571288
#> 3    2.16153860 10.52401108
#> 4    1.17464066 22.09936357
#> 5    3.57885793  1.31682916
#> 6    1.52217764 23.73005761
#> 7   -0.32192640 17.01346614
#> 8    2.53237509  8.19116720
#> 9    0.93750865  6.41580626
#> 10   2.08965492  5.70854625
#> 11   2.72815043  8.47483818
#> 12   0.11182068 21.97212682
#> 13  -0.36759255 16.78749688
#> 14  -0.11826839 15.34044616
#> 15   1.01104489  5.58359285
#> 16   2.02685805 23.23612335
#> 17   0.17980748 11.44129776
#> 18   0.23210587 19.14179887
#> 19   2.55804879  3.21052635
#> 20   2.62842687  3.30259080
#> 21   1.35522089 12.22887835
#> 22   1.12740982 17.41106481
#> 23   1.01316607 19.47106824
#> 24   2.15088089  2.73963038
#> 25   0.13356133  9.71293857
#> 26   1.02032229 11.36173159
#> 27   1.51072205 10.69408756
#> 28   0.24597304 23.68163669
#> 29  -0.63573154 14.51451352
#> 30  -0.07815083 16.67082030
#> 31   0.66729331 12.89232066
#> 32   0.04283136 19.60675151
#> 33   3.37958034  6.76004015
#> 34   0.73438415 10.74892807
#> 35   3.60396138  8.69557090
#> 36  -0.07784254 22.00318519
#> 37   1.25559039  1.86268838
#> 38  -1.04226983 17.66202470
#> 39  -1.95305275 13.03596403
#> 40   1.29364001 21.96232105
#> 41   1.27336344 21.64566537
#> 42   1.68340624  0.82433294
#> 43   1.01085366 12.43080606
#> 44   1.79933421 22.31275945
#> 45   3.00478150  0.02808377
#> 46   1.93780321  6.97205392
#> 47   0.72568468 14.99481243
#> 48   2.75393391 23.10316006
#> 49   1.98964409  8.37769422
#> 50   2.02142854  5.33794940
#> 51   2.51151849  4.17811676
#> 52  -0.88628498 18.31945889
#> 53   1.11179752 16.24995165
#> 54  -0.55771772  9.02580534
#> 55   3.14081912  4.16820506
#> 56   1.67725908 22.92193938
#> 57  -0.18428145 11.27979584
#> 58  -0.06058857 17.15244743
#> 59   1.16785608  7.89391480
#> 60   0.25831931 17.86119243
#> 61   1.47528374  9.86083468
#> 62   0.53749432  8.75198044
#> 63   0.52480596 19.61402505
#> 64  -1.86384351 15.96774772
#> 65   1.16205251  9.96996590
#> 66   1.43490573  1.99288768
#> 67   0.82735171  4.86111439
#> 68   0.61182745 12.64774132
#> 69   3.20919464  5.37039934
#> 70   1.18607871 11.96502780
#> 71  -0.65619956 16.31153767
#> 72  -1.18832646 15.90717291
#> 73   0.91296540  6.11796697
#> 74   0.75543688  9.95160216
#> 75   2.15460102  6.83639208
#> 76   1.70762943  3.15822781
#> 77   2.82194741 22.64047334
#> 78  -1.46055323 14.05755672
#> 79   0.88392246  8.51152605
#> 80   2.24347962  8.12969364
#> 81   2.52232115  2.93746962
#> 82   1.95313945  5.18508621
#> 83   2.42476714 23.39566810
#> 84   2.37623464  4.65945578
#> 85   0.63818856 14.03526106
#> 86   0.25495673 23.94180095
#> 87   1.66111880 21.99872962
#> 88   2.11817164  6.65126823
#> 89   2.66311974  0.30513568
#> 90  -0.79821037 16.55150657
#> 91   1.37693363  1.40482190
#> 92  -0.76629125 18.95449033
#> 93   2.81206878  5.27510747
#> 94   0.65036492 23.23284996
#> 95   2.51019523 23.87930107
#> 96  -0.49549376 17.85171257
#> 97   0.69806397 14.30348875
#> 98   1.59887900 20.79391354
#> 99  -0.08790712  8.67795136
#> 100  1.49334852 23.04843567