Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1.

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1).

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>                Y       times
#> 1    0.373744123  0.50220771
#> 2    1.663144552 19.67397891
#> 3    0.433820626 23.36200851
#> 4    1.599866681  5.32925092
#> 5   -0.227741662  5.07924845
#> 6    0.917040148  6.73490007
#> 7   -1.022690551  8.18668648
#> 8    0.310657981  7.06768633
#> 9    0.454988970  5.16183733
#> 10   0.796469657  1.75917265
#> 11   1.737588442  7.12078566
#> 12  -0.053112236  9.71293676
#> 13   0.332640549 23.29457883
#> 14   2.303431430 19.43943285
#> 15   1.706630125 12.50346740
#> 16   0.238119741 23.71058006
#> 17   0.208182210 11.77998418
#> 18   2.024470153  3.90208706
#> 19   2.094447416 14.20892959
#> 20  -0.343568007 11.74907150
#> 21   0.751382537  8.62239672
#> 22   1.709217853 14.29079818
#> 23   0.389448802  6.44392058
#> 24   1.343388915 21.87095962
#> 25   2.231171401 23.17913891
#> 26   0.670223255  0.11367478
#> 27   1.203260535  3.68789166
#> 28   0.743437710  6.33707769
#> 29  -1.570834502  7.07122972
#> 30  -0.360681799  4.88583789
#> 31   2.413972166 16.82160630
#> 32   1.005293238 22.06299886
#> 33   2.733960348  3.14528093
#> 34   0.079982636  7.29922256
#> 35  -0.406951890  2.89515861
#> 36  -0.939728484 10.75211131
#> 37   0.827926506 23.32137456
#> 38   0.944513443 16.32888160
#> 39  -0.057935469  5.34540854
#> 40   1.671329227 10.99333027
#> 41   1.386511777 11.66833878
#> 42   1.309013134 21.81629374
#> 43  -0.725504862  8.09253088
#> 44   0.966915802 10.23705432
#> 45   1.491023447 18.41231452
#> 46   2.387779407 23.78675058
#> 47   3.043146965 17.91419291
#> 48   0.008992075  5.14208261
#> 49  -0.284110382  8.72858012
#> 50   2.756397721 18.21953654
#> 51   0.747848871  0.99242388
#> 52   0.261237472 10.11801767
#> 53   0.966499950 23.36539486
#> 54  -0.751705908  3.77302562
#> 55  -0.425118528  6.34002852
#> 56  -0.022375624 19.64060306
#> 57   1.962400929 18.03725348
#> 58   2.213884988  0.09136062
#> 59   4.056124252  1.47456934
#> 60   0.050634074  7.33705584
#> 61   2.794365579  0.06442857
#> 62   1.542408477 21.04744122
#> 63   0.484975179 13.14828109
#> 64   0.505679489  3.70737887
#> 65   0.951632044 17.07575274
#> 66   2.208290048 16.90746228
#> 67   2.181619367 19.76550170
#> 68  -0.440606442  7.50189427
#> 69   0.502556905 20.90651071
#> 70   2.431443840 16.81894897
#> 71   0.284242407 10.09620878
#> 72   2.358843275 19.33017096
#> 73   1.883881290 12.92832027
#> 74  -0.116662466 13.88190970
#> 75   1.100033015 22.61591693
#> 76   2.163548474  6.16565538
#> 77   2.771004131 20.20376801
#> 78  -0.462184590  6.07813331
#> 79  -0.846810009  7.63716719
#> 80   1.890975267 16.60391681
#> 81   0.660248409  5.39214312
#> 82   0.119895650 11.64169608
#> 83  -0.828865986 11.83059589
#> 84   0.626337585 14.81014335
#> 85   1.426045235 13.04721326
#> 86  -0.871884673  2.58803648
#> 87  -1.280345247  7.51295247
#> 88   2.921986580 21.89244055
#> 89  -1.167728167  9.38185066
#> 90   3.587443290 19.91646462
#> 91  -0.178605097 13.42503565
#> 92  -0.737792267 12.93061982
#> 93   2.400328407 22.89483880
#> 94   0.211963093 23.27115861
#> 95  -1.096380637  6.67738480
#> 96   1.927108904  7.46614842
#> 97  -1.517280117  5.40616443
#> 98   1.783386472 22.26643411
#> 99   1.335186718 14.77088890
#> 100 -0.106569436 10.11694261