Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times group
#> 1   -0.07023461 18.6634575     0
#> 2   -1.00286591 15.7104504     0
#> 3    1.91785196  2.7218680     0
#> 4    1.85576804  4.4645817     0
#> 5    0.68096470  5.7467189     0
#> 6    2.59181763 20.7541607     0
#> 7    1.04039729 20.5139008     0
#> 8   -0.90528717 22.3778627     0
#> 9    0.64859216  9.6508845     0
#> 10   1.88806818  5.0094773     0
#> 11   0.63878599  7.4241222     0
#> 12   1.63772226 22.7070402     0
#> 13   2.53035523 23.4327948     0
#> 14   0.94584481  9.0751986     0
#> 15   1.96038052  7.3732131     0
#> 16   1.84044421  9.5051632     0
#> 17   1.86156472  5.1837951     0
#> 18   0.14076418  7.7073680     0
#> 19   0.45496957 19.7811240     0
#> 20   1.00647001 11.0374298     0
#> 21  -0.07941486  8.7596320     0
#> 22   0.50247904 16.6882240     0
#> 23   2.65969568  4.0160105     0
#> 24  -0.97113340 18.0281521     0
#> 25   1.06352623 13.3561891     0
#> 26   0.75488027  5.7791929     0
#> 27   1.42419616 20.8926410     0
#> 28   1.33390716 14.3486388     0
#> 29   1.36735012 20.1624421     0
#> 30   0.31868291 16.9463072     0
#> 31   1.14690344  9.1877023     0
#> 32  -0.71491836 16.1809860     0
#> 33   0.25342447 17.4136102     0
#> 34   1.91401636 10.8637360     0
#> 35   2.24937548  8.3487293     0
#> 36   1.75544341 18.7907819     0
#> 37  -0.12625444  6.1449176     0
#> 38   1.21292179  0.2063153     0
#> 39   0.77979642 13.7049311     0
#> 40   0.80245708 20.2946002     0
#> 41   1.99958337 22.4025900     0
#> 42   1.34874679 15.8067986     0
#> 43   3.15914168  5.5990834     0
#> 44   0.46479029 11.0595878     0
#> 45  -0.01669896 19.7029031     0
#> 46   2.20385382  4.7145476     0
#> 47   1.06773171 22.7911498     0
#> 48   1.63161512  1.3839768     0
#> 49  -0.17182423 19.2466123     0
#> 50   1.49785155  3.0081086     0
#> 51   1.30537375  5.8482212     0
#> 52   0.69045232 23.6857390     0
#> 53   1.40821097  5.7281644     0
#> 54   0.40851519 21.5661268     0
#> 55   1.27630885 19.9161332     0
#> 56  -0.03745484 18.1833037     0
#> 57   0.97931643  6.4927249     0
#> 58  -0.93500251 14.4300207     0
#> 59   0.64107475 19.1396031     0
#> 60   2.77565963  5.2568404     0
#> 61   0.10341872 15.3500654     0
#> 62   1.02620096  3.5372720     0
#> 63   0.41717727 18.8185087     0
#> 64   0.24132690 20.4509656     0
#> 65   2.44395757  2.7578064     0
#> 66  -0.24106311 16.7065284     0
#> 67   2.69459113  2.4899491     0
#> 68   1.98820039  8.0583492     0
#> 69  -0.16513079 15.1165679     0
#> 70   2.15103985 10.4485708     0
#> 71  -0.85763951 11.7085067     0
#> 72   0.34319617 10.6810529     0
#> 73   1.37810460 10.7022833     0
#> 74   2.23432490 21.3952251     0
#> 75   0.87403895 16.3365940     0
#> 76   1.82616256  3.0544231     0
#> 77   2.60315659 23.6721340     0
#> 78   1.97492298 14.9319122     0
#> 79   1.36081121  4.0034330     0
#> 80   0.13036466 10.5053437     0
#> 81   0.30168569 15.6882105     0
#> 82  -0.02616701 15.2430743     0
#> 83   0.75296650 12.3567258     0
#> 84   0.20735338 17.3894925     0
#> 85   0.77830971 13.6362653     0
#> 86   0.13769035 20.0959818     0
#> 87   3.10935270  5.7687880     0
#> 88  -0.73988379 20.5120200     0
#> 89   0.82828243  8.5693775     0
#> 90   0.59747211  0.6836844     0
#> 91   1.20648005 10.8381932     0
#> 92   1.88966712  9.6007963     0
#> 93  -0.28221327 11.6912643     0
#> 94   1.20216149  1.4506571     0
#> 95  -1.01816903 12.0615183     0
#> 96   2.04265728 23.8312750     0
#> 97  -0.40874741 13.9632696     0
#> 98  -0.82470337 16.6912981     0
#> 99  -0.67457763 19.3967900     0
#> 100 -0.34005269 21.1608997     0