Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1.

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1).

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>                 Y      times
#> 1    2.8572129521 17.8473062
#> 2    1.0925547103  5.1667682
#> 3    2.3129474913  9.3358089
#> 4    0.6999279415  6.2748065
#> 5    2.0065882186 22.9601104
#> 6    0.0097919326  6.3174305
#> 7    2.5933947296 20.8498430
#> 8    1.7150363480  2.5360407
#> 9    1.2046165726 12.4681526
#> 10   0.2430507111 10.0311441
#> 11  -0.3812671810  1.6807960
#> 12   0.4434480895  5.3837301
#> 13   1.5333457725 14.8644148
#> 14   2.9118093838 19.5364625
#> 15   0.3034173621  4.8049818
#> 16   1.4053482055  2.8912029
#> 17   2.2895491714 21.4713298
#> 18   0.3581385177 20.1256136
#> 19  -1.7720924421  7.3798730
#> 20   1.8128862539 21.9842038
#> 21  -1.4571977597  8.6106296
#> 22   2.4869397238 14.0060534
#> 23   1.9033533435 23.0417696
#> 24   2.7427501107 14.0893611
#> 25   2.0018705361 23.7031732
#> 26   3.3958036929 17.5607114
#> 27  -1.0118889073 13.2460046
#> 28   1.4178060609 17.5357532
#> 29   1.7582067292 22.2906064
#> 30   0.8036823353 23.3657002
#> 31   1.1764629489 12.0133914
#> 32   2.0504738000 23.8088251
#> 33   2.9044911592 22.2953592
#> 34   0.1667009244 12.0569077
#> 35   0.4784020313 20.1435734
#> 36   2.7093938963 18.2602326
#> 37  -0.2695039128 18.7025760
#> 38   0.8089018562 19.6965949
#> 39  -0.3090451761  7.9529567
#> 40   2.2273902351 22.8342833
#> 41   0.9087893592  3.1535380
#> 42   1.3069872210 20.1540036
#> 43   2.4323876773 20.1648396
#> 44   0.8565825055  6.9186572
#> 45   1.8093697208 10.6158649
#> 46   0.0004815369  4.3111898
#> 47  -0.9237272523  9.9462206
#> 48   0.5806524061  4.0632040
#> 49   2.0696927770 22.6302538
#> 50   1.2178689416 20.8969412
#> 51   1.6496697371  2.8375277
#> 52  -1.1506275848  5.7097439
#> 53   0.8081074719 21.5572574
#> 54   0.8043131489 11.0424700
#> 55   1.3221616688 18.9854318
#> 56   1.4282728885 16.9373341
#> 57   1.2870892341 10.1161310
#> 58   1.7552650513 23.4181428
#> 59   0.6653743453  4.9655872
#> 60   2.7257145357 20.1637143
#> 61   0.6455860096  1.9946037
#> 62   1.0979811887  3.7015605
#> 63   0.2159904830  4.4949348
#> 64  -0.9217729466  6.3274455
#> 65   1.5228573962 16.1615295
#> 66   0.5075191609  5.4636477
#> 67   2.2848425486 13.4230217
#> 68   1.0040237463 22.6688643
#> 69  -0.7638994675  5.0014704
#> 70   2.0494422266 23.0175106
#> 71   1.6348590347 20.4180767
#> 72   1.3440065702 12.4043639
#> 73   0.5374109037 18.7329323
#> 74   0.5982479203  1.6836368
#> 75   2.2281005511 17.2083603
#> 76  -1.5686811371  4.9961419
#> 77   3.0578330835 23.4526247
#> 78   1.8129648255 20.7635684
#> 79   0.1585605405  6.1189062
#> 80  -0.1733549567  8.6003488
#> 81   2.1143005529 19.2182412
#> 82  -1.1600463501 10.0631816
#> 83   2.9020415691  2.1743566
#> 84  -0.0285356956  9.4462877
#> 85   0.9005017926 11.4928031
#> 86  -0.1334452226 11.9926256
#> 87  -0.6164170942 10.4698564
#> 88  -0.8318205948 11.1806202
#> 89   1.2674402347  7.6589933
#> 90   1.7661769471  1.0982890
#> 91   2.2194762970  0.3700335
#> 92   1.3884977911  7.5751900
#> 93   0.9670136612  4.6202363
#> 94  -0.1884593268  8.8987374
#> 95   2.9116186796  0.5960381
#> 96  -2.2562983109 11.3472987
#> 97   1.0664784840 13.1614411
#> 98   1.8748406090 10.6264802
#> 99   3.6523095241 21.1554253
#> 100  1.7012297547  6.7392388