Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1.

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1).

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times
#> 1    1.97385823 22.4836849
#> 2    1.03906477 19.0109321
#> 3   -1.16586298  7.8480897
#> 4    3.63911397  0.8556786
#> 5    2.24213483  1.5502046
#> 6    0.38062996  2.5892186
#> 7   -0.28925909 13.0375084
#> 8    1.77098613  7.4345559
#> 9    2.63038434 18.7149237
#> 10  -1.92185570  7.4672191
#> 11   3.46697580 20.2899936
#> 12   2.99558188 21.3433830
#> 13   0.12251749  4.3745575
#> 14  -0.09762685  4.7251590
#> 15   1.12464448 14.6651769
#> 16   0.52379618  7.2247866
#> 17   1.31931263 10.6382897
#> 18   1.32545932 18.8195975
#> 19   1.30922342 14.8892802
#> 20   1.95589626 19.5514877
#> 21   0.21074055  3.9620460
#> 22   1.76723020 14.9957118
#> 23   3.06946326 17.7918209
#> 24   0.38976239 16.7403664
#> 25   0.79708090 13.8976253
#> 26   2.01725838  2.6842466
#> 27   1.02038029 21.7542279
#> 28   0.13577168  4.9680721
#> 29   0.90375492 16.9563366
#> 30   0.87068252  4.6200485
#> 31   1.52178951 12.5693881
#> 32   0.18198859 16.8075111
#> 33   0.77278014  4.9658962
#> 34   0.70167918  0.7474343
#> 35  -0.70047591  9.7207363
#> 36   0.84492398  5.3547257
#> 37   0.96289898  8.9549947
#> 38   3.06994186 17.9177266
#> 39   0.43476611 21.8403187
#> 40   2.38001352  1.1952283
#> 41  -0.72751710  7.9096040
#> 42   3.43218510 20.7188128
#> 43  -0.44735943 10.2252151
#> 44   1.29079142 16.1863357
#> 45  -0.02624746  9.4912412
#> 46  -0.08802857 19.2063007
#> 47   2.28825030 14.0475163
#> 48   1.40690451 19.4973623
#> 49   1.83499037 14.2978077
#> 50   2.88284530 18.7931945
#> 51   0.49245336  3.7754138
#> 52  -0.99040396  7.6478235
#> 53   0.52214112  1.0733053
#> 54   2.17544024 20.1692786
#> 55   1.26076457  3.9567216
#> 56   0.82269097 13.5049745
#> 57   3.41690437 20.9951577
#> 58   0.81887651  0.6729879
#> 59   0.16519181  3.9884673
#> 60   1.98637516 11.6945308
#> 61   2.61851218 22.8632075
#> 62  -0.53335334  4.4355883
#> 63   1.20383907 13.1799718
#> 64   3.05679717 23.2619357
#> 65   2.76959414 17.9435977
#> 66   1.56249454 20.0249357
#> 67   1.76770686  2.6600145
#> 68   3.93319844 21.8768612
#> 69   1.83543767 16.3382585
#> 70   1.67299448 15.1884758
#> 71  -1.47841851  1.8546599
#> 72   0.25045640  6.4802777
#> 73   0.14986085  4.4978533
#> 74   1.29647597 13.8341464
#> 75   2.00333704 18.7855559
#> 76  -1.58021348 10.6662287
#> 77   3.01468396  0.4403982
#> 78   2.23004043  0.8442433
#> 79   2.25862259 17.3695305
#> 80  -0.69763954 11.1722086
#> 81   1.65773059 17.0822056
#> 82  -0.27558393  8.2202617
#> 83   1.28755070 21.1266987
#> 84   1.28194532 14.8899450
#> 85   2.98849395 22.8381044
#> 86   0.79108685 11.1905201
#> 87  -1.48605925  8.8756463
#> 88  -0.31273575  6.0596539
#> 89   1.38913731 22.8624362
#> 90   1.73372861 21.5582191
#> 91   1.38559161  9.1957110
#> 92   0.91516935 14.5363014
#> 93   1.25332757  0.4349930
#> 94  -0.01036989  8.6682826
#> 95   1.72373868  1.3567814
#> 96  -0.80304322  8.4594861
#> 97   0.94784220 14.2858229
#> 98   0.48963823  2.2480242
#> 99  -0.46753412  3.6771376
#> 100 -0.04615964 10.2906730