Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The MESOR is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y       times group
#> 1    0.29350651 21.00588492     0
#> 2    1.26167553 11.56035979     0
#> 3   -1.66872858  6.88746364     0
#> 4    0.08242046 15.01150518     0
#> 5    1.45479278  9.09101121     0
#> 6    2.68208904 21.03829264     0
#> 7   -0.14801033 21.22032521     0
#> 8    0.73726736 20.78559894     0
#> 9    1.02863342  4.18921458     0
#> 10   0.13318174 16.04689595     0
#> 11  -0.33822380 11.07477751     0
#> 12   2.62548560  1.80173644     0
#> 13   0.43551807  4.16787795     0
#> 14   3.04028434  0.69055558     0
#> 15   4.25032842  5.66534117     0
#> 16  -0.22810691 13.56604527     0
#> 17   0.75660399  2.32750756     0
#> 18  -1.07851446 22.02482296     0
#> 19   2.11883394  7.42441978     0
#> 20   1.56593999  7.98442283     0
#> 21   2.01008542  0.31952985     0
#> 22   2.04913050 22.59875843     0
#> 23   0.04704450 23.99082369     0
#> 24   3.25560713  4.22186897     0
#> 25   2.51401390 20.80866045     0
#> 26  -0.06895875 19.93056797     0
#> 27  -0.72114431  0.09869301     0
#> 28   0.71545286 11.14882108     0
#> 29   0.75702842 23.31277243     0
#> 30  -0.42440411 14.47703654     0
#> 31   1.87914115  0.88556311     0
#> 32  -0.56381490 11.86038559     0
#> 33  -0.90973686 17.65568040     0
#> 34   2.00246422  0.77201825     0
#> 35   2.29904127  7.35477206     0
#> 36  -0.48926359 17.80861908     0
#> 37   1.97587084 22.75104382     0
#> 38   0.58748860  3.52070532     0
#> 39   2.29116661 19.91778376     0
#> 40   1.27580062 22.23907656     0
#> 41   1.37844723  9.85805408     0
#> 42   1.00586344  0.18010892     0
#> 43   3.76781523  2.86992024     0
#> 44  -0.95550614 14.61968511     0
#> 45  -0.01903482 11.82077534     0
#> 46   3.92043949  1.83807012     0
#> 47   0.58801626 14.90725312     0
#> 48   0.49530342 11.23086194     0
#> 49   0.77680918  0.82533775     0
#> 50   0.42554680  3.25247043     0
#> 51   1.10311780 12.91697594     0
#> 52   0.45659506 23.27412769     0
#> 53   2.36010150  4.12096893     0
#> 54   1.55999446  9.59404951     0
#> 55  -0.15734252 21.50377681     0
#> 56   1.61746774 22.93829463     0
#> 57   2.54871541 23.22650024     0
#> 58   0.62656226 15.55233180     0
#> 59   0.66505718 20.12583438     0
#> 60  -0.88214819 14.00743225     0
#> 61   0.56160011 18.51508191     0
#> 62   2.21881553 11.60578032     0
#> 63  -0.04045614 11.26516589     0
#> 64   2.35465112 11.48624763     0
#> 65   2.30786718  0.94893069     0
#> 66   1.40742597  9.90052957     0
#> 67   1.44597005 23.02362443     0
#> 68   1.26061895  8.19142556     0
#> 69  -0.50770853 18.77260960     0
#> 70   0.06636035 17.56882804     0
#> 71   2.66227981 20.56134612     0
#> 72   0.86373388 16.49680287     0
#> 73   1.33564790 21.36884917     0
#> 74   1.23256291 22.88023263     0
#> 75   0.84855157 10.35879430     0
#> 76   1.83540831 23.98659096     0
#> 77   1.93309970 20.00528162     0
#> 78  -1.16693545 13.61925349     0
#> 79   1.46704085  8.07255290     0
#> 80   3.28808246 21.04788585     0
#> 81   1.55198102 22.27679159     0
#> 82   1.40516161 21.65991958     0
#> 83   3.69494155  4.64037465     0
#> 84   3.90622812  6.57791100     0
#> 85   0.13859722 20.76249752     0
#> 86   2.61858139  4.79183253     0
#> 87   0.36647371 22.71467116     0
#> 88   0.38995800 21.81419894     0
#> 89  -0.54474260 23.31077413     0
#> 90   1.03435495 11.76343597     0
#> 91   1.83509900 10.35230395     0
#> 92  -0.41196092 17.58218185     0
#> 93   0.37538573  8.67654222     0
#> 94   1.87397229 21.28112532     0
#> 95   1.27320921  9.96422203     0
#> 96   2.33004086  0.69468282     0
#> 97   1.69400901  8.30178056     0
#> 98   1.32544006 12.36937124     0
#> 99   1.72765235  1.08999219     0
#> 100  1.94024563 21.65798367     0