Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The MESOR is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times group
#> 1    0.71295579 12.0075317     0
#> 2    1.40769059 12.2174919     0
#> 3   -0.38892974 13.4013263     0
#> 4   -1.46470057 14.8045726     0
#> 5    1.12322533  9.7331777     0
#> 6    0.65416826 21.8326545     0
#> 7    2.51302793  4.5013995     0
#> 8    3.32296936 23.3898819     0
#> 9    1.90280492 20.7043999     0
#> 10   3.05799443  5.2790038     0
#> 11  -0.19946485 16.4061154     0
#> 12  -1.59486225 17.1481776     0
#> 13   1.44253390 13.9549518     0
#> 14   1.36592953  1.7219728     0
#> 15   0.01670182 15.5132054     0
#> 16   1.54930910 14.1534887     0
#> 17   0.60757155 13.2084139     0
#> 18  -0.73300030 13.6240974     0
#> 19  -0.06328319 18.7738538     0
#> 20   0.93188128 17.3182680     0
#> 21   1.55007881  6.4002462     0
#> 22  -0.08370007 17.6583822     0
#> 23   1.47754556 22.9641539     0
#> 24   2.80821626  2.9782328     0
#> 25   2.25020443  3.1195208     0
#> 26   2.80737281  4.8384041     0
#> 27   1.58758024 22.8766263     0
#> 28  -0.23586376 11.1010259     0
#> 29   3.64507974  7.9775960     0
#> 30   0.89928917 11.2622396     0
#> 31   2.16930604 20.0823844     0
#> 32  -1.46129639 11.5625533     0
#> 33   1.66318701  0.4763489     0
#> 34   0.15071727 17.9592528     0
#> 35   1.43982635  5.3834049     0
#> 36   2.12094121 22.5893144     0
#> 37  -0.68020931 18.1404785     0
#> 38   2.33340424  6.9491123     0
#> 39   1.36866940  7.3711918     0
#> 40   2.48809961  5.1567581     0
#> 41  -0.73326955 14.3278089     0
#> 42   0.49690457 11.4634210     0
#> 43   0.09907887 12.3334373     0
#> 44   1.59326841 16.1967656     0
#> 45  -1.15867168  6.8767620     0
#> 46   1.68998881 22.4538753     0
#> 47   1.45548908  3.4863018     0
#> 48  -0.46165412 16.5839934     0
#> 49   0.62778480  4.7429729     0
#> 50   0.26945955 10.9889919     0
#> 51  -0.31343496 17.5101951     0
#> 52   2.51762110  7.7249031     0
#> 53   1.39503312  9.6781192     0
#> 54  -1.22015925 12.5533845     0
#> 55  -0.78901080 14.5054705     0
#> 56   1.56395739  4.9394569     0
#> 57   2.83241690  4.7097840     0
#> 58  -1.09354708 18.9259983     0
#> 59  -0.23201108 12.9667946     0
#> 60   1.50201641 19.2316485     0
#> 61   2.62549239  8.0263336     0
#> 62  -0.68236825 10.7341693     0
#> 63   1.00138370 17.9376061     0
#> 64   0.76027173  7.4895190     0
#> 65   0.74258123 10.0266582     0
#> 66  -0.06519523 15.2594762     0
#> 67  -0.39831023 16.8952380     0
#> 68   1.51679349 23.0983027     0
#> 69   1.94841591 13.9309275     0
#> 70   1.54781356 11.6308433     0
#> 71   0.39725032 11.7115498     0
#> 72  -0.13699357 14.8308827     0
#> 73   0.71616044  2.0679885     0
#> 74  -1.06070602 17.9289592     0
#> 75  -0.55434484 15.3409213     0
#> 76   0.27160826 10.1342414     0
#> 77  -1.31742299 13.1186501     0
#> 78   0.64016108  2.2234162     0
#> 79   1.14724368  0.3874232     0
#> 80   0.08112804 12.6228581     0
#> 81  -0.53333255 20.2663810     0
#> 82   1.58608664 22.6946997     0
#> 83   0.50795088  4.8855246     0
#> 84   0.91103066  6.1572412     0
#> 85   0.91368544 20.7105447     0
#> 86   1.13899428 20.2995566     0
#> 87  -0.02062804 10.1540557     0
#> 88  -0.20573779 15.8092633     0
#> 89   3.22681615  2.7189906     0
#> 90   0.24821760 22.1591790     0
#> 91   1.13201911 20.6982140     0
#> 92   3.54010711  1.2569429     0
#> 93   2.84392267  5.5031378     0
#> 94  -0.77576797 19.3448764     0
#> 95   1.14955331  2.0990377     0
#> 96   2.14085262 20.8818809     0
#> 97  -0.97431044  7.8152432     0
#> 98   4.87219320  3.4503356     0
#> 99   0.79194414 23.1475726     0
#> 100  0.15550691  7.2080538     0