Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times group
#> 1    2.17299005 17.7310296     0
#> 2    1.33271249 13.2274655     0
#> 3   -0.57674194 16.9726731     0
#> 4   -0.33596287 10.5614719     0
#> 5    0.52712194 13.7254576     0
#> 6   -0.19732833 11.9457759     0
#> 7   -0.80151630 14.3610162     0
#> 8    2.60691998  2.8817759     0
#> 9    2.62101967  2.6275937     0
#> 10  -1.78045952 19.8587211     0
#> 11  -0.39530370 21.0488931     0
#> 12   0.49584543 15.8397135     0
#> 13  -0.06702672 13.4116343     0
#> 14  -0.68475184 16.6572528     0
#> 15   3.13094783  8.0326099     0
#> 16   0.41204081 14.3664602     0
#> 17  -1.32907803 14.3348750     0
#> 18   1.88778861  3.8201732     0
#> 19   0.42607327 10.0044117     0
#> 20   0.71433087 17.3736131     0
#> 21   0.67067719 13.1600664     0
#> 22  -0.19523641  0.1556886     0
#> 23   2.71946038  2.7181006     0
#> 24  -1.12847807 16.0391453     0
#> 25   1.11276424  6.9998490     0
#> 26   0.71505432  7.8573430     0
#> 27   0.39129708 23.0121684     0
#> 28   1.01676023  9.3588014     0
#> 29   2.54708329  4.7947518     0
#> 30   0.48987860 12.5603533     0
#> 31   1.14339819  0.2432166     0
#> 32   2.02581289  5.6500687     0
#> 33   0.60859116 19.1362184     0
#> 34   1.71726176  2.7757707     0
#> 35   1.24923568  4.5621405     0
#> 36   0.68772874 15.2918392     0
#> 37   0.89569912  1.0676071     0
#> 38   1.06131405 21.9755400     0
#> 39   1.05460460 15.1070323     0
#> 40   1.01025031  3.2186732     0
#> 41   1.17996964 11.3172698     0
#> 42   2.26701188 23.0990188     0
#> 43   1.95459282  8.1500539     0
#> 44   0.42066147  2.6904253     0
#> 45   2.33918367  1.6632933     0
#> 46   1.35901229  4.2851914     0
#> 47   3.60683750  5.7841705     0
#> 48   3.32073772  7.6941908     0
#> 49  -1.72183414 10.4349331     0
#> 50   0.76531354 12.0746817     0
#> 51   2.83629649  7.1859914     0
#> 52   1.20247830 17.8903018     0
#> 53  -0.07925080 14.3289632     0
#> 54   3.93460491  4.6397889     0
#> 55   0.51984443  8.9623155     0
#> 56  -0.23274107 10.3520097     0
#> 57   1.07515810 10.9620419     0
#> 58   2.61821120 21.5510661     0
#> 59   2.21442636 23.0980761     0
#> 60   1.06758218 10.4657108     0
#> 61  -0.74529730 21.6895392     0
#> 62   2.01428912  2.2745192     0
#> 63  -0.02320560 12.4243631     0
#> 64   0.38303162 12.0727354     0
#> 65   0.75077986  1.0962593     0
#> 66   0.51261284 22.4499881     0
#> 67  -0.36961285  0.9792571     0
#> 68   2.93164570 23.9358816     0
#> 69  -1.41333930 15.4415878     0
#> 70  -0.53631359 19.9834218     0
#> 71  -0.26825100  3.7506491     0
#> 72   0.64580480 13.4081908     0
#> 73   3.59827393  5.1949248     0
#> 74  -0.01614665  9.6345140     0
#> 75   3.01927509 22.0343496     0
#> 76   3.62561471  9.3777840     0
#> 77   1.38495261  3.6388771     0
#> 78   1.99281415 23.2212187     0
#> 79   1.98772984 22.6212640     0
#> 80  -0.22779046 16.0427794     0
#> 81   1.01528378  5.0294023     0
#> 82   2.74654558  1.5683908     0
#> 83   2.29780165  7.1452343     0
#> 84   1.66692859 22.3986711     0
#> 85   0.29833394 15.0335233     0
#> 86   1.64322174  4.6031833     0
#> 87   2.19217250  0.8037958     0
#> 88   1.18116405  1.2692509     0
#> 89  -0.72344220 10.7821377     0
#> 90   1.84458376  1.7928195     0
#> 91   2.12455978 22.1850580     0
#> 92  -0.04076058  2.0907293     0
#> 93  -0.40019027 14.7945078     0
#> 94   0.18214798  3.1299873     0
#> 95   1.17463435  3.2509773     0
#> 96   0.68743650  8.6292843     0
#> 97   1.21358673  1.0512822     0
#> 98   0.41626607  6.7007161     0
#> 99   1.14062423 21.5064163     0
#> 100 -0.25690804 21.8532848     0