Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The MESOR is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times group
#> 1    1.37811489 19.5777482     0
#> 2    1.62271768  7.3288216     0
#> 3    1.32373704  5.9984852     0
#> 4    1.86312687  2.5910919     0
#> 5   -0.99185807 16.8816321     0
#> 6    0.69160379 18.6480998     0
#> 7    0.52170853 21.6389615     0
#> 8    2.27938352  9.4630965     0
#> 9    0.21757216 13.4355274     0
#> 10   1.55938710  5.0322678     0
#> 11   0.67084543 12.9220830     0
#> 12  -1.59101265 13.4867266     0
#> 13   1.99862999  6.2811515     0
#> 14   1.33870323  1.9922251     0
#> 15  -0.45077811 16.4689572     0
#> 16   2.30022281  4.2657021     0
#> 17   2.76256269 23.5771732     0
#> 18   0.21127031 14.7878503     0
#> 19   0.72270503 18.0058084     0
#> 20   0.37271752  3.9737325     0
#> 21  -0.06391291 19.3854567     0
#> 22   0.92661026 22.2070942     0
#> 23   1.91797715  2.7154638     0
#> 24   0.15678787 12.7879086     0
#> 25  -0.30727295 16.7497804     0
#> 26   1.90340235 10.0428438     0
#> 27   1.02146531 10.5139524     0
#> 28   2.14451743  1.3803583     0
#> 29   1.25944605 10.9769740     0
#> 30  -1.41233241  9.6456827     0
#> 31   1.10249686 21.1665027     0
#> 32   0.87764985  8.2468778     0
#> 33   0.43537496  0.1211787     0
#> 34   0.65079564  7.9095797     0
#> 35   3.06022279  1.4424346     0
#> 36   0.05847099 12.3431829     0
#> 37   2.47648942  7.7425975     0
#> 38   0.83849070 11.6824011     0
#> 39  -0.23910245 17.8633408     0
#> 40   0.98222702 21.3173009     0
#> 41  -0.78517119  4.1571827     0
#> 42   0.25098269  8.5396679     0
#> 43   3.28468094  0.2852937     0
#> 44   1.53980243  5.6120202     0
#> 45   2.07454443 23.5772611     0
#> 46   1.85715158 12.3944074     0
#> 47   0.40049887 11.3527519     0
#> 48   3.55584251  5.8003463     0
#> 49   1.58666072  7.5035602     0
#> 50   1.87076841 11.4132032     0
#> 51   0.36286446 11.5211080     0
#> 52   2.76533874  7.4625643     0
#> 53  -0.37696165 17.6468938     0
#> 54   1.55380097  0.7780707     0
#> 55  -1.04165309 16.6316390     0
#> 56   1.21327682 12.0893213     0
#> 57  -0.16005376  3.2459081     0
#> 58   0.49875919 18.5330219     0
#> 59   0.64255650 20.8847087     0
#> 60  -2.15179866 10.1666248     0
#> 61   2.56985470  3.7162143     0
#> 62   1.50814980  1.0108012     0
#> 63   0.30826514 16.8658683     0
#> 64  -0.62224850 21.2171042     0
#> 65  -0.56576880 16.0275308     0
#> 66   0.37908186 18.9928339     0
#> 67   1.01308891  9.3020618     0
#> 68   1.29369781  8.7779360     0
#> 69   0.03992782  8.7935167     0
#> 70   2.23344392  6.6627057     0
#> 71   2.46742466 16.9019614     0
#> 72   0.26254674 14.3643439     0
#> 73  -0.52774046 17.9881688     0
#> 74  -0.11607306 17.1420716     0
#> 75  -1.06679606 17.1628566     0
#> 76   2.97431497  7.9779973     0
#> 77  -0.18642833  9.9487557     0
#> 78   2.69425228  6.0350208     0
#> 79   0.83903282  5.6680941     0
#> 80  -0.92913427 12.4567422     0
#> 81   0.72136786 12.5124273     0
#> 82  -0.55848226 13.6064656     0
#> 83   2.16757877 11.9279574     0
#> 84   2.63671635  1.1564152     0
#> 85   0.93874254  3.3192190     0
#> 86  -1.20044235 15.8855502     0
#> 87   1.02885418 16.0734350     0
#> 88  -0.56960724 13.6696518     0
#> 89   0.16638143  6.4078774     0
#> 90   1.38013476 23.4444257     0
#> 91   0.74649011 14.1895464     0
#> 92   0.83107308 11.7414646     0
#> 93   0.62148785 19.4706077     0
#> 94   1.54242978  2.9985503     0
#> 95   1.10118832 22.6788505     0
#> 96   0.09674560 11.2185209     0
#> 97   1.30254020 20.8398177     0
#> 98   1.40343648 16.1566186     0
#> 99   1.07470071  1.6274899     0
#> 100 -0.89913719 19.5391502     0