Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1.

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1).

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times
#> 1    0.83856793 23.4323840
#> 2   -0.53254946 10.0431292
#> 3    2.32157274  2.4829489
#> 4    1.99103447  1.4568646
#> 5    0.89637360 13.9082182
#> 6    0.77232951  8.3421402
#> 7    0.82192408 14.4156658
#> 8    0.61607718 13.0096781
#> 9    1.61106553  0.8234519
#> 10   3.41159817  0.3413456
#> 11  -0.26333260  7.4854569
#> 12  -0.95194648 18.8367725
#> 13   0.67579182  5.3976397
#> 14  -0.01242311  9.6063467
#> 15   1.11644904 19.3967235
#> 16   0.32684719 17.5946894
#> 17   0.59322015 23.5378387
#> 18   2.47813964  1.4220682
#> 19   3.36415270 23.3918037
#> 20  -0.62996040  0.3808705
#> 21   2.63237090  0.7333367
#> 22   1.49902889 20.7804095
#> 23   0.17778831 17.9330584
#> 24   0.41287480 15.5621903
#> 25   0.67882284 12.5364020
#> 26   0.43019874 20.5582001
#> 27   2.74106429  2.7778483
#> 28   0.11342874 22.9423251
#> 29   0.74449212 22.5733327
#> 30   0.70413092 18.9502096
#> 31   0.85405619  7.8125802
#> 32   2.73296129  5.8321115
#> 33   1.68742268 13.9147632
#> 34   1.47137402  8.8276666
#> 35   0.31195669 17.9843273
#> 36  -1.84437166 12.7700108
#> 37   2.74708061  4.1743346
#> 38   0.03697583 19.5331110
#> 39  -1.84657058 15.9571639
#> 40   0.87504956 21.6935370
#> 41   2.24803773  2.2708859
#> 42   1.03268445  3.5691411
#> 43   1.33093149  9.7134907
#> 44   1.30256139  0.8621311
#> 45   3.34494059  5.0903138
#> 46  -0.04956368  8.5340343
#> 47  -0.82445390 16.8198630
#> 48   1.39153838  3.5618019
#> 49   2.38898387  1.3303970
#> 50   1.92266261  5.0871905
#> 51  -1.82407011 15.8844047
#> 52   0.33795796 13.6843499
#> 53   0.64178601 14.5272687
#> 54   0.96940789 21.8425992
#> 55  -1.38500343 16.7971888
#> 56   1.06318112  7.5685526
#> 57  -0.16091571 13.4157597
#> 58  -0.21647312  9.6488479
#> 59  -0.49776157 16.8407094
#> 60   2.62292117  6.4395633
#> 61  -1.18262087 14.3951376
#> 62  -0.07384740 21.7055246
#> 63  -1.34806099 16.0817876
#> 64   1.77848838 11.4133815
#> 65   1.58343750  2.1020495
#> 66  -1.37057524 17.0754335
#> 67   1.86996981 20.4779041
#> 68   4.11740435  8.3584446
#> 69   0.91874566 13.8334596
#> 70   2.50651219  7.3806851
#> 71   1.18517659  5.5934571
#> 72   1.21279577 23.2386848
#> 73   2.48042743 19.9751485
#> 74   2.94525283 23.5637009
#> 75   2.55208835  5.4016900
#> 76   0.21094427  1.1518559
#> 77  -0.29294950 14.4291617
#> 78   1.04178502  0.7703872
#> 79  -1.13097248 22.3079335
#> 80  -1.24243999 16.2708253
#> 81   1.38085217  5.3634013
#> 82   0.97528656 12.2177380
#> 83   0.97102540  8.5030553
#> 84  -0.66378622 14.6866395
#> 85  -0.33286866 12.5590120
#> 86   0.81433104  9.4132513
#> 87   1.08616934 18.6611225
#> 88   3.05024057  7.8167579
#> 89   2.97940817  4.8165974
#> 90   0.63579354  0.4444750
#> 91   1.93536013  1.3500267
#> 92   0.36659786 14.2212124
#> 93  -0.96838952 16.2229040
#> 94   2.42868917  1.3237267
#> 95  -0.79308812 14.4600396
#> 96  -0.37616461 19.9618140
#> 97  -2.15103068 13.0349057
#> 98   1.88528391 21.2057882
#> 99   1.76970364 22.6950252
#> 100  0.08969966 13.9254615