Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1.

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1).

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>                Y      times
#> 1   -0.079565050  8.8198120
#> 2   -1.338596175 10.1526796
#> 3   -0.925778110  4.4145266
#> 4    0.949398508  4.4377050
#> 5   -0.904433468  0.7332044
#> 6    1.877325076 13.1311841
#> 7   -1.161280386 14.1726981
#> 8   -0.191734517  3.7030128
#> 9   -0.245017305  1.6378328
#> 10   0.978491289 20.8355562
#> 11   1.507486286 14.3938416
#> 12   1.700969868  6.4837521
#> 13   1.499703183  6.6704556
#> 14   0.642331622  4.4210535
#> 15   0.618154132  1.8152237
#> 16   2.513015387 20.7010124
#> 17   2.995224535 19.0462793
#> 18   1.238854822  8.1212486
#> 19   1.341440850 15.8228112
#> 20   0.770046833  6.8622418
#> 21   0.714055774 18.8543697
#> 22   2.116383745 23.6422656
#> 23   0.001753167 13.6076586
#> 24  -0.255828571  1.2524928
#> 25  -1.178784618 11.5119625
#> 26   1.288558319 20.9664481
#> 27   0.535955560 13.1184406
#> 28  -0.256411942 11.0041627
#> 29   1.820995612 15.7010571
#> 30   2.433641149  6.1890388
#> 31  -3.044257159  9.2313558
#> 32   3.107988020 18.0746954
#> 33  -1.156355654  0.5840389
#> 34   1.014606601 17.0518924
#> 35   1.149069751  2.6915681
#> 36   3.074902627 14.7291144
#> 37   0.398936759 13.9844323
#> 38   0.626047926 14.4513727
#> 39   0.342331902  4.4668681
#> 40   3.071482841 21.2246836
#> 41   0.430388808 22.0487778
#> 42   1.553518036 17.4698310
#> 43  -0.247460924  9.5871048
#> 44   2.375056851  0.5088839
#> 45  -0.727383957  5.5802273
#> 46   1.114918973 13.5419976
#> 47   1.322938801 23.6810934
#> 48   1.833561670  8.0470216
#> 49   0.740287581  5.0671389
#> 50   0.807573225 15.4233113
#> 51  -1.630172550  1.1639606
#> 52   1.355837629 21.2623154
#> 53   3.233702647  3.4503155
#> 54  -1.046232323 12.0619755
#> 55   3.563463393 21.3189363
#> 56  -0.102798699  4.7789293
#> 57   1.045607150  8.0294976
#> 58  -1.167240759  6.9291931
#> 59   3.085715113 22.3898743
#> 60   2.676947696 14.6412153
#> 61  -0.178421141 10.0062379
#> 62   1.264067562 11.4862329
#> 63   2.760930559 21.3217441
#> 64   2.051065444 17.3178094
#> 65   0.768472633 17.8849948
#> 66   3.298021091 19.1341551
#> 67   1.797996219 20.6192694
#> 68  -0.858090581  8.2583230
#> 69  -1.292795653  4.4609095
#> 70  -0.508320212  8.4980469
#> 71  -0.442765989 13.5701839
#> 72   2.064883067 16.0768469
#> 73   2.489581910  1.6155335
#> 74   2.642759986 17.9108230
#> 75   2.607831219 15.7516103
#> 76   0.855762577  4.7491149
#> 77  -0.469872750  2.9288279
#> 78  -0.442374674 11.7028252
#> 79  -1.021702128 13.7653169
#> 80   2.199120334 19.4411536
#> 81   0.085378485 12.0823152
#> 82   1.093192687 11.3585479
#> 83   1.329600959 21.2918734
#> 84   2.034130646 18.5981288
#> 85   0.770236605  8.6826865
#> 86  -2.947406486  2.3105218
#> 87   0.258958947 16.1015576
#> 88   1.068296732 21.1756043
#> 89   3.105037791 19.9555126
#> 90   0.706652876  9.3922366
#> 91   2.624280661 22.5837358
#> 92   1.258846566  8.6263086
#> 93  -0.750252250 11.0609476
#> 94   1.976034246 17.2512680
#> 95   2.210974332 17.7399235
#> 96   2.241576766 16.4750277
#> 97  -0.678489580  9.5585508
#> 98   1.579864764 22.4737261
#> 99   2.273176824 22.8833666
#> 100  2.360522389 19.6413121