Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1.

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1).

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times
#> 1    1.18508622  0.1393372
#> 2    1.14240911 17.3445936
#> 3   -0.78212188 18.7300982
#> 4    2.17501476 22.4721833
#> 5    0.96019032  8.1848475
#> 6    1.90394755 17.8087148
#> 7   -0.39853981  6.5323861
#> 8    1.93977210 17.0880844
#> 9    2.23364640  2.4023270
#> 10   0.91934130 13.8482956
#> 11   0.77158539 21.1139822
#> 12   0.79722684  5.4678201
#> 13   3.07616303 12.3967875
#> 14   2.26905686  0.3597959
#> 15   1.14400514  5.2197626
#> 16   0.09484171  7.3003759
#> 17   1.16544607 15.9341006
#> 18  -0.11018714  4.7864312
#> 19   2.97873180 23.7089943
#> 20   0.09110264  4.9329944
#> 21  -2.24730847  5.5251462
#> 22   2.74053250 19.0001902
#> 23   1.80214614 23.0086690
#> 24   1.90498998 12.3873595
#> 25   0.50438644 15.9540846
#> 26   0.75521177  3.5721719
#> 27   2.41563857 16.4579742
#> 28   0.01320610  7.2045965
#> 29   0.44285129  3.9272012
#> 30   2.82606993 15.1761971
#> 31   1.97222569  4.3243069
#> 32   2.85990803 17.1811749
#> 33   3.45079074 19.0360302
#> 34  -0.10156983 12.7804258
#> 35  -0.11359879  7.5466184
#> 36   3.34939229 18.4091164
#> 37   0.91226699 15.6508592
#> 38   1.28593471  8.0764171
#> 39   2.27200094 22.4953857
#> 40   1.57910117 14.6295089
#> 41  -1.19615681  8.2469272
#> 42   2.42453938 11.9517813
#> 43   1.66225015 14.4112133
#> 44   1.91029331 16.7795809
#> 45  -0.12203267  5.2430934
#> 46   4.15880330 21.3020720
#> 47   0.76579231 13.3280481
#> 48   1.24656369 22.4470893
#> 49   1.51136599 13.6867849
#> 50   0.99783783 17.0924600
#> 51   2.66113770  3.0586672
#> 52   1.35329390 17.8351649
#> 53   0.15984802  4.9874294
#> 54   0.94449151  3.4838056
#> 55   1.84090982 18.6172795
#> 56  -3.16813665  5.9640387
#> 57   0.54651727  1.6695652
#> 58   0.37249098 10.9973096
#> 59   0.89604070 14.9860989
#> 60   0.66477789  8.5007613
#> 61   2.10456319 13.8733323
#> 62  -0.31835058 11.8627691
#> 63   3.34483317 21.1804490
#> 64   2.17836317 15.6595193
#> 65   1.29379921 23.7104033
#> 66   1.05675100 17.4813061
#> 67   0.26666893  6.9491809
#> 68   2.49458143 14.2735299
#> 69   1.40110493 15.1043834
#> 70   2.08333794 20.5996874
#> 71   3.40228963  1.0177136
#> 72   1.89767487 18.2548536
#> 73   0.54925446 21.3722290
#> 74   2.19071639 23.9017593
#> 75   1.53618164 10.3709959
#> 76   0.58026167  3.8910281
#> 77  -1.05745031  6.9688725
#> 78   3.66349527 22.8882767
#> 79   2.92422296 22.0425361
#> 80  -0.11506064  5.5073311
#> 81   1.65241017 23.7350561
#> 82   0.56442075  6.8559899
#> 83   0.73888369 20.6245281
#> 84   2.73412736 20.3132598
#> 85   0.27699526 15.8450365
#> 86   1.68518346 19.0095163
#> 87  -0.26991936  5.5017684
#> 88   0.66357445  8.2220050
#> 89  -0.70252143 16.1091371
#> 90   0.46562175  0.2840311
#> 91  -0.29361316  2.1166886
#> 92   1.47847380 16.1140812
#> 93   0.12737533  7.5502498
#> 94   0.67282495  3.6458966
#> 95   1.94693551 21.5139596
#> 96   3.12698788 19.1881842
#> 97  -0.32296087  4.7186634
#> 98   1.69016854 17.0151253
#> 99   0.11231536 15.2779652
#> 100  3.27401670 20.4182743