Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The MESOR is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>                Y      times group
#> 1    1.681814569  3.7051646     0
#> 2    0.231463277 15.1639814     0
#> 3    1.715598658  5.4085067     0
#> 4    1.919610767 21.8133373     0
#> 5    2.652971953  4.2049352     0
#> 6    2.867118870  0.5275215     0
#> 7    0.595732292 21.1077397     0
#> 8    0.760843362 18.8344984     0
#> 9   -0.181735415 13.0629315     0
#> 10  -0.337426344  9.2089728     0
#> 11   1.175621836  1.9200373     0
#> 12   2.677506060  4.1505730     0
#> 13   0.419116290 19.6981769     0
#> 14   1.633664621 15.4836984     0
#> 15   0.580224485 16.8664556     0
#> 16  -0.772443427 13.6798330     0
#> 17   0.803854797  1.7516294     0
#> 18   2.379038117 21.1967590     0
#> 19   2.967762560  0.3774481     0
#> 20   2.064913939  4.1235584     0
#> 21   0.119151807 20.8790173     0
#> 22   1.428075450 12.3288759     0
#> 23   0.179724510 13.7109142     0
#> 24   0.423407900 14.2355763     0
#> 25   0.897331491 14.3405858     0
#> 26   1.353029489 23.7537837     0
#> 27   0.341256457  9.0654509     0
#> 28   0.999507271 22.3870666     0
#> 29   0.759092132 18.8884965     0
#> 30   1.843793104 23.7304730     0
#> 31   1.910155498  4.8996240     0
#> 32   2.377902467  6.8487277     0
#> 33   2.102589712  4.1336279     0
#> 34   0.448938664  8.7435619     0
#> 35  -1.819402803 14.5152835     0
#> 36   0.175016221 19.6518921     0
#> 37   0.762078672  6.6002325     0
#> 38  -0.055833708 12.4053069     0
#> 39   0.935798696 14.6608902     0
#> 40   1.667028088 22.5509738     0
#> 41   1.007160574 10.7266640     0
#> 42   1.760309429  2.7010086     0
#> 43   2.286991315 16.7020744     0
#> 44  -1.053701913 13.7065995     0
#> 45  -0.252362635 14.3904370     0
#> 46   0.363280835 15.2445013     0
#> 47   0.565025705  1.7228226     0
#> 48   3.419310842  4.9852452     0
#> 49   0.934397677 21.6943298     0
#> 50   0.004340101 18.1713614     0
#> 51   2.502091638  7.5395495     0
#> 52   1.354125899 21.0316376     0
#> 53   0.129061048 20.3445895     0
#> 54   1.645618501 17.0215233     0
#> 55   1.323026742 23.7510314     0
#> 56   0.304225167 10.7460678     0
#> 57   3.146019810  7.0852828     0
#> 58   2.107226382 10.8500342     0
#> 59   0.490408832 17.7792063     0
#> 60   1.650369920 11.0117701     0
#> 61  -0.710746158 13.9952206     0
#> 62   2.771170869  7.8410582     0
#> 63  -0.791291622 11.3331885     0
#> 64   2.565153172  8.5154553     0
#> 65  -0.912771815 18.6866602     0
#> 66   0.863988554 20.9547207     0
#> 67  -1.251481535 18.2391993     0
#> 68   2.571746804  3.4447962     0
#> 69   0.682023571 15.1352537     0
#> 70   3.331899219  1.5836585     0
#> 71   0.574440951 22.0060360     0
#> 72  -0.175815352 12.0642889     0
#> 73   1.616127390 23.6034535     0
#> 74   0.766131969 20.5175134     0
#> 75   0.392396919 13.0146870     0
#> 76   1.028519476 16.3087855     0
#> 77   0.743394808  8.0521640     0
#> 78   2.441496087  4.7538985     0
#> 79   0.518256741  7.9668581     0
#> 80   0.140758611 16.2246640     0
#> 81   0.315269342 17.1310757     0
#> 82   1.970076075  1.2708468     0
#> 83   2.963300079  4.0914820     0
#> 84   0.405588626 19.9326547     0
#> 85  -0.463778981 18.5307854     0
#> 86   1.100453985  1.7500346     0
#> 87  -0.761140225 17.4209330     0
#> 88   1.418616184 13.0070798     0
#> 89   2.024554029  8.6000817     0
#> 90   1.827283236 21.0770566     0
#> 91   1.680288861  6.2696104     0
#> 92   1.838763587 16.3546714     0
#> 93   1.430598714 18.1479926     0
#> 94  -0.826390409 13.0266803     0
#> 95   1.577344761  6.6419870     0
#> 96   2.462475493  6.5840813     0
#> 97   1.687427789 19.7194451     0
#> 98   1.880295641  1.1790924     0
#> 99   1.290881778  3.1028581     0
#> 100  0.487699983 18.8878125     0