Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1.

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1).

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times
#> 1   -0.56151538 10.9205965
#> 2    1.19241904  1.1463386
#> 3    0.94714723  3.8842865
#> 4    1.65379533 12.5374759
#> 5    0.50546479 13.9092806
#> 6    1.61782261 20.2677466
#> 7   -0.41246893  6.2534597
#> 8    1.69723632 14.9845999
#> 9    0.66012238  9.4698049
#> 10   1.70129065 19.6818974
#> 11   0.86983952 17.3260843
#> 12   2.34996166 16.1532688
#> 13   0.95872182 13.9453471
#> 14   1.30666682  3.3213749
#> 15   1.82689410 22.9138074
#> 16   1.16841002 22.0826641
#> 17   3.11456170 19.4490795
#> 18   0.55829303  8.2371502
#> 19   1.78188744  8.6158626
#> 20   0.18821395 14.0203525
#> 21   2.93014434 21.9396142
#> 22   1.46064799 19.1390841
#> 23   0.59319039 11.0288252
#> 24   1.25570237  9.2780886
#> 25   2.48422411 23.3290943
#> 26   2.21516113 21.8001266
#> 27   0.02898113  9.5416554
#> 28   2.77571080 18.7165016
#> 29   0.91976683 18.0670677
#> 30   2.07855023 23.5757025
#> 31   1.81729275 19.3290947
#> 32   0.29525024  0.8021512
#> 33   2.07851482 22.8353334
#> 34   4.20404405 19.5514604
#> 35   1.99797148  0.2980619
#> 36   0.76582949 10.0138546
#> 37   2.54012116 18.1500248
#> 38   1.20798125 23.8903848
#> 39   0.96063626 14.0515811
#> 40  -1.50165637 10.1942796
#> 41   1.94824200 15.4073580
#> 42   0.14420431  8.6554043
#> 43  -0.07134168  6.7623389
#> 44   2.12676727  2.2518018
#> 45   1.57301783 17.2045852
#> 46   1.40164154 22.6060483
#> 47   2.04499580 23.7337594
#> 48   0.89055675  4.8658853
#> 49   0.25989620 11.2505519
#> 50   1.80448265 22.0778072
#> 51   3.12055234 17.8673562
#> 52   2.00500657 19.7389310
#> 53   2.11002508 14.7598118
#> 54   1.02900070  0.1022923
#> 55   0.27104972 11.1520443
#> 56   1.44519725 15.9276076
#> 57  -1.54496128  3.8490027
#> 58  -0.42971615  6.1338137
#> 59   1.47162959 16.0656611
#> 60   1.69320401 10.5901744
#> 61  -1.21226226  5.3129975
#> 62   0.32062238 22.7680470
#> 63   1.22847414  3.9847320
#> 64   0.88109714  8.2299906
#> 65  -0.73942140  2.2908545
#> 66  -0.96041515  5.3152273
#> 67   1.37702401  0.1958619
#> 68   0.68928750  2.5524116
#> 69  -0.29275765 12.0732115
#> 70   1.48796717  8.4627619
#> 71   2.34217065 17.5364428
#> 72   2.79120534  1.4910619
#> 73  -0.91139232  6.1173226
#> 74   0.47698489  5.8094157
#> 75   2.53755815 16.7282778
#> 76  -0.15040365  9.3489435
#> 77   1.78523973 13.9622842
#> 78   2.13820446  1.9326404
#> 79   1.00982552 22.9974260
#> 80  -0.86969031  9.0606629
#> 81   1.24919929  5.8497571
#> 82   1.13706957 23.8689208
#> 83   2.07097299 19.2380923
#> 84   0.36998077  7.5732739
#> 85   2.07090796  7.7386554
#> 86   2.28728205 22.7896527
#> 87   0.86207444 17.5093411
#> 88   0.65892101 15.2967790
#> 89  -0.18908256  9.0428584
#> 90  -0.70918297 10.4584740
#> 91   0.90677116 17.5687146
#> 92  -1.99248372  8.3521728
#> 93   1.36763361 16.8751586
#> 94   0.82646007 22.3796251
#> 95   1.73690639 23.8509103
#> 96   2.34662601 18.9637115
#> 97   0.90778051 10.4439101
#> 98   1.09576372  8.9393936
#> 99  -0.31383990  7.0660813
#> 100  1.79676997 10.5478392