Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1.

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times group
#> 1   -0.30268423 12.3006328     0
#> 2    1.14953701  5.8830670     0
#> 3    1.60482295 21.8401409     0
#> 4    2.68390895  1.1963163     0
#> 5   -0.98992479 13.0337031     0
#> 6    0.50842372  2.7004643     0
#> 7    1.13111117  9.1141307     0
#> 8    2.65671456  5.3975548     0
#> 9   -0.04867831 16.7624395     0
#> 10  -0.17854475  5.4129673     0
#> 11  -1.46047571 15.2911099     0
#> 12  -0.72860753 18.1868833     0
#> 13   2.33401945  1.2133763     0
#> 14  -0.38598150 20.7471450     0
#> 15   2.26228698  2.0650492     0
#> 16   1.42324024  9.1679892     0
#> 17   1.68441302  1.4495426     0
#> 18   0.87720411 11.8016390     0
#> 19   2.23707524  5.2656340     0
#> 20   0.55629798 18.3301234     0
#> 21   0.82529190 11.9197961     0
#> 22  -0.60346742 10.1722127     0
#> 23  -0.44729239 10.4306090     0
#> 24  -0.66292667 12.0699381     0
#> 25   1.21888404  1.2167036     0
#> 26   0.73207691 23.6932551     0
#> 27   0.68085886 12.8303608     0
#> 28   0.56473760  0.5778742     0
#> 29   0.05754972 19.4193179     0
#> 30   1.91065950 10.0606189     0
#> 31   0.89166293  0.8200037     0
#> 32   0.33787319  9.8099280     0
#> 33   0.33595087 13.2850706     0
#> 34   0.81506268 20.1424597     0
#> 35  -0.61156916 18.7583142     0
#> 36  -0.39621948 17.4028615     0
#> 37   1.94249995  4.0564747     0
#> 38  -0.60731948 14.0439813     0
#> 39  -0.18118604  8.0922272     0
#> 40   1.13417839 10.8127037     0
#> 41  -1.23660151 16.8108200     0
#> 42   1.66363344  4.6291683     0
#> 43   1.51372540 19.0429486     0
#> 44  -0.14062355 14.9163950     0
#> 45   1.61602755  5.7196101     0
#> 46  -0.75928257 11.7935190     0
#> 47  -1.54904572 16.0336864     0
#> 48   2.19215236 18.7674281     0
#> 49   2.02636571  7.9128321     0
#> 50   0.61589497 10.1320333     0
#> 51   2.51695025  4.5484749     0
#> 52   1.52456435 10.2992705     0
#> 53   2.66747719 20.3838421     0
#> 54   0.76764266  0.6603358     0
#> 55   2.80282826  2.6094499     0
#> 56   0.52243389 15.3756490     0
#> 57   0.71859928 17.3465574     0
#> 58  -1.12360884 14.5180047     0
#> 59   0.54426917  0.8844305     0
#> 60  -0.88734399 16.1305637     0
#> 61   1.18264428  0.2592093     0
#> 62   2.12561864  2.0733443     0
#> 63   1.13593329 18.7833549     0
#> 64  -0.46333383 16.1178874     0
#> 65   3.22200195 22.0119149     0
#> 66   0.53072959 20.5433886     0
#> 67   0.69515755 19.4405313     0
#> 68   1.03276386 14.1121142     0
#> 69   0.77883844 20.9374773     0
#> 70   0.95840536 12.3482227     0
#> 71   0.24745993 15.8606857     0
#> 72   1.54431678 20.5922458     0
#> 73   0.49469427 17.5414530     0
#> 74   0.08712559  9.3577785     0
#> 75   0.47585138 21.1329922     0
#> 76   0.54431517 18.0793253     0
#> 77   0.11567984 22.5999792     0
#> 78  -0.23642131 14.4833280     0
#> 79   3.02201138  4.2933394     0
#> 80  -0.45272146 14.0035023     0
#> 81   1.37671708  7.9195412     0
#> 82   1.77181547  1.1674553     0
#> 83   1.21474801 15.9944803     0
#> 84   3.31353427  7.3516528     0
#> 85   3.10489053  5.0713896     0
#> 86   1.51384566  8.8119240     0
#> 87  -0.32532980 12.1473492     0
#> 88  -1.00942793  6.0352448     0
#> 89  -0.04696569 16.4546915     0
#> 90   1.91367982 13.0335812     0
#> 91  -1.50116208 20.8475137     0
#> 92   1.70391362 22.4755564     0
#> 93  -0.29984346 15.5238635     0
#> 94   0.68308610  9.1332531     0
#> 95   0.92501832  3.4448456     0
#> 96  -0.35080411 15.3001188     0
#> 97   1.69372572  1.5186753     0
#> 98   1.51694875  3.5678334     0
#> 99   2.21698332 19.0200220     0
#> 100  0.60511358 23.5917407     0