Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y       times group
#> 1    0.64252713 18.17485478     0
#> 2    0.32318553 22.94233914     0
#> 3    2.79681022 23.64302318     0
#> 4    0.46195105 21.58915770     0
#> 5    0.77094726 10.16873612     0
#> 6    0.34957822 14.18574849     0
#> 7   -0.96206489 23.58815048     0
#> 8   -0.15080787 15.28093570     0
#> 9    3.62066816  6.02132389     0
#> 10   0.81045531 11.62845379     0
#> 11  -0.09346240 19.00095786     0
#> 12   0.79696773  0.16565501     0
#> 13   3.88903976  3.61592846     0
#> 14   2.07232704 14.09132091     0
#> 15   1.09547835  7.77954545     0
#> 16   1.48372438 17.34060357     0
#> 17   1.83517960 11.77855961     0
#> 18   2.27404456  1.43237967     0
#> 19   1.70035935  0.21279311     0
#> 20   2.31203094 11.62616104     0
#> 21   3.58202582  3.05130895     0
#> 22   0.91978365  1.16924831     0
#> 23   2.50314358 21.75088336     0
#> 24   0.51947826 14.07566461     0
#> 25   0.61298533 17.12843852     0
#> 26   2.08170145  4.19262184     0
#> 27  -1.38015695 20.16747207     0
#> 28   0.50712201 22.39116774     0
#> 29   0.62029369  9.87833064     0
#> 30   1.16707000 10.94105827     0
#> 31   0.06178254 20.63573803     0
#> 32   1.88987167 20.11037314     0
#> 33   0.33868380 19.64508749     0
#> 34   2.78341277  6.27012339     0
#> 35   0.28500753 13.87152487     0
#> 36   2.20389962  4.34006279     0
#> 37   2.15684701  5.26188381     0
#> 38  -0.01450989 21.33821262     0
#> 39   2.11065344  4.93979938     0
#> 40   3.66174639  4.98525031     0
#> 41   0.93556333 16.98689813     0
#> 42   1.60310784  5.40277795     0
#> 43   3.55785506  1.61843952     0
#> 44   1.92149979  4.70931413     0
#> 45   1.72126026 16.14795163     0
#> 46   0.68756446 22.63569570     0
#> 47   2.21622615 23.28884208     0
#> 48  -1.47305136 18.02306954     0
#> 49  -0.49926823 15.39133437     0
#> 50   1.79085682 12.96722567     0
#> 51  -0.59934247 20.41487785     0
#> 52   1.47153667  6.22159398     0
#> 53   2.32136225  5.09906408     0
#> 54   2.33534876 22.56450130     0
#> 55   2.25358782  7.13830881     0
#> 56   0.71591281  3.29676026     0
#> 57  -0.13500537 13.21974677     0
#> 58   1.30261420  7.57127540     0
#> 59   2.66302429  4.10843259     0
#> 60   2.07495700  6.50961681     0
#> 61   0.01264446 19.87604077     0
#> 62   0.76437875  5.19782721     0
#> 63   0.29065785  0.67280416     0
#> 64   2.72118113  4.10833055     0
#> 65   0.08445911 19.39744374     0
#> 66   0.39543954 13.37131921     0
#> 67   0.66349343  0.06206816     0
#> 68   2.23610125 23.30009691     0
#> 69   1.48402107 16.07496048     0
#> 70  -0.56878400 14.20181924     0
#> 71  -0.16685453 14.62541810     0
#> 72   0.82103548  9.63735026     0
#> 73   0.13609746 11.50313836     0
#> 74   0.41241172  0.28605589     0
#> 75   1.80596135  7.28696610     0
#> 76   0.65046372  1.43281101     0
#> 77  -1.13401617 18.25587140     0
#> 78   2.57322093  1.18557783     0
#> 79   3.34861250  7.17071218     0
#> 80   0.26198982  9.96748502     0
#> 81   1.09979923  0.45743387     0
#> 82   0.33404289  9.60526187     0
#> 83   0.93616188  1.06058222     0
#> 84   2.89659090  5.99805016     0
#> 85  -0.95889196 17.30415168     0
#> 86   1.57678729  1.38730907     0
#> 87   1.22944342 12.10264613     0
#> 88   3.24257268  8.63313555     0
#> 89   1.95813212  2.53978798     0
#> 90   1.71374762  6.06620518     0
#> 91  -1.05040874 15.58503652     0
#> 92   1.84632247 23.82675221     0
#> 93  -0.83870377 13.65208060     0
#> 94  -0.42168352  0.30131005     0
#> 95  -0.53023314 16.18033100     0
#> 96   0.03140530 17.57800245     0
#> 97  -0.24521378 16.76918959     0
#> 98   0.69657369  6.10194881     0
#> 99   0.92898876  3.44378916     0
#> 100 -0.25352733 14.01092479     0