Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1.

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1).

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y        times
#> 1    0.14286066 23.401308144
#> 2    2.57268565  0.266399106
#> 3    1.15588382 20.654934140
#> 4   -0.72416424 15.937090991
#> 5   -1.61754012 13.371957013
#> 6    2.47423050  7.192169001
#> 7   -0.10883670 19.707923865
#> 8   -0.29100967 15.041171271
#> 9    1.18588837  2.947519837
#> 10   0.27203468 12.187357128
#> 11  -0.21998425 15.795302762
#> 12   0.48548131  0.642959073
#> 13  -1.78125280 18.491148787
#> 14   1.51739043 16.652583411
#> 15   0.09397848 15.871594833
#> 16   0.26622762 19.998906998
#> 17   0.66024623 17.139323356
#> 18   4.20355542  1.166332634
#> 19  -0.29011879 19.457044423
#> 20   2.38256076  0.708303871
#> 21   3.64909758  0.387452401
#> 22   0.10159665 18.638236865
#> 23   3.45859189  5.690949613
#> 24   0.64322788 10.498087155
#> 25   2.11584560  2.245506853
#> 26   1.54079829  5.507131236
#> 27   2.20806512  1.564067580
#> 28   1.59175881  6.025772469
#> 29   2.18069475  7.483210925
#> 30   2.22562630 22.725069730
#> 31   0.34730913 23.244446615
#> 32   1.80792692  2.879310938
#> 33   0.35653209  4.534872796
#> 34   1.54585234  9.901188992
#> 35   0.89606027  6.414962271
#> 36   2.80063927 23.375638017
#> 37   0.47056782 19.985011898
#> 38   1.57799636 21.945546169
#> 39   1.19729929  1.600483669
#> 40   3.47895805  5.736994447
#> 41   2.11536952  2.951756587
#> 42   2.33738141  2.167246867
#> 43   1.74399294  9.029601038
#> 44   2.59989035  7.976278882
#> 45   0.80246053 17.597100684
#> 46   1.32911105 10.138945101
#> 47   2.19598204  9.223400289
#> 48   0.22884324 16.692803157
#> 49   2.08287888 23.951071166
#> 50   0.59458037 14.643582607
#> 51   1.69956138  8.032191250
#> 52   0.58477006  0.785578866
#> 53   0.67667473 10.413964016
#> 54   0.42699992  4.956987821
#> 55   1.16336692 11.535472853
#> 56  -0.45651970  0.550434340
#> 57   0.47352478 14.672169780
#> 58  -0.57526493 23.540377239
#> 59  -0.24961941 17.642887183
#> 60   0.60326605  8.096312514
#> 61   2.03365036  1.616226688
#> 62   1.48724749  6.956583822
#> 63   0.85901497 13.062710524
#> 64   0.83877368  8.472640973
#> 65   2.05108565  7.654940207
#> 66   1.48177813  0.004146785
#> 67   1.66393523  5.106514178
#> 68   1.57781712 12.908235267
#> 69  -0.14174242 19.712959360
#> 70   0.71345063  6.379353320
#> 71   2.56544229 10.851690475
#> 72   2.74760220 10.354440015
#> 73   0.20761931 22.978753388
#> 74   2.93605818  8.577567268
#> 75   2.43995233 12.084341398
#> 76  -0.44602555 14.250678794
#> 77   0.69194968 17.567591686
#> 78   0.31444470 11.118637953
#> 79  -0.33479645 17.555728337
#> 80   0.12930931 17.211382672
#> 81  -0.19835126 14.308218403
#> 82   1.55763077  8.123411369
#> 83   1.61962960 11.986434722
#> 84   1.72118479  3.683336953
#> 85  -1.55776024 12.328936782
#> 86   2.06633736  3.752727326
#> 87   0.70270562  6.971976779
#> 88   0.99360202 21.580540588
#> 89   2.53264978  8.321143661
#> 90   0.25084258  6.661622493
#> 91  -1.95072511 14.759206815
#> 92   1.94436313  8.132283568
#> 93  -0.40857191 12.093219465
#> 94   1.34615294  9.184366437
#> 95   1.85786523 23.547649859
#> 96  -1.75174179 10.894316934
#> 97   2.22451003  5.959679414
#> 98   0.26530259 12.522769310
#> 99   0.92816035  2.450589566
#> 100  1.18013629 16.038637826