Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>                Y      times group
#> 1    2.018213755  3.4719313     0
#> 2    1.404584693  9.9207755     0
#> 3    1.391846011 11.9696746     0
#> 4    3.478117666 23.8989283     0
#> 5    2.481379602 20.0994309     0
#> 6    1.886261809  8.3878138     0
#> 7    2.200929748 22.9272161     0
#> 8    1.135883849  5.0161748     0
#> 9   -1.087174750 14.8631981     0
#> 10  -0.148958876 19.5436456     0
#> 11   0.945205979 17.6801438     0
#> 12   0.420375022 16.8064472     0
#> 13   0.790018230 18.5674204     0
#> 14   0.323707638 17.6647052     0
#> 15   0.145855259  3.6795767     0
#> 16   0.382394110 13.2226605     0
#> 17   0.784021375  5.7864376     0
#> 18   1.353018529  0.2063340     0
#> 19   1.445334727  0.8940460     0
#> 20   2.565485615 23.9491989     0
#> 21   1.213937760  6.1544244     0
#> 22   3.772355031  2.4912483     0
#> 23   1.244630334  7.8062163     0
#> 24   0.844082759 12.3667428     0
#> 25   1.814926162 20.6384253     0
#> 26   0.557929931  8.3840687     0
#> 27   1.195784691 23.1436254     0
#> 28  -0.359412191 19.6896187     0
#> 29   1.155980719  3.9648252     0
#> 30  -0.097163166 15.5305683     0
#> 31   0.589697768  7.1165745     0
#> 32  -0.760554013 20.0025150     0
#> 33  -1.089732189 23.6060259     0
#> 34  -0.024594036 18.7386679     0
#> 35   0.798125309 23.6882100     0
#> 36  -0.433498061 14.7184927     0
#> 37  -0.022854418 21.9296089     0
#> 38   2.780542314  4.1703223     0
#> 39  -1.172606664 19.8549607     0
#> 40   2.008525059  2.8236523     0
#> 41   3.141618359  5.8743856     0
#> 42   0.172806753 14.5994261     0
#> 43   1.825913805  4.7224464     0
#> 44   2.514874505  5.2288000     0
#> 45  -0.069698415  7.5155813     0
#> 46   0.328141216 15.8019331     0
#> 47   2.239176487 23.9799006     0
#> 48   1.998472596  7.9092595     0
#> 49  -0.060611638  7.7687912     0
#> 50   0.408656615  2.9939010     0
#> 51   1.162073289 20.7874622     0
#> 52  -0.399362692 19.7276187     0
#> 53  -0.696143467 19.1091624     0
#> 54   0.113885280 22.4450495     0
#> 55   2.519405860  7.0152544     0
#> 56   0.170716350 23.5083790     0
#> 57   0.136528408 16.2768196     0
#> 58   0.614164923 10.4538233     0
#> 59   1.958646949  2.9574334     0
#> 60  -0.556515469 16.1890628     0
#> 61   0.982173510  3.4978659     0
#> 62   1.532198627 10.0150712     0
#> 63   1.465703311 23.3533668     0
#> 64   3.180945266  9.0033642     0
#> 65   0.173485046 13.1705753     0
#> 66   1.814956005  0.1329444     0
#> 67   0.799576275 12.4905147     0
#> 68   2.106358834  2.1452029     0
#> 69  -0.009992178 21.0634626     0
#> 70   0.060655381 11.5429145     0
#> 71   1.921132795 22.1791340     0
#> 72   1.242194766 12.0292702     0
#> 73   1.332021142 23.4528896     0
#> 74  -0.463723375 16.8958859     0
#> 75   1.025592418 14.4790045     0
#> 76   2.407486571  6.7481220     0
#> 77   1.140727922 22.8965562     0
#> 78   1.381381832 23.3973692     0
#> 79   1.858806458 12.7023886     0
#> 80   0.722086108  7.4694696     0
#> 81   2.015748790  3.9179849     0
#> 82   2.920150019  6.5061079     0
#> 83   1.949394116  7.1261045     0
#> 84  -0.792212171 15.4362112     0
#> 85   0.851147961 16.1513123     0
#> 86   0.467396810  1.5286637     0
#> 87   2.187972908  9.3647421     0
#> 88   1.415799368  4.7303254     0
#> 89   1.983364741  4.1636327     0
#> 90   0.285831182  8.2240700     0
#> 91   0.105219259 11.5198563     0
#> 92   1.095402196 18.3544516     0
#> 93  -0.051029467  1.7021882     0
#> 94   2.484555886 17.7375372     0
#> 95   0.953696228  3.9033722     0
#> 96   0.298636161  9.8426226     0
#> 97  -1.393008443 11.7662262     0
#> 98  -0.708918668 11.6873773     0
#> 99  -0.597627576 13.6313613     0
#> 100 -0.419779954  8.4598820     0