Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>              Y      times group
#> 1   -0.1131128 16.5076306     0
#> 2   -0.4479698 16.8734752     0
#> 3    1.2954353 23.6656411     0
#> 4    3.5165941  5.4836261     0
#> 5    3.3249083  7.8755719     0
#> 6    1.2286266 16.0827954     0
#> 7   -1.8149415 20.0467024     0
#> 8    0.6102014 11.3399077     0
#> 9    2.0286205  1.1533666     0
#> 10   0.9083101  0.1162071     0
#> 11   3.1613940 22.6233390     0
#> 12   1.2885062  5.0799311     0
#> 13   0.3998265 14.9831852     0
#> 14  -0.8184134 16.9785762     0
#> 15  -0.9103698 16.0902900     0
#> 16   0.8086505 13.8863896     0
#> 17   1.4934030  0.9291062     0
#> 18   2.0730014 23.7460939     0
#> 19   0.2967598 21.6893583     0
#> 20  -0.9296788 16.5440198     0
#> 21  -1.1144626 19.4907661     0
#> 22   2.7128191  3.8076610     0
#> 23   1.2393953 20.7093648     0
#> 24   1.8940341  2.6121445     0
#> 25   2.5021068  0.1810591     0
#> 26   2.5475000  2.3650384     0
#> 27   1.2778744 19.6170442     0
#> 28   3.3419671  4.6761883     0
#> 29   1.2410247 12.6478683     0
#> 30   2.7199340  0.6976246     0
#> 31   1.3104492 21.2436979     0
#> 32  -0.2494875 21.2089217     0
#> 33  -0.7555088 14.4410225     0
#> 34   2.2232642  1.9027816     0
#> 35   1.1701494 22.6322051     0
#> 36   0.8263742 18.4821317     0
#> 37   1.6145560  2.0087105     0
#> 38  -0.4267742  9.6473315     0
#> 39   0.3823806 21.2280963     0
#> 40  -0.1010174 14.5699269     0
#> 41   0.4096288 13.0358719     0
#> 42   3.0841321 22.3603252     0
#> 43   0.5291829 19.1189196     0
#> 44   1.6756614 12.0771666     0
#> 45   4.3558163  5.8829783     0
#> 46   1.0381005 15.3516760     0
#> 47   2.1631899 15.2623682     0
#> 48   1.4352066  5.0582403     0
#> 49   1.6418128 11.8891821     0
#> 50   4.1652866  4.4874835     0
#> 51   0.2400658  3.8020062     0
#> 52   1.5582736 18.5535702     0
#> 53  -0.7488261 15.7345638     0
#> 54   1.0261689 19.8979387     0
#> 55   2.4038547  8.8862387     0
#> 56   0.6276230  1.3766402     0
#> 57   3.2302168  9.6355619     0
#> 58   2.0556626 23.1199749     0
#> 59   1.2799503  6.8795512     0
#> 60   1.0694715  5.1020786     0
#> 61   0.8784765 23.1525410     0
#> 62  -1.6595128 18.8309259     0
#> 63   0.1162735 23.9716376     0
#> 64   3.3883741  4.7458115     0
#> 65  -0.3754787 11.8899323     0
#> 66   1.8711500 21.0177086     0
#> 67   2.6426528  3.6110786     0
#> 68   3.2341150 23.4980624     0
#> 69   0.9296323 21.5619815     0
#> 70   0.9271419  5.0591899     0
#> 71   2.3521859  1.4755590     0
#> 72   4.7332918  3.2157093     0
#> 73   0.7983179  8.2769372     0
#> 74   0.2346267  2.8753608     0
#> 75  -1.0188353 16.2625375     0
#> 76   0.7769532  2.7704830     0
#> 77   1.5595330 12.0970596     0
#> 78  -0.3395855 20.7587924     0
#> 79  -0.2389788  7.1815923     0
#> 80  -0.4856101 14.3408414     0
#> 81   0.8913344 23.7848214     0
#> 82   2.8010092  3.3545818     0
#> 83   2.6591089  2.5666894     0
#> 84   0.1212417 23.5861559     0
#> 85  -0.6858228 14.8951010     0
#> 86   0.1765225 14.6513933     0
#> 87   1.0181889 10.9295202     0
#> 88  -0.5276214 14.1189574     0
#> 89   0.1408894 16.4551878     0
#> 90   0.8514692  1.0608510     0
#> 91   2.0287118 10.8931013     0
#> 92  -0.8802655 16.4458688     0
#> 93   2.9998805  2.1078623     0
#> 94   0.1127848 16.0692248     0
#> 95   1.5885272  1.8175837     0
#> 96  -1.0354524 15.9085300     0
#> 97   0.7061582  2.9806850     0
#> 98   0.6094312 21.7418939     0
#> 99   1.4571377 23.0049715     0
#> 100  2.4964869  4.9723092     0