Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The MESOR is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>                Y      times group
#> 1    0.463781924 17.0416445     0
#> 2    2.439149760  8.5058156     0
#> 3    1.275749385  3.2649309     0
#> 4    3.737928628 23.4872510     0
#> 5    1.233837594 11.5206256     0
#> 6    1.017124995 16.6314498     0
#> 7    3.109520992  0.8625911     0
#> 8    2.635067093  3.1215301     0
#> 9   -1.057484658 16.7177860     0
#> 10   1.104916446 18.4279400     0
#> 11  -0.894799291 17.3431588     0
#> 12   1.123277189 19.0673784     0
#> 13   1.723090819 19.4240706     0
#> 14   0.240242702 14.9043867     0
#> 15   0.200500376 10.7156924     0
#> 16  -2.837761400 15.6135776     0
#> 17   0.450956116 10.5939570     0
#> 18   1.410175190 10.3062445     0
#> 19  -0.800107431 15.2987951     0
#> 20   2.691215586  8.3868251     0
#> 21   2.330158369 21.3098786     0
#> 22   1.743572421  9.5050511     0
#> 23   1.228407317 21.9730083     0
#> 24  -1.044833658 20.1764943     0
#> 25   1.677325511 16.1474885     0
#> 26   2.439940185 23.6398042     0
#> 27   2.320486309 23.8569998     0
#> 28   1.295052372 10.2736890     0
#> 29   1.915064898  4.1287053     0
#> 30   1.638949935  9.9050332     0
#> 31   2.150137975  5.5687638     0
#> 32  -1.179676283 16.2717021     0
#> 33  -1.013233273 17.9772279     0
#> 34   0.302050412 12.6844739     0
#> 35   0.561652298  8.8307081     0
#> 36   0.002417245 13.5628856     0
#> 37   0.701006110 21.5634680     0
#> 38  -0.925550034 12.8258814     0
#> 39  -1.838258403 16.3973945     0
#> 40   2.046705129  8.2154652     0
#> 41   1.125632056  6.0927729     0
#> 42   1.105806134 21.5635770     0
#> 43   1.277917847 17.6045151     0
#> 44   1.429667848 15.1162455     0
#> 45   1.804519326  7.5209915     0
#> 46   1.422303518  6.8486644     0
#> 47   0.981928311  6.8047699     0
#> 48   0.943105106 13.2311629     0
#> 49   2.017242214  4.1307031     0
#> 50   1.980381313  1.6985725     0
#> 51   2.027418241  2.5445588     0
#> 52  -0.678366127  6.7039931     0
#> 53  -0.449021859 17.6313980     0
#> 54  -0.469226156 19.7190570     0
#> 55   1.049303783  0.4792044     0
#> 56  -1.299558539 11.8185721     0
#> 57   0.441280224 18.6061362     0
#> 58   0.712450303 14.0558795     0
#> 59   2.868321791 23.6285764     0
#> 60   2.003827562  2.1590354     0
#> 61  -0.250324678 11.4434292     0
#> 62   1.509350995  4.8095741     0
#> 63   1.750248376  1.0633655     0
#> 64   1.858985196 22.1105416     0
#> 65   3.573923938  2.7330082     0
#> 66   0.783492584  5.7954717     0
#> 67   1.065257679 19.2093865     0
#> 68  -0.736959047 16.1479742     0
#> 69  -0.189082674  4.7768416     0
#> 70   3.521251927 23.7057574     0
#> 71   1.007809492  0.2073320     0
#> 72   0.215498679  8.1549942     0
#> 73   0.874946148 12.0288300     0
#> 74   0.323938633  6.6117305     0
#> 75  -0.123058307 16.1077368     0
#> 76   1.256534724 18.3580638     0
#> 77  -0.030548789 21.1226693     0
#> 78   2.705160870  3.2523725     0
#> 79  -1.384130965 19.4447572     0
#> 80   2.050492893  5.2445916     0
#> 81   0.530303493 13.1816800     0
#> 82   0.549457495 12.1455900     0
#> 83   4.075185972  6.2533781     0
#> 84   0.762551569 18.5861560     0
#> 85   2.213298056 10.3832119     0
#> 86   1.025541739 20.3269613     0
#> 87   0.261215703 15.6512879     0
#> 88  -2.029494421 17.8948149     0
#> 89   1.894093709  9.1929579     0
#> 90  -1.039547132 16.6565072     0
#> 91   0.954505343 16.8757345     0
#> 92   2.053839876 23.9355375     0
#> 93   2.145205292  9.7740825     0
#> 94   3.083316271  2.0352116     0
#> 95  -1.640585651 16.3076330     0
#> 96   2.070197255 14.5993120     0
#> 97  -0.383390638 17.2614226     0
#> 98  -1.005879778 14.0332917     0
#> 99   1.396247222  3.1799339     0
#> 100  0.191820499 12.0636719     0