Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y       times group
#> 1    0.97047916  9.20534213     0
#> 2    0.99109509 22.65488227     0
#> 3    2.08083753  2.48624749     0
#> 4    3.18605621  1.16728063     0
#> 5    3.05986971  6.91009407     0
#> 6   -1.09573038 20.65562275     0
#> 7   -0.92911784 19.66067223     0
#> 8    0.45922303 22.02106783     0
#> 9    1.22197019  1.60284574     0
#> 10   1.86007096  4.42054973     0
#> 11   0.66828689 15.23963942     0
#> 12   1.96654501 16.82360567     0
#> 13   2.81049721  3.58167755     0
#> 14  -0.91697714 19.27574747     0
#> 15   2.49640017  4.75145556     0
#> 16   0.93702715 16.64257803     0
#> 17  -0.65610689 16.54790399     0
#> 18   2.03959625 17.71057880     0
#> 19   1.09688523  1.65572841     0
#> 20   2.75902126  1.25605117     0
#> 21   2.92228178  3.97064004     0
#> 22   1.03578350  9.46552692     0
#> 23   2.07644981 23.28489300     0
#> 24   2.55102616  7.52691384     0
#> 25  -1.50161655 13.93599520     0
#> 26   0.48276066 18.25204970     0
#> 27   2.73203513  6.60694995     0
#> 28  -1.16020647 16.32606400     0
#> 29   1.26368052 20.44091261     0
#> 30   1.95231541  2.09918802     0
#> 31   1.72910721  1.84747447     0
#> 32   2.53545367 23.45314424     0
#> 33   2.09310399  0.37452908     0
#> 34   0.47787255 18.27307640     0
#> 35   1.12512130  2.92517658     0
#> 36   1.01019122  0.16274735     0
#> 37   1.84930491  1.76132621     0
#> 38   3.57383231  9.67087345     0
#> 39   2.00058340  5.39509598     0
#> 40   2.11917726  9.71606607     0
#> 41  -1.48256313 10.92833993     0
#> 42   1.65557985 12.66765338     0
#> 43  -0.48160454 15.77431460     0
#> 44   2.08123151  4.89866610     0
#> 45   2.43503159  8.14023481     0
#> 46   0.78971679 16.09065738     0
#> 47  -0.41541830 12.12468245     0
#> 48  -0.06874590 12.58730570     0
#> 49   0.64822753  0.63825306     0
#> 50   1.00754959 15.74702901     0
#> 51   2.12693728 13.73221830     0
#> 52   1.91180010  4.58080249     0
#> 53   1.70191199 10.96524116     0
#> 54   2.97072934  1.12339331     0
#> 55  -1.24663291 14.49389304     0
#> 56   1.14918106  8.40456202     0
#> 57   1.84940625 21.96399189     0
#> 58  -0.59701559  8.93963880     0
#> 59   0.30126804  4.13051557     0
#> 60   2.56684656  0.02684762     0
#> 61  -0.69961313 13.33138197     0
#> 62   0.03614723 21.16531237     0
#> 63   2.54851344  0.24363607     0
#> 64   2.48577653  5.41302523     0
#> 65   1.91046395  4.28243823     0
#> 66   3.05593488  3.38391352     0
#> 67   0.79882987 12.69314759     0
#> 68   0.15489700 17.29829339     0
#> 69   2.74812883  5.90448939     0
#> 70   0.71042033  9.43623345     0
#> 71   0.69265791 20.06805190     0
#> 72  -1.80877094 16.68446496     0
#> 73   2.68657003  5.15374509     0
#> 74   2.82922369  3.07711888     0
#> 75   2.67982819  1.79553575     0
#> 76  -0.05897165 18.41250128     0
#> 77   2.34895894 12.77920210     0
#> 78   0.06413234 11.57828112     0
#> 79   1.32117643  1.06653991     0
#> 80   3.00014076  3.25294207     0
#> 81  -0.29946261 18.82217421     0
#> 82   0.79530997 11.23804605     0
#> 83   2.18607397  2.08848509     0
#> 84   0.03991835 18.76070550     0
#> 85   0.72186436 15.55335538     0
#> 86  -0.34915336 17.32292410     0
#> 87  -0.13792328 21.50370458     0
#> 88   0.99972880 22.73948045     0
#> 89   2.36812645 23.15563534     0
#> 90   0.26868896 10.52463470     0
#> 91   2.04937128 21.44629821     0
#> 92   0.72041144 21.72101481     0
#> 93   2.32652721 10.71584716     0
#> 94  -0.61996922 12.63091153     0
#> 95   1.01639473  6.54673120     0
#> 96   1.91750951 11.25824752     0
#> 97   1.76617769 12.66885313     0
#> 98   2.81747705 23.59158673     0
#> 99   0.69416783 20.26884990     0
#> 100  0.87094539 17.88311171     0