Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The MESOR is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y       times group
#> 1   -1.38762067 12.10325108     0
#> 2   -0.90264587 14.73574001     0
#> 3    2.81254405  9.08764706     0
#> 4    0.68703470  8.95411880     0
#> 5    1.38092329  2.95529508     0
#> 6    1.42974712 10.95154390     0
#> 7    1.12863470  2.17333920     0
#> 8    0.46344531 10.26944630     0
#> 9    2.92249750 22.19281060     0
#> 10   2.48388117  4.37651950     0
#> 11   0.95746628 12.88885167     0
#> 12   0.42343984  0.01411989     0
#> 13   0.16994029 20.85737392     0
#> 14   1.03350195  6.53273501     0
#> 15   2.81421344  3.09054364     0
#> 16   0.98706084  2.71908090     0
#> 17   1.27875278 11.58562633     0
#> 18   1.18374941  8.66380371     0
#> 19   2.15052895 10.30528634     0
#> 20   0.29195163 14.37784626     0
#> 21   1.21118425  6.44854902     0
#> 22   0.73322123  8.08790400     0
#> 23  -0.85377648 12.98061850     0
#> 24   1.40621851 11.15241965     0
#> 25   0.20803445 11.99941920     0
#> 26   2.63109632  5.64473954     0
#> 27   0.57022432  3.87905693     0
#> 28   0.70319532 12.55489030     0
#> 29   0.55543459 16.93458088     0
#> 30   0.49140658 13.36645136     0
#> 31   0.11775925 16.83726608     0
#> 32   0.49978380 19.72929370     0
#> 33   0.84548667 21.98693108     0
#> 34   0.78848079 12.92074605     0
#> 35   0.14120781 13.93529687     0
#> 36   0.70324443 21.42810463     0
#> 37   1.16207527 11.50695480     0
#> 38  -0.61987347 21.46172166     0
#> 39  -1.13464217 15.32438846     0
#> 40   3.54983447  2.65247268     0
#> 41   2.23209322 18.18481689     0
#> 42   0.90029987  9.07750235     0
#> 43   1.84079886  0.70145873     0
#> 44   2.30516286 22.51778364     0
#> 45  -0.49008206 16.80430625     0
#> 46  -0.93240868 10.28073757     0
#> 47   1.33889411  3.96944690     0
#> 48   1.88250001 16.61815219     0
#> 49   1.51321740  8.57714412     0
#> 50   2.64056046  9.25743182     0
#> 51   0.32057519 21.92554187     0
#> 52  -0.47970200 20.10644322     0
#> 53   1.48571992  4.50170735     0
#> 54   1.12826588 15.56100688     0
#> 55   1.04377988 18.55747814     0
#> 56   2.55600733  6.49583475     0
#> 57   1.22314082 12.95746725     0
#> 58   1.21370630 22.26362511     0
#> 59   1.66206335  8.27252797     0
#> 60   1.43865402 12.93691206     0
#> 61   2.10115362  1.42096708     0
#> 62   0.06372369  5.08383583     0
#> 63   1.86925554  9.64873294     0
#> 64   1.76951148  5.49115480     0
#> 65  -1.34047250 15.34436703     0
#> 66   3.97874720  7.53698264     0
#> 67   0.77895415  8.39254790     0
#> 68   0.61247998 20.41976250     0
#> 69  -1.99391710 17.35502681     0
#> 70   0.32373000 19.16303745     0
#> 71   2.81681383  5.07934495     0
#> 72   2.64675043  0.34090319     0
#> 73   2.87129006  1.69424108     0
#> 74   0.44765515 10.87960332     0
#> 75  -0.20153318 12.64326735     0
#> 76   0.52305097 12.97332074     0
#> 77   2.43211112  5.95848327     0
#> 78   0.29119709 19.19508889     0
#> 79   1.90307119  6.18126867     0
#> 80   2.06071694 22.62475762     0
#> 81   0.63347638 11.54216580     0
#> 82  -0.96801797 12.79746506     0
#> 83   2.18441186  7.59676406     0
#> 84  -0.42355122 17.20097411     0
#> 85   1.27346596 11.13893782     0
#> 86  -0.61922788 23.17295966     0
#> 87   1.66992610 18.70752719     0
#> 88   2.81810494  5.88201610     0
#> 89   0.25988502 10.38891977     0
#> 90   0.18793032 18.98878704     0
#> 91   1.80138339 22.38295949     0
#> 92   1.32901319 21.04378675     0
#> 93   1.48136108 20.86865840     0
#> 94   1.42850938  5.38864269     0
#> 95   0.71677809 19.25263555     0
#> 96  -0.49852536 15.59903977     0
#> 97   0.89203528 19.65374422     0
#> 98   1.09141175 11.13826151     0
#> 99   1.51091665 19.26103254     0
#> 100  0.22807591 12.50483133     0