Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1.

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1).

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times
#> 1   -0.01552746 13.9347178
#> 2    1.23864859  4.1640915
#> 3    2.98382658 15.5944886
#> 4    1.44764907 14.1722934
#> 5    0.70207974  2.7373868
#> 6    1.93439492 23.6619721
#> 7   -1.44736376 18.7360130
#> 8    2.15601147 18.7177663
#> 9    1.21851828 15.3546405
#> 10   1.47645883  6.0958294
#> 11   2.62513636 22.2938881
#> 12  -0.78825255  1.5737123
#> 13  -0.66590298  6.8439576
#> 14   1.85304003 17.3850959
#> 15   0.31911320 15.2195832
#> 16   1.22807283 12.9137336
#> 17   1.58289574  4.8669961
#> 18   1.33147966  2.0500516
#> 19   0.14983761  3.8587728
#> 20  -0.25797567  5.5778207
#> 21   1.48493251  3.0993148
#> 22   3.92991836 18.3024033
#> 23   2.25858418 11.4937169
#> 24   1.69100346 12.0031652
#> 25   2.06470840  0.9889126
#> 26   0.28484279  2.0397212
#> 27   0.46726313 14.9480532
#> 28   0.53331659 14.0330759
#> 29  -0.20926072 10.1786109
#> 30   0.59098833  9.6391456
#> 31   0.41112950 16.7279533
#> 32   0.64144818 14.5846642
#> 33   0.50457921  6.0944177
#> 34   2.95916452 21.5260032
#> 35   3.50422516 18.8074961
#> 36   1.07058811  5.3975300
#> 37   1.13453365 14.3523229
#> 38   1.16349107 10.0757019
#> 39   1.12584757 13.2686997
#> 40   0.18503226 15.6569050
#> 41   1.44923373 17.1706995
#> 42   3.12188399 19.2164584
#> 43   1.13979252 17.9319224
#> 44   1.97426381 22.3699680
#> 45  -2.13501830  9.7336555
#> 46  -0.48790398 20.4428761
#> 47   2.50540013 20.0561192
#> 48  -0.95152950 10.6704431
#> 49   1.35941270  6.8237775
#> 50   0.79252966  6.3380493
#> 51   0.84170214 15.5046462
#> 52  -0.25568907  8.3724684
#> 53   2.21722058  3.9037291
#> 54   1.18244227  8.4197737
#> 55   0.06374564  6.7367691
#> 56   3.18567501 10.9816009
#> 57   2.64855458 21.9658805
#> 58  -1.16709134 11.4123302
#> 59  -1.38028533  6.7487078
#> 60   2.02178288 21.9133149
#> 61   2.70618590 21.0645560
#> 62   1.77659887  2.4773977
#> 63   0.59412625  9.3078949
#> 64   3.27022599 22.2230087
#> 65  -0.73864165  9.9887809
#> 66   0.73943342 11.6013826
#> 67   0.48509834 14.0684419
#> 68   2.39876180 16.6429775
#> 69   1.44818049 19.1888725
#> 70   2.78849604 21.2597269
#> 71   0.62132586  5.7545287
#> 72   0.62275586  1.9709505
#> 73   1.05953948  1.7971511
#> 74   3.10865876  1.6381982
#> 75   2.74083918 19.3865804
#> 76  -1.79316317  7.1339484
#> 77   0.35532559  9.9660480
#> 78   1.56276078  4.4236474
#> 79   1.03247857 21.3442066
#> 80   0.06728588 12.9345724
#> 81   1.53074411  5.8015406
#> 82   1.26191885  1.6513632
#> 83   2.48167928 20.4240307
#> 84  -1.47348475 10.9870156
#> 85  -0.65529161 13.0159099
#> 86   1.81792429 20.5826428
#> 87  -1.38634132 11.6574089
#> 88   1.81516090 13.7730374
#> 89   0.26741162  9.8128607
#> 90   0.25435667  3.4235187
#> 91   0.19213041  8.7082140
#> 92  -0.16226154  2.3363323
#> 93   1.44035204 13.6622654
#> 94   1.25388450 21.4105117
#> 95   0.61492733  1.6852738
#> 96  -0.73800056  5.9968549
#> 97   1.29238939 12.5973267
#> 98  -1.15050610  8.5346963
#> 99   0.83559043 17.2679414
#> 100  2.00154120 19.1514644