Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

Amplitude parameter term for group = 0.If multiple components, specify the parameters for each component as a vector. E.g: amp = c(1,2) for two components.

acro

Acrophase parameter for group = 0. In units of radians.If multiple components, specify the parameters for each component as a vector. E.g: acr = c(1,2) for two components.

period

The period of data. If multiple components, specify the parameters for each component as a vector. E.g: period = c(12,6) for two components.

n_components

The number of components in the model. This must match the length of the parameter inputs.

beta.group

A logical argument. TRUE if you want to simulate second group dataset. If FALSE, beta. arguments will be ignored.

beta.mesor

Amplitude parameter term for group = 1.If multiple components, specify the parameters for each component as a vector. E.g: amp = c(1,2) for two components.

beta.amp

Acrophase parameter for group = 1. In units of radians.If multiple components, specify the parameters for each component as a vector. E.g: acr = c(1,2) for two components.

beta.acro

Mesor parameter for group = 1.If multiple components, specify the parameters for each component as a vector. E.g: mesor = c(1,2) for two components.

n_period

is the number of periods that are simulated.

family

family of simulated dataset expressed as a string. Can take values: "poisson", "binomial", "gamma", "gaussian".

...

Extra arguments, such as alpha parameter for the gamma simulation, or sd (standard deviation) for gaussian simulation.

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y       times
#> 1   -0.37424698  8.72637409
#> 2    1.62284712 19.83400925
#> 3    1.40665321 22.85510704
#> 4    2.02422496 20.46899880
#> 5   -0.11710413  6.15582861
#> 6    0.44734075 14.89307308
#> 7    1.60329820  4.28899761
#> 8    0.34519122  0.50265054
#> 9    0.86878644  7.55212776
#> 10  -0.67858112  2.07703274
#> 11   3.82707562 23.26450924
#> 12   0.11133377 16.45548676
#> 13  -0.24628326 13.33439406
#> 14   1.19814436 18.32972664
#> 15   3.54084153 17.90272956
#> 16  -0.46091387  4.24465633
#> 17  -0.07189566 11.58296049
#> 18   2.03373318  1.52679594
#> 19  -0.34307951 11.54102684
#> 20   2.75468729 23.01004608
#> 21   0.74141738 21.99234888
#> 22   1.57338883  0.46814693
#> 23   0.07979789 14.55224040
#> 24   2.39951774 14.89717231
#> 25   2.28203920 14.80765850
#> 26   1.01836421  6.88168467
#> 27   1.48187746  5.41880771
#> 28   0.73425762 15.74122906
#> 29   0.80326714  2.86596947
#> 30   3.35195662 22.49539344
#> 31   1.16932301 16.86709980
#> 32   0.79253404  2.27464464
#> 33   2.23327123 20.24484312
#> 34   0.34973231  4.98971146
#> 35   2.62066571 23.88757581
#> 36  -0.13663597 16.66749817
#> 37   2.57372731  2.04040145
#> 38   0.61163368  4.71574885
#> 39   1.95201113 17.73301714
#> 40   1.69187948 21.73236495
#> 41   0.94168375  5.70059463
#> 42   1.42313976  8.90303840
#> 43   2.60279392 12.17102176
#> 44   1.16410015  9.86501969
#> 45   1.19112452 20.64119783
#> 46   0.69673856 20.66182762
#> 47   1.94468017 22.96869048
#> 48   0.80635456 22.64427810
#> 49  -0.62167902  5.32447623
#> 50  -0.58635808  9.94667211
#> 51   1.34385027 16.34788399
#> 52   2.45824111 20.88994823
#> 53   3.06594147 14.60887846
#> 54   1.06900423  4.31209879
#> 55  -0.17135459  6.27258103
#> 56   0.96213048  8.21711513
#> 57  -0.03287551 15.97285651
#> 58   0.13689830  9.59454553
#> 59   1.93160117 16.90050580
#> 60   3.08388422 18.33657395
#> 61   1.49396480 14.63362600
#> 62   3.59668126 18.18899049
#> 63  -1.27864737  5.39949065
#> 64  -0.77829075  8.99729490
#> 65  -1.07911537  6.79787441
#> 66   1.60221849 21.20029802
#> 67   1.32230993  1.43313222
#> 68   2.49886917 22.32398948
#> 69   1.32051354  9.85734432
#> 70   0.58219253 12.20342921
#> 71   1.01593900 16.24955392
#> 72   3.04278413 17.58390649
#> 73   0.81657129  8.66961274
#> 74   1.00939145 15.33372257
#> 75   1.06940703 21.20350370
#> 76   0.72685938  9.69053796
#> 77   1.52899990 21.49510743
#> 78   1.26618055 12.75745891
#> 79   1.96091822  4.59558725
#> 80   3.02706428 22.83299075
#> 81  -1.11713560  9.91064303
#> 82   0.68207055 10.25556109
#> 83   1.00791469  8.17716572
#> 84   0.46250848  0.16496600
#> 85  -0.88901502  7.54551733
#> 86  -0.41289529 11.92061098
#> 87   1.91178716  3.53331637
#> 88   0.56188224 13.96501947
#> 89   1.69111852 23.74438744
#> 90   2.24683020 20.33154569
#> 91   0.21166945 14.35569531
#> 92   0.49948985  1.33539641
#> 93   3.17829545  1.25039974
#> 94   1.98615998  3.35993759
#> 95   0.94769790 18.18525413
#> 96   1.13695638 23.86123565
#> 97   1.63857928  0.08175283
#> 98   3.17229706 19.35607472
#> 99   0.79891834  6.23764461
#> 100 -0.26833587  2.10463803