Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The MESOR is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times group
#> 1    0.97197466 13.4590653     0
#> 2   -0.16985953 12.6700984     0
#> 3    1.23455988  8.9602163     0
#> 4    1.39769883  7.4659377     0
#> 5    1.96046943  5.2167394     0
#> 6    1.47971737  6.6737800     0
#> 7   -0.44641813 12.4909568     0
#> 8   -0.31290213  9.7069211     0
#> 9    1.22866557 14.3112464     0
#> 10   1.81046142 14.6539459     0
#> 11   0.62803364  7.6890803     0
#> 12   1.24808965 20.2972271     0
#> 13  -0.41334359 18.5865174     0
#> 14  -0.31586212 16.0832851     0
#> 15   1.23633338 16.4986001     0
#> 16   2.02012242 23.3126399     0
#> 17   0.47824801  5.7342429     0
#> 18   1.92842358  9.9275690     0
#> 19   2.18682982  9.7011061     0
#> 20   3.56198203  5.4213973     0
#> 21  -0.65502769 14.3910474     0
#> 22   3.30897866  2.1930394     0
#> 23   2.08346125 20.2050513     0
#> 24  -0.26742096 18.1987240     0
#> 25  -0.67600552 15.7302050     0
#> 26   1.08444017  6.6179846     0
#> 27   2.44759980 20.8975379     0
#> 28   1.94751274 22.6768126     0
#> 29   0.01786028 19.2130673     0
#> 30   0.68943085 21.4845066     0
#> 31   1.14055400 10.6325440     0
#> 32   2.36254304  4.1696383     0
#> 33   0.78576814 12.9551909     0
#> 34   0.72759674 15.7200045     0
#> 35  -0.70811631 17.1937730     0
#> 36   1.77799618  5.4272532     0
#> 37   0.75511912 14.0829762     0
#> 38   0.99701449  8.3866517     0
#> 39   2.99052716  4.1859282     0
#> 40   1.12256544 12.2516107     0
#> 41   2.42487695  1.5716866     0
#> 42   1.13106371  8.8034543     0
#> 43   1.91590759  0.4270879     0
#> 44   1.34892239  8.0705808     0
#> 45   2.31343866  9.1340636     0
#> 46   1.73983913  5.6292467     0
#> 47  -0.24962986  3.6595662     0
#> 48   0.42482954 16.6765305     0
#> 49  -0.64224323 16.4412897     0
#> 50   3.43543874 23.9541611     0
#> 51   1.62453202  7.0704828     0
#> 52  -1.99850265 14.7843310     0
#> 53  -0.95085125 16.2282699     0
#> 54   2.03309866  6.4762761     0
#> 55   2.00552940 11.2880483     0
#> 56   2.04127949  9.0612532     0
#> 57  -0.10822030 22.8495489     0
#> 58  -0.38123972 11.6178118     0
#> 59   2.34237336  1.1863854     0
#> 60  -2.22438301 23.1360391     0
#> 61   0.14058949 17.3326163     0
#> 62  -2.89260638 17.5284393     0
#> 63   2.08437218  4.4624404     0
#> 64   1.15383317  7.5335656     0
#> 65   1.79244864 18.8994734     0
#> 66   1.13385556 16.3020898     0
#> 67   0.60966931  6.8185261     0
#> 68   1.16444572 20.2479495     0
#> 69   3.86353804  7.4471334     0
#> 70   0.32887748  5.6930132     0
#> 71  -1.37529254 15.9684179     0
#> 72   1.48071520  1.6110190     0
#> 73  -0.22034708 17.9155627     0
#> 74  -1.36905127 18.9910093     0
#> 75  -0.14448867  9.8333963     0
#> 76   1.11912295 19.2354314     0
#> 77   0.25565923 15.0352469     0
#> 78   1.05608116 20.6094814     0
#> 79   1.64442607  8.9844460     0
#> 80   1.97669472  9.9781259     0
#> 81  -0.20699976 21.9569419     0
#> 82   1.57025951  5.3780991     0
#> 83   1.00726853 21.9779082     0
#> 84  -0.24694092 10.6389135     0
#> 85   2.40017243  1.7864997     0
#> 86   3.38185845  4.4518273     0
#> 87  -1.02387620 10.3940905     0
#> 88  -0.28069483  7.7648511     0
#> 89   0.60746325 20.3463396     0
#> 90   1.24546580 21.7634091     0
#> 91   0.72128318  1.5954302     0
#> 92   2.98494334 11.5477428     0
#> 93   0.09600790 22.3430136     0
#> 94  -0.17750921 14.2241189     0
#> 95   0.79318429 23.7448188     0
#> 96   2.33681855  3.2589050     0
#> 97   3.03327342 18.3140716     0
#> 98   2.69478435 21.7680123     0
#> 99  -0.01859236 14.0049142     0
#> 100  3.90110005  6.3001784     0