Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times group
#> 1    0.22464986 11.1614985     0
#> 2    2.36713701  5.3583947     0
#> 3    0.79657504  2.9509659     0
#> 4    1.19091913  9.5558790     0
#> 5    0.21428040 20.1887532     0
#> 6    2.67646585  0.9806916     0
#> 7   -0.08451624 14.5890519     0
#> 8   -0.33126098 19.3815990     0
#> 9    0.89676895  6.2659256     0
#> 10  -1.25589017 18.4823533     0
#> 11   0.35552489 11.8543826     0
#> 12   1.18145416 14.0002930     0
#> 13   1.11466461  4.3083609     0
#> 14  -0.27138068 19.3358216     0
#> 15  -0.43179577 12.2890705     0
#> 16   0.45633473 21.5757380     0
#> 17   0.86448569 12.3339271     0
#> 18   2.60517440  1.5946586     0
#> 19  -0.27644403 12.9079882     0
#> 20   0.63930238 14.5926759     0
#> 21   2.01441269 20.8904663     0
#> 22   1.01960176  2.7020210     0
#> 23   1.67347437 20.0079191     0
#> 24   1.99260344 22.3004147     0
#> 25   0.56875880 14.3833774     0
#> 26   1.97466172  1.7047419     0
#> 27   1.06791769 23.7334585     0
#> 28   0.59664356  7.9306059     0
#> 29  -1.04830810 16.2334478     0
#> 30  -0.23708832 14.7798360     0
#> 31   1.43625799 10.1012202     0
#> 32   0.28372904  8.2738622     0
#> 33   1.48443107 13.4530815     0
#> 34   1.80885816  7.0218402     0
#> 35  -0.01784789 14.7998194     0
#> 36  -0.53436251 14.2536649     0
#> 37   1.78906044 14.7426446     0
#> 38   2.39156909 19.6822615     0
#> 39  -1.61902147 11.5693579     0
#> 40   0.37155568  6.8400212     0
#> 41   1.44658173  1.4942867     0
#> 42   1.23142233  5.4075807     0
#> 43   1.21766427 14.9772448     0
#> 44   0.86747269  1.0118070     0
#> 45  -0.09966670 16.1262463     0
#> 46   1.48610548 20.1053701     0
#> 47   1.84092363  8.5586650     0
#> 48  -0.10277318 11.4416548     0
#> 49   0.27088749 16.8418033     0
#> 50   0.02361586 11.7511065     0
#> 51   2.13391749  0.3144710     0
#> 52  -1.08591271 14.7753535     0
#> 53   0.75054033 23.4299809     0
#> 54   1.27873175  5.5389585     0
#> 55   0.96782778 13.1982580     0
#> 56   1.90411460  8.5236501     0
#> 57   0.73045276 15.2247064     0
#> 58   1.08825771  8.1925066     0
#> 59   0.42961989 19.2767794     0
#> 60   1.00496540 23.9994194     0
#> 61   0.52555045 13.2709578     0
#> 62   0.21067268 21.8336992     0
#> 63   1.60785840  2.0370126     0
#> 64  -0.12788882 21.5004219     0
#> 65  -0.05770785 14.9797290     0
#> 66   1.67996128 16.8863156     0
#> 67   3.04540521  2.9312296     0
#> 68   1.65888756  4.8215720     0
#> 69   0.87033491 22.3528634     0
#> 70   2.14016311  1.4629497     0
#> 71   0.83559931 16.3365475     0
#> 72  -1.19599589 11.8159936     0
#> 73   1.77818670  6.7050542     0
#> 74   1.25994410 11.5745615     0
#> 75   0.76896665 10.1729588     0
#> 76   1.69535556  5.9635922     0
#> 77   2.01801479  5.3100242     0
#> 78   1.19898818 16.6264951     0
#> 79   0.72608079 16.9208536     0
#> 80   1.43973134 21.4347916     0
#> 81   1.88568209  8.0861485     0
#> 82   1.61154538 21.0878436     0
#> 83   1.30696202  6.1167827     0
#> 84   1.60863636  1.6044428     0
#> 85   0.45821448 10.5384897     0
#> 86   0.78086773  5.1660538     0
#> 87   2.58005792 19.9963987     0
#> 88   0.74634669 20.4170671     0
#> 89   0.38642528 18.7327437     0
#> 90   0.63573184  7.4535329     0
#> 91   1.04139450  1.1563788     0
#> 92   1.61470014 11.1930068     0
#> 93  -1.15648922 18.0369681     0
#> 94   1.25943723  1.7822149     0
#> 95   1.49322003  2.2339130     0
#> 96   1.51147879  2.8075899     0
#> 97   2.09704250  6.8796820     0
#> 98   1.11234745 21.6502371     0
#> 99   0.98860227  6.7764406     0
#> 100  1.05791365  4.2925940     0