Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times group
#> 1    1.32681462  9.2804071     0
#> 2    2.27031022  1.1985154     0
#> 3    1.44650182 18.2006291     0
#> 4    0.85840617  0.5934994     0
#> 5    1.31937570  6.6635338     0
#> 6    0.45775171 14.6229340     0
#> 7    0.32261559 10.8414858     0
#> 8   -1.51344173 15.3775729     0
#> 9   -0.10428827 13.8921018     0
#> 10   1.25218410 22.2623332     0
#> 11  -0.83575586 11.4779494     0
#> 12  -0.44722497 20.5934804     0
#> 13  -0.28442150 23.9721511     0
#> 14   0.41163700 12.0534407     0
#> 15  -0.46773178 14.0769733     0
#> 16   3.57512618  2.8561555     0
#> 17   1.14624574 19.1229108     0
#> 18   0.25561266  8.9255936     0
#> 19   1.25347691 18.1690123     0
#> 20   0.72478295  0.5756451     0
#> 21  -0.20301596 14.2568219     0
#> 22   0.61164852 15.4539844     0
#> 23   0.47721430 21.8368626     0
#> 24   1.78599710  9.6971686     0
#> 25   0.08373114  7.5778712     0
#> 26   3.36528562  2.6513893     0
#> 27  -1.60880337 15.9292739     0
#> 28   0.94739652 10.0440563     0
#> 29   0.84273920 20.0592827     0
#> 30   0.31263715 12.7029034     0
#> 31  -0.27063766 15.8787186     0
#> 32   0.05787712 10.7786120     0
#> 33   0.81188085 12.0884063     0
#> 34  -0.28843590 10.8352672     0
#> 35   1.05108506  5.2606939     0
#> 36  -1.15988556 17.5362220     0
#> 37   0.55259312  0.9580655     0
#> 38   1.24827531  1.6905079     0
#> 39   1.46001943  4.8195056     0
#> 40   1.69878573 21.0167876     0
#> 41   0.46713875 16.4137397     0
#> 42  -0.98419231 17.0338215     0
#> 43   1.05121295  9.3810462     0
#> 44   0.15860935 15.7103502     0
#> 45   1.65065922  2.6443829     0
#> 46   1.57135408  6.9409618     0
#> 47   0.12720792  3.6065270     0
#> 48   2.03605905  5.7469259     0
#> 49   0.02378326 17.9633265     0
#> 50   0.37958444 23.7441116     0
#> 51   2.30261186  8.8582494     0
#> 52   1.49997294  1.1193424     0
#> 53  -1.37069196 17.8448173     0
#> 54  -0.69410247 13.0514428     0
#> 55   2.31141821  4.2063053     0
#> 56   0.04823585 15.6056139     0
#> 57   1.99014222  3.8415229     0
#> 58   0.92086714  2.2310352     0
#> 59  -1.51835547 14.7462260     0
#> 60  -1.29441881 15.8551883     0
#> 61   0.65559367 16.7669859     0
#> 62   1.75443849  7.7350907     0
#> 63   0.79866811 11.9473897     0
#> 64   2.95699063 20.3091367     0
#> 65   3.00298588  7.0421347     0
#> 66   0.37903088 10.7210098     0
#> 67  -0.28925928 23.8300981     0
#> 68   0.89497415 18.6620137     0
#> 69   0.90079777  6.6434084     0
#> 70   1.40281838  3.1700955     0
#> 71   2.12879186  8.0976426     0
#> 72   1.96709742  4.4056868     0
#> 73   0.57752983  0.3757766     0
#> 74   1.63060478  7.6687239     0
#> 75   1.42170747  3.9382496     0
#> 76   1.00582122  6.1085287     0
#> 77  -0.05124817  1.1809688     0
#> 78   3.27069372  2.2167454     0
#> 79   1.70732213  4.5095385     0
#> 80   2.12686443  5.7897092     0
#> 81   0.74653969  1.7513846     0
#> 82  -0.19877147 21.4220352     0
#> 83   0.40827451 13.9812664     0
#> 84  -0.65925855 13.3332744     0
#> 85   1.99028641  6.0961487     0
#> 86  -0.13198403 18.0996662     0
#> 87  -0.38703895 16.2811649     0
#> 88   3.06372054  5.2270063     0
#> 89   2.22893642  3.9564215     0
#> 90   2.70201103  2.3275670     0
#> 91   1.49314515 10.3171987     0
#> 92   0.97391695  4.5928646     0
#> 93   3.07330273  4.2422016     0
#> 94   0.73417579 20.7078575     0
#> 95   1.80807435  2.0152717     0
#> 96   2.56613571 13.3397077     0
#> 97   1.71951959  4.8977659     0
#> 98   1.54893969  7.1942803     0
#> 99   0.13322233 18.2031105     0
#> 100 -0.18477140 17.8664463     0