Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1.

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>                Y      times group
#> 1    1.596912058  2.0281174     0
#> 2    1.710958161  7.8692484     0
#> 3    2.330203886 12.5414064     0
#> 4    1.741663399 20.7310958     0
#> 5    0.909095163 18.4827131     0
#> 6   -1.138114165 13.6331715     0
#> 7   -1.640926691 13.7790060     0
#> 8    1.732064806  8.6098594     0
#> 9   -0.629501546 17.1385230     0
#> 10  -0.171734375 22.7075860     0
#> 11   1.510914960 23.3738221     0
#> 12  -0.054185692 16.8360638     0
#> 13   1.905602867  7.0544810     0
#> 14   1.526610830  1.9128197     0
#> 15  -0.119105204  1.3078729     0
#> 16   0.803964683 10.7443365     0
#> 17   2.235369291  3.5887698     0
#> 18  -0.001968847 17.3262816     0
#> 19   1.601471570  0.5001834     0
#> 20   1.041122708 11.5371074     0
#> 21   1.344544031  0.5096702     0
#> 22   2.164575517  4.9134509     0
#> 23  -0.404260020 16.7000186     0
#> 24   1.234290052  4.2514223     0
#> 25  -1.620934377 12.7594147     0
#> 26   2.571769711  3.9814001     0
#> 27  -0.830727162 13.7719565     0
#> 28   0.405465855 16.1383757     0
#> 29   2.016871912  3.8499310     0
#> 30   1.475436776 14.6434502     0
#> 31   1.322017323 20.8961194     0
#> 32   1.747512724 23.3077013     0
#> 33   0.946141186  9.6535440     0
#> 34   1.146250867  4.9113968     0
#> 35   1.697258373 21.7268836     0
#> 36   2.653350799  7.4294483     0
#> 37   0.970976044  7.6088404     0
#> 38   1.359581601  8.1746848     0
#> 39   1.810488280 18.2991744     0
#> 40   1.770589457  6.0658797     0
#> 41  -1.524115250 16.7415367     0
#> 42   3.153855898  6.1693112     0
#> 43   1.347370007 22.1626512     0
#> 44   1.140448891  6.8560516     0
#> 45  -0.258004122 15.7307575     0
#> 46   2.355014460  1.9880003     0
#> 47   2.666125138  2.6153337     0
#> 48   1.142973475  7.6464891     0
#> 49   0.583736015 18.6703481     0
#> 50   1.691861316 23.1011233     0
#> 51  -0.995847603 18.2945407     0
#> 52   3.086624066  2.9318427     0
#> 53   0.589908340  2.4810183     0
#> 54   2.959396242  3.1088801     0
#> 55   1.636469684  5.1424131     0
#> 56   0.963408661 21.7961158     0
#> 57   0.918297185 22.1470928     0
#> 58   0.650032911 14.0441111     0
#> 59   2.474749272  2.9552860     0
#> 60  -0.205248471 20.3072581     0
#> 61  -0.126737095 11.3588549     0
#> 62   3.317562798  5.9633064     0
#> 63   1.557875347  9.6394362     0
#> 64  -0.435297967  9.4464832     0
#> 65   2.689372730  6.0926396     0
#> 66   0.456574317 17.8934323     0
#> 67   1.578543076  4.2488578     0
#> 68   2.461449104  5.0830156     0
#> 69   1.997009005 20.0782063     0
#> 70   1.395811774 18.4028485     0
#> 71  -1.782220027 15.3132314     0
#> 72   1.539223913  1.4456286     0
#> 73   1.307052907 13.7991668     0
#> 74   0.347965971 17.1016940     0
#> 75   0.594203789 17.6491480     0
#> 76   3.653944905 21.1979642     0
#> 77  -1.041728682 19.7360368     0
#> 78   1.402313364  6.7329399     0
#> 79  -0.169178721 16.0164875     0
#> 80   0.475251966 13.3757347     0
#> 81   1.393060917  4.7460677     0
#> 82   1.526946272 16.7257565     0
#> 83   1.086751734 22.4172081     0
#> 84   2.109505255 22.3609513     0
#> 85   2.359762828  1.0402511     0
#> 86  -0.934016460 16.2940517     0
#> 87  -0.248612413 19.4443347     0
#> 88   2.151846473  4.0936937     0
#> 89   0.514271407 19.9205285     0
#> 90   1.873822185 10.8245990     0
#> 91  -0.741731810 18.6024118     0
#> 92   2.115484051  1.6492339     0
#> 93   2.727937592  3.0489350     0
#> 94  -1.823294525 17.6226601     0
#> 95   1.428343107 19.3426545     0
#> 96   1.680171985 22.6670924     0
#> 97   2.921883658  8.9206732     0
#> 98   1.241318718  1.6591742     0
#> 99  -0.316476818  3.2495632     0
#> 100  0.008700711 22.0171622     0