Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The MESOR is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times group
#> 1    0.40641083 17.5149785     0
#> 2    0.66477614 23.0031813     0
#> 3    0.42008694  6.1699171     0
#> 4    2.19210329  4.0416985     0
#> 5    1.46118628 23.1404772     0
#> 6   -0.88727055 16.0683839     0
#> 7    0.69076821 11.5141095     0
#> 8   -1.54476204 15.1611250     0
#> 9    3.78017648  3.4039405     0
#> 10   1.13544155  8.5632214     0
#> 11   1.63432715 13.5128211     0
#> 12  -0.37005457 17.3651183     0
#> 13   0.51530366 13.9000239     0
#> 14  -0.92787886 13.5444317     0
#> 15  -0.37390628 17.0226147     0
#> 16   2.35141996  9.2991733     0
#> 17   0.14147532  4.4635404     0
#> 18   0.90132317  6.1314000     0
#> 19   0.26019643 22.6000652     0
#> 20   0.20860922 16.5349769     0
#> 21  -1.07931493 15.8627025     0
#> 22   0.65535379 10.5083357     0
#> 23   0.25770108 19.2342139     0
#> 24   0.67751868 10.0018495     0
#> 25   1.16726321 23.5438386     0
#> 26   0.85307637 12.6232675     0
#> 27   1.52239096 20.0975451     0
#> 28   0.67379176 20.9189824     0
#> 29   1.57323993 17.1418654     0
#> 30   1.40760804  5.9787532     0
#> 31   0.66206170  3.1211875     0
#> 32  -0.42045444 16.0416437     0
#> 33   2.08808662  1.8167201     0
#> 34   0.26652358 23.8255898     0
#> 35  -1.37942043 17.8330879     0
#> 36   0.91180279  8.6445419     0
#> 37   2.39942859 23.8254969     0
#> 38  -0.83994117  1.1201110     0
#> 39   1.20269705  7.4810739     0
#> 40   1.82521703 22.6608236     0
#> 41  -0.76568057 13.8963719     0
#> 42   0.73689649  8.3895425     0
#> 43   1.78800717  7.5111689     0
#> 44   0.49579025 15.9313665     0
#> 45   1.24295252  2.9060391     0
#> 46   0.37887476 16.8603900     0
#> 47   1.90827213  3.0784524     0
#> 48  -0.01809592 10.8847489     0
#> 49   2.23771700 20.7686285     0
#> 50  -0.32694713 16.1167784     0
#> 51   0.73928796 17.3628627     0
#> 52  -0.02154172 16.9579551     0
#> 53  -1.01280848 11.0905445     0
#> 54   1.18150270 19.0941372     0
#> 55   0.36924312  5.6954819     0
#> 56   1.78386898  6.2214874     0
#> 57  -0.10522577 16.7107719     0
#> 58  -0.01458683 18.3879036     0
#> 59   0.35165788 14.0896653     0
#> 60   3.05899764 17.8794144     0
#> 61   0.67040796 22.4895010     0
#> 62   0.42014547 19.4323426     0
#> 63  -0.47859738 18.7287462     0
#> 64   1.94159404  9.3769935     0
#> 65   4.57154475  4.1080742     0
#> 66   0.66357527  5.3154766     0
#> 67  -0.10583418 22.5384049     0
#> 68   1.08288504  1.7671800     0
#> 69   0.95108106  8.0690430     0
#> 70   1.69299599  0.9142790     0
#> 71  -0.12056685  8.5365101     0
#> 72   2.30223027  2.6242727     0
#> 73  -0.22268896 15.4885884     0
#> 74  -0.69844874  9.4660053     0
#> 75   0.29712106  6.9071511     0
#> 76   2.02137323  7.3854029     0
#> 77  -0.51814721  2.3963320     0
#> 78   0.79032991 17.0785710     0
#> 79   2.31125849  7.0884579     0
#> 80  -0.28904127 13.4834535     0
#> 81   3.14101994  2.2351107     0
#> 82   2.40719171  7.7283235     0
#> 83   1.75918996 22.2852706     0
#> 84   2.44219756  1.3682077     0
#> 85   0.69360182  1.0327829     0
#> 86  -2.21795768 17.3304210     0
#> 87   0.83004292 22.3804380     0
#> 88   2.65090960  7.3729415     0
#> 89   0.76980279 16.1811833     0
#> 90   0.33864401 15.2291065     0
#> 91   0.94011266 20.8715051     0
#> 92   0.21850320 17.2904151     0
#> 93   2.22781689 10.5752427     0
#> 94   0.32319253 23.2572323     0
#> 95   1.34868797 17.4448362     0
#> 96   2.50167879  7.5319520     0
#> 97   4.66104281 13.5100901     0
#> 98   0.05450224 10.2607454     0
#> 99   0.80248986  9.7465426     0
#> 100  0.83959273  0.6901682     0