Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1.

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1).

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>                Y      times
#> 1    2.610796460 18.3994991
#> 2    0.529061414  5.2709851
#> 3    2.454499705  1.7530528
#> 4   -0.077960210  3.8422743
#> 5    0.190595645  0.8203762
#> 6   -0.290503991  0.4554946
#> 7    1.150234955  1.0301303
#> 8    2.237875797  0.5230229
#> 9    0.654555605 19.9196324
#> 10   1.076480756 19.6761414
#> 11   1.788070813  5.0628113
#> 12   0.019655387 20.5443546
#> 13   1.250779706  1.5958344
#> 14  -0.939563720 12.1694290
#> 15   2.584671494  3.0874689
#> 16   0.301251348  0.4527873
#> 17   1.660345899 11.2350924
#> 18   0.519741676 17.5099854
#> 19   0.810342733  5.4754282
#> 20   3.695892790 19.8088015
#> 21   2.645543520 20.0873250
#> 22   3.293116627 20.7655868
#> 23  -0.255093362  6.3617193
#> 24   0.858452605  9.2218552
#> 25   0.652973841  8.0227931
#> 26   2.683944335 22.1962835
#> 27   0.773875810  8.6390096
#> 28   3.547250950 18.3089127
#> 29   1.958246303 23.2167863
#> 30   0.505453860  7.8732621
#> 31   1.163893214  2.8788435
#> 32   0.957276176 17.2226994
#> 33   1.090701006 17.5936790
#> 34  -1.859031671  8.4218930
#> 35  -0.375073173  6.9281728
#> 36   0.560995226  5.4480305
#> 37  -1.121301124 10.3649556
#> 38   3.155251853 17.4566091
#> 39   1.843585194 13.4279570
#> 40   3.074950984 20.3094503
#> 41   2.559321947  0.1017348
#> 42   3.429796734 18.5738841
#> 43   0.889744657 22.4999757
#> 44   0.922627793  4.8376612
#> 45  -0.420928443  6.9800608
#> 46  -2.410782734  9.4614413
#> 47   2.721281320 20.7039758
#> 48   1.987240044 15.2315003
#> 49   1.542685782  1.1273990
#> 50   0.303725440  1.7791426
#> 51   0.673331431  8.9144196
#> 52   1.332962636 11.3962109
#> 53   1.226436814 17.3179441
#> 54  -0.105426200 16.0843009
#> 55   0.362945694 14.1667565
#> 56   3.840129693 17.8309944
#> 57   1.567856337 17.4741983
#> 58   1.880838020 23.1629740
#> 59   1.430947853 13.2447788
#> 60   0.881128225 17.5922172
#> 61   1.109996236 15.9864878
#> 62   0.328802194 21.5986014
#> 63   0.555475340  9.5440813
#> 64  -0.941260849 10.7130488
#> 65   2.749997022 23.3354915
#> 66   0.824501329 13.8641961
#> 67  -0.003895306 11.1228142
#> 68   0.799825010 12.2821840
#> 69   3.115753897 19.3842138
#> 70   3.926266247 16.1760919
#> 71   0.288374577 10.1840349
#> 72  -1.891653283  9.2742413
#> 73   2.957468662 22.7032618
#> 74   0.985388386 15.7437703
#> 75   0.982372751 21.5052985
#> 76   0.260325007 19.0352954
#> 77   2.125625524  6.2892128
#> 78   1.843099465  3.7326871
#> 79   0.075285942  1.1683888
#> 80  -0.344959150  3.5453564
#> 81   1.043318577  8.1521188
#> 82   1.357460758 11.9855515
#> 83   0.858095838 17.7275002
#> 84   1.565365594 10.2182404
#> 85   1.234887081  0.1641455
#> 86  -0.470858639  5.8976230
#> 87  -0.324101801 12.5206117
#> 88   0.380768331  1.6101964
#> 89   1.842307328 15.8348537
#> 90  -2.639823171  8.0233233
#> 91   0.065861174  3.8227624
#> 92   0.410392244  3.1408038
#> 93   1.672362383 23.1864750
#> 94   1.476231832 23.8455154
#> 95   1.846919013 20.6573046
#> 96  -2.132121082  8.2221030
#> 97   3.215728793  0.9050694
#> 98   0.551010260 11.7954217
#> 99   1.810196757  1.8545260
#> 100  1.204215277  0.1912148