Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1.

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1).

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y       times
#> 1    0.16503403  3.82964350
#> 2    1.16131259  0.06379133
#> 3    1.59156768 12.78986825
#> 4   -0.02779549  7.03271075
#> 5    1.11835175  0.48689328
#> 6    0.77090686  2.80784068
#> 7    0.51940227 20.80547320
#> 8    2.41400889 20.64334945
#> 9    1.45741821 23.38088535
#> 10   0.16529324 15.93540891
#> 11  -1.88070527 11.56506713
#> 12   0.35614027 16.63967021
#> 13   0.86600462 10.22067246
#> 14   1.15371352 16.21773542
#> 15   1.55309565  4.48301771
#> 16  -0.35328245  7.63940325
#> 17   1.61401795 10.83061242
#> 18   0.98257695 11.86400921
#> 19   0.56625208 23.60128737
#> 20   0.94124201 12.83488036
#> 21   2.11420308 20.67857198
#> 22   3.05697218 22.48899420
#> 23   0.31264451 21.79392038
#> 24   2.55114790  5.87668105
#> 25   1.10907894  2.23111168
#> 26   0.52709464  0.97391717
#> 27   1.08782195 17.76038939
#> 28   1.82840733 23.74025749
#> 29   0.66378241 10.14760972
#> 30  -0.08550583  2.26623672
#> 31   1.07143496  3.45527900
#> 32   3.28515650 18.01079021
#> 33   1.32775343  2.40323206
#> 34   0.04493763 12.53554705
#> 35   2.08383223 15.48428753
#> 36   0.88696179 10.67083311
#> 37   0.54049637  2.30906886
#> 38  -0.01771717  4.24688419
#> 39   2.38668763 17.40163860
#> 40  -0.24458927  9.87141516
#> 41   1.34700504  2.93487078
#> 42   1.88767258 13.75574561
#> 43   0.50896870 19.23139436
#> 44   1.15798749 13.82902389
#> 45   2.17411163 16.95895595
#> 46   0.62996705 18.24722357
#> 47  -0.42082769  8.47415774
#> 48   1.36878029 16.20053596
#> 49   0.66877047 11.57029568
#> 50   0.23075800  3.17376029
#> 51   0.98635144 17.83775025
#> 52  -0.16404181 13.33540974
#> 53  -0.64971695 12.57291365
#> 54   1.99891287 17.49681074
#> 55   2.49679253 17.71451685
#> 56   0.30665614  1.94635474
#> 57  -0.16808018  8.24116978
#> 58   0.72666203 15.56410353
#> 59   1.93778862  0.70948396
#> 60  -1.60793971 10.14614442
#> 61   0.20175576 11.09970844
#> 62   0.68924536 23.74667403
#> 63   1.05761140  3.18482293
#> 64   2.23878794 18.85238579
#> 65  -0.69625054 16.50031100
#> 66  -2.01323666 10.88526947
#> 67  -1.16533831  9.84440497
#> 68   3.57384460 13.79469055
#> 69  -0.21008209 12.69874356
#> 70   1.21712669 13.23069094
#> 71  -0.01568270  8.92053816
#> 72  -0.84928452  6.72416329
#> 73  -1.03346060  7.77586092
#> 74   0.74383229  2.95589935
#> 75   0.35572719  5.15534427
#> 76   0.47688453 10.46313879
#> 77  -1.31861926  7.81898272
#> 78   0.32328755 20.72948122
#> 79  -0.81292210  2.92092334
#> 80   0.90637069 22.34864356
#> 81   0.83793937 13.52310831
#> 82   1.41066665  2.88665293
#> 83   1.63664770 18.22134606
#> 84   3.24582420 21.77655311
#> 85   0.45727894  6.00500976
#> 86   0.61432286 18.58607600
#> 87  -0.66331513  5.91498182
#> 88   2.46684880  2.19948842
#> 89   2.42686806 16.73449444
#> 90   2.45787538 17.73918884
#> 91  -0.51613036 11.72919544
#> 92   1.08732460 11.19697086
#> 93   2.10598934  0.98212375
#> 94   2.93430516 21.04685292
#> 95   2.81721958 17.36005695
#> 96   0.83098067 12.99265274
#> 97  -0.52231072  6.39917427
#> 98   1.06519155  1.10433979
#> 99   1.49721795 15.61550793
#> 100  3.43339105 21.97045475