Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times group
#> 1    2.41031055  0.7441893     0
#> 2    0.53831141 13.9881823     0
#> 3    2.58036761 22.3328980     0
#> 4   -0.74023311 17.5032077     0
#> 5   -0.76419326 14.3371666     0
#> 6    0.70682022  1.7902735     0
#> 7    1.11575181 21.7613106     0
#> 8    2.65930327  9.4038487     0
#> 9    2.25437571  6.5742439     0
#> 10   3.23640051  0.2728275     0
#> 11  -0.66550270 13.7850818     0
#> 12   1.06371753 10.7206557     0
#> 13   0.26045085 16.1866856     0
#> 14   0.15196272 15.8858980     0
#> 15   0.01863245 10.4037332     0
#> 16   1.88673412 17.0468993     0
#> 17   2.76891641  5.8848953     0
#> 18  -1.74041015 10.5982733     0
#> 19   0.62476639  2.7131692     0
#> 20  -1.12115340 12.5512549     0
#> 21   0.64381890 21.6141685     0
#> 22   1.01463462 21.8932942     0
#> 23   0.42478347 18.0853715     0
#> 24   0.66490935 18.1507252     0
#> 25   0.29988084 23.2548228     0
#> 26   0.47034739 19.0023351     0
#> 27   0.07753373 12.2190638     0
#> 28   0.32533215  7.4023635     0
#> 29   0.88309800 11.3363257     0
#> 30   3.34555393  3.4256886     0
#> 31  -1.06226212  8.2591667     0
#> 32   0.42951025 11.4730956     0
#> 33   0.97459074 22.6102863     0
#> 34   0.52817828 20.0871557     0
#> 35  -0.54239157 19.5225839     0
#> 36   0.57495875 11.2245939     0
#> 37  -0.31459404 22.7006671     0
#> 38   2.31960292  5.6691069     0
#> 39   2.30616621  6.0545001     0
#> 40  -1.67750696 20.4650704     0
#> 41   0.64191849  9.5552634     0
#> 42   2.09475303  7.1550819     0
#> 43   1.87087146  1.7991561     0
#> 44   0.80731039 17.6377293     0
#> 45   0.85180163  2.6126048     0
#> 46   2.97409337  6.0423021     0
#> 47   0.26032332 15.3265401     0
#> 48   1.26366486 16.7341473     0
#> 49   1.14339476  5.5134362     0
#> 50   0.81802129  2.7391108     0
#> 51   2.59920556  3.0299278     0
#> 52   0.01143446 17.6400028     0
#> 53   0.05801154 19.2517281     0
#> 54   0.39243830  8.8681062     0
#> 55  -0.89118671 20.8284513     0
#> 56   0.99381462 10.6070960     0
#> 57   3.19395489  5.8017718     0
#> 58   4.62421227  5.6273850     0
#> 59   0.34631933 12.8767130     0
#> 60   0.50400684 14.1429265     0
#> 61   2.84655402  3.0704056     0
#> 62   2.80841147  4.2799704     0
#> 63   2.59481214  6.8502592     0
#> 64   0.49416096 11.4537861     0
#> 65   0.79022797 21.9500548     0
#> 66   2.15573706  0.1955235     0
#> 67  -1.02386248 19.8928267     0
#> 68   0.50950002 17.6994469     0
#> 69  -0.67400088 21.7333819     0
#> 70   1.28936252  7.0635380     0
#> 71   0.80551768 15.5965878     0
#> 72   0.42227798 12.1760724     0
#> 73  -1.60066064 17.1539862     0
#> 74   1.76624009  3.8407820     0
#> 75   1.74990590  6.6215835     0
#> 76   1.78819303  8.8613402     0
#> 77   1.15517637 21.6459186     0
#> 78   3.21498706  6.0445458     0
#> 79   1.66269750 22.6155742     0
#> 80  -0.23090779 17.6289566     0
#> 81   1.34662439  4.7706326     0
#> 82   1.51079540  1.8413722     0
#> 83   2.50310516  5.2163267     0
#> 84   1.22449370 19.5153589     0
#> 85   1.13961442  0.4558677     0
#> 86  -0.04892934 17.1813870     0
#> 87   2.01649959 23.4761677     0
#> 88   0.25551831  8.7420499     0
#> 89  -0.80086455 19.8305701     0
#> 90  -0.88313820 12.3534583     0
#> 91   0.61983745  9.5745788     0
#> 92   0.31757448 11.2399815     0
#> 93   1.14372870  5.5157605     0
#> 94  -0.82644810 15.4957915     0
#> 95   1.28575742 19.3063133     0
#> 96   1.68737740 23.9819362     0
#> 97   0.65051876 13.7660750     0
#> 98  -2.30634694 12.7022209     0
#> 99   2.38402542  1.2101788     0
#> 100  0.32041182 17.6367124     0