Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>                 Y       times group
#> 1    2.9425637453  2.80099143     0
#> 2    0.3306243435 11.92479663     0
#> 3    0.0927790104 15.36480475     0
#> 4    2.1370491620 22.20776345     0
#> 5    2.7688291106  0.84737966     0
#> 6    1.8474874638  8.84265870     0
#> 7    1.4355435938 12.85624475     0
#> 8    0.6794573833  5.89006818     0
#> 9    2.2011775432 22.54915109     0
#> 10   1.2935693947 14.51923621     0
#> 11   0.1460540086 22.18685689     0
#> 12  -0.4467828452 22.05853228     0
#> 13   0.6898233040 12.94751638     0
#> 14  -1.0222387239 12.19774680     0
#> 15   0.5799935080 12.31933939     0
#> 16   0.1787597613  5.13629478     0
#> 17   1.3180758548 17.61084298     0
#> 18   1.9091801141  3.70861275     0
#> 19   1.6174281200  5.99441490     0
#> 20  -0.9975698867 21.63833709     0
#> 21   2.2309563536 21.59491329     0
#> 22   1.5273722219 22.61903720     0
#> 23   3.0733079053  3.35286251     0
#> 24   0.0001720458 15.31592814     0
#> 25   1.7658162491  2.35502963     0
#> 26   0.2781539015 13.95838741     0
#> 27   1.4701560277  7.97823464     0
#> 28   1.9478805222  4.85751700     0
#> 29   1.6653584084 22.16001205     0
#> 30   2.2380029267  3.87810485     0
#> 31   0.6718037035  6.08567281     0
#> 32   2.5139146313  0.08291176     0
#> 33   3.2042931637  5.25122205     0
#> 34   1.9978638100  2.48812518     0
#> 35   0.5940964822 10.18800195     0
#> 36   0.5876071003 17.14220282     0
#> 37   3.9168195196  3.51136756     0
#> 38  -1.1337478475 16.27460537     0
#> 39   2.5396369868  7.07157869     0
#> 40   1.7521499538  0.06884329     0
#> 41   0.5436742342  1.48139958     0
#> 42  -0.8852198395 14.92599633     0
#> 43   2.5882408909 22.96708573     0
#> 44   1.0178385063  4.67184125     0
#> 45   2.5960859512  6.11452645     0
#> 46   0.9502552357 20.13591563     0
#> 47   0.7814272549  9.31347574     0
#> 48   0.6713589227 21.89577018     0
#> 49   1.0713900440 20.50454950     0
#> 50  -1.9124369967 14.41442844     0
#> 51   2.0899341048  0.04512219     0
#> 52  -0.7954292804 16.99108551     0
#> 53  -0.8892409951 21.59405383     0
#> 54   0.5547896380 20.73670824     0
#> 55  -0.1315262180 13.83826348     0
#> 56   1.8024494224  2.07186318     0
#> 57   1.6188328331 20.43535630     0
#> 58   1.2594939080 21.78315296     0
#> 59   0.5612179607  6.93450637     0
#> 60   2.2665866389 10.39108409     0
#> 61   0.8106523295 21.90849989     0
#> 62   1.7800575580 13.71736054     0
#> 63   0.3991507139 19.02290474     0
#> 64   0.9096003038 17.56948702     0
#> 65  -0.5909479436 16.75919087     0
#> 66  -0.6562725524 15.37968698     0
#> 67   0.2513743880  4.93557872     0
#> 68   0.8182686696 16.08990160     0
#> 69   2.8374883020  5.07717472     0
#> 70  -0.3460381869 21.18877940     0
#> 71   2.0438649354  1.32330451     0
#> 72   0.0185614296 11.42519246     0
#> 73   2.2706666640  4.09557251     0
#> 74  -0.6148232771 14.93198613     0
#> 75  -1.0738644896 19.13942992     0
#> 76   3.7891944672  2.45364785     0
#> 77   2.2425479867  8.28193748     0
#> 78  -0.3166596285 21.51501393     0
#> 79   1.0931865390  3.01866707     0
#> 80   1.1180145306  9.38776477     0
#> 81   0.7624362414 22.14289340     0
#> 82   1.5404974479  1.06250722     0
#> 83   1.8372883290  0.75205354     0
#> 84  -1.4850101683 13.65077019     0
#> 85   1.7615447377  3.44130552     0
#> 86  -0.3464028598 13.74419896     0
#> 87   0.4927045080 23.61099086     0
#> 88   3.6368134868  5.14296962     0
#> 89   1.0918281602 13.44886717     0
#> 90   1.3866391959  0.94851831     0
#> 91   2.0846260634  6.84830342     0
#> 92   2.4907649362 23.74851610     0
#> 93  -0.2864718217 15.62204037     0
#> 94   0.9417420396 12.74978742     0
#> 95   1.1233364739 13.21081692     0
#> 96   2.0440529895 22.32800162     0
#> 97   2.0739035531  4.16922603     0
#> 98   2.9702133218  5.78986183     0
#> 99  -0.7536836464 16.70902712     0
#> 100  1.8001744544  1.09354610     0