Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times group
#> 1    0.52758298  1.5440839     0
#> 2    1.10283139 11.0925041     0
#> 3    1.13332736 10.8945918     0
#> 4    0.88918440  0.5924902     0
#> 5   -1.18627461 20.4307770     0
#> 6   -0.10940311 12.8459111     0
#> 7    1.30051643  4.9998166     0
#> 8    1.18829406 19.6807499     0
#> 9   -0.15547334 17.8154530     0
#> 10  -0.83735586 16.0975779     0
#> 11   0.87508024 12.7884684     0
#> 12   2.77727048 23.9087796     0
#> 13   1.32200731 23.2992346     0
#> 14  -0.36579116  6.9925862     0
#> 15  -0.53856788 13.0939149     0
#> 16   1.75799300 23.2138923     0
#> 17   1.54037079  7.4686573     0
#> 18   1.80397960 10.1011933     0
#> 19   3.81104273  1.7440568     0
#> 20   1.10526083 16.8072187     0
#> 21   1.87267691 14.9361847     0
#> 22   0.67374448 14.4661994     0
#> 23   0.67186570 10.9567281     0
#> 24   0.77929383 16.9677024     0
#> 25   0.20499936  8.9462328     0
#> 26   0.32486329 12.3422528     0
#> 27   2.50301156  6.3260589     0
#> 28   0.87027011 23.1958474     0
#> 29   2.20598770 19.7838905     0
#> 30   3.17609664  3.0339010     0
#> 31  -0.46910333 19.2786183     0
#> 32   2.54389664 12.9420204     0
#> 33  -0.32039055 18.3328059     0
#> 34   1.17443467 21.7460363     0
#> 35   3.04598633  5.2924673     0
#> 36  -0.65969091 13.3969989     0
#> 37   1.29553830  8.1235814     0
#> 38  -0.04454872 19.6154124     0
#> 39   1.52593319  7.9301474     0
#> 40   0.38670073  8.4103690     0
#> 41   0.22068370 16.1199907     0
#> 42  -0.06486137 15.8113061     0
#> 43   1.13330997  8.0203182     0
#> 44  -0.01030693 13.5459284     0
#> 45   0.20738159 13.1756378     0
#> 46   1.44251117  8.2183626     0
#> 47  -1.01867879 13.0380838     0
#> 48   0.82433635 13.4312538     0
#> 49  -0.37737853 17.5260769     0
#> 50   0.38783694 17.4453755     0
#> 51   0.48393180 13.0197805     0
#> 52   3.12223904  5.4642863     0
#> 53   1.61404176  4.8632312     0
#> 54   2.98680342  1.9782711     0
#> 55   1.56942039 10.4771695     0
#> 56   2.28063803  1.5938384     0
#> 57   1.28496576 13.4212793     0
#> 58   0.04522488 15.4220280     0
#> 59   0.93878346 13.7997949     0
#> 60   1.95605601  7.2222052     0
#> 61   2.06403262  3.3452009     0
#> 62  -0.05911593 17.4597747     0
#> 63   0.55842250  7.0571275     0
#> 64   0.96357862 19.8121501     0
#> 65   2.34227836  0.3233937     0
#> 66  -0.59680364 14.5836265     0
#> 67   2.73261782 23.0243769     0
#> 68   0.80071612 20.9668530     0
#> 69   1.39927160 23.7439947     0
#> 70  -2.00142300 17.5509903     0
#> 71   0.96516698  6.8859540     0
#> 72   0.80357757 19.2122023     0
#> 73   1.71782866  4.4839716     0
#> 74   1.48913174 11.7169536     0
#> 75   1.62807839  4.1196586     0
#> 76   0.46988210  7.1307571     0
#> 77   0.70060285 22.0767681     0
#> 78   3.51291591  5.6571748     0
#> 79   2.58693144  6.8562081     0
#> 80   2.07516171  6.5999380     0
#> 81   0.65984001 16.8742333     0
#> 82   0.48752187 10.8446960     0
#> 83   0.66803161 12.6856968     0
#> 84   1.80342284  5.4784489     0
#> 85   1.34713758  2.3504279     0
#> 86   1.03612256  9.5249127     0
#> 87   2.20138179 13.8423576     0
#> 88  -0.89159622 15.4982694     0
#> 89  -0.13292170 14.9621761     0
#> 90   0.73456931  1.9354904     0
#> 91   1.10879558  1.9964826     0
#> 92   1.32633149  2.6537575     0
#> 93   3.09118809  7.2187851     0
#> 94  -0.93624637 12.9509284     0
#> 95   1.85884708  8.9127806     0
#> 96   2.98890873  0.6753976     0
#> 97  -1.45823100 19.7030705     0
#> 98  -0.18309104 16.0279693     0
#> 99   1.74085411  8.8490493     0
#> 100  0.88350164  7.7371031     0