Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times group
#> 1   -0.91888473 15.7095080     0
#> 2    0.87022610 18.1924277     0
#> 3    1.70251806  1.5400785     0
#> 4   -0.22353929 22.2791450     0
#> 5    1.51956054  6.5973014     0
#> 6    3.06021333  1.4314695     0
#> 7    0.54039083  0.4799471     0
#> 8    1.98609357  7.7668902     0
#> 9   -1.55328286 15.6224524     0
#> 10   2.40726473  0.9885760     0
#> 11   1.74671514  7.1414410     0
#> 12   1.43881564 12.3844054     0
#> 13   0.88381013  6.4110600     0
#> 14   2.12549987  2.8988965     0
#> 15   1.79398385 12.4081197     0
#> 16  -1.22518329 16.3267175     0
#> 17   0.06954025 12.0488853     0
#> 18   1.56911407 17.0977580     0
#> 19   1.13872614 22.2331179     0
#> 20   2.09980717  6.6229266     0
#> 21   1.55826479 21.6490009     0
#> 22   2.26605961 12.7845237     0
#> 23  -0.91643842 18.1242279     0
#> 24   2.80952146 20.9296749     0
#> 25   1.00610970 21.5688280     0
#> 26  -1.37851419 11.0274904     0
#> 27   0.46415798 21.4511779     0
#> 28   1.51597890 21.0700151     0
#> 29  -0.35971363 15.4186765     0
#> 30   0.87815899 22.7648849     0
#> 31   2.49439152 13.4162812     0
#> 32   0.30701446 23.6184641     0
#> 33   2.06690932  6.1669978     0
#> 34   1.15978611  9.1890448     0
#> 35   0.18513500 21.5399611     0
#> 36   1.81324100 12.5726444     0
#> 37  -0.57552619 18.1861911     0
#> 38   0.84513411 13.0897506     0
#> 39  -0.08592878  9.1309275     0
#> 40   1.20135550 22.7256645     0
#> 41   0.60603440  1.9576371     0
#> 42   0.70920549  8.2232479     0
#> 43  -0.28275721 13.8658495     0
#> 44   1.26526605  1.3944174     0
#> 45  -0.27847203 19.6306647     0
#> 46  -0.07188098 20.6367198     0
#> 47   0.70090262  3.2659391     0
#> 48   1.80234338  2.1676531     0
#> 49   1.48738395 22.8264269     0
#> 50   2.00105860  4.1335690     0
#> 51  -0.52749298 13.9236478     0
#> 52  -1.84118487 14.3064295     0
#> 53   1.76314515  6.4571850     0
#> 54  -0.20940514 15.8712900     0
#> 55   1.74796100 21.4635444     0
#> 56   0.68474053 12.2085813     0
#> 57   0.86782367 14.6195801     0
#> 58   0.26576722  9.7946859     0
#> 59   1.50280101 18.2952670     0
#> 60   0.59679223 19.0647852     0
#> 61   2.43113470  8.6090174     0
#> 62   1.72861589 21.3044382     0
#> 63   0.79538333 10.3628295     0
#> 64  -1.37222236 12.3244138     0
#> 65   1.86026491 20.0978331     0
#> 66   1.38705552 17.3684912     0
#> 67   0.29959283 21.8364671     0
#> 68   2.47214310  1.6417032     0
#> 69   2.64269924  3.8292016     0
#> 70   2.03569301 20.2259521     0
#> 71   0.24246163 21.4748575     0
#> 72   1.94700726  0.7893344     0
#> 73   0.43729142 13.7654802     0
#> 74   2.46133039 22.7123371     0
#> 75  -0.27040548  9.2415240     0
#> 76   0.80635621 20.7359294     0
#> 77   0.71517145 22.1828962     0
#> 78   1.64404087 19.6044766     0
#> 79   0.83250889 17.5433885     0
#> 80   2.38951468  6.3099942     0
#> 81   1.25319759 21.7127529     0
#> 82   1.48486510  3.2677569     0
#> 83   0.46221749  4.7624624     0
#> 84   3.13384498  3.1864116     0
#> 85  -0.77493836 16.4335063     0
#> 86  -0.51414015 17.8831287     0
#> 87   0.60319435 14.6697641     0
#> 88  -0.82503391 15.0670091     0
#> 89   0.06599942 14.1620725     0
#> 90   2.68498355 23.4075910     0
#> 91  -0.28267900 13.8399631     0
#> 92  -0.32011205  7.4428263     0
#> 93   1.19850794  8.9980062     0
#> 94   2.84965465 23.5378647     0
#> 95   1.85608632  1.0434820     0
#> 96   0.79942575 14.5440787     0
#> 97   1.04006621 19.4586031     0
#> 98   1.51483331 22.7975302     0
#> 99   2.63037152  6.0441700     0
#> 100  1.89416716 23.4895343     0