Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1.

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1).

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times
#> 1    2.20876649  3.0268791
#> 2    0.75058427  0.8340153
#> 3    3.70610357 20.9664902
#> 4    0.90018319  1.3971701
#> 5   -0.91182574  9.5353529
#> 6    1.39371255  7.5906168
#> 7    0.57068795 21.7200243
#> 8    1.43717384 18.0497115
#> 9    2.59388571 16.8384372
#> 10   1.63588371 15.2178476
#> 11  -0.65311684 14.7868411
#> 12  -0.29701751  7.9449414
#> 13  -0.51835454  0.8177245
#> 14  -0.25587447  7.2807533
#> 15  -0.07702768 13.4250893
#> 16   0.15289526  1.8358594
#> 17  -1.59866282 11.2798262
#> 18   0.18662160  4.9682029
#> 19   0.13552385  4.9610273
#> 20   1.01792431  1.0649816
#> 21  -0.42574954  7.1743569
#> 22   0.08073258  5.3761252
#> 23  -0.32703919  5.6936951
#> 24  -1.60249353  9.3012696
#> 25  -0.03166421  5.5096282
#> 26   1.26571542 22.9252292
#> 27   2.67955893 20.1744446
#> 28   1.76091145 13.9453407
#> 29   2.42825629 21.1845309
#> 30  -0.08558112 15.1460083
#> 31   0.92503486  5.5934930
#> 32   0.73327942  2.0449328
#> 33  -1.63424115  7.7907309
#> 34   2.80498831  3.0969887
#> 35   0.73484887  9.7320378
#> 36   0.43092488 14.7381691
#> 37   1.70836850 20.5199502
#> 38   0.51424720 19.6308485
#> 39   0.57089314 14.9224990
#> 40   2.07933554 19.3472885
#> 41   3.57246401 22.7412432
#> 42   1.05128801  9.3184912
#> 43   1.69068935  2.9105924
#> 44   3.21660344 21.2802945
#> 45   0.51967176  3.9134420
#> 46   1.95487609 16.7925718
#> 47   1.42679672 18.8081786
#> 48   0.91626934  0.7221341
#> 49   0.41506870 16.8083545
#> 50   0.36387644  0.8880817
#> 51   0.83697983  7.2809411
#> 52   3.01067228  1.0038231
#> 53  -0.04683150  5.5318372
#> 54   1.16804461  1.2742376
#> 55   0.92218586  4.7219533
#> 56   1.74632192 23.7089575
#> 57   2.46424378 16.5522050
#> 58   1.05624295 13.3613669
#> 59   1.90258117 13.2767007
#> 60   3.43763354 16.9880947
#> 61   2.15885337  3.7444546
#> 62   1.80526822 18.8248380
#> 63   0.76784726  6.4986212
#> 64   1.81576021 23.7729536
#> 65  -0.72200768 10.4635441
#> 66   4.93058874 17.6351591
#> 67   0.45656406 14.7926496
#> 68   0.01168458  7.8547246
#> 69   0.22217774 14.0707843
#> 70   0.95316224  2.3354719
#> 71   1.10112466 13.7051956
#> 72   1.25407429 22.9990065
#> 73   0.05547035 12.9950132
#> 74  -0.11742436  9.6254102
#> 75   0.06258483 15.6217522
#> 76  -0.19366284  2.7677107
#> 77  -0.11176349  7.8847247
#> 78   0.70310362 18.3403106
#> 79   0.19407259  2.6842549
#> 80   0.91072140 18.8017725
#> 81  -0.79784926 13.4058846
#> 82   0.71642309  5.5345855
#> 83   1.07283975 23.8555204
#> 84   0.80383644 15.8909224
#> 85   0.01927143  4.7112854
#> 86   1.41665570 15.3674331
#> 87   1.80918955 15.1398676
#> 88   0.33553697  6.6888190
#> 89   1.25865316 22.9111958
#> 90   1.43632085  7.4900820
#> 91   2.02530125  1.6208238
#> 92   1.82921956  6.0688293
#> 93   0.67091910 14.4716324
#> 94   1.40462732  1.9379774
#> 95   0.53155189 11.0557209
#> 96   2.42819567 20.6956457
#> 97   0.34186415 10.3786821
#> 98   0.77552675  6.2490326
#> 99   0.45540926 12.8394578
#> 100  1.39221509 16.0878037