Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>                 Y      times group
#> 1    0.8050792734 23.8428951     0
#> 2   -1.6394486558 16.4402913     0
#> 3   -0.8459927188  0.3630476     0
#> 4    0.4382935618 22.8301271     0
#> 5    0.1879294147 21.7362642     0
#> 6    0.6281241725 13.3170036     0
#> 7   -1.9364214098 13.4538707     0
#> 8    1.4787931438 22.7015418     0
#> 9    1.7721279015  7.5948975     0
#> 10  -0.8032938965 17.3972473     0
#> 11   1.0382400247  1.5865743     0
#> 12  -0.9326621679 22.1368540     0
#> 13   0.3039724781  6.3552707     0
#> 14   2.4040993992  4.4516443     0
#> 15   0.2844005306 12.9663181     0
#> 16   3.0218011716  2.8463678     0
#> 17   0.4837126412 12.5472425     0
#> 18   0.1101110044 23.2338967     0
#> 19  -0.4339681677 17.3798353     0
#> 20   1.3042841203 20.4267365     0
#> 21   2.1923731931 23.9008799     0
#> 22   3.7333624959  0.4260288     0
#> 23   1.4387844548 21.0252382     0
#> 24  -0.1180847549 18.3695557     0
#> 25   2.0467474605 16.9029918     0
#> 26   1.7344187143  2.2078707     0
#> 27   1.8009558173 10.1604037     0
#> 28   0.5480188559 20.8064538     0
#> 29   0.9116225218 23.3131855     0
#> 30   2.6037986757  2.6348963     0
#> 31   0.5765217177 20.4931890     0
#> 32   0.1816316010 18.6206606     0
#> 33   1.7986937705 23.1003073     0
#> 34   0.7333580032 14.8216521     0
#> 35   2.9861115761  9.9610371     0
#> 36   2.8954793534  7.8005153     0
#> 37   1.0366186932 21.1118353     0
#> 38   0.1014269023 15.0417433     0
#> 39   1.0403708834 17.9757104     0
#> 40   1.7174161454 11.9041285     0
#> 41   1.6578225700 23.1404188     0
#> 42  -1.2559457912 18.2080081     0
#> 43  -0.6155159312 20.9985316     0
#> 44   0.2737666891 21.0849126     0
#> 45  -0.1506282826  5.6453605     0
#> 46  -1.1178916088 13.0981875     0
#> 47   0.0148015877 21.7485721     0
#> 48  -0.9033161472 16.2396443     0
#> 49   0.6216677082 22.7071246     0
#> 50  -1.0705343288 17.2925213     0
#> 51   3.3190937435  3.0835421     0
#> 52   2.1627137644  9.8061248     0
#> 53   0.1328148256 21.8772534     0
#> 54   0.3196523333 12.3803998     0
#> 55  -0.9304608605 16.0681570     0
#> 56   1.7486349130 11.3123669     0
#> 57   0.4337657872 18.3267030     0
#> 58   1.9818331454 11.1640734     0
#> 59   2.9607735338  8.8012747     0
#> 60   0.4855396557 23.6504841     0
#> 61   1.6936569655 23.3674478     0
#> 62  -0.0170120328  9.1215375     0
#> 63   0.5583736148 19.1825052     0
#> 64   0.1610039603  7.9224768     0
#> 65   2.3915199574 16.6828083     0
#> 66   1.6213473094  2.6557766     0
#> 67  -0.3556808828 15.7630696     0
#> 68   2.8007387643  3.8134607     0
#> 69  -1.3997699620 18.0194775     0
#> 70  -1.3897949842 17.6811763     0
#> 71  -1.2771277702 19.0103090     0
#> 72   0.6507445832 10.0856224     0
#> 73   1.9205128398  8.5131187     0
#> 74   1.9060804834  4.2543374     0
#> 75   0.8811101497 19.9197522     0
#> 76   1.5402157748  6.0740033     0
#> 77   1.7764876230  5.1341504     0
#> 78   0.0899593398 10.3225425     0
#> 79   1.0217253787 10.7696467     0
#> 80  -0.8004183511 11.7138676     0
#> 81  -0.1692575721 23.9666295     0
#> 82  -1.3130322071 12.4828435     0
#> 83   3.3825566974  1.0378030     0
#> 84   3.3042503269  4.4392609     0
#> 85  -1.2607058083 14.5516341     0
#> 86   0.2691391920  5.2618277     0
#> 87   0.2037289182 23.5333446     0
#> 88   0.5195723600 18.3915806     0
#> 89   2.0844786234 10.5587048     0
#> 90   0.0005665648  8.3312541     0
#> 91   2.1974771560  2.1165751     0
#> 92   1.2370528723  0.6651207     0
#> 93   1.4827455219 17.8780603     0
#> 94   1.5709590182  1.4023257     0
#> 95   4.2414025356  5.7979085     0
#> 96  -1.5938908610 13.9238255     0
#> 97   1.4585819911  7.5428024     0
#> 98   1.1630054108 20.4196597     0
#> 99  -1.0533948548 19.8296132     0
#> 100 -0.4636147759 19.0119458     0