Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The MESOR is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times group
#> 1    3.09469736  7.7354611     0
#> 2   -1.42966840 16.4398873     0
#> 3    0.60532975 23.9478152     0
#> 4   -1.03816947 21.2318223     0
#> 5    1.88381102  0.9930210     0
#> 6    0.26876234 17.5933107     0
#> 7    2.46268456  8.1112164     0
#> 8    1.59278184  3.5524897     0
#> 9   -0.48684270 17.5166312     0
#> 10   0.37482118 16.4951914     0
#> 11   1.25092417 15.8204848     0
#> 12  -0.47039510 19.0643460     0
#> 13  -0.04334657 21.4253507     0
#> 14   1.29422944 18.4081241     0
#> 15  -0.58187845 14.1167605     0
#> 16  -1.82914438 13.4363445     0
#> 17  -0.56746410 15.8879436     0
#> 18   0.80743316 18.9923252     0
#> 19  -0.38105974 14.8236503     0
#> 20  -1.48700253 14.3590173     0
#> 21   1.60509806 22.1944372     0
#> 22   0.33499108 21.9914101     0
#> 23   2.43096118  0.3222712     0
#> 24   1.89205139  4.2347957     0
#> 25   2.17759185  1.9751738     0
#> 26   2.70914967  9.5458871     0
#> 27   1.85027274  5.4872576     0
#> 28   2.40215783  3.6328603     0
#> 29  -1.11939872 19.4506466     0
#> 30  -0.59798922 12.7070132     0
#> 31   2.08451563 16.9656893     0
#> 32   2.11138755 22.9976489     0
#> 33  -0.08432665 18.9702144     0
#> 34   0.99247005 22.1660418     0
#> 35   1.18070719  8.7437476     0
#> 36   0.97557087 14.7972874     0
#> 37   1.51474651  9.1650891     0
#> 38   0.45075687 18.3418653     0
#> 39   2.27175023  9.5714784     0
#> 40   0.93060363 23.3269800     0
#> 41   0.28146868  8.7856465     0
#> 42  -0.01678964 17.5656302     0
#> 43   1.66135056 16.7323554     0
#> 44   0.21015830 15.0680910     0
#> 45   1.47813582  0.4713450     0
#> 46   0.20458507 12.8277732     0
#> 47   3.20646613  4.2037705     0
#> 48   2.52306850 20.1764780     0
#> 49   0.95442758  6.9107257     0
#> 50   0.90332813 22.8210102     0
#> 51   0.60070656 11.5653159     0
#> 52   1.63624143  7.7783647     0
#> 53   0.66498274 11.3464906     0
#> 54   0.82466348 17.3394296     0
#> 55   1.47734485  0.2872322     0
#> 56   0.03530458 18.4133755     0
#> 57   0.65944935  0.8648708     0
#> 58   0.62192593 23.4486995     0
#> 59   0.35208771  0.1581129     0
#> 60   0.73366493 23.0850156     0
#> 61   2.40271482  3.9126738     0
#> 62   0.80773535 22.5480437     0
#> 63   1.20927776 12.1444633     0
#> 64   0.73112954 11.4770989     0
#> 65   2.04499209 23.4700103     0
#> 66   0.98998765 10.2712795     0
#> 67   1.57696726  8.2500533     0
#> 68   1.99387995  5.0954285     0
#> 69   3.11903187  3.6322752     0
#> 70   0.63958861 11.6336852     0
#> 71  -1.74848523 12.9523422     0
#> 72   0.82178930 15.4899082     0
#> 73  -0.70437526 14.5039946     0
#> 74   0.40947365  8.6522036     0
#> 75   2.45930449  8.4024158     0
#> 76   1.39565973  7.3609775     0
#> 77   1.82312370  1.1162645     0
#> 78   2.57099064  8.1742323     0
#> 79   1.48206833  3.2220927     0
#> 80   2.09099388  5.3665953     0
#> 81  -0.97923343 21.1195573     0
#> 82   0.92520018 15.8936963     0
#> 83   0.25898554 12.2378861     0
#> 84  -0.89956216 13.9859960     0
#> 85   1.02496025  5.0069794     0
#> 86  -0.34230886 16.4143010     0
#> 87  -1.06543617 12.8140453     0
#> 88   1.36513571  2.3616709     0
#> 89  -0.21025689 16.0664308     0
#> 90   3.29714575  0.5758314     0
#> 91  -0.39717771 13.3478434     0
#> 92   3.17065361  3.8972887     0
#> 93   0.78299035 19.8732264     0
#> 94  -0.73241809 16.6271229     0
#> 95  -1.10243357 15.1114389     0
#> 96   1.87410963  4.4539883     0
#> 97   1.10467857 23.5557729     0
#> 98   0.46382795 19.8896269     0
#> 99   0.65106609 11.4182491     0
#> 100  2.43503111  3.4640392     0