Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>                 Y      times group
#> 1   -2.082575e-01 19.4625233     0
#> 2    3.345499e+00 10.5455015     0
#> 3    8.488213e-01 18.9275556     0
#> 4    5.922959e-01  7.1302004     0
#> 5    1.036940e+00 11.2244017     0
#> 6    5.457584e-01 20.5367241     0
#> 7    9.553494e-01 20.8032116     0
#> 8    9.968852e-01 19.1886569     0
#> 9    2.515386e-01  7.4154929     0
#> 10   6.865877e-02 12.5746782     0
#> 11   1.592190e-01 12.7453234     0
#> 12  -2.016440e+00 14.5821311     0
#> 13   1.779907e-01 15.1575298     0
#> 14   1.128025e+00  4.0045664     0
#> 15   2.635241e+00  3.3875060     0
#> 16   9.711754e-01  5.0689396     0
#> 17  -1.121117e+00 16.0088899     0
#> 18   6.006418e-01 11.3099405     0
#> 19   8.227198e-02 23.9225794     0
#> 20   1.606333e+00  1.7027425     0
#> 21   4.341568e-01  0.8720552     0
#> 22  -1.235911e+00 15.3317598     0
#> 23   1.292499e+00  9.0735736     0
#> 24   3.498060e-01  4.3240639     0
#> 25   8.937527e-01  8.6443183     0
#> 26   1.389353e+00  7.3205768     0
#> 27  -2.678712e-02 14.4899182     0
#> 28   3.664057e+00  7.2517358     0
#> 29   2.331210e-02 19.4485522     0
#> 30   2.694889e-01  1.7332254     0
#> 31   1.685003e-01 16.6469676     0
#> 32   2.357782e+00 20.4030867     0
#> 33   2.167038e+00  1.2674795     0
#> 34  -6.233734e-01 20.7304870     0
#> 35   1.770441e+00  0.6532280     0
#> 36   4.180658e-01  9.0383545     0
#> 37  -6.981058e-01  8.5442915     0
#> 38   1.176316e+00  1.5572203     0
#> 39   2.641428e+00  5.8409535     0
#> 40  -3.297094e-01 17.5777310     0
#> 41   2.391936e-01 13.2445865     0
#> 42   2.143269e+00  4.1851883     0
#> 43   2.673193e+00  1.4937935     0
#> 44   6.295976e-01 18.5159446     0
#> 45   3.276117e-01  7.6883562     0
#> 46  -3.379998e-01 22.2690558     0
#> 47   2.447219e+00  4.0995581     0
#> 48  -7.947288e-01 10.1792115     0
#> 49  -3.335673e-01 21.4166068     0
#> 50   1.085609e+00 23.1707834     0
#> 51   1.707664e+00  8.8088391     0
#> 52  -1.255649e+00 18.8936309     0
#> 53   2.851674e+00  1.4567890     0
#> 54   2.033946e+00 16.7868229     0
#> 55   2.850497e-01  4.6645708     0
#> 56   1.212667e+00 22.7473078     0
#> 57  -7.218981e-01 15.8180763     0
#> 58   1.886483e+00  6.2493947     0
#> 59  -7.445287e-01 14.3507763     0
#> 60   4.210222e-01 10.9872722     0
#> 61   1.193172e+00 16.2729919     0
#> 62   6.631194e-01 21.7865250     0
#> 63   2.368504e+00 11.4229482     0
#> 64   8.836160e-01 10.2030121     0
#> 65   1.386127e+00 11.2357999     0
#> 66   1.379406e+00 10.1268518     0
#> 67   2.530812e-01 14.5950839     0
#> 68   7.209858e-01 23.9209610     0
#> 69   7.673185e-01  8.7310864     0
#> 70   4.092953e-01 20.2851336     0
#> 71   5.163749e-02 17.6140567     0
#> 72   2.335885e+00 21.0969389     0
#> 73   5.009429e-01 20.5742031     0
#> 74  -1.109589e+00 12.8818203     0
#> 75   1.747934e-02  9.4613665     0
#> 76  -9.494856e-01 16.6472438     0
#> 77   9.912965e-01 23.7960451     0
#> 78   3.518202e+00 21.8824752     0
#> 79  -1.043432e-01 17.5598044     0
#> 80   1.452283e+00 23.4547729     0
#> 81  -2.197880e-01 16.5725363     0
#> 82   7.652954e-02 11.9382553     0
#> 83   1.598369e+00  1.0992544     0
#> 84   1.407843e+00  0.5961303     0
#> 85   2.477125e+00  1.0297615     0
#> 86   1.061760e+00  0.4971725     0
#> 87   1.197005e+00  4.1940102     0
#> 88   2.229616e+00  2.9415548     0
#> 89  -1.455095e-01 15.5471697     0
#> 90   2.561389e+00  9.5647771     0
#> 91   1.400812e+00 23.8641790     0
#> 92   8.579918e-05 16.0078933     0
#> 93   1.042243e+00  6.0182437     0
#> 94   2.597095e+00  3.8766237     0
#> 95   3.511123e-01  9.1028825     0
#> 96   6.406506e-01  8.1530772     0
#> 97   1.858882e+00  7.7550102     0
#> 98   6.389578e-01 16.8395924     0
#> 99   1.759920e+00 11.2806496     0
#> 100 -7.240690e-01  9.3802703     0