Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1.

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1).

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times
#> 1    2.86523876  2.8862030
#> 2    1.28679902  4.8140550
#> 3    1.09878668 22.2132441
#> 4    1.32623212 15.0321668
#> 5    0.97067619 23.9431360
#> 6    1.62412990 22.4534799
#> 7    2.29186189  3.3118288
#> 8    0.41979986 16.0814829
#> 9    1.23338379  5.4957921
#> 10  -0.02202932  9.5576071
#> 11   2.08924265  3.7481016
#> 12   1.95809518  8.1002467
#> 13   0.56764126 13.4832418
#> 14   0.77543537  3.3868721
#> 15   0.76361255 14.0443518
#> 16   2.48447053  2.2022076
#> 17   1.90244841 11.4733817
#> 18   2.10130897  3.8113887
#> 19   3.54708262  5.8107439
#> 20   2.33390369  1.2086109
#> 21   2.54940116  7.0867081
#> 22   2.12831709  5.7312031
#> 23   0.92772433 12.7946225
#> 24  -0.93997314 21.5662044
#> 25   2.00943003  1.8794690
#> 26   3.98841090  6.0748290
#> 27   1.13979800 22.5610830
#> 28  -0.79818922 17.6651633
#> 29   1.06313215  2.0528980
#> 30   0.71910993 13.1390850
#> 31   3.49711515  5.4982686
#> 32   1.98861458 22.4724509
#> 33   0.14560512 13.8933580
#> 34   0.54348064  3.2918349
#> 35   4.22143826  6.4892229
#> 36   0.90667161 20.8447219
#> 37   0.44342527 17.2877428
#> 38   0.72224103 18.3139318
#> 39   0.76516893  2.8310163
#> 40   0.25378585  3.2301720
#> 41   1.09837923  8.7694188
#> 42   1.35783903 20.8268347
#> 43   1.78660571  3.9656829
#> 44   1.60332407  2.5190233
#> 45   3.05129474  2.2695999
#> 46   0.29644724 17.8677342
#> 47   0.94006751 19.2980323
#> 48   0.90078191 15.6964401
#> 49   1.53612489 23.2557058
#> 50   2.33873233 23.9821256
#> 51   1.87487275 23.4036481
#> 52   1.55503872 23.7049403
#> 53   0.71887829 13.2127610
#> 54   0.45187053  6.2851045
#> 55   1.92212245  5.0097696
#> 56   4.57405956  2.1180078
#> 57   3.54835535  6.1734891
#> 58   1.65577617 21.3466951
#> 59  -0.47660551 11.1683977
#> 60  -0.01619607 15.4974643
#> 61  -1.37823862 17.7383501
#> 62  -0.78961859 13.0302231
#> 63   1.71253315 23.8320604
#> 64   0.83361574 14.4008665
#> 65   1.70986967  6.5536189
#> 66   0.66167909 14.1103863
#> 67   1.35478766  7.6997542
#> 68   1.17316249  0.2758401
#> 69   2.32095169 21.4147869
#> 70   1.25314525  1.5018626
#> 71   1.86406881  8.3878604
#> 72  -0.03693516 10.7930015
#> 73   2.74860909  4.3348803
#> 74   2.89307895  4.5477084
#> 75   0.09808448  7.1876347
#> 76  -0.03067616 17.0788024
#> 77   3.17010820  7.6613593
#> 78   3.33676266  8.2410815
#> 79   1.24795913 17.5626467
#> 80   1.97060598  1.2369819
#> 81   3.25215979  1.9501247
#> 82   0.22641207 20.3952161
#> 83  -0.33048806 17.4619546
#> 84   2.09079117  6.9136406
#> 85   0.57479608 11.5029673
#> 86   0.23434431  8.3407958
#> 87  -1.21492014 15.7033650
#> 88   1.05528000 16.3731304
#> 89   1.58251105 21.1555223
#> 90  -0.27744074 18.3648520
#> 91   0.29801470 15.6447457
#> 92  -0.35082643  8.7679615
#> 93   1.24053995 20.5698930
#> 94   3.87174634  3.2999673
#> 95  -1.13869750 17.7277391
#> 96   2.76970378  4.4136420
#> 97   0.46016748 20.8167354
#> 98   0.17151856 23.4257083
#> 99   1.87045456 12.6711723
#> 100  1.06193580  5.8088578