Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1.

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1).

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y       times
#> 1    0.52769712 12.88458013
#> 2    1.75207136 18.16244683
#> 3    2.72120225 23.12141121
#> 4    2.13974357 21.49209013
#> 5    0.69387312  4.19763939
#> 6    0.00126531 11.60076849
#> 7    2.27801978 14.00124440
#> 8    2.04834680 13.60770182
#> 9   -1.08745740  6.74193986
#> 10   1.17531353 18.98469557
#> 11   0.46317613 15.13759622
#> 12   0.92781879  3.65687496
#> 13   1.49489293  9.98920392
#> 14  -1.04239136  5.11060618
#> 15   0.75334878 23.07499579
#> 16  -0.70254159  6.85692750
#> 17   1.14448621  5.61892430
#> 18   1.84248652 14.75971983
#> 19  -0.82415392  8.99877615
#> 20   0.63337831  6.76691023
#> 21   1.07384276 23.56648934
#> 22   1.65755841  8.07944329
#> 23   1.31201271 16.87114450
#> 24   0.96792511  6.92544259
#> 25   1.90337986  4.98806713
#> 26   2.16583193 10.70657414
#> 27  -0.70686044 11.52325214
#> 28   1.67885753 23.05316732
#> 29   1.76660068 16.67531586
#> 30   2.45188660 21.18866055
#> 31   0.09775470 11.18072436
#> 32   0.36394028  2.67119156
#> 33   3.20539650 20.94238511
#> 34  -0.18219551 18.96940804
#> 35   3.48574468 20.76075112
#> 36  -0.03059837  3.28971500
#> 37   1.40464138 20.39790509
#> 38   1.41764465 22.83933719
#> 39   0.45584129 19.59853747
#> 40   0.34701115  4.14721775
#> 41   1.78726238 20.90334473
#> 42   1.53861773 23.90108753
#> 43   2.19789703  1.91879340
#> 44  -0.35529171 11.81561709
#> 45   1.48802297  4.48378339
#> 46   1.50664626 21.85798097
#> 47   1.03799158 16.95121210
#> 48   0.47861121  2.61565408
#> 49   0.61334714  6.10212761
#> 50   0.34265070  9.96907829
#> 51   0.07897269  8.23834622
#> 52   0.92258790 13.52705235
#> 53   1.38254745  4.07859783
#> 54   0.49387096 13.97465431
#> 55  -0.19803931  9.29204725
#> 56  -2.90554813  8.29078628
#> 57   2.05695673 21.20897313
#> 58   2.16208713 20.04516579
#> 59   1.09004106 17.48487634
#> 60   0.52123703 23.17544064
#> 61   2.13015145 20.18039499
#> 62   0.90280744 12.47746188
#> 63  -0.42669462  9.63377976
#> 64  -1.24788753  8.60094911
#> 65   1.47684147 16.59325894
#> 66   0.67412133  1.18902630
#> 67   0.56416762  9.67094519
#> 68  -0.24604923  6.57861533
#> 69   0.42291287 14.52177751
#> 70   0.80633674  9.57163486
#> 71   0.20116946 10.30543992
#> 72   0.05153340  9.14753484
#> 73  -0.92840586  7.97450900
#> 74   2.94209587 18.46349172
#> 75  -1.39357319 10.04109280
#> 76   1.48611493 19.16880117
#> 77   0.34859066  5.64898437
#> 78   2.92423551 17.23773203
#> 79   1.70668956 17.53037336
#> 80   1.72931079 18.78642261
#> 81  -0.80136487  7.37182808
#> 82   0.47085183 10.68318537
#> 83   3.00037746  0.04701335
#> 84   0.38624945  0.26238089
#> 85  -0.28654326 10.66211447
#> 86   1.85274765  0.08409439
#> 87   0.78749898  5.16813744
#> 88   0.81338846  7.01823094
#> 89   1.73537272 22.18467835
#> 90   0.71274249  3.50636380
#> 91   0.73651747  6.19998172
#> 92  -0.28185013  9.72458312
#> 93   1.24384402  9.61305111
#> 94   2.84364423 19.73660854
#> 95   1.07091682  2.75648869
#> 96   2.72602438 14.71399021
#> 97   2.02065604 18.97474732
#> 98   0.77827039 13.51481965
#> 99   1.25198098  1.29285740
#> 100  1.28725981  3.34644560