Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1.

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1).

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>                Y       times
#> 1    2.169930431  4.10305777
#> 2    2.551218995  1.41930733
#> 3    1.471350227  3.42497903
#> 4    0.821447312 16.62272421
#> 5    0.379853210 15.53425293
#> 6   -2.015669452 14.54742326
#> 7    3.067714167  1.61586541
#> 8    3.137619747  5.86056896
#> 9    2.860489055  3.96150465
#> 10   0.575910791 16.38432484
#> 11   1.753800975  1.86188142
#> 12   2.515030065  4.04824815
#> 13   0.199314613 17.28196608
#> 14  -0.008232745  8.91344228
#> 15   1.951448260  1.28966150
#> 16   1.909376247 10.53815687
#> 17  -0.574545626 22.08646687
#> 18   1.578373042  4.09846100
#> 19   0.249085948 11.87541606
#> 20   0.621387520  7.61987995
#> 21   1.018282568  9.31663572
#> 22  -1.996658567 13.59533455
#> 23   1.440181914 10.95073805
#> 24   0.189569044  0.95813153
#> 25   1.423657048  9.95223061
#> 26   0.187842830 19.10410198
#> 27   1.946329407  2.91519218
#> 28   0.699725455 22.34456548
#> 29  -0.401275964  8.90484483
#> 30   0.404950652 15.66347877
#> 31   2.375496190  5.07439487
#> 32   1.023828633 23.13649094
#> 33   1.871005043 10.43483488
#> 34  -0.418158192 11.39254689
#> 35   1.807205740  3.06968915
#> 36   1.028730662  8.63990484
#> 37   2.263433335 23.32164286
#> 38   3.856779404  0.52680335
#> 39   2.924108861  5.81751260
#> 40  -0.490736839 22.25577094
#> 41   2.496190602  6.16128407
#> 42   0.602968675  6.32899558
#> 43   1.021481920 13.96780016
#> 44   0.567786800  6.60575633
#> 45  -0.812695005 18.25885297
#> 46   0.938185169 16.98980805
#> 47   2.786887555  0.92232095
#> 48   1.495091842  8.14637066
#> 49   1.242357287 20.71037428
#> 50   1.114062333  1.15743467
#> 51  -0.390914223 11.35947632
#> 52   0.045957571 23.81002432
#> 53  -0.532319025 16.30819821
#> 54   1.454625930  1.61128053
#> 55   0.900552572  8.97750719
#> 56   0.077263848 19.58724162
#> 57  -0.410471295 23.31174093
#> 58   1.517132761 14.78592868
#> 59   1.064145764 22.28447668
#> 60   1.841322975 18.45969144
#> 61   2.162220603 20.51082966
#> 62  -1.517002311 14.29840078
#> 63   1.707331117 11.34431748
#> 64   1.045050412 17.20425009
#> 65   1.090962954  8.47589402
#> 66   1.743385681 19.04680891
#> 67   2.029788391  4.03668300
#> 68   0.192391692 14.96202725
#> 69   2.038343716 10.42724949
#> 70   0.307225405 19.19835101
#> 71   0.981559014  1.48093833
#> 72   1.058903620 21.20970977
#> 73   2.503908736  3.22668402
#> 74   0.787857781 18.39249623
#> 75   0.628886098  9.70297405
#> 76  -1.557071868 15.53243545
#> 77   0.643656089 16.21287959
#> 78   2.078239751  9.91094148
#> 79   2.446748806  0.55411971
#> 80   1.023876597  6.56912988
#> 81   1.405118785 12.17093602
#> 82   2.826978671  0.19566707
#> 83  -0.757874755 13.55894904
#> 84   0.845376477 20.99528088
#> 85   0.873051990  4.52661992
#> 86   2.068554220  8.64045651
#> 87  -1.273352767 13.84799933
#> 88   2.644562280  0.06909702
#> 89  -2.140866979 17.03056560
#> 90   0.895547129 18.68524268
#> 91   0.911908964 16.69130148
#> 92  -0.531174192 11.08652502
#> 93   0.416921429  0.94885541
#> 94   0.998232163  1.08283274
#> 95   0.867056132  3.73472836
#> 96   1.516912012  7.43093022
#> 97   2.467032889  6.11507585
#> 98   1.315627519 23.86248042
#> 99   1.021169338 20.73714446
#> 100  0.374117031 16.82330652