Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>                Y      times group
#> 1    3.061454021  3.7020828     0
#> 2    1.438013493  6.0130236     0
#> 3    0.704287309 22.7833049     0
#> 4    3.147468833  2.4564192     0
#> 5    0.419918249  0.9289118     0
#> 6    1.975713935 10.9220501     0
#> 7    0.023931661 16.0483081     0
#> 8    1.689661872  1.8635905     0
#> 9    1.519763623  5.9869246     0
#> 10   2.556049701  4.7027401     0
#> 11   0.945522477  0.1391709     0
#> 12  -1.162101109 16.7272039     0
#> 13   0.531735361 10.4918905     0
#> 14   1.224256757 15.6417029     0
#> 15   1.168575080  5.3832274     0
#> 16   3.286704094  2.1928314     0
#> 17   0.534158056 23.5296343     0
#> 18   3.758560381  6.2199371     0
#> 19  -0.175608943  9.9743000     0
#> 20   1.085180828 13.9979724     0
#> 21   2.132996168  5.9820092     0
#> 22   1.085006129  9.9771262     0
#> 23   2.392617433  9.5666228     0
#> 24   0.452651429 12.7112942     0
#> 25  -1.094910060 11.2555991     0
#> 26  -0.111342803 18.7415067     0
#> 27   0.662037653 21.3858531     0
#> 28  -0.116385479 20.4700805     0
#> 29  -1.163031534  2.2201329     0
#> 30   1.195039871  5.1201077     0
#> 31  -0.464773308 15.4270787     0
#> 32   1.695014560  4.6826475     0
#> 33  -0.434953089 17.6841526     0
#> 34  -0.510866032  9.7795948     0
#> 35   0.780518718  9.3039011     0
#> 36   2.656286371  4.0644214     0
#> 37   1.501111388 21.6709911     0
#> 38   0.002651748 12.2634764     0
#> 39  -0.237867439 23.1667262     0
#> 40   2.156642001  8.4769013     0
#> 41   1.332929866 21.8900267     0
#> 42   2.961076137  5.2109472     0
#> 43   1.298939485  3.7833569     0
#> 44   2.272015662  0.5824963     0
#> 45   0.286185788 12.9174405     0
#> 46   0.823070317  6.9132759     0
#> 47   1.685545657  7.0547750     0
#> 48   0.728780391  0.6038324     0
#> 49   1.419318343  0.7452488     0
#> 50   1.360065956  7.3660424     0
#> 51   0.529213720 21.0539509     0
#> 52   1.535206211  8.1580408     0
#> 53   0.568714226 15.1187052     0
#> 54   0.554293254 20.9707057     0
#> 55   0.604880488 19.8045175     0
#> 56   3.652655799  5.1678718     0
#> 57   2.594761814  0.5515148     0
#> 58  -1.103323683 19.5067175     0
#> 59   0.927153213 10.7827734     0
#> 60   0.306123311 13.7558065     0
#> 61   0.140984084  8.8695033     0
#> 62   2.128291024  5.1369299     0
#> 63   0.246799828  7.1475980     0
#> 64   1.123826129 18.7538807     0
#> 65  -0.916580413 15.0179106     0
#> 66   0.736447374  4.2747939     0
#> 67   0.782308864 12.3360748     0
#> 68   0.478258081  0.7156687     0
#> 69  -1.003297696 15.2044917     0
#> 70   1.852040341  5.0240038     0
#> 71   1.143859334 11.4223080     0
#> 72   3.805812784  1.6749135     0
#> 73   1.220964639  3.0265750     0
#> 74  -0.118004830 18.5834555     0
#> 75   0.712521606 18.4736718     0
#> 76   0.250612865  7.2199376     0
#> 77   2.328731374  7.8574884     0
#> 78   1.944793739  4.8284017     0
#> 79   1.548463686  9.8818256     0
#> 80   2.661757636  1.4676682     0
#> 81   1.006912651 19.9699789     0
#> 82   0.447192122 19.7833515     0
#> 83   2.506630025  1.7362144     0
#> 84   2.180204938  2.8719669     0
#> 85   1.654392271 15.8670809     0
#> 86   3.490840295  4.2912442     0
#> 87   0.060693808 11.2051741     0
#> 88   1.083049501 15.7522778     0
#> 89   1.381853385 22.7806684     0
#> 90   3.109387929  7.9984425     0
#> 91   0.044521833 14.0803808     0
#> 92   1.853118563  0.5377747     0
#> 93   2.982927218  4.9303849     0
#> 94   1.111035035  4.9521405     0
#> 95   0.752902937 21.1266527     0
#> 96   3.146523643 23.6337157     0
#> 97   0.373514131 20.9694037     0
#> 98   3.557988690  5.2690466     0
#> 99  -1.043077212 16.5420422     0
#> 100  0.689028758  6.9987811     0