Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>                 Y      times group
#> 1    1.1776981763  8.6484617     0
#> 2    0.5577403481  2.5736862     0
#> 3    0.8132390403  4.0620303     0
#> 4    1.1830260661  8.5114963     0
#> 5    1.4440035455 20.9241647     0
#> 6    3.1329153209  6.8149994     0
#> 7    1.8473235584 20.7531274     0
#> 8    1.6820372601 22.1828535     0
#> 9    1.8973132306  1.3813581     0
#> 10   0.4461391537 11.3288810     0
#> 11   2.9526150334 20.3564884     0
#> 12   3.0413459977  3.2511165     0
#> 13  -0.6807920645 15.0275291     0
#> 14   0.8381063440  6.5235592     0
#> 15   2.5610127872  3.5792031     0
#> 16   0.9502914153 12.5503648     0
#> 17  -0.0482119056 12.8477291     0
#> 18   1.7720342584  7.2429926     0
#> 19   1.1675848293 12.2149624     0
#> 20   0.8000500185 19.8341345     0
#> 21   0.5184547175 13.7633150     0
#> 22   0.6520288096 12.2824546     0
#> 23  -0.4458843266  2.7882736     0
#> 24   2.4252508846  4.0433812     0
#> 25   0.8206000669  9.0740672     0
#> 26   0.2165122874  5.3921266     0
#> 27   3.9827561240  1.9522684     0
#> 28   2.8688248684  1.9954489     0
#> 29   0.6567013702 11.3703014     0
#> 30  -0.1874949325  1.7616580     0
#> 31   0.5374086588 11.9437508     0
#> 32   0.8868265238  1.6625675     0
#> 33  -0.1361516007 13.1294048     0
#> 34   2.2923639920  0.1224352     0
#> 35  -0.3083240152 20.9549490     0
#> 36   0.5329323917 14.8535946     0
#> 37   0.6351419344 18.8102671     0
#> 38   1.0460194582 12.4480018     0
#> 39   1.1846128288 12.4740697     0
#> 40  -1.4686759205 18.7637636     0
#> 41   1.6217258076  1.3598266     0
#> 42   0.7826734171 10.7189196     0
#> 43   1.2265577603 13.3714930     0
#> 44   0.5449792871 15.4582397     0
#> 45   2.4646475029  7.5466944     0
#> 46   0.4221083211 18.6832416     0
#> 47  -1.1768601441 15.1819706     0
#> 48   1.7565224626  5.3095726     0
#> 49   1.1307759245 18.8492732     0
#> 50   0.3644256888 13.4903794     0
#> 51   1.8307609423 19.0748110     0
#> 52   0.8823448992 17.2532405     0
#> 53  -0.0982307788 18.9567575     0
#> 54   0.9363811027 13.4946441     0
#> 55   1.9777943280 12.0242315     0
#> 56   0.1589082666  9.2583155     0
#> 57   3.4967794285  2.1545051     0
#> 58  -2.1041504492 16.6338194     0
#> 59   0.7997486635  0.5563794     0
#> 60   0.9802289623  8.5503375     0
#> 61   0.7219538759  6.4590799     0
#> 62   2.2523864349  0.3412291     0
#> 63  -1.0155373457  5.8320002     0
#> 64   1.6474750463  4.5144155     0
#> 65   1.8911937187  8.2184827     0
#> 66   1.2654019232 23.2741082     0
#> 67   0.3888077880 12.0987292     0
#> 68   1.1613603616 15.1912014     0
#> 69   1.9633428671 22.1937885     0
#> 70   3.1242696705 23.3742090     0
#> 71   2.9130122291  5.2777703     0
#> 72   0.3351121696 20.7238423     0
#> 73   0.1190764708 11.0803442     0
#> 74   0.5998576542  7.4932754     0
#> 75   3.1558792710  1.9908815     0
#> 76  -0.1773628435 19.1917047     0
#> 77   0.6024747970  6.2093759     0
#> 78   1.4709754128  7.5648395     0
#> 79   0.5915879042  8.4543980     0
#> 80   0.7139548698 18.1956708     0
#> 81   0.8045669011 12.4049820     0
#> 82   2.1878558651  4.0982961     0
#> 83   2.8119906158  0.6731368     0
#> 84   0.4495003093 23.9897980     0
#> 85   0.5655053869  8.4075901     0
#> 86   0.2664566453 11.4871596     0
#> 87   0.0002359948 14.9899281     0
#> 88   2.6349266483 10.0712752     0
#> 89   1.2048110949  1.1163929     0
#> 90   1.6302872324 22.7742490     0
#> 91  -0.6414473473  1.0424252     0
#> 92   0.4578737660  6.0890951     0
#> 93   2.4084572337  0.9867581     0
#> 94   3.6769868688  5.5881197     0
#> 95   0.0551635494 14.9682759     0
#> 96  -0.4718281666 17.3195801     0
#> 97   2.1483180908 20.1563695     0
#> 98   3.8712613929  5.7464976     0
#> 99   1.3092234953 10.4790964     0
#> 100 -0.5851927899 18.7732801     0