Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The MESOR is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times group
#> 1    1.63964181  8.8107601     0
#> 2    2.68820794  9.0986899     0
#> 3   -2.47290567 17.1843485     0
#> 4    0.90881015 10.7797485     0
#> 5   -0.32690495 12.3991224     0
#> 6    1.81978208  7.0812950     0
#> 7    1.08854337 13.9387189     0
#> 8    0.72960799  3.7509875     0
#> 9    1.09718384 19.9214754     0
#> 10   1.56890266  4.4163082     0
#> 11   1.73543289  4.9849902     0
#> 12   2.21773268 23.0962244     0
#> 13   1.29297271  0.3832781     0
#> 14   1.13662938 19.4848578     0
#> 15  -0.71601211 16.6861651     0
#> 16   3.14793959  1.3650412     0
#> 17   4.56678313  4.4001503     0
#> 18   3.52593054  0.4851789     0
#> 19   1.53872583  0.4053079     0
#> 20  -0.99406668 20.8940554     0
#> 21  -1.55744048 14.3452877     0
#> 22   0.87319531 15.9818172     0
#> 23   0.32471996  3.5993205     0
#> 24   0.45510665  5.6240300     0
#> 25   1.10164589  0.6109901     0
#> 26   0.42488303 17.2491709     0
#> 27  -0.47846474 19.3619487     0
#> 28   1.66515869  5.8978818     0
#> 29   1.43558841 22.5540847     0
#> 30   1.17029956  9.9054767     0
#> 31   2.16978368  2.6166329     0
#> 32   0.82473233 16.9806361     0
#> 33  -1.24866432 19.8730171     0
#> 34  -1.53792673 18.8963344     0
#> 35   2.52502805 19.2794194     0
#> 36   1.24157827  9.1026041     0
#> 37  -0.53751309 19.0400020     0
#> 38   2.85616076  7.1051215     0
#> 39   1.66036161  7.1985972     0
#> 40   2.05509558  1.6283929     0
#> 41   0.26986330 21.8784001     0
#> 42   1.35060795  5.3825028     0
#> 43   1.09755386 15.8626665     0
#> 44   0.16154365 14.4917846     0
#> 45   1.96568033  5.9266888     0
#> 46   2.65463331  5.3267565     0
#> 47   1.20456031  8.6432804     0
#> 48   1.47170350 19.2622820     0
#> 49   2.48071776  1.8482708     0
#> 50   1.48609961  7.7951755     0
#> 51   2.95560799 22.0886224     0
#> 52  -0.07676649 18.4898855     0
#> 53  -0.67892183 15.5995224     0
#> 54   1.42000037 21.0419582     0
#> 55   1.28610758  4.2764941     0
#> 56   1.40844929 23.7044838     0
#> 57  -0.14099702  3.4727462     0
#> 58   1.28774657  8.8142892     0
#> 59  -0.20361525 14.2265886     0
#> 60   1.55512833 14.4601609     0
#> 61  -1.27843690 16.9338871     0
#> 62   2.83532453  1.8326922     0
#> 63   0.51534273  5.5456842     0
#> 64   1.78151083  3.8508364     0
#> 65   0.05501898 12.9725718     0
#> 66   0.34678886 12.0689674     0
#> 67  -1.82612311 19.1292816     0
#> 68   0.81769316 18.4778984     0
#> 69   2.22285425  2.7587747     0
#> 70   0.22607311  8.2494269     0
#> 71   2.06263073 23.4054475     0
#> 72   3.44519693  5.8967603     0
#> 73  -0.23914437  1.7979065     0
#> 74  -0.89047698 12.4368425     0
#> 75  -1.79126065 11.5962643     0
#> 76   0.88677972  6.0168625     0
#> 77   0.35768868 13.6147374     0
#> 78  -0.65970298 11.8032354     0
#> 79  -0.98021513 15.9257547     0
#> 80   1.33376223  0.8203211     0
#> 81   0.62122153 13.8047868     0
#> 82  -0.22681992 14.4695813     0
#> 83   1.51757960 17.7632982     0
#> 84   3.48245192  3.0150084     0
#> 85   1.90803736  2.5456084     0
#> 86   2.34538267 14.9424638     0
#> 87   1.74141632 10.5813451     0
#> 88   2.64339774 10.7630573     0
#> 89   0.83561481  7.1985225     0
#> 90   1.33837483  2.4531392     0
#> 91   3.67245562  8.8421890     0
#> 92   0.85702097 19.5520891     0
#> 93   1.18855792 22.8578635     0
#> 94   1.43885954  3.5708418     0
#> 95  -0.61729355 18.4659822     0
#> 96   2.23860393  0.9903427     0
#> 97   1.45298778 22.9881358     0
#> 98   1.79498424 17.6231387     0
#> 99  -0.78684009 16.1106102     0
#> 100  0.74241009 22.2702462     0