Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y       times group
#> 1    2.24391328 12.42148035     0
#> 2    2.12431669  1.03559295     0
#> 3   -0.21211885 14.92994645     0
#> 4    2.00836974  1.84342532     0
#> 5   -0.67824960 15.19840207     0
#> 6    1.58580733  0.65935864     0
#> 7    2.13051197 19.29212083     0
#> 8    0.34832600 20.34143048     0
#> 9    1.34477215  1.88543387     0
#> 10   1.85895236  9.52244109     0
#> 11   1.28728038 20.72216119     0
#> 12   0.07064117 19.87245410     0
#> 13   0.86107521 20.88924600     0
#> 14   0.14764460 14.56025276     0
#> 15   1.99724718  8.64299113     0
#> 16  -0.02116693 20.97411557     0
#> 17   1.08002240 10.69516047     0
#> 18   3.03851009  5.10745381     0
#> 19   0.19913230 15.57498117     0
#> 20   3.16627732  1.77323515     0
#> 21   1.20574787 13.68824192     0
#> 22   0.49854792 11.40075186     0
#> 23   1.62669306  6.61954581     0
#> 24  -0.91284128 14.95698370     0
#> 25   0.38111026 17.80062878     0
#> 26   0.61594540  7.09953619     0
#> 27   0.48588588  1.27083777     0
#> 28  -0.01581450 11.31927956     0
#> 29   0.06004591 15.61746101     0
#> 30   3.27160382  5.64711228     0
#> 31   1.11161728 19.15131822     0
#> 32   0.72639639 21.44734188     0
#> 33  -0.36088550 21.51936079     0
#> 34   1.64690888 14.07305519     0
#> 35   1.85498839 10.98392205     0
#> 36   0.67681219 22.56925445     0
#> 37   1.32540304  9.40045581     0
#> 38   2.97454056  2.15822936     0
#> 39   0.72410010 21.94622763     0
#> 40   0.98781805  0.31875820     0
#> 41  -0.30171336 19.49636474     0
#> 42   2.78421587 23.75254051     0
#> 43   1.92838897  1.08515768     0
#> 44   3.09503650  9.40462180     0
#> 45   3.13039786  2.08943332     0
#> 46   0.83247671 12.52024153     0
#> 47  -0.45047912 11.89337652     0
#> 48   0.67924420 20.87246427     0
#> 49   3.84660212  2.41581515     0
#> 50  -0.71864709 12.96330854     0
#> 51   1.41121743  0.05358522     0
#> 52   1.55064361  1.41637577     0
#> 53   1.35551519  4.57570055     0
#> 54   0.39246795  0.61172073     0
#> 55  -1.10837003 10.85939570     0
#> 56   0.58779543 14.93952042     0
#> 57  -0.04912417 13.85698120     0
#> 58   1.68402131  7.57255003     0
#> 59   0.96599364 10.61591726     0
#> 60   1.83741448 22.88163433     0
#> 61   1.01132966  7.36190778     0
#> 62   0.75810633  9.28936224     0
#> 63  -0.29345797 22.41358213     0
#> 64   2.73948702  8.14136638     0
#> 65   0.51063921 22.48705131     0
#> 66  -0.72469729 12.32014748     0
#> 67   1.93407964  0.10948570     0
#> 68   0.74347294  9.69155494     0
#> 69   2.44291197  2.60079056     0
#> 70   0.63543807  7.09788738     0
#> 71   2.53791608 20.40412877     0
#> 72   1.81262633 18.58737169     0
#> 73  -0.59610465  8.77379253     0
#> 74   2.23290819 20.66026899     0
#> 75   1.65925630  3.17319528     0
#> 76   2.89209602  5.23086836     0
#> 77   2.15628693  6.66349903     0
#> 78   1.89823131  9.25211192     0
#> 79   1.55703669  2.33579586     0
#> 80   1.09573267  9.26028383     0
#> 81   2.34865771  7.33932603     0
#> 82   0.67424428 17.90967285     0
#> 83   1.14977615  6.29877943     0
#> 84   1.81700241 18.85802188     0
#> 85  -1.20132800 16.15986173     0
#> 86   1.85680528  8.27729696     0
#> 87   1.53388956  7.70783435     0
#> 88  -0.49014953 17.83004668     0
#> 89   2.36737323  0.08361608     0
#> 90   1.00627120  3.35978171     0
#> 91   0.18547626 16.52085982     0
#> 92   0.06537150 16.91674660     0
#> 93   1.96303491 21.41590025     0
#> 94   0.66381580 11.33533159     0
#> 95  -0.56160645 19.56688112     0
#> 96   1.20675432 23.57843536     0
#> 97   0.96108966 21.50942649     0
#> 98   3.07550770  6.97241604     0
#> 99   1.89757272 17.62784208     0
#> 100  1.67174374 11.63896530     0