Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1.

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1).

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times
#> 1    1.42939896  1.5281910
#> 2    0.42812321  7.2521133
#> 3   -1.26765771  2.5704792
#> 4    0.50992354  5.9564656
#> 5    1.26960597 18.5756325
#> 6    0.48406399  9.8856631
#> 7   -0.31648711  6.2448407
#> 8    0.02804811 10.9865356
#> 9   -0.76746920  3.2937121
#> 10   0.71008750  2.6137060
#> 11   2.79833446 23.5998095
#> 12   0.82292635 20.3397245
#> 13   2.14232654 17.2357816
#> 14   0.76169323  6.3316418
#> 15   0.83220187 23.2139639
#> 16   0.63454789  5.4580595
#> 17   1.87663017 20.3074638
#> 18   0.29830843  2.8312839
#> 19  -0.56646419  5.8776593
#> 20  -0.90115180  8.6060032
#> 21  -0.54482400  6.6630493
#> 22  -0.65869213 10.1665947
#> 23  -0.55463949 13.0417874
#> 24   1.84514269 16.0355986
#> 25   0.13591135  9.7191892
#> 26  -0.83183673 10.7688581
#> 27   0.87707359  4.2255516
#> 28   0.08264944  1.8377183
#> 29   3.55988773 15.6568167
#> 30  -0.56045283  6.0589336
#> 31   0.11417179 12.8178172
#> 32  -0.49287996  3.1928187
#> 33   1.60093301 10.8590528
#> 34   2.32344130 21.7916278
#> 35   0.88925426  8.6294085
#> 36   0.01509479 14.4204434
#> 37   1.35528227  5.2134655
#> 38   0.28028088 11.9217618
#> 39   1.13934592  1.4025087
#> 40   2.20958498 20.1078809
#> 41   0.09355392  5.0015741
#> 42   0.07661285 13.3437591
#> 43   1.07816983 20.2683920
#> 44   0.49692499  0.7650470
#> 45   1.78445798 14.2208726
#> 46   0.16889201  5.4181517
#> 47   1.17994067 14.7881777
#> 48   1.62147110  2.8946894
#> 49   1.67351839 14.7687236
#> 50   1.60470372 22.7150606
#> 51   1.17233266  0.5947024
#> 52   1.03428264 19.8553795
#> 53   0.14880637  8.4051611
#> 54   1.86641386 16.3920813
#> 55  -0.27589276  0.9032910
#> 56   1.23083958 14.1752065
#> 57   2.46626674 21.8028735
#> 58   2.60864575 18.0542135
#> 59  -0.06790454  8.2766995
#> 60  -0.59538891 11.4781192
#> 61   1.96167876  3.1038202
#> 62  -0.02586711  8.7814248
#> 63   1.46226945 21.9840592
#> 64  -1.67872522  6.4480080
#> 65  -1.29735720  5.5204398
#> 66   2.48021854 14.6571725
#> 67  -0.31460106 10.1007773
#> 68   2.83561513 16.9011719
#> 69   1.40492742 13.6491501
#> 70   1.86136745 18.8469759
#> 71   1.06268077  8.9910068
#> 72   1.71883303 20.8371888
#> 73   2.22902621  0.4143562
#> 74   0.55814710  4.5483086
#> 75   0.24888990  5.3496785
#> 76   0.71164302 22.9886714
#> 77   0.80870159 14.1604111
#> 78   2.54692078 22.7441462
#> 79   3.41469518 16.8385307
#> 80   0.15752813  8.2814688
#> 81   2.42237633 22.8048482
#> 82  -2.65548999  7.6075958
#> 83   1.98948564  4.2288042
#> 84   0.56635573 15.3189270
#> 85   2.99263202 20.7435712
#> 86   2.27383908  4.7604419
#> 87   1.32908227 17.2520891
#> 88   1.86298283  1.3199376
#> 89   0.82153713 22.9082993
#> 90   1.62057451 20.0088207
#> 91   1.53711410 23.3117049
#> 92   3.94906413 19.4660818
#> 93   1.53632560  1.6912637
#> 94   1.34212860  4.9621239
#> 95   1.62161393  1.2090870
#> 96   0.78121838  6.7209760
#> 97   0.39931434  3.8990103
#> 98  -0.39462231 17.4748146
#> 99   1.86092150 20.8567360
#> 100  0.97551967  5.9885725