Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1.

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1).

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times
#> 1    0.68062712  9.0342558
#> 2    0.68814761  8.5922858
#> 3    2.17151830  4.5564045
#> 4    0.74608731 14.6078225
#> 5    0.82800308 13.3871351
#> 6    0.39464419 13.5281538
#> 7    3.05414904 16.0334863
#> 8    0.59635669  7.4026025
#> 9    1.49343023 15.0098335
#> 10   2.39905857 21.4308232
#> 11   1.61620758 16.7627966
#> 12   1.09492390 16.1650307
#> 13   0.84937683 12.7536319
#> 14  -0.20943812  3.0622103
#> 15  -0.83252857  6.7329060
#> 16  -1.18735357 10.2515965
#> 17   2.05650493 16.7635342
#> 18   2.63303967 13.9908965
#> 19   0.15187196 12.9721206
#> 20   3.41131559 20.5280032
#> 21   0.64626950 15.0386633
#> 22   1.03268705  9.5837964
#> 23   3.30852940 19.3564950
#> 24  -0.32113541  1.4962261
#> 25   2.12584534 16.0939812
#> 26   0.24707195  8.8049411
#> 27   1.23063597  3.8313929
#> 28   0.44401037 21.9934577
#> 29   2.66324937  2.9704469
#> 30   2.18836537 11.5476477
#> 31   0.53033779  4.1579926
#> 32   2.05238862 19.2296748
#> 33   0.62029909  2.9823043
#> 34   0.37339750  7.9045499
#> 35   2.75371486 19.3350881
#> 36   0.81402395  4.1080163
#> 37   1.54493086 17.1190573
#> 38   2.13727275  8.4702313
#> 39   0.16404697 11.4677452
#> 40   2.59516994 22.9009527
#> 41   2.32467123 21.3056153
#> 42   1.53155997 15.8490600
#> 43   0.12963065  5.3278458
#> 44   3.28613797 19.5825632
#> 45   0.57618417 23.1920349
#> 46   2.77507718 15.7966036
#> 47   1.23174475  7.3792016
#> 48   2.14669813 17.8624621
#> 49   3.39738700 15.6348309
#> 50   0.33639443  4.1889250
#> 51   1.49034247  3.8203957
#> 52   2.58164960 20.1859265
#> 53  -0.50608836 13.3574865
#> 54   3.18386961 23.4421191
#> 55   0.14067061  5.6074598
#> 56  -0.71056340 12.1196282
#> 57   0.61001377  3.0144372
#> 58   0.59131413  4.6622336
#> 59   2.18838757  0.6931466
#> 60   2.02329278 18.7435908
#> 61  -0.85138794  3.1053508
#> 62   3.15365364 15.9271810
#> 63   0.48451513 10.0571904
#> 64   0.06898804 13.6344110
#> 65   1.80700900  0.1082161
#> 66   0.94078732 15.7037384
#> 67   0.14929498  5.4841868
#> 68   1.58084879 23.1620301
#> 69   0.74847538 16.7147905
#> 70   0.23116762  3.4173781
#> 71   2.06024133 20.9185074
#> 72   0.49351374  1.6705631
#> 73   2.98030983 21.4059641
#> 74   2.17621637 15.3782046
#> 75   0.34257106  8.5997291
#> 76   2.65717865 21.6772321
#> 77  -0.15032457  0.1708863
#> 78   0.45251304  9.4151902
#> 79   1.96290497 22.3813718
#> 80   0.36131473 13.3608880
#> 81   0.01195694  3.1569135
#> 82   1.49009025 19.4396298
#> 83   0.62118254  8.7963551
#> 84   1.43427577  4.0841407
#> 85   0.11151479 22.1395052
#> 86   1.78131774  3.7987687
#> 87   1.52711547  2.1211036
#> 88  -0.63511923 14.5661607
#> 89   3.45614644 19.9054798
#> 90   1.21471934 22.5767870
#> 91   3.09647655 17.3440809
#> 92   0.79048311 13.5090769
#> 93   1.67541920 23.3160538
#> 94   0.69275329  1.7684449
#> 95   1.94995755  1.3173555
#> 96   3.30227179 21.3277061
#> 97  -1.01391379 10.5631444
#> 98   1.78496361  0.5273674
#> 99   1.48026837 21.2435076
#> 100 -0.73075140  6.8949445