Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times group
#> 1    2.91271624  3.3886942     0
#> 2    0.88958860  9.5686023     0
#> 3    1.82689138  2.6643616     0
#> 4   -0.14456632 13.3725667     0
#> 5    1.94963924  9.1681429     0
#> 6    2.89832126  3.4549813     0
#> 7    0.67324068 20.8488560     0
#> 8    2.24632282  8.0867623     0
#> 9    2.94317535  5.3805227     0
#> 10  -0.47369220 15.7770313     0
#> 11   2.55115118  5.5432563     0
#> 12   1.34685458  9.4704138     0
#> 13   0.92634596 16.2189041     0
#> 14   0.48983880  0.8568682     0
#> 15   1.45911114 18.0738514     0
#> 16  -0.23964273 14.1358855     0
#> 17   1.72665375  0.6201438     0
#> 18   2.98379543 23.4766342     0
#> 19   0.24495003 20.6219088     0
#> 20  -0.55160073 11.3767356     0
#> 21  -1.15594455 20.4165212     0
#> 22   0.35703294 22.0295999     0
#> 23   1.56936919 20.2070820     0
#> 24   0.63432859 15.3686003     0
#> 25   1.64731564 15.4439899     0
#> 26   1.82015065  2.9985531     0
#> 27   0.20348584 16.2199674     0
#> 28   0.58898668 20.7311604     0
#> 29   1.91239671 22.2332261     0
#> 30  -0.26195618 18.2083825     0
#> 31   0.09402473 15.5165724     0
#> 32   0.27868210 22.0874992     0
#> 33   1.96779960 20.5085003     0
#> 34   0.11498345 11.2865328     0
#> 35   1.48320455  4.9717224     0
#> 36   0.96639927  3.1735958     0
#> 37   0.24616194 11.5979560     0
#> 38   0.89536742 13.6079233     0
#> 39  -1.20343322 17.8655216     0
#> 40   0.37503768 17.3622100     0
#> 41  -0.58959597  3.3925471     0
#> 42   3.37640358 23.9318205     0
#> 43   0.96516738 13.0071920     0
#> 44   0.51105393  2.4445285     0
#> 45   0.01832283 11.2513770     0
#> 46   1.12092065 14.2453603     0
#> 47   0.64857602  4.8074113     0
#> 48   1.30281548  7.7549156     0
#> 49   3.65682837  7.1477960     0
#> 50  -1.09647612 18.3492842     0
#> 51   4.61108530  3.7595983     0
#> 52   0.94552614 19.7013960     0
#> 53   2.00339558 11.6245887     0
#> 54   1.41649968 20.8559208     0
#> 55   0.47945039  0.2483599     0
#> 56  -0.63908826 10.8068652     0
#> 57   0.32744830 17.5079409     0
#> 58  -0.79768101 20.8268049     0
#> 59   0.94290003 13.5898467     0
#> 60   1.33203728 12.3753052     0
#> 61   1.07557466 18.0981246     0
#> 62   2.36380370  0.1624943     0
#> 63   0.05077056 13.5791367     0
#> 64   0.93045224 13.7261123     0
#> 65   0.65427079  0.9576384     0
#> 66   0.69882000 20.1594972     0
#> 67   2.29600972  7.9314613     0
#> 68   0.98105084 16.2188440     0
#> 69   2.39601003  1.9962006     0
#> 70  -0.80212360 19.3846576     0
#> 71   1.31511848 17.7697050     0
#> 72   0.04488587 10.0555100     0
#> 73   3.35934885  2.8995505     0
#> 74   1.71775985  8.5237902     0
#> 75   1.05344615  8.8798875     0
#> 76   3.80679723  7.5952249     0
#> 77   2.10387133 14.4776660     0
#> 78   2.45085530  9.0591614     0
#> 79   1.03353086  6.2796490     0
#> 80  -1.03476212 17.7942208     0
#> 81   0.99452140  2.8498747     0
#> 82   1.50154998  0.5016519     0
#> 83   3.10920626  4.1017398     0
#> 84   1.44709873  9.8567312     0
#> 85  -0.73129674 16.1727980     0
#> 86   2.29543792  0.3049634     0
#> 87   1.02417429  8.6699511     0
#> 88   2.75500239  7.3960304     0
#> 89   2.82367540  7.6578895     0
#> 90   3.30246419  1.3002441     0
#> 91  -1.55376216 10.4881199     0
#> 92   0.06209714 21.9557004     0
#> 93   0.16926328 19.2412821     0
#> 94  -0.31745221 20.1984048     0
#> 95   2.16936936 23.0436049     0
#> 96   0.61743274 10.6255370     0
#> 97   0.24629494 20.6474511     0
#> 98   1.52203371 16.2615800     0
#> 99  -0.16967738 11.2775886     0
#> 100 -0.95618026 14.1693181     0