Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1.

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1).

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times
#> 1    1.40082385 21.1496749
#> 2    3.81557218 17.2559817
#> 3    1.19151102 19.7876860
#> 4    0.03444139  0.9275841
#> 5    0.62513804  3.1113309
#> 6    1.52579748 23.9084056
#> 7    1.26617404 17.8050811
#> 8    0.39144937  4.1038184
#> 9    0.70255585 23.4693158
#> 10  -0.81963302  8.3884839
#> 11   0.87317221 17.0336920
#> 12  -0.96981925  6.0145818
#> 13   3.22357241  2.2596149
#> 14   1.22946789  1.8173300
#> 15   0.56333890  9.3192691
#> 16  -0.37421522  5.9935303
#> 17   1.92266473 17.1924436
#> 18  -1.52836079  3.5474679
#> 19   1.27499623  6.9964990
#> 20   1.02404453  5.0105776
#> 21   2.82134415 19.7133128
#> 22   0.85587439 10.1085737
#> 23  -0.41529407  9.8498182
#> 24   2.52831436  3.7295288
#> 25   2.57240054 22.7428247
#> 26   2.18946177 15.7407126
#> 27   0.72884609 22.9775517
#> 28   0.63655740  9.2340747
#> 29   2.15242903 16.7520289
#> 30   0.33770657 20.9201941
#> 31   1.72984464 23.4281601
#> 32   3.62251069 18.4056727
#> 33   2.23852457  1.0268676
#> 34   1.22672395 13.9880804
#> 35   0.63092970 18.2883433
#> 36  -1.15162381  9.2698654
#> 37  -0.13045704  7.1932981
#> 38   2.80897887 21.6898212
#> 39   2.35966594  1.0437218
#> 40  -1.12752706  5.1954505
#> 41   0.99324731  5.0220140
#> 42   1.93669330 16.3458481
#> 43  -0.08706135  6.1768379
#> 44   2.36894596 22.1693608
#> 45   0.65286826  1.8174582
#> 46  -0.44123882 13.3190627
#> 47   1.62682149 19.7009821
#> 48   3.39585965 19.5476918
#> 49   0.75727041  2.5156017
#> 50  -1.67961915  9.0634964
#> 51   3.14885678 22.8438498
#> 52   2.39701058 16.8944618
#> 53   0.29470336  4.2992745
#> 54  -0.63932364 10.3782774
#> 55   1.16334641 19.9712347
#> 56   0.57270919 10.4225115
#> 57  -0.42649924  2.7235816
#> 58   0.51246922 11.1438901
#> 59   1.29602979 17.2447665
#> 60   2.63587463 22.3052236
#> 61   0.16366249 10.4998548
#> 62  -0.75271584 11.4517670
#> 63   0.85294913 11.0817072
#> 64   1.83268103 20.3455337
#> 65   0.38653140  2.8199757
#> 66   1.68637610 16.3359198
#> 67   1.76464293 17.5818798
#> 68   2.05426532  4.2830171
#> 69   1.88551445 18.1202947
#> 70   1.38149981 21.4067817
#> 71   0.22846435 11.1956780
#> 72  -0.92790069  8.8354659
#> 73   1.36557585  6.8502746
#> 74  -0.79876615  7.1819291
#> 75   3.54703771 21.0642672
#> 76   0.97902606 15.0123119
#> 77   1.74738842 18.0945630
#> 78   1.91472328 13.7903889
#> 79   1.13531259 22.6644716
#> 80  -0.85295396  3.0891194
#> 81   0.65760749  3.2901968
#> 82  -0.33111911 13.3637356
#> 83  -0.37209015 15.8733539
#> 84   0.88487642  5.7291664
#> 85  -0.87183358  8.7355218
#> 86   1.61311804  5.7097572
#> 87  -0.73494517  9.6251035
#> 88   0.35740082 12.2647289
#> 89  -0.13215513  2.2203730
#> 90   1.43538523  6.6109891
#> 91   2.74408313 15.3815231
#> 92   1.85696563 23.4099074
#> 93   1.34432953  1.4316415
#> 94  -1.55341560  8.8611253
#> 95   1.15332507 16.5193991
#> 96  -0.75137053  5.7214726
#> 97  -0.99837092  4.6788749
#> 98   0.82920673 11.9167397
#> 99  -0.70258133 13.4927252
#> 100 -1.39908021 16.9250998