Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1.

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1).

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>                 Y      times
#> 1    4.2688293195 17.2288150
#> 2    1.6818491334  5.2525167
#> 3   -0.2517775961  8.8000251
#> 4    1.0360034693 14.6649680
#> 5    0.4872870228 20.1922528
#> 6    0.9998149123 18.8697613
#> 7    0.7216231753  9.9467005
#> 8    0.1429559933  3.1336101
#> 9    0.7804193380 22.9527046
#> 10   0.2595730894 11.0324499
#> 11  -0.3597790451  4.8309345
#> 12   0.2232009410  1.8624937
#> 13   0.8135872472  3.7533978
#> 14   2.1324363620  0.8707680
#> 15   2.1989423837 16.4874298
#> 16   3.3533731224 21.5686055
#> 17  -0.8522111672  3.9884765
#> 18  -1.1602560775  5.3342178
#> 19   0.9921341142 21.8468273
#> 20   1.3213376430 13.6892928
#> 21   0.0126668365  2.2062431
#> 22   2.5172092272 18.7322383
#> 23   2.4853196568 16.8699642
#> 24  -0.8308694989  7.0329799
#> 25   1.2265956245 16.1288073
#> 26   3.4926468419 22.5727188
#> 27  -0.8275472519 10.4862562
#> 28   0.1472505455 17.9893752
#> 29   1.7496816681  0.1813299
#> 30   0.6311064349 21.1847702
#> 31   2.3861913054 20.5060502
#> 32  -0.0192670574 15.8262198
#> 33   1.3593268241 20.0389902
#> 34   2.4118257166  3.4537280
#> 35   0.9408265194  4.1687490
#> 36   0.9000283936  9.2530539
#> 37   2.2760560326 20.5844088
#> 38  -0.2481273143  3.7831920
#> 39  -0.3115916697 20.9488546
#> 40  -0.4090329356 11.3580538
#> 41  -0.1730328376 12.9546184
#> 42   2.1720626030 18.3799161
#> 43   0.1626459335 23.6003008
#> 44   1.7391397715 22.5386822
#> 45  -0.0109224280  9.0088703
#> 46   0.2600939478  7.8775172
#> 47  -0.4506880903  6.2748993
#> 48   3.4215007707  1.9440212
#> 49   1.0269409108 23.9324972
#> 50   0.7319419382 21.9688752
#> 51   3.5491297766 15.0362266
#> 52  -0.3571685203  8.3885653
#> 53   3.8715074305 20.1688531
#> 54   0.2449807594  2.9425205
#> 55  -0.8577316957  8.4349448
#> 56   1.3376673787  4.0749825
#> 57   2.0983027166 19.4708564
#> 58   0.0953791521 23.3390150
#> 59   2.2159007700 20.2677056
#> 60   0.7927346482  1.3616308
#> 61   0.9998008829 21.4985468
#> 62   1.7255413530 13.0093825
#> 63   3.1471107257 20.4505012
#> 64   0.9397431484 12.2501189
#> 65   1.2104831922  8.6613292
#> 66   1.3574981610 22.6690348
#> 67   0.6620661327  6.1634068
#> 68   0.0005080319 23.8954093
#> 69   0.2331595175 16.2498457
#> 70   1.9783373744 17.5771716
#> 71  -0.8992845084  7.9387937
#> 72   0.2085931135 13.7765755
#> 73   0.3724229472 23.3290304
#> 74  -0.0964633212 13.1981470
#> 75   0.5215144797  0.8915501
#> 76   1.9993542262 13.3591193
#> 77   3.5883559655 23.2292497
#> 78   2.5134346506 22.9086037
#> 79  -0.3050623476 11.7130299
#> 80   2.1145232738  0.4949615
#> 81   0.4069229458 16.5335002
#> 82  -0.6061094274  8.3229314
#> 83   1.0070837474  0.5463268
#> 84  -0.1903115510 11.2174250
#> 85  -1.0841080361  8.0572705
#> 86  -0.5833081141  7.2211145
#> 87   0.9213179325 12.9879923
#> 88   2.1618469003 20.8618059
#> 89   0.3365782618  8.2923057
#> 90  -0.5889962896 10.3346512
#> 91   2.2224091758 21.7542562
#> 92   1.1478536936 21.2529025
#> 93   0.5309897897  1.7073013
#> 94   1.4124363679 15.8378247
#> 95   1.4431051619  3.6011022
#> 96  -0.0611348614 10.8706397
#> 97   0.5026019560 13.9061535
#> 98  -0.3971719664  5.5037745
#> 99  -0.0811877901  9.9426680
#> 100  1.4402844924  0.2688369