Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1.

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>                Y      times group
#> 1    0.274758241 15.3421286     0
#> 2    0.718451393 23.2673244     0
#> 3    0.970897596  2.0412981     0
#> 4    2.398399017  2.2195591     0
#> 5    0.886111215 21.8088163     0
#> 6    1.964702889 23.6437828     0
#> 7   -0.725331595 16.8895529     0
#> 8    1.230215018 21.3734063     0
#> 9    0.760774193 11.2931291     0
#> 10   2.877342041  7.2445366     0
#> 11   1.599053443 11.3982336     0
#> 12   1.545815341  2.5459928     0
#> 13  -0.578546601  8.8017838     0
#> 14   2.466202794 13.9973667     0
#> 15   1.244535541  0.3842577     0
#> 16   2.420626610  6.4220912     0
#> 17   3.192554119  3.0423156     0
#> 18   1.476852208  7.0056789     0
#> 19   0.047564848 12.2632079     0
#> 20  -0.937820905  8.9042913     0
#> 21   1.794725913  1.4315503     0
#> 22   2.253672610 23.4471227     0
#> 23  -0.616978712 12.5115471     0
#> 24  -1.572470752 15.0577192     0
#> 25   0.788826499 11.7558425     0
#> 26   2.642703914 23.0134604     0
#> 27   1.365185354 19.6787178     0
#> 28   1.115243333 23.6834657     0
#> 29   1.334254311  1.9443832     0
#> 30   3.437149039  0.2572209     0
#> 31   2.921776438  8.3996085     0
#> 32   1.362029871  5.3590970     0
#> 33   2.747114013 21.6584925     0
#> 34   2.850105850  3.1002174     0
#> 35   1.512598262  3.0003247     0
#> 36   1.910858376 20.3955671     0
#> 37   1.186317847 14.8648948     0
#> 38   0.228535774 15.6608713     0
#> 39   2.493838775  1.4446387     0
#> 40   0.126043287 10.7645623     0
#> 41   0.239884320 19.9035281     0
#> 42   0.988259522 23.3775872     0
#> 43   0.535161246 22.7621265     0
#> 44   0.700218425 11.6506060     0
#> 45   2.967375923  7.7245047     0
#> 46   0.459015772 22.1793356     0
#> 47   0.168933970 20.1746203     0
#> 48  -0.024077652 14.7616081     0
#> 49   0.896481821 22.4770978     0
#> 50   1.456437141  8.0394388     0
#> 51   1.320897435  7.4395546     0
#> 52   0.289739093 16.7503152     0
#> 53   1.803059779  6.6154274     0
#> 54   2.526249983  5.1384392     0
#> 55   0.475096276  5.9656759     0
#> 56   1.514886192  2.4707092     0
#> 57  -1.988420151 13.3521667     0
#> 58  -0.566664977 15.5616589     0
#> 59   1.014341247 11.9213983     0
#> 60   1.525234326  4.2036310     0
#> 61   1.882218815  5.4139233     0
#> 62   1.267318241 10.9096973     0
#> 63   2.805210721 23.4348810     0
#> 64   2.970407013  2.0226215     0
#> 65   3.098073289  3.0305318     0
#> 66   1.495453590  8.5658772     0
#> 67   3.502793746  3.7200716     0
#> 68   1.262264416 14.7507023     0
#> 69   0.864962888 18.4147109     0
#> 70  -0.291442331 13.4737173     0
#> 71   2.063389875 21.6229468     0
#> 72   0.265168764 16.7705824     0
#> 73   0.688359308 21.4194432     0
#> 74  -0.251469769 22.2010863     0
#> 75  -2.193291836 15.7846259     0
#> 76   0.839241789  1.5360029     0
#> 77   3.062331694  0.9596721     0
#> 78   1.333554242 19.5158620     0
#> 79   1.640207466  5.4371211     0
#> 80   0.443497182 23.7481457     0
#> 81   2.672344346 21.5727830     0
#> 82  -1.070007276 13.9852468     0
#> 83   0.796201853  6.9058657     0
#> 84   0.238732964 19.9656213     0
#> 85   1.594571729  5.4823148     0
#> 86   2.254018344 10.8362056     0
#> 87   1.578436042  0.8735154     0
#> 88   2.333357627  4.8834826     0
#> 89   0.704808205  8.9310531     0
#> 90   0.527789047 11.8521056     0
#> 91   2.407516984  3.4814219     0
#> 92   0.938642145  0.1220997     0
#> 93   0.754164451 23.6442395     0
#> 94   1.388121388  5.5162182     0
#> 95  -0.123104422 15.3183966     0
#> 96  -0.091249515  3.6817488     0
#> 97   1.911466191  4.1716303     0
#> 98   1.300684217  3.8054340     0
#> 99   0.772119878  9.0824666     0
#> 100  0.006245105  8.8964773     0