Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y       times group
#> 1    3.52866518  8.50106841     0
#> 2    1.87916667  5.98588266     0
#> 3    2.87461827  6.06746967     0
#> 4    0.71678086 14.41538827     0
#> 5    0.12063292 11.39798644     0
#> 6   -0.39748105 20.18869035     0
#> 7   -0.47944893  9.91122012     0
#> 8    1.07004843  1.13879680     0
#> 9    2.28734656  3.63431117     0
#> 10   0.02750506 15.73536984     0
#> 11   2.04404006  6.53863749     0
#> 12   1.90653418 18.78181498     0
#> 13   2.69174969  7.08674121     0
#> 14   2.16282171  5.27440883     0
#> 15  -0.04767956 16.46349163     0
#> 16   2.06864669  0.99919974     0
#> 17  -0.10141946 12.19311381     0
#> 18   0.49965045 12.03481498     0
#> 19  -0.10738147 14.27445431     0
#> 20   1.57740582  5.49763162     0
#> 21   0.37112857 19.10136804     0
#> 22   0.95805523  1.60303492     0
#> 23  -0.31136947 16.30779001     0
#> 24   0.71867908 16.86785947     0
#> 25   1.71655148 22.21631028     0
#> 26  -1.86628619 14.35608329     0
#> 27   0.95739062  3.56310112     0
#> 28   1.57435912  5.54838561     0
#> 29  -0.11173460 18.07975650     0
#> 30   0.85987935 19.64282566     0
#> 31  -0.15044056 18.46174817     0
#> 32   1.61634000  6.73306849     0
#> 33   1.05308706  8.39467172     0
#> 34   0.37361235 19.69402384     0
#> 35   0.07158302 13.55756108     0
#> 36   1.19604640 16.70481559     0
#> 37   0.01260498 22.41096784     0
#> 38   0.47147432 21.47527808     0
#> 39   2.45728343  7.24689370     0
#> 40   0.25878025 15.99485045     0
#> 41  -0.67488752 17.52882927     0
#> 42   0.57396061 20.54185344     0
#> 43   2.95815372  3.09645630     0
#> 44   2.47270842 22.22607588     0
#> 45   1.13116646 23.96167258     0
#> 46  -0.12410087 15.97279226     0
#> 47   2.26895395 23.00868221     0
#> 48  -0.37611269  2.42056303     0
#> 49   0.45275512  1.43502074     0
#> 50   0.77724849  3.74847582     0
#> 51   2.50914083  4.73351821     0
#> 52   1.06651168 19.36662536     0
#> 53   0.59381306  0.08616113     0
#> 54   0.96595926 18.66998377     0
#> 55  -1.04943596 19.39843527     0
#> 56  -0.81542676 20.73234254     0
#> 57   1.62487077  4.03444563     0
#> 58   0.23363847 23.55880649     0
#> 59   2.19466371 14.17662390     0
#> 60   2.59539115  8.33264163     0
#> 61  -1.41899364 17.15648232     0
#> 62   2.55042619  0.81458534     0
#> 63   0.09597045 15.95617142     0
#> 64   1.17546052 10.59279033     0
#> 65   1.63877125  8.53703271     0
#> 66   1.64222762  0.15528730     0
#> 67   1.19026112 14.05566649     0
#> 68   0.36204401  4.26781010     0
#> 69  -0.26799049 14.21054097     0
#> 70  -0.54509301 10.56696339     0
#> 71   0.59981012 19.75984835     0
#> 72   1.18247910 22.80203923     0
#> 73   0.63635220 11.38496212     0
#> 74   0.14933067 12.13519325     0
#> 75   1.20998462 23.25679810     0
#> 76   1.29016198 11.57444507     0
#> 77   1.89272092 21.44984387     0
#> 78   0.84144895 20.27768658     0
#> 79   0.10478242 19.69891142     0
#> 80   0.72055876  6.71117035     0
#> 81   1.30506754 22.04322875     0
#> 82   1.56314176 11.72070118     0
#> 83  -2.83467135 15.32153576     0
#> 84   0.55746409 15.13818847     0
#> 85   0.71674293 17.77965817     0
#> 86   2.34610252  3.38714213     0
#> 87   3.09373452  6.37878479     0
#> 88  -1.80715540 16.14894213     0
#> 89  -0.43494617 11.39661258     0
#> 90   1.50396856 19.05430542     0
#> 91  -1.82249724 11.94566408     0
#> 92   1.07732032 15.09079495     0
#> 93   1.22927266 11.26724203     0
#> 94   0.30713045  4.19747306     0
#> 95   0.37520326 23.03261678     0
#> 96  -0.57768684 16.86097289     0
#> 97   1.70113414  6.22136840     0
#> 98   3.61198467  0.35222200     0
#> 99   0.88523794  2.48398089     0
#> 100  0.38416502 17.39434390     0