Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1.

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1).

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>                Y      times
#> 1    0.042421011 23.2674751
#> 2    0.247707257 11.9691386
#> 3    4.320607240  5.4660081
#> 4   -0.541594717 13.0057542
#> 5    1.671084356  0.8652330
#> 6    1.612930490  7.8086823
#> 7    1.387034247 13.1615708
#> 8    2.158313710  2.1737078
#> 9   -0.193669236 15.4333781
#> 10  -0.065509320 12.6376671
#> 11   0.683957741  8.8529389
#> 12  -0.403343844  2.2032051
#> 13   1.713055468  1.6170148
#> 14   1.399540096 10.2690685
#> 15   0.005961623  9.3986142
#> 16   1.158393670  2.4999552
#> 17   0.909993227 16.7721893
#> 18   1.139164873 10.1321486
#> 19  -0.249092941 11.9902496
#> 20   0.221481615 19.3518393
#> 21  -1.524278699 12.8016614
#> 22   0.447029662 21.1582859
#> 23   0.088776450  3.9532394
#> 24   0.582048608 22.8009182
#> 25  -0.340542994 15.0635127
#> 26   0.551991252  8.9356349
#> 27   3.222917864  4.4336585
#> 28  -1.786115003 11.5792298
#> 29  -1.757248624 15.1308962
#> 30   2.731539797  0.7314071
#> 31  -0.828175744 12.3196198
#> 32   0.599271686 10.0257729
#> 33   0.878307075 14.9624969
#> 34   0.257468564 13.5064722
#> 35   0.737394267  7.8520722
#> 36   2.582735096  9.1222790
#> 37  -1.269555494 18.4416563
#> 38   1.070947145  2.7966146
#> 39   0.397049357 10.0772817
#> 40  -1.854962066 13.3730697
#> 41   0.777585471  2.1327335
#> 42   1.634521521 22.9260330
#> 43   1.219444534  4.8395591
#> 44  -2.079717445 17.9835271
#> 45   0.763345277 11.7734296
#> 46   2.038265875  8.8003517
#> 47   0.301420221 16.0411043
#> 48  -0.684828485 17.0194035
#> 49   1.523038447 22.7205654
#> 50   1.720621098  1.1857618
#> 51   2.641843201  0.8322255
#> 52   1.299866695  1.9250899
#> 53   0.720047774  9.7436744
#> 54  -0.361389558 11.4791237
#> 55   1.172358008  9.4379077
#> 56   1.175041845 11.5103062
#> 57   3.339085739  5.5912794
#> 58   1.352345189 20.4133804
#> 59  -0.681565436 13.4972836
#> 60  -1.007438276 13.2519820
#> 61   2.078993054  6.3349990
#> 62   1.652283360 11.0531943
#> 63   1.737997564 22.7523826
#> 64   3.794429549  4.9604490
#> 65   1.345276592 21.7276637
#> 66   2.351670867 18.5617232
#> 67   0.014031261 11.4850828
#> 68   1.702711172  3.0973527
#> 69   0.936131104  1.8249545
#> 70   0.080545096 19.7265408
#> 71   0.845833006  7.6844129
#> 72  -0.698276691 14.8910582
#> 73   0.424950441 22.7260289
#> 74  -1.549122951 19.8806646
#> 75  -0.958808664 13.6052844
#> 76  -1.014924012 13.0146111
#> 77   1.555788134 17.8033485
#> 78   1.021364994 12.7814742
#> 79   0.792087809 22.5199398
#> 80   3.176816480  1.8485552
#> 81   1.117666204 14.9437208
#> 82   2.679058456  3.6053977
#> 83   2.552697410  3.8069175
#> 84   2.217555942  0.5749809
#> 85   1.657425907 14.3326089
#> 86   3.172755504  5.9794947
#> 87   0.198989323  9.8171537
#> 88   1.995226968  3.6790993
#> 89   3.218901207  6.4997471
#> 90  -0.421371691 18.9879720
#> 91   1.960839293 23.6140658
#> 92   1.416000107 22.4459654
#> 93   2.777761837  4.5603096
#> 94   0.599111061  9.4198712
#> 95   0.135629008 20.3258891
#> 96   1.822281957 23.3971870
#> 97  -0.560835516 12.4901635
#> 98   0.173922145 15.8491190
#> 99   0.704091235 17.3191385
#> 100 -0.418009555 12.1233943