Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times group
#> 1    3.14925636 10.4159981     0
#> 2   -0.53922231 12.1138141     0
#> 3    1.20126347 16.9104403     0
#> 4    1.45514472  9.0822543     0
#> 5   -1.03477403 20.0013846     0
#> 6    1.22989044  7.7174091     0
#> 7   -0.44209453 20.7183164     0
#> 8    1.62832477  2.1498048     0
#> 9    2.11413781  1.4638197     0
#> 10   1.69508710  6.2312577     0
#> 11   1.08569001 18.6057773     0
#> 12   1.07474577 17.3905003     0
#> 13   0.67732755 13.7092074     0
#> 14  -0.31056431 21.0947283     0
#> 15   0.49914226 20.6231558     0
#> 16   2.52387343 12.0960393     0
#> 17   1.93833069  3.5063548     0
#> 18   2.55798220 22.0118769     0
#> 19  -1.93241272 16.1280593     0
#> 20   1.44332695 22.6906183     0
#> 21  -0.08277527 23.0335133     0
#> 22  -0.24094896 20.1549066     0
#> 23   2.81143609  7.8592130     0
#> 24   1.59456356  7.2208415     0
#> 25  -0.77254789 15.4699781     0
#> 26   2.21171557  3.7357371     0
#> 27   1.45871894  6.9134037     0
#> 28   0.75454301 23.1783033     0
#> 29   2.53182798 21.8246079     0
#> 30   2.24405450 18.0078229     0
#> 31   1.40637307 19.6035211     0
#> 32   1.28623212 11.0510710     0
#> 33   0.85181356  0.5021185     0
#> 34   0.99351720 13.8893227     0
#> 35   2.02149792  0.8455015     0
#> 36   0.20534090  8.9150601     0
#> 37  -0.05926085 18.8706403     0
#> 38   2.22682113 20.4812514     0
#> 39   1.36925787  3.3518540     0
#> 40   2.13391526 22.2680516     0
#> 41   2.45203012  3.5304114     0
#> 42   1.65360696 21.6638247     0
#> 43   1.69008824 19.2410349     0
#> 44   1.85921632  2.6129701     0
#> 45   3.59040218  3.4165565     0
#> 46   1.01105015 13.2605551     0
#> 47   0.59429218 22.3051718     0
#> 48   1.37668315 21.9266579     0
#> 49   0.59991623 22.9953539     0
#> 50   1.12023917 10.4254210     0
#> 51  -0.41633971  8.5703039     0
#> 52   0.85490710  7.0913977     0
#> 53   1.33135247  2.8460059     0
#> 54   0.98743566  8.6669105     0
#> 55   0.27025270 19.0964834     0
#> 56   3.21058610  4.9424822     0
#> 57   2.09379741  5.1123644     0
#> 58   1.95263248  8.1722863     0
#> 59   0.63219531  8.8550719     0
#> 60   1.70972014  2.7446453     0
#> 61   1.25402643  5.3132348     0
#> 62   0.57358572 10.9373489     0
#> 63   2.78353454  7.8235436     0
#> 64   0.37311882 12.1619844     0
#> 65  -0.48655114 22.8150521     0
#> 66   1.67846639  9.4853007     0
#> 67   1.83873255  8.6367679     0
#> 68   0.37372358 21.1319956     0
#> 69   1.40067609  2.7227514     0
#> 70  -0.79380760 17.1753848     0
#> 71  -0.30056922 14.9459969     0
#> 72   1.34006030 20.0782152     0
#> 73   1.96104487  6.6768258     0
#> 74   0.90339223  3.0409296     0
#> 75   1.37257914  2.3631432     0
#> 76   2.07690013  4.0025536     0
#> 77   0.55598695  9.8483762     0
#> 78   2.24350051  6.9290372     0
#> 79  -1.28861503 17.7341299     0
#> 80   1.89070396  9.3351169     0
#> 81   3.91512384  5.4852878     0
#> 82   1.88384401  4.0528015     0
#> 83   0.99032320 16.5098970     0
#> 84   0.76673492  3.9486305     0
#> 85   1.22605600 12.3660899     0
#> 86  -0.72886787 17.7769045     0
#> 87   0.22963756 18.8978420     0
#> 88  -0.12881334  8.5864141     0
#> 89   2.98831231  9.3841297     0
#> 90   1.16705778 11.7713715     0
#> 91   1.09755238 21.1178905     0
#> 92   2.95090257  4.7898464     0
#> 93   2.52202995  5.4163123     0
#> 94  -0.11533956 13.3021234     0
#> 95   2.65611528  2.3797550     0
#> 96   1.63298794 22.7260520     0
#> 97   1.36638176  2.9997292     0
#> 98   1.45611934 23.0558882     0
#> 99  -1.05707525 12.9498996     0
#> 100  2.24375193  8.4779428     0