Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times group
#> 1    2.26768849 10.2529619     0
#> 2    1.03410559  3.0339353     0
#> 3    0.45998062 20.2016881     0
#> 4    2.17726823  2.2533327     0
#> 5   -0.28373081  4.8501247     0
#> 6    0.79533219  9.5395233     0
#> 7    1.53146641 21.7292851     0
#> 8    0.29782133  4.6386098     0
#> 9    1.43283633 23.2021231     0
#> 10   1.08610015  1.0374150     0
#> 11   0.42939684 18.3630111     0
#> 12   2.51437863 23.2117650     0
#> 13   1.34686623 16.8494651     0
#> 14   2.07031248  2.8749082     0
#> 15   1.49222582 21.7283800     0
#> 16   0.30686651 10.4135905     0
#> 17   0.33762797  0.4798946     0
#> 18   0.03548335  7.7317004     0
#> 19   3.05180343  5.3679966     0
#> 20   1.62551550 20.5122782     0
#> 21   0.72714124 15.5712557     0
#> 22   0.16939980 13.3587880     0
#> 23   3.46731652  5.7123630     0
#> 24   0.34817113 18.5006946     0
#> 25   0.36139846 10.1096699     0
#> 26   0.37389312 19.0092207     0
#> 27   0.05590381 11.0743296     0
#> 28   1.97779978 11.9696524     0
#> 29   2.67723171  4.5504414     0
#> 30   1.45178596 22.1908883     0
#> 31   1.02283838  9.2759061     0
#> 32   1.36004737 13.4950565     0
#> 33   4.35805314  3.8044472     0
#> 34   1.46873786  4.6942775     0
#> 35   1.72474137 13.5003384     0
#> 36   0.72266902  4.4704258     0
#> 37   1.12333790  7.7438339     0
#> 38   0.98737473  8.3318992     0
#> 39   1.98059528  2.4229411     0
#> 40   0.75420097  4.7661914     0
#> 41  -0.68236980 21.3099720     0
#> 42   0.31870239 17.5986383     0
#> 43   0.21514543 18.4509035     0
#> 44   0.18791184 21.5934702     0
#> 45   1.09415088 12.1309539     0
#> 46   0.21512845 18.7383337     0
#> 47   0.59296505 18.8351121     0
#> 48   0.99365691 12.1278451     0
#> 49   1.41738469  9.2321380     0
#> 50   0.30671514 16.9815693     0
#> 51   1.68177371  9.3162544     0
#> 52   0.77810980 17.7891409     0
#> 53   1.97483766  1.6011010     0
#> 54   1.22389233  1.1209613     0
#> 55   3.24675962  3.0308837     0
#> 56  -0.84332893 18.7294142     0
#> 57   0.39748496 21.0078159     0
#> 58   2.12042588 20.0589477     0
#> 59   3.69882248  1.1175324     0
#> 60   1.89102835 18.7191883     0
#> 61   1.89787291  6.4359035     0
#> 62  -0.52957966 12.6407407     0
#> 63  -1.04032742 13.4149173     0
#> 64   1.59627090  5.2774741     0
#> 65   4.76692516  3.5892143     0
#> 66   1.18047059 20.4708022     0
#> 67  -0.34429606 18.6763027     0
#> 68   2.19740967  4.2047201     0
#> 69   0.77951786 15.4752873     0
#> 70   0.33903610 18.1389458     0
#> 71   1.36229183 23.0591032     0
#> 72   0.10787629 12.3954471     0
#> 73   2.36069585 19.8376675     0
#> 74  -0.77385188 17.6528795     0
#> 75   0.83038487 11.3457138     0
#> 76   0.78133077  9.0748533     0
#> 77   1.58433710  0.5633003     0
#> 78  -0.22837219 16.0253375     0
#> 79   2.19655902  4.4773527     0
#> 80   2.71526994  7.7787229     0
#> 81   1.44592291  3.9321139     0
#> 82  -0.57141800 21.6926247     0
#> 83   0.96749760 20.0684494     0
#> 84   0.40135050 16.6032388     0
#> 85   1.83870298 22.6386116     0
#> 86  -0.28273847  2.2696962     0
#> 87  -0.15681568 15.5370773     0
#> 88  -1.06089410 18.0397769     0
#> 89   0.44809443  7.5868361     0
#> 90  -0.40942333 14.2984103     0
#> 91   2.69203509  6.2886484     0
#> 92   2.13019657 10.0397076     0
#> 93   1.15942080 13.7182609     0
#> 94  -0.31328644 11.3827177     0
#> 95  -0.62024186 15.9626613     0
#> 96   0.13540737 21.9863575     0
#> 97   3.68897893  7.1488041     0
#> 98  -0.13955838 14.3075334     0
#> 99   1.84092430  1.3546670     0
#> 100  0.52161646  6.2741946     0