Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>               Y      times group
#> 1    2.13507929  5.3499310     0
#> 2    3.46156214  3.9362295     0
#> 3    0.28162019 20.1893181     0
#> 4   -0.38387252 23.8597412     0
#> 5   -0.17527836 17.0637021     0
#> 6    1.56802900 11.7702682     0
#> 7    1.95863058 17.3557234     0
#> 8    0.22302499 12.7415078     0
#> 9    1.03917635 23.4448966     0
#> 10   3.42842471  0.2046968     0
#> 11   1.46307653  3.8131311     0
#> 12  -0.52833831 12.7337579     0
#> 13   1.35271048 12.3795517     0
#> 14   3.53320350  5.6209686     0
#> 15   1.73621431  0.7048780     0
#> 16   2.06210394 16.1685293     0
#> 17  -0.11437979 14.8855712     0
#> 18   3.25207120 23.6426102     0
#> 19   1.62454313  2.3956954     0
#> 20   2.00558731 20.5834031     0
#> 21   3.36435881  5.8969481     0
#> 22  -0.43776474 11.3217966     0
#> 23  -0.34391675  9.8637984     0
#> 24   1.08795936  6.2072087     0
#> 25   1.20084704  7.8169567     0
#> 26   1.44008565  4.9600925     0
#> 27   1.00332423 17.5520801     0
#> 28   2.86886943  1.5280950     0
#> 29  -0.72943331  0.3914155     0
#> 30   1.58695774 20.8581276     0
#> 31   0.84685705  4.4206809     0
#> 32  -0.84023422  9.2538751     0
#> 33   1.96199318  0.0873492     0
#> 34   0.57811392 11.1803215     0
#> 35   0.77973472 20.1185683     0
#> 36   1.06338386 12.4478826     0
#> 37   1.55174515  8.5375196     0
#> 38   2.50012025  8.4902734     0
#> 39   0.90073311  3.0798121     0
#> 40   1.30343163 12.9173617     0
#> 41   0.29245108 13.7007999     0
#> 42   2.96433502  0.9282066     0
#> 43   0.62082837 17.3356940     0
#> 44   3.00814374  4.8285149     0
#> 45  -0.02683945 14.6121384     0
#> 46   1.92989906  6.2908354     0
#> 47   0.38812822  7.4013899     0
#> 48  -2.22025677 10.9978390     0
#> 49   2.91856233  3.5120331     0
#> 50   1.19216466 15.2265440     0
#> 51   1.17017501 12.7374905     0
#> 52   1.28735490  6.2759493     0
#> 53   0.43417685  9.5536375     0
#> 54   1.74745920  8.5021668     0
#> 55  -0.87180287 13.6624066     0
#> 56   0.26402384 19.1377419     0
#> 57   1.60175435 11.9458832     0
#> 58   0.40226353 20.5258294     0
#> 59   1.68305205  2.5257381     0
#> 60  -0.10374015 10.9665804     0
#> 61   1.82348469  0.5498497     0
#> 62   2.04517733  7.1272359     0
#> 63   1.31798309 20.7281287     0
#> 64   2.90812279  5.8804794     0
#> 65   2.75460149  8.2809988     0
#> 66   1.34417816 10.4840768     0
#> 67   1.47942093 17.5517965     0
#> 68   0.41219462 18.3385847     0
#> 69   1.03253363 10.0192739     0
#> 70   2.50575211  0.4399229     0
#> 71   3.75704933  2.0476670     0
#> 72   1.51811196 10.2275526     0
#> 73   0.34831299 20.1660020     0
#> 74   1.25553244  7.5051606     0
#> 75   1.78813777 19.8243563     0
#> 76   2.86311709  5.8663748     0
#> 77   2.35282036 15.0257787     0
#> 78   1.04571126 12.2176459     0
#> 79   0.17403697 11.0821554     0
#> 80   1.68827212 19.3125409     0
#> 81  -1.16191259 12.1279477     0
#> 82  -0.94105472 16.9853622     0
#> 83   1.81390205  7.7916205     0
#> 84   4.05743178  5.2636335     0
#> 85   0.40427913  2.6780586     0
#> 86   1.78700790  6.6447142     0
#> 87   1.02634450 14.8437788     0
#> 88  -0.48286054 16.8074455     0
#> 89   0.46091835 18.9844699     0
#> 90   1.66139796 23.4063178     0
#> 91   1.83415860  0.5248101     0
#> 92   0.98941792  7.7832289     0
#> 93  -0.22903540 20.2730871     0
#> 94   2.22723505  6.4798876     0
#> 95   1.81292833 18.6933546     0
#> 96   1.60310423  2.5580471     0
#> 97   1.38943568  7.9735374     0
#> 98   1.10970024 16.9588848     0
#> 99   1.47449775  2.3845091     0
#> 100  1.90617611  8.8122576     0