Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>                Y       times group
#> 1    1.553321328  5.39389350     0
#> 2   -0.800423248 12.36482154     0
#> 3    0.763978754 16.81095739     0
#> 4    0.151335215 21.45935161     0
#> 5   -1.026230298 18.36420570     0
#> 6    0.447136270 11.64430918     0
#> 7    0.055552664 10.52745234     0
#> 8   -0.210831732 14.22996601     0
#> 9    0.351328153 13.31582528     0
#> 10   1.055480453 21.81605466     0
#> 11  -0.222907125 21.61896798     0
#> 12  -0.036841955 21.00394431     0
#> 13   1.142301098 11.70630537     0
#> 14   2.571616532  2.46274912     0
#> 15  -0.305959561 13.89247292     0
#> 16   2.523368809  4.69407602     0
#> 17   3.000039776  6.34564534     0
#> 18   1.955260592 23.36379895     0
#> 19   1.160715070 12.01895213     0
#> 20  -0.006311457 16.39006671     0
#> 21   1.151179037 12.93364506     0
#> 22   1.592192869  0.46974337     0
#> 23   1.685981718 14.92905321     0
#> 24  -1.824279070 12.98281530     0
#> 25   1.613929438  6.99612031     0
#> 26  -0.559717656 21.35482995     0
#> 27   1.131094530 18.23733650     0
#> 28   0.983859860 17.74304262     0
#> 29   2.168690604  5.52785945     0
#> 30   1.978378814  3.07150198     0
#> 31  -0.459563021 18.13111930     0
#> 32   1.721988472  7.17647089     0
#> 33   1.758068232 22.87876715     0
#> 34   0.603127584  1.55396278     0
#> 35   0.868455684  3.31385184     0
#> 36   1.010466979  8.79397242     0
#> 37   0.365871072 13.36990318     0
#> 38   2.086579218  6.97228266     0
#> 39   1.637433326 10.03749788     0
#> 40   1.658946187  6.64275318     0
#> 41   2.574147063  5.27646043     0
#> 42   3.900318026  3.08621620     0
#> 43  -0.336912960 10.40568360     0
#> 44   0.671541595  0.45971624     0
#> 45   0.409075270 13.44409618     0
#> 46   0.877081038 19.14594605     0
#> 47   1.746182566 22.90905917     0
#> 48   1.554263457  4.68409765     0
#> 49   1.496950637 20.04200798     0
#> 50  -0.305843567 16.68517508     0
#> 51  -0.542465771  4.19599878     0
#> 52   0.411406530  2.34164030     0
#> 53   1.094715245  3.36157167     0
#> 54   0.351276981 13.27399028     0
#> 55  -0.583196680 11.70995068     0
#> 56   1.616901949  2.04631258     0
#> 57   2.023823045 19.66861081     0
#> 58   2.682100943  9.58331702     0
#> 59   1.549549795  2.69384122     0
#> 60  -0.247571993  9.41527437     0
#> 61   1.928370765  1.77928615     0
#> 62   2.581473637  5.10128142     0
#> 63   0.268834074 21.97831055     0
#> 64   0.419691157 15.20111726     0
#> 65   2.443959133  3.94252834     0
#> 66  -0.593752656 11.50598824     0
#> 67   2.239204351  5.50088286     0
#> 68   1.761833908 14.86841189     0
#> 69   4.490551039  0.45332227     0
#> 70   1.617108295  3.05299560     0
#> 71   0.829299944 17.68345038     0
#> 72   2.069559065  0.07604026     0
#> 73   0.914393266  7.28063509     0
#> 74   2.252947121  5.95782330     0
#> 75   3.373991815  2.05072759     0
#> 76   0.583463044  8.64608230     0
#> 77   0.441875163 21.18011208     0
#> 78   1.607184688 12.01300163     0
#> 79   0.249687622 15.70486921     0
#> 80   0.807432092 17.22711706     0
#> 81   0.463666996 18.52686806     0
#> 82   1.728851484 19.29033830     0
#> 83  -0.311783708 23.08949665     0
#> 84   0.068139308  7.16071937     0
#> 85   0.189094428 10.91799179     0
#> 86   2.163057529 11.73350713     0
#> 87   2.383839739  1.33306946     0
#> 88   0.381619225 19.00841477     0
#> 89   2.077079993  6.27689575     0
#> 90  -0.781340889 12.78812770     0
#> 91   0.327945791  9.30584271     0
#> 92   1.811383345 11.77468607     0
#> 93  -1.442173989 16.35672200     0
#> 94   2.228879150  7.24212774     0
#> 95   0.589027271 19.14187826     0
#> 96   1.027114677  9.93168069     0
#> 97   0.074754705 16.53560917     0
#> 98   0.329727615 19.60299734     0
#> 99  -0.378576302 15.39384883     0
#> 100  0.863221434 12.07844594     0