Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The mesor is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1.

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1).

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>                Y      times
#> 1    2.861242736 22.4611999
#> 2    5.105168513 21.2844862
#> 3    0.594068422 15.9528255
#> 4    1.582495229 21.4621757
#> 5   -0.076659257 21.1379272
#> 6   -0.046252471 10.8379990
#> 7    0.363513576  8.2183227
#> 8   -1.158613674  6.6976290
#> 9   -0.634550417  4.8758329
#> 10   4.841744751 17.2238507
#> 11   0.735175407  0.6210934
#> 12   0.563768199  7.7924697
#> 13   0.207884300 13.8593794
#> 14  -1.184798282  5.0100655
#> 15   1.292001886 16.3158452
#> 16   0.886353479  6.4340209
#> 17   0.688676506 12.2629693
#> 18   0.828498337  0.4385468
#> 19   1.851044635  1.4168102
#> 20   1.488230076 18.6395490
#> 21   2.693704783 20.1148295
#> 22   2.503301886 15.0720448
#> 23  -0.014577385  8.5076091
#> 24   2.773533751 18.9572557
#> 25  -0.122318474  9.3751051
#> 26   1.791356273 15.3362180
#> 27  -1.236330525  2.6329187
#> 28   1.379965504  9.4382513
#> 29   0.022758203  7.6719482
#> 30  -0.501579509  7.3764922
#> 31   0.021704389  8.9086568
#> 32   2.038906032 19.4109765
#> 33   1.300498850  1.3415513
#> 34   0.096889873 12.0095239
#> 35   2.017854968 22.5176513
#> 36   1.290832586 11.0734249
#> 37  -0.103230834 16.0547680
#> 38   1.935654824  5.8794891
#> 39   1.003827730  9.4058590
#> 40   1.619065104 16.3096573
#> 41   2.577161013 22.8552927
#> 42   1.483312514 18.2736064
#> 43   0.004391518 21.2712728
#> 44   0.841908380 17.5064958
#> 45   0.144561739  4.5619869
#> 46   2.392734970 21.3908333
#> 47   0.260139044  8.5311104
#> 48   1.517551285 21.1234732
#> 49   0.846121374 11.2480079
#> 50   2.711429416 22.2723312
#> 51   2.512197791 22.1991049
#> 52   1.468555026 21.1450409
#> 53   0.153442714  3.7913024
#> 54  -0.454621477 10.7426297
#> 55   1.826339760  6.6638568
#> 56   1.941610234 21.6651242
#> 57   0.948608738 18.6327779
#> 58   0.387764060  6.7477292
#> 59   1.655592091 14.0933832
#> 60  -1.166369044  5.5335112
#> 61   2.595890362 19.9796404
#> 62   1.977129904  0.5460569
#> 63   1.498782744  9.3036653
#> 64   1.927335825  1.4049820
#> 65  -0.191355514 12.8717505
#> 66   1.239234738 18.0370911
#> 67  -0.250046556 23.0882723
#> 68   1.965664694 11.2705026
#> 69   2.212073419 15.6931431
#> 70  -0.374044321 10.6208535
#> 71   2.578827460  2.0547409
#> 72   2.093658656  3.9396555
#> 73  -0.287839048  8.0536194
#> 74   0.277559815 20.7213834
#> 75   0.753482775  5.8737725
#> 76   2.371512351 18.6041612
#> 77  -1.014295113  7.9366520
#> 78  -0.515511476  6.7404148
#> 79   0.509554877 14.4263716
#> 80   1.537865941  0.6418450
#> 81   3.516358963 20.2751761
#> 82  -0.088429748 11.8685046
#> 83   0.741682016 23.5138870
#> 84   2.658783692 22.7521953
#> 85   2.070875639 20.5097354
#> 86   1.730360936 23.6892231
#> 87   1.129038062 14.6849322
#> 88   0.385109557  3.4715728
#> 89   0.391881499  0.9237091
#> 90   1.822583021 12.0111538
#> 91   0.940328247  5.9925400
#> 92   0.903042093  9.2643918
#> 93  -0.535295783  9.1667506
#> 94   2.375395654 16.2383741
#> 95   0.930100584  4.6189723
#> 96   1.271576851 17.4230016
#> 97   0.604797510 12.8134097
#> 98  -1.923108615  5.7870467
#> 99   0.924015979 17.6773259
#> 100  2.082347542 17.2120482