Simulate data from a cosinor model

This function simulates data from a cosinor model with a single covariate, where the time scale is month, and optionally allows for single covariate effects on the mean, amplitude, and acrophase.

Usage

simulate_cosinor(
  n,
  mesor,
  amp,
  acro,
  period = 24,
  n_components,
  beta.group = FALSE,
  beta.mesor,
  beta.amp,
  beta.acro,
  n_period = 1,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  ...
)

Arguments

n

The sample size. An integer greater than 0.

mesor

A numeric. The MESOR (midline estimating statistic of rhythm) for group = 0. The MESOR is independent of the cosinor components, so only one value is allowed even if there are multiple components in the data being simulated.

amp

A numeric. The amplitude value (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(5, 10).

acro

A numeric. The acrophase value in radians (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(0, pi) for two components.

period

The period of the rhythm data (for group = 0 if grouped data are being simulated (beta.group = TRUE)). If simulating data with multiple components, specify a vector with values for each component. E.g: period = c(12, 6) for two components.

n_components

The number of components in the model. This must match the length of the inputs for amp and acro.

beta.group

A logical. If TRUE a second group of data will be simulated and included in the returned data set. If FALSE, beta.acro, beta.mesor, and beta.amp arguments will be ignored.

beta.mesor

A numeric. The MESOR value term for group = 1

beta.amp

A numeric. The amplitude value for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: amp = c(2, 8).

beta.acro

A numeric. The acrophase value in radians (for group = 1. If simulating data with multiple components, specify a vector with values for each component. E.g: acr = c(2, 5) for two components.

n_period

A numeric. The number of cycles of the rhythm to be simulated.

family

A character. The family (see ?family) of the simulated dataset. Can handle values in c("poisson", "binomial", "gamma", "gaussian").

...

Extra arguments, including alpha that controls the shape argument when sampling from a gamma distribution (when family = "gamma"; default is 1), and sd (standard deviation) which is used when sampling from a normal distribution (when family = "gaussian"; default is 1). To specify these parameters for the beta (treatment) group, use beta.alpha and beta.sd

Value

Returns simulated data in a data.frame.

Examples

simulate_cosinor(
  n = 100,
  mesor = 1,
  amp = 1,
  acro = 1,
  period = 24,
  family = "gaussian"
)
#>                Y      times group
#> 1    2.965523000  2.2318140     0
#> 2    1.048415407  5.9439546     0
#> 3    1.367336388 23.1234423     0
#> 4    1.733544939 16.9297311     0
#> 5    3.283904812  5.7811518     0
#> 6    2.080901656  4.3352919     0
#> 7   -1.222385441 18.8332017     0
#> 8    2.483794805  0.3592800     0
#> 9    1.029566426 18.0920280     0
#> 10   1.208509546  8.6103543     0
#> 11   0.462657460 16.3331718     0
#> 12   2.009576035  4.6324877     0
#> 13   2.881526705 13.6017555     0
#> 14   2.901115077 10.4658184     0
#> 15   3.032320866  4.0638690     0
#> 16   0.076884581 18.1469024     0
#> 17   1.799377202 15.7426357     0
#> 18   0.962057040 10.0884454     0
#> 19  -1.021014061 17.9449822     0
#> 20   2.345757257 14.1566542     0
#> 21   1.071967561  8.1804982     0
#> 22   0.406692657  1.4493726     0
#> 23  -0.766512967 10.7887138     0
#> 24  -0.614438367 10.8593344     0
#> 25  -0.995381199 16.6769106     0
#> 26  -0.544133751 19.6633035     0
#> 27   0.959806932  5.1919400     0
#> 28  -0.076708168  8.5793941     0
#> 29   1.779971896  3.5831666     0
#> 30   1.960414941 23.2036635     0
#> 31  -0.227644097 12.8007249     0
#> 32   2.323926920  3.3264483     0
#> 33  -0.376435708 15.5893424     0
#> 34   3.374669594  5.3247857     0
#> 35   0.190347131  5.2057703     0
#> 36   1.580314023  7.8787533     0
#> 37   1.499027507  2.8120087     0
#> 38   2.435205499  4.0429922     0
#> 39  -0.005062952 16.8115273     0
#> 40  -0.588159033 16.7620203     0
#> 41   0.372456885 14.5065639     0
#> 42   0.516086705  5.6930597     0
#> 43  -1.406700357 12.3033375     0
#> 44   3.432815618  3.4026277     0
#> 45   2.440789705 22.2105731     0
#> 46   1.350310366  1.8338882     0
#> 47   0.595081762 21.3209444     0
#> 48  -0.190811420 20.5499278     0
#> 49   0.555860372 17.2294785     0
#> 50   1.322088982 10.1574861     0
#> 51  -0.903732602 17.0088800     0
#> 52   1.762485052  1.8969115     0
#> 53  -0.289987141 22.5419905     0
#> 54   3.627774669  4.5453556     0
#> 55   0.987044660 18.7970706     0
#> 56   0.666858255 19.2513696     0
#> 57   1.831353393  5.7642174     0
#> 58   0.690556610 12.0117846     0
#> 59   0.942193882 17.2166847     0
#> 60   0.065866211  6.1714717     0
#> 61   0.779411818  1.7654045     0
#> 62   0.677156332 23.7335284     0
#> 63   1.551650823  9.8194965     0
#> 64  -0.950048493 16.5950114     0
#> 65   0.745856414 20.2003676     0
#> 66   2.279403630  6.1174084     0
#> 67   0.258124331 17.8791421     0
#> 68   2.163053542  9.5465726     0
#> 69   0.499574092 18.8976580     0
#> 70  -0.734889882 14.8801474     0
#> 71   0.082365589 20.9719387     0
#> 72  -1.608932607 16.1644290     0
#> 73   0.958297421  7.4588692     0
#> 74   0.780883901 21.4892780     0
#> 75  -0.733620060 13.0583236     0
#> 76  -0.895794994 13.4463818     0
#> 77  -1.706767554 18.2202314     0
#> 78  -0.671271479 20.2899607     0
#> 79   2.751965174  5.1368182     0
#> 80  -0.427279710 17.6472480     0
#> 81   1.529088119 18.4901954     0
#> 82  -1.900291845 12.6788735     0
#> 83  -0.464586318 21.9512254     0
#> 84   1.102946519 23.8274738     0
#> 85   0.985093219  8.2361548     0
#> 86   0.854027331 21.7304123     0
#> 87   0.596042915  5.0319416     0
#> 88   0.550484734  8.1514369     0
#> 89   1.046457842 11.0563013     0
#> 90   1.042692822  3.9897472     0
#> 91   0.834194622  8.5298407     0
#> 92  -0.001332418 14.0130591     0
#> 93   0.246606908 21.1747858     0
#> 94   0.655312358 18.9595581     0
#> 95   0.859582567 19.8610068     0
#> 96   2.762559260  2.6671497     0
#> 97   0.910581646  7.1215590     0
#> 98   1.121583570  0.7092134     0
#> 99   1.463835957 21.2028926     0
#> 100 -0.888308113 13.1419527     0